\name{aal}
\alias{aal}
\docType{data}
\title{aal}
\description{
A data frame label numbers and names for the neuroanatomical AAL labels.
}
\usage{data(aal)}
\format{
  A data frame listing the following variables.
  \describe{
    \item{\code{label_num}}{ Numerical label value. }
    \item{\code{label_name}}{ Shorthand anatomical value. }
  }
}
\references{
\url{http://en.wikipedia.org/wiki/Automated_Anatomical_Labeling}
}
\examples{
data(aal)
}
\keyword{datasets}
