% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alff.R
\name{alffmap}
\alias{alffmap}
\title{alffmap}
\usage{
alffmap(x, flo = 0.01, fhi = 0.1, tr = 1, detrend = TRUE)
}
\arguments{
\item{x}{input vector for the time series of interest}

\item{flo}{low frequency, typically 0.01}

\item{fhi}{high frequency, typically 0.1}

\item{tr}{the period associated with the vector x (inverse of frequency)}

\item{detrend}{detrend the input time series}
}
\value{
vector is output showing ALFF and fALFF values
}
\description{
Amplitude of Low Frequency Fluctuations (ALFF; Zang et al., 2007) and
fractional Amplitude of Low Frequency Fluctuations (f/ALFF; Zou et al., 2008)
are related measures that quantify the amplitude of low frequency
oscillations (LFOs).  This function outputs ALFF and fALFF for the input.
}
\examples{

mat <- matrix(rnorm(3000),ncol=50)
fallf = apply( mat, FUN=alffmap, MARGIN=2 )

}
\author{
Avants BB
}

