% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ants_set_seed.R
\name{ants.set.seed}
\alias{ants.set.seed}
\title{Set seed for ANTsR functionality}
\usage{
ants.set.seed(seed)
}
\arguments{
\item{seed}{Seed to set to environment variable
\code{ANTS_RANDOM_SEED}}
}
\value{
Logical indicating the variable was set
}
\description{
Set seed for ANTsR functionality
}
\examples{
ants.set.seed(1)
ants.set.seed("")
}
