% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/antsAffineInitializer.R
\name{antsAffineInitializer}
\alias{antsAffineInitializer}
\title{antsAffineInitializer}
\usage{
antsAffineInitializer(...)
}
\arguments{
\item{...}{see ants function}
}
\value{
none
}
\description{
Wrapper for the ANTs funtion antsAffineInitializer
}
\author{
Avants BB
}

