\name{antsApplyTransforms}
\alias{antsApplyTransforms}

\title{
A simplified (or full) interface to antsApplyTransforms.
}

\description{
Apply a transform list to map an image from one domain to another.
}

\usage{
myoutputimg<-antsApplyTransforms( fixed=imgf , moving=imgm , transformlist=txlist , interpolator=c("Linear","NearestNeighbor","MultiLabel") )
}

\arguments{
\item{fixed}{
fixed image defining domain into which the moving image is transformed.
}

\item{movingImage}{
moving image to be mapped to fixed space.
}

\item{transformlist}{
list of transforms generated by antsRegistration where each transform is a filename. 
}

\item{interpolator}{
choose the Linear, NearestNeighbor, MultiLabel or any other interpolator available in antsApplyTransforms.
}

}

\value{
an antsImage is output.
1 -- Failure
}

\author{
Shrinidhi KL, Avants BB
}

\examples{
\dontrun{
# will give the full form of help
antsApplyTransforms( "-h")
# example 1 - simplified
deformedimage<-antsApplyTransforms(fixed=img1,moving=img2,
  transformlist=list(rev(c("my0GenericAffine.mat","my1Warp.nii.gz")))
}
}
