% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsTransform_class.R
\name{antsApplyTransformToImage}
\alias{antsApplyTransformToImage}
\title{antsApplyTransformToImage}
\usage{
antsApplyTransformToImage(transform, image, reference,
  interpolation = "linear")
}
\arguments{
\item{image}{antsImage to transform}

\item{tx}{antsTransform}

\item{ref}{antImage giving the reference output space}
}
\value{
antsImage
}
\description{
Apply transform to spatial point
}
\examples{
img <- antsImageRead(getANTsRData("r16"))
tx = new("antsTransform", precision="float", type="AffineTransform", dimension=2 )
antsTransformSetParameters(tx, c(0,-1,1,0,,0) )
img2 = antsApplyTransformToImage(tx, img, img)
}

