% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsTransform_class.R
\name{antsApplyTransformToPoint}
\alias{antsApplyTransformToPoint}
\title{antsApplyTransformToPoint}
\usage{
antsApplyTransformToPoint(tx, point)
}
\arguments{
\item{tx}{antsTransform}
}
\value{
array of coordinates
}
\description{
Apply transform to spatial point
}
\examples{
tx = new("antsTransform")
params = antsTransformGetParameters(tx)
antsTransformSetParameters(tx, params*2)
pt2 = antsApplyTransformToPoint(tx, c(1,2,3))
}

