% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsTransform_class.R
\name{antsApplyTransformToVector}
\alias{antsApplyTransformToVector}
\title{antsApplyTransformToVector}
\usage{
antsApplyTransformToVector(transform, vector)
}
\arguments{
\item{transform}{antsTransform}

\item{vector}{array of vector coordinates}
}
\value{
array of coordinates
}
\description{
Apply transform to spatial vector
}
\examples{
\dontrun{
vec2 = antsApplyTransformToVector(transform, c(1,2,3))
}
}

