% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/antsAverageImages.R
\name{antsAverageImages}
\alias{antsAverageImages}
\title{averages an image list with or without intensity normalization}
\usage{
antsAverageImages(imageList, normalize = FALSE)
}
\arguments{
\item{imageList}{must be an antsImage list}

\item{normalize}{boolean determines if image is divided by mean before
averaging}

\item{reference}{must be an antsImage of the same size as those in the list}
}
\description{
calculate the mean of a list of antsImages imageList: can contain anything that be cast via as.array()
}
\author{
Avants BB, Pustina D
}

