% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/antsBOLDNetworkAnalysis.R
\name{antsBOLDNetworkAnalysis}
\alias{antsBOLDNetworkAnalysis}
\title{antsBOLDNetworkAnalysis}
\usage{
antsBOLDNetworkAnalysis(bold = NA, mask = NA, labels = NA, motion = NA,
  gdens = 0.25, threshLo = 1, threshHi = 90, freqLo = 0.01,
  freqHi = 0.1, winsortrim = 0.02, throwaway = 10)
}
\arguments{
\item{bold}{input 4D image}

\item{mask}{antsImage defines areas of interest}

\item{labels}{antsImage defines regions of interest}

\item{motion}{motion parameters}

\item{gdens}{graph density}

\item{threshLo}{lower threshold for the label image}

\item{threshHi}{upper threshold for the label image}

\item{freqLo}{lower frequency cutoff}

\item{freqHi}{upper frequency cutoff}

\item{winsortrim}{winsorize the bold signal by these values eg 0.02}

\item{throwaway}{this number of initial bold volumes}
}
\value{
list of outputs
}
\description{
An implementation of a network analysis framework for BOLD data.
}
\examples{
# none yet - this is not very well tested with recent ANTsR
}
\author{
BB Avants
}

