% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flags.R
\name{antsCompileFlags}
\alias{antsCompileFlags}
\title{return ants compile flags}
\usage{
antsCompileFlags()
}
\description{
call this to get the compilation flags used for ants and dependent software
}
\examples{

antsCompileFlags()

}
\author{
Avants BB
}
