% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/antsCopyImageInfo.R
\name{antsCopyImageInfo}
\alias{antsCopyImageInfo}
\title{Copy header info}
\usage{
antsCopyImageInfo(reference, target)
}
\arguments{
\item{reference}{Image object of S4 class 'antsImage' to get values from.}

\item{target}{Image object of S4 class 'antsImage' to copy values to.}
}
\value{
Matrix of image indices
}
\description{
Copy header info from one 'antsImage' to another
}
\examples{
img<-makeImage(c(10,10),rnorm(100))
img2<-makeImage(c(10,10), rnorm(100))
img2<-antsCopyImageInfo(img, img2)
}

