% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsCopyImageInfo.R
\name{antsCopyImageInfo2}
\alias{antsCopyImageInfo2}
\title{Copy header info with different input order}
\usage{
antsCopyImageInfo2(target, reference)
}
\arguments{
\item{target}{image object of S4 class \code{antsImage} to copy values to.}

\item{reference}{image object of S4 class \code{antsImage} to get values from.}
}
\value{
Target image with reference header information.
}
\description{
Copy origin, direction, and spacing from one \code{antsImage} to another.
Opposite order of input relative to \code{antsCopyImageInfo}
}
\examples{

img <- makeImage(c(10,10),rnorm(100))
img2 <- makeImage(c(10,10), rnorm(100))
img2 <- antsCopyImageInfo(img2, img)
testthat::expect_error(antsCopyImageInfo(img, 1))

}
