% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flags.R
\name{antsDir}
\alias{antsDir}
\title{return ANTs binary directory information}
\usage{
antsDir()
}
\description{
call this to get the binary paths to ANTs executables.
}
\examples{

antsDir()

}
\author{
Avants BB
}
