% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsImage_class.R
\docType{class}
\name{antsImage-class}
\alias{antsImage-class}
\alias{show,antsImage-method}
\alias{initialize,antsImage-method}
\title{An S4 class for an antsImage}
\usage{
\S4method{show}{antsImage}(object)

\S4method{initialize}{antsImage}(
  .Object,
  pixeltype = "float",
  dimension = 3,
  components = 1,
  isVector = FALSE,
  filename = ""
)
}
\arguments{
\item{object}{input object to convert}

\item{.Object}{input object to convert}

\item{pixeltype}{usually float, can be other types unsigned char, int, double
etc noting that short is not supported}

\item{dimension}{usually 2 or 3 but can be 4}

\item{components}{number of pixel components}

\item{isVector}{logical indicator of the image is a vector}

\item{filename}{character filename if the data was read in, otherwise
""}
}
\description{
C++ type used to represent an ITK image pointed to by 'pointer'. the
actual image is of C++ type 'itk::image< pixeltype , dimension >::Pointer'
}
\section{Slots}{

\describe{
\item{\code{pixeltype}}{usually float, can be other types unsigned char, int, double
etc noting that short is not supported}

\item{\code{dimension}}{usually 2 or 3 but can be 4}

\item{\code{components}}{number of pixel components}

\item{\code{pointer}}{the memory location}

\item{\code{isVector}}{logical indicator of the image is a vector}

\item{\code{filename}}{character filename if the data was read in, otherwise
""}

\item{\code{pixeltype}}{usually float, can be other types unsigned char, int, double
etc noting that short is not supported}

\item{\code{dimension}}{usually 2 or 3 but can be 4}

\item{\code{components}}{number of pixel components}

\item{\code{isVector}}{boolean}

\item{\code{filename}}{character filename if the data was read in, otherwise
""}
}}

