% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsImage4DtoList.R
\name{antsImage4DtoList}
\alias{antsImage4DtoList}
\title{antsImage4DtoList}
\usage{
antsImage4DtoList(img)
}
\arguments{
\item{img}{input 4D image of class antsImage, 4th dimension has volumes}
}
\value{
output list filled with 3D volumes of class antsImage
}
\description{
Resample image by using another image as target reference.
This function uses antsApplyTransform with an identity 
matrix to achieve proper resampling.
}
\examples{
## Do not run
## my4Dimage = antsImageRead('/path/to/image.nii.gz)
## my3Dlist = antsImage4DtoList(my4Dimage)
## End do not run

}
\author{
Pustina D
}

