% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/antsImage_class.R
\name{antsImageArith}
\alias{*.antsImage}
\alias{+.antsImage}
\alias{-.antsImage}
\alias{/.antsImage}
\alias{\%\%.antsImage}
\alias{^.antsImage}
\alias{antsImageArith}
\title{arith.antsImage}
\usage{
\method{+}{antsImage}(x, y)

\method{-}{antsImage}(x, y)

\method{/}{antsImage}(x, y)

\method{*}{antsImage}(x, y)

\method{^}{antsImage}(x, y)

\method{\%\%}{antsImage}(x, y)
}
\arguments{
\item{x}{antsImage}

\item{y}{antsImage or numeric}
}
\description{
Atomic arithmetic operators for antsImages
}
\examples{
r16 <- antsImageRead(getANTsRData('r16'), 2)
r64 <- antsImageRead(getANTsRData('r64'), 2)
r16 + r64
r16 + 5
r16 / 10
}

