% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsImageClone.R
\name{antsImageClone}
\alias{antsImageClone}
\title{Clone an \code{antsImage}.}
\usage{
antsImageClone(in_image, out_pixeltype = in_image@pixeltype)
}
\arguments{
\item{in_image}{image object of S4 class \code{antsImage} to be cloned.}

\item{out_pixeltype}{C++ datatype to be used to represent the pixels in the
output image. Allowed values: \code{double}, \code{float},
\code{unsigned int}, \code{unsigned char}.}
}
\value{
object of class \code{antsImage}
}
\description{
Clone an image object of S4 class \code{antsImage}. N.B.: You cannot use
\code{a <- img} because the R assignment operator does not deal with the underlying
C++ pointers.
}
\examples{

img <- antsImageRead(getANTsRData("r16"), 2)
img2 <- antsImageClone(img)
img.int <- antsImageClone(img , "unsigned int")

}
\author{
Shrinidhi KL
}
