% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/antsImage_class.R
\name{antsImageGetSet}
\alias{antsGetDirection}
\alias{antsGetOrigin}
\alias{antsGetSpacing}
\alias{antsImageGetSet}
\alias{antsSetDirection}
\alias{antsSetOrigin}
\alias{antsSetSpacing}
\title{antsImageGetSet}
\usage{
antsGetSpacing(x)

antsSetSpacing(x, spacing)

antsGetOrigin(x)

antsSetOrigin(x, origin)

antsGetDirection(x)

antsSetDirection(x, direction)
}
\arguments{
\item{x}{Image to access}

\item{spacing}{Desired spacing as vector of size dim(x)}

\item{origin}{Desired origin as vector of size dim(x)}

\item{direction}{Desired direction as vector of size dim(x)}
}
\value{
For \code{get} methods, vector of size \code{dim(x)}.
For \code{set} methods, 0 to indicate success.
}
\description{
Get and set methods for image header information
}
\examples{
img <- makeImage(c(5,5), rnorm(25))
antsSetSpacing(img, c(2.0, 2.0))
antsSetOrigin(img, c(0.5, 0.5))
}

