% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/antsImageHeaderInfo.R
\name{antsImageHeaderInfo}
\alias{antsImageHeaderInfo}
\title{Read file info from image header}
\usage{
antsImageHeaderInfo(filename)
}
\arguments{
\item{filename}{name of image file to scan for info}
}
\value{
outputs a list containing:
\itemize{
  \item{pixelclass: }{Type of pixel (scalar, vector, etc).}
  \item{pixeltype: }{Type of pixel values (int, float, etc).}
  \item{nDimensions: }{Number of image dimensions.}
  \item{nComponents: }{Number of pixel dimensions.}
  \item{dimensions: }{Size of image dimensions.}
  \item{spacing: }{Pixel resolution.}
  \item{origin: }{Spatial origin of image}
  \item{pixelclass: }{Spatial directions of image axes.}
}
}
\description{
Read file info from image header
}
\examples{
antsImageHeaderInfo( getANTsRData("r16") )
}
\author{
Duda JT
}

