% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsImageIterator_class.R
\name{antsImageIterator}
\alias{antsImageIterator}
\title{antsImageIterator}
\usage{
antsImageIterator(x)
}
\arguments{
\item{x}{antsImage}
}
\value{
antsImageIterator
}
\description{
Get image iterator
}
\examples{
img <- makeImage(c(5, 5), rnorm(25))
it <- antsImageIterator(img)
}
