% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsImageIterator_class.R
\docType{class}
\name{antsImageIterator-class}
\alias{antsImageIterator-class}
\alias{show,antsImageIterator-method}
\alias{initialize,antsImageIterator-method}
\title{An S4 class for an antsImageIterator}
\usage{
\S4method{show}{antsImageIterator}(object)

\S4method{initialize}{antsImageIterator}(.Object, pixeltype = "float",
  dimension = 3, components = 1)
}
\arguments{
\item{object}{input object to convert}

\item{.Object}{input object to convert}

\item{pixeltype}{string e.g. "float" "unsigned char" "int"}

\item{dimension}{dimensionality of the image}

\item{components}{number of components per pixel}
}
\description{
C++ type used to represent an ITK image iterator. the
actual iterator is of C++ type
'itk::ImageRegionIteratorWithIndex< itk::image< pixeltype, dimension > >'
}
\section{Slots}{

\describe{
\item{\code{pixeltype}}{usually float, can be other types unsigned char, int, double
etc noting that short is not supported}

\item{\code{dimension}}{usually 2 or 3 but can be 4}

\item{\code{components}}{number of pixel components, currently only 1 is suppored}

\item{\code{pointer}}{to the memory location of the itk object}
}}

