% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsImageIterator_class.R
\name{antsImageIteratorGet}
\alias{antsImageIteratorGet}
\title{antsImageIteratorGet}
\usage{
antsImageIteratorGet(x)
}
\arguments{
\item{x}{antsImageIterator}
}
\value{
pixel value
}
\description{
Get value at current position
}
\examples{
img <- makeImage(c(5,5), rnorm(25))
it <- antsImageIterator( img )
value <- antsImageIteratorGet( it )
}

