% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsImageIterator_class.R
\name{antsImageIteratorGetIndex}
\alias{antsImageIteratorGetIndex}
\title{antsImageIteratorGetIndex}
\usage{
antsImageIteratorGetIndex(x)
}
\arguments{
\item{x}{antsImageIterator}
}
\value{
image index
}
\description{
Get index at current position
}
\examples{
img <- makeImage(c(5,5), rnorm(25))
it <- antsImageIterator( img )
index <- antsImageIteratorGetIndex( it )
}

