% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsImageIterator_class.R
\name{antsImageIteratorGoToBegin}
\alias{antsImageIteratorGoToBegin}
\title{antsImageIteratorGoToBegin}
\usage{
antsImageIteratorGoToBegin(x)
}
\arguments{
\item{x}{antsImageIterator}
}
\value{
antsImageIterator
}
\description{
move iterator to begining
}
\examples{
img <- makeImage(c(5,5), rnorm(25))
it <- antsImageIterator( img )
it <- antsImageIteratorNext( it )
it <- antsImageIteratorGoToBegin( it )
}

