% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsImageIterator_class.R
\name{antsImageIteratorGoToReverseBegin}
\alias{antsImageIteratorGoToReverseBegin}
\title{antsImageIteratorGoToReverseBegin}
\usage{
antsImageIteratorGoToReverseBegin(x)
}
\arguments{
\item{x}{antsImageIterator}
}
\value{
antsImageIterator
}
\description{
move iterator to begining of reverse data
}
\examples{
img <- makeImage(c(5,5), rnorm(25))
it <- antsImageIterator( img )
it <- antsImageIteratorGoToReverseBegin( it )
}

