% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsImageIterator_class.R
\name{antsImageIteratorIsAtReverseEnd}
\alias{antsImageIteratorIsAtReverseEnd}
\title{antsImageIteratorIsAtReverseEnd}
\usage{
antsImageIteratorIsAtReverseEnd(x)
}
\arguments{
\item{x}{antsImageIterator}
}
\value{
boolean indicating position
}
\description{
test if iterator is at end of reverse data
}
\examples{
img <- makeImage(c(5,5), rnorm(25))
it <- antsImageIterator( img )
it <- antsImageIteratorGoToReverseBegin( it )
it <- antsImageIteratorPrevious( it )
flag <- antsImageIteratorIsAtReverseEnd( it )
}
