% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsImageIterator_class.R
\name{antsImageIteratorNext}
\alias{antsImageIteratorNext}
\title{antsImageIteratorNext}
\usage{
antsImageIteratorNext(x)
}
\arguments{
\item{x}{antsImageIterator}
}
\value{
antsImageIterator
}
\description{
advance iterator forward
}
\examples{
img <- makeImage(c(5, 5), rnorm(25))
it <- antsImageIterator(img)
it <- antsImageIteratorNext(it)
}
