% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsImageIterator_class.R
\name{antsImageIteratorPrevious}
\alias{antsImageIteratorPrevious}
\title{antsImageIteratorPrevious}
\usage{
antsImageIteratorPrevious(x)
}
\arguments{
\item{x}{antsImageIterator}
}
\value{
antsImageIterator
}
\description{
advance iterator backward
}
\examples{
img <- makeImage(c(5,5), rnorm(25))
it <- antsImageIterator( img )
it <- antsImageIteratorPrevious( it )
}

