% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsImageIterator_class.R
\name{antsImageIteratorRemaining}
\alias{antsImageIteratorRemaining}
\title{antsImageIteratorRemaining}
\usage{
antsImageIteratorRemaining(x)
}
\arguments{
\item{x}{antsImageIterator}
}
\value{
boolean indicating if data remains
}
\description{
test if iterator is at end of data
}
\examples{
img <- makeImage(c(5, 5), rnorm(25))
it <- antsImageIterator(img)
it <- antsImageIteratorNext(it)
flag <- antsImageIteratorRemaining(it)
}
