% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsImageIterator_class.R
\name{antsImageIteratorSet}
\alias{antsImageIteratorSet}
\title{antsImageIteratorSet}
\usage{
antsImageIteratorSet(x, value)
}
\arguments{
\item{x}{antsImageIterator}

\item{value}{pixel value to set}
}
\value{
TRUE'
}
\description{
Set value at current position
}
\examples{
img <- makeImage(c(5, 5), rnorm(25))
it <- antsImageIterator(img)
flag <- antsImageIteratorSet(it, 2.0)
}
