% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsImageIterator_class.R
\name{antsImageIteratorSetIndex}
\alias{antsImageIteratorSetIndex}
\title{antsImageIteratorSetIndex}
\usage{
antsImageIteratorSetIndex(x, index)
}
\arguments{
\item{x}{antsImageIterator}

\item{index}{index to move to}
}
\value{
TRUE'
}
\description{
move interator to a given index
}
\examples{
img <- makeImage(c(5, 5), rnorm(25))
it <- antsImageIterator(img)
flag <- antsImageIteratorSetIndex(it, c(2, 3))
}
