% Generated by roxygen2 (4.0.2): do not edit by hand
\name{antsImageMutualInformation}
\alias{antsImageMutualInformation}
\title{mutual information between two images}
\usage{
mival<-antsImageMutualInformation( i1, i2 )
}
\arguments{
\item{img}{antsImage}

\item{img}{antsImage}
}
\value{
mutual information value
}
\description{
compute mutual information between two images
}
\examples{
fi<-antsImageRead( getANTsRData('r16') ,2)
mi<-antsImageRead( getANTsRData('r64') ,2)
mival<-antsImageMutualInformation(fi,mi)
}
\author{
Brian B. Avants
}
\keyword{image}
\keyword{information}
\keyword{mutual}
\keyword{similarity,}

