% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_Operators.R
\docType{methods}
\name{antsImage-summary}
\alias{antsImage-summary}
\alias{Summary,antsImage-method}
\alias{mean.antsImage}
\alias{Summary,antsImage-method}
\alias{sd.antsImage}
\title{Summary for antsImage Objects}
\usage{
\S4method{Summary}{antsImage}(x, ..., na.rm = FALSE)

\method{mean}{antsImage}(x, ...)

sd.antsImage(x, ...)
}
\arguments{
\item{x}{is an object of class \code{antsImage}.}

\item{...}{further arguments passed to summary methods}

\item{na.rm}{logical: should missing values be removed?}

\item{x}{is an object of class \code{antsImage}.}
}
\description{
Overloaded Summary for antsImage objects

Overloaded Mean for antsImage objects
}
\details{
#' @rdname antsImagemath
#' @aliases xor,antsImage-method
setMethod(f = "xor", signature(x = "antsImage",
                             y = "antsImage"), definition = function(x) {
  xx = as.array(x)
  yy = as.array(y)
  xor(xx, yy)
  !a2
})
}
