% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsImageTypeCast.R
\name{antsImageTypeCast}
\alias{antsImageTypeCast}
\title{Cast images to either a passed pixeltype or the most precise of the inputs}
\usage{
antsImageTypeCast(imageList, pixeltype = NA)
}
\arguments{
\item{imageList}{list of images}

\item{pixeltype}{to cast to}
}
\value{
list of antsImages
}
\description{
Cast images to either a passed pixeltype or the most precise of the inputs
}
\examples{

img <- antsImageRead(getANTsRData("rand")) \%>\% resampleImage(c(32, 32))
img2 <- antsImageRead(getANTsRData("rand"), pixeltype = "double") \%>\%
  resampleImage(c(32, 32))
allDoubles <- antsImageTypeCast(list(img, img2))
floats <- antsImageTypeCast(list(img))
pixeltype(img) <- "unsigned int"
ints <- antsImageTypeCast(list(img))

}
\author{
J Duda
}
