% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/antsImageWrite.R
\name{antsImageWrite}
\alias{antsImageWrite}
\title{Image Write}
\usage{
antsImageWrite(image, filename)
}
\arguments{
\item{image}{Image object of S4 class 'antsImage' to be written.}

\item{filename}{Name of the file to write the image to.}
}
\value{
0 -- Success\cr 1 -- Failure
}
\description{
Write an image object of S4 class 'antsImage' to a file.
}
\examples{
\dontrun{
# write an image 'img' of class 'antsImage' to a file named 'imagefile.nii'
antsImageWrite( img , "imagefile.nii" )
}
}
\author{
Shrinidhi KL
}

