% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flags.R
\name{antsIncludes}
\alias{antsIncludes}
\title{return ANTs installation information}
\usage{
antsIncludes()
}
\description{
call this to get the include path for ANTs headers and libraries
}
\examples{

antsIncludes()

}
\author{
Avants BB
}
