% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flags.R
\name{antsLibs}
\alias{antsLibs}
\title{return ants library location}
\usage{
antsLibs()
}
\description{
call this to get the path to ants static libaries to which you will link
}
\examples{

antsLibs()

}
\author{
Avants BB
}
