% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsMatrix.R
\docType{class}
\name{antsMatrix-class}
\alias{antsMatrix-class}
\alias{initialize,antsMatrix-method}
\title{antsMatrix Class}
\usage{
\S4method{initialize}{antsMatrix}(.Object, elementtype)
}
\arguments{
\item{.Object}{input object to convert}

\item{elementtype}{string e.g. "float"}

\item{x}{input object to convert}
}
\description{
An S4 class to hold an antsMatrix imported from ITK types
 C++ type used to represent an element of the matrix pointer
 to the actual image
 C++ type 'itk::image< pixeltype , dimension >::Pointer'
}
\section{Slots}{

\describe{
\item{\code{elementtype}}{string of the type of storage of the matrix e.g. "float"}

\item{\code{pointer}}{the memory location}
}}

