% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscaleSVDxpts.R
\name{antspymm_nuisance_names}
\alias{antspymm_nuisance_names}
\title{return nuisance variable strings}
\usage{
antspymm_nuisance_names()
}
\description{
these strings can be used with getNamesFromDataFrame or multigrep to 
either include or exclude nuisance variables.
}
