% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscaleSVDxpts.R
\name{antspymm_predictors}
\alias{antspymm_predictors}
\title{Generate Predictors from ANTsPyMM Imaging Data}
\usage{
antspymm_predictors(demog, doasym = FALSE, return_colnames = FALSE)
}
\arguments{
\item{demog}{A dataframe containing demographic and imaging data.}

\item{doasym}{boolean}

\item{return_colnames}{boolean}
}
\value{
A dataframe with processed demographic and imaging data.
}
\description{
This function generates a list of variable names to be used as predictors
in a model, based antspymm tabular version of imaging data.
It filters and processes the data to create meaningful predictors. LRAVG
}
\examples{
# predictors <- antspymm_predictors(demographic_data)
}
