% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscaleSVDxpts.R
\name{antspymm_qc_names}
\alias{antspymm_qc_names}
\title{Get Quality Control (QC) Metric Names}
\usage{
antspymm_qc_names()
}
\value{
A character vector containing the names of QC metrics.
}
\description{
This function returns a vector of quality control (QC) metric names used in the \code{antspymm} package. 
These metrics include volume measures, reflection errors, PSNR (Peak Signal-to-Noise Ratio), CNR (Contrast-to-Noise Ratio), 
and various metrics related to motion correction in rsfMRI and DTI.
}
\examples{
qc_names <- antspymm_qc_names()
print(qc_names)
}
