% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANTsR-package.R
\docType{package}
\name{ANTsR}
\alias{ANTsR}
\alias{ANTsR-package}
\title{Advanced Normalization Tools in R}
\description{
ANTsR interfaces state of the art image processing with R statistical methods. The project grew out of the need, at University of Pennsylvania, to develop large-scale analytics pipelines that track provenance from scanner to scientific study. ANTsR wraps an ANTs and ITK C++ core via Rcpp to access these frameworks from within R and support reproducible analyses. Specialized functionality in ANTsR includes image segmentation and registration along imaging specific variations of principal component and canonical correlation analysis.
}
\author{
\strong{Maintainer}: Brian B Avants \email{stnava@gmail.com}

Other contributors:
\itemize{
  \item Benjamin M Kandel \email{NA} [contributor]
  \item Jeff T Duda \email{NA} [contributor]
  \item Philip A Cook \email{NA} [contributor]
  \item Nicholas J Tustison \email{NA} [contributor]
  \item Dorian Pustina \email{NA} [contributor]
  \item Nicholas C Cullen \email{nickcullen31@gmail.com} [contributor]
}

}
\keyword{internal}
