% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscaleSVDxpts.R
\name{antsr_rgcca_baseline}
\alias{antsr_rgcca_baseline}
\title{Multi-view Regularized CCA baseline using RGCCA}
\usage{
antsr_rgcca_baseline(data_matrices, tau = 0.5, scheme = "factorial", ncomp = 3)
}
\arguments{
\item{data_matrices}{A named list of subject × feature matrices}

\item{tau}{Regularization parameter (0 = no regularization, 1 = PCA limit)}

\item{scheme}{RGCCA scheme ('horst', 'factorial', 'centroid')}

\item{ncomp}{Number of components}
}
\description{
Multi-view Regularized CCA baseline using RGCCA
}
