% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsRegistration.R
\name{antsrGetPointerName}
\alias{antsrGetPointerName}
\title{Return the antsImage pointer string}
\usage{
antsrGetPointerName(img)
}
\arguments{
\item{img}{image whose pointer we want}
}
\value{
string
}
\description{
This is a low-level function that may be useful for debugging.  It will
return the pointer name itself from an antsImage object.
}
\examples{

img = antsImageRead( getANTsRData("r16") )
antsrGetPointerName( img )
antsrGetPointerName( antsImageClone( img ))

}
\author{
Avants BB
}
