% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsrimpute.R
\name{antsrimpute}
\alias{antsrimpute}
\title{antsrimpute}
\usage{
antsrimpute(mydat, FUN = mean, ...)
}
\arguments{
\item{mydat}{Data frame to be imputed.}

\item{FUN}{Method to be used for imputation (defaults to mean).}

\item{...}{Additional parameters to pass to FUN.}
}
\value{
Returns mydat with NA's replaced by imputed numbers.
}
\description{
Impute NA's on data frame.
}
\details{
Imputes NA's on data frame (column-wise), using a user-specified function
(mean is default).  also works on a vector.
}
\examples{

mydat <- data.frame(A=c(1,2,4,5), B=c(1,NA,4,5))
mean.impute <- antsrimpute(mydat)
median.impute <- antsrimpute(mydat, median)

}
\author{
Kandel BM, Avants BB
}

