% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsrMetric_class.R
\name{antsrMetricGetDerivative}
\alias{antsrMetricGetDerivative}
\title{antsrMetricGetDerivative}
\usage{
antsrMetricGetDerivative(metric)
}
\arguments{
\item{metric}{an 'antsrMetric'}
}
\value{
image similarity value
}
\description{
get derivative of image metric
}
\examples{
x <- antsImageRead(getANTsRData("r16"))
y <- antsImageRead(getANTsRData("r30"))
metric <- antsrMetricCreate(x, y, type = "MeanSquares")
tx <- createAntsrTransform(precision = "double", type = "AffineTransform", dimension = 2)
setAntsrTransformParameters(tx, c(0, -1, 1, 0, 0, 0))
setAntsrTransformFixedParameters(tx, c(128, 128))
antsrMetricSetMovingTransform(metric, tx)
metricValue <- antsrMetricGetDerivative(metric)
}
