% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsrMetric_class.R
\name{antsrMetricGetValue}
\alias{antsrMetricGetValue}
\title{antsrMetricGetValue}
\usage{
antsrMetricGetValue(metric)
}
\arguments{
\item{metric}{an 'antsrMetric'}
}
\value{
image similarity value
}
\description{
get current value of metric
}
\examples{
x =  antsImageRead( getANTsRData( 'r16' ))
y =  antsImageRead( getANTsRData( 'r30' ))
metric = antsrMetricCreate(x,y,type="MeanSquares")
metricValue = antsrMetricGetValue(metric)
}
