% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsrMetric_class.R
\name{antsrMetricSetFixedTransform}
\alias{antsrMetricSetFixedTransform}
\title{antsrMetricSetFixedTransform}
\usage{
antsrMetricSetFixedTransform(metric, transform)
}
\arguments{
\item{metric}{an 'antsrMetric'}

\item{transform}{an 'antsrTransform'}
}
\description{
set transform for fixed image
}
\examples{
x =  antsImageRead( getANTsRData( 'r16' ))
y =  antsImageRead( getANTsRData( 'r30' ))
metric = antsrMetricCreate(x,y,type="ANTSNeighborhoodCorrelation")
tx <- createAntsrTransform( precision="double", type="AffineTransform", dimension=2)
setAntsrTransformParameters(tx, c(0,-1,1,0,0,0))
setAntsrTransformFixedParameters(tx, c(128,128))
antsrMetricSetFixedTransform(metric, tx)
antsrMetricGetValue(metric)
}
