% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsrMetric_class.R
\name{antsrMetricSetMovingImageMask}
\alias{antsrMetricSetMovingImageMask}
\title{antsrMetricSetMovingImageMask}
\usage{
antsrMetricSetMovingImageMask(metric, image)
}
\arguments{
\item{metric}{an 'antsrMetric'}

\item{image}{the moving 'antsImage'}
}
\description{
set moving image for image to image metric
}
\note{
After calling this, must call antsrMetricInitialize(metric)
}
\examples{
x =  antsImageRead( getANTsRData( 'r16' ))
y =  antsImageRead( getANTsRData( 'r30' ))
metric = antsrMetricCreate(x,y,type="MeanSquares")
z = getMask(y)
antsrMetricSetMovingImageMask(metric, z)
}
