% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsrMetric_class.R
\name{antsrMetricSetNumberOfHistogramBins}
\alias{antsrMetricSetNumberOfHistogramBins}
\title{antsrMetricSetNumberOfHistogramBins}
\usage{
antsrMetricSetNumberOfHistogramBins(metric, nBins)
}
\arguments{
\item{metric}{an 'antsrMetric' of type 'MattesMutualInformation' or 'JointHistogramMutualInformation'}

\item{nBins}{number of bins (minimum is 5 even for binary data)}
}
\description{
set histogram bins image to image metric
}
\examples{
x =  antsImageRead( getANTsRData( 'r16' ))
y =  antsImageRead( getANTsRData( 'r30' ))
metric = antsrMetricCreate(x,y,type="MattesMutualInformation")
antsrMetricSetNumberOfHistogramBins(metric,12)
}
