% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsrMetric_class.R
\name{antsrMetricSetRadius}
\alias{antsrMetricSetRadius}
\title{antsrMetricSetRadius}
\usage{
antsrMetricSetRadius(metric, radius)
}
\arguments{
\item{metric}{an 'antsrMetric' of type 'ANTSNeighborhoodCorrelation'}

\item{radius}{radius of neighborhood}
}
\description{
set neighborhood radius
}
\examples{
x =  antsImageRead( getANTsRData( 'r16' ))
y =  antsImageRead( getANTsRData( 'r30' ))
metric = antsrMetricCreate(x,y,type="ANTSNeighborhoodCorrelation")
antsrMetricSetRadius(metric,5)
}
