% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsrMotionCalculation.R
\name{antsrMotionCalculation}
\alias{antsrMotionCalculation}
\title{antsrMotionCalculation}
\usage{
antsrMotionCalculation(img, fixed = NA, mask = NA,
  typeofTransform = "Rigid", verbose = FALSE)
}
\arguments{
\item{img}{antsImage, usually ND where D=4.}

\item{fixed}{Fixed image to register all timepoints to.  If not provided, mean image is used.}

\item{mask}{mask for image (ND-1).  If not provided, estimated from data.}

\item{typeofTransform}{One of \code{"Affine"}, \code{"Rigid"}, or
\code{"SyN"}.}

\item{verbose}{enables verbose output.}
}
\value{
List containing:
\itemize{
 \item{moco_img}{ Motion corrected time-series image.}
 \item{moco_params}{ Data frame of translation parameters.}
 \item{moco_avg_img}{ Average motion-corrected image.}
 \item{moco_mask}{ Mask used to calculate framewise displacement.}
 \item{fd}{ Time-series mean and max displacements.}
 \item{dvars}{ DVARS, derivative of frame-wise intensity changes.}
}
}
\description{
Correct time-series data for motion.
}
\examples{
set.seed(120)
simimg<-makeImage(rep(5,4), rnorm(5^4))
antsrMotionCalculation(simimg,moreaccurate=0)
}
\author{
Brian B. Avants, Benjamin M. Kandel
}

