% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsrSurf.R
\name{antsrSurf}
\alias{antsrSurf}
\title{Render a function onto a surface.}
\usage{
antsrSurf(x, y, quantlimits = c(0.1, 0.9), colormap = "jet",
  inflationFactor = 25, rotationParams = c(270, 0, 270), verbose = FALSE)
}
\arguments{
\item{x}{input antsImage defining the surface on which to render}

\item{y}{input antsImage list defining the function to render
on the surface. these image(s) should be in the same space as \code{surf}.}

\item{quantlimits}{lower and upper quantile limits for overlay}

\item{colormap}{character, one of: grey, red, green, blue, copper, jet, hsv,
spring, summer, autumn, winter, hot, cool, overunder, custom}

\item{inflationFactor}{number of inflation iterations to run}

\item{rotationParams}{3 Rotation angles expressed in degrees}

\item{verbose}{prints the command used to call \code{antsSurf}}
}
\value{
no output
}
\description{
Use a system call to \code{antsSurf} in order to render a function onto
a surface.  This requires \code{antsSurf} to be in the enviornment as well
as \code{ConvertScalarImageToRGB}, both available in Advanced Normalization
Tools software on github.
}
\examples{

\dontrun{
fn = 'ADNI_137_S_0158_MR_MPR__GradWarp__N3__Scaled_Br_20070306171702344_S20209_I42985BrainSegmentation.nii.gz'
img = antsImageRead( fn ) # see antsSurf on github for data
wm   = thresholdImage( img, 3, 4 )
# just the surface
antsrSurf( x=wm, rotationParams = c(270, 0, 90) )
# surface and one overlay
wm = thresholdImage(img, 3, 4) \%>\% iMath("FillHoles")
wms = smoothImage( wm, 1.0 )
wmk = weingartenImageCurvature( wms, 1.5, 0 )
antsrSurf( x=wm, y = list( wmk ), rotationParams = c(270, 0, 90), quantlimits=c(-0.5,0.5) )
# surface and two overlays
blob = antsImageRead( "manualBlob.nii.gz" ) \%>\% smoothImage( 1 )
antsrSurf( x=wm, y = list( wmk, blob ), colormap=c("red","blue"),
  rotationParams = c(270, 0, 90),  quantlimits=c(0.1,0.9))

}

}
\author{
Avants BB, Tustison NJ
}

