% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsrTransform_class.R
\name{antsrTransformFromDisplacementField}
\alias{antsrTransformFromDisplacementField}
\title{getAntsrTransformParameters}
\usage{
antsrTransformFromDisplacementField(field)
}
\arguments{
\item{field}{deformation field (multiChannel image)}
}
\value{
antsrTransform'
}
\description{
Convert deformation field (multiChannel image) to antsrTransform
}

