% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsTransform_class.R
\name{antsTransformCompose}
\alias{antsTransformCompose}
\title{antsTransformCompose}
\usage{
antsTransformCompose(transformList)
}
\arguments{
\item{transformList}{a list of antsTransforms in the reverse order they should be applied}
}
\value{
antsTransform of type "CompositeTransform"
}
\description{
compose multiple transforms
}
\examples{
tx = new("antsTransform", precision="float", type="AffineTransform", dimension=2 )
antsTransformSetParameters(tx, c(0,-1,1,0,0,0) )
tx2 = new("antsTransform", precision="float", type="AffineTransform", dimension=2 )
antsTransformSetParameters(tx2, c(0,-1,1,0,0,0) )
tx3 = antsTransformCompose( list(tx, tx2) )
}

