% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsTransform_class.R
\name{antsTransformCreate}
\alias{antsTransformCreate}
\title{antsTransformCreate}
\usage{
antsTransformCreate(type = "AffineTransform", precision = "float",
  dimension = 3, matrix = NA, offset = NA, center = NA,
  translation = NA, parameters = NA, fixed.parameters = NA,
  displacement.field = NA, supported.types = FALSE)
}
\arguments{
\item{type}{type of transforms}

\item{precision}{numerical precision}

\item{dimension}{spatial dimension of transform}

\item{matrix}{matrix for linear transforms}

\item{offset}{offset for linear transforms}

\item{center}{center for linear transforms}

\item{translation}{translation for linear transforms}

\item{parameters}{array of parameters}

\item{fixed.parameters}{array of fixed parameters}

\item{displacement.field}{multichannel antsImage for non-linear transform}

\item{supported.types}{flag that returns array of possible transforms types}
}
\value{
antsTransform or list of transform types
}
\description{
Create and initialize a antsTransform
}
\examples{
trans= c(3,4,5)
tx = antsTransformCreate( type="Euler3DTransform", translation=trans )
}

