% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsTransform_class.R
\name{antsTransformGetParameters}
\alias{antsTransformGetParameters}
\title{antsTransformGetParameters}
\usage{
antsTransformGetParameters(transform)
}
\arguments{
\item{transform}{antsTransform}
}
\value{
array of parameters'
}
\description{
Get parameters of transform
}
\examples{
tx = new("antsTransform")
params = antsTransformGetParameters(tx)
}

