% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsTransformPoints.R
\name{antsTransformIndexToPhysicalPoint}
\alias{antsTransformIndexToPhysicalPoint}
\title{Get Spatial Point from Index}
\usage{
antsTransformIndexToPhysicalPoint(x, index)
}
\arguments{
\item{x}{image object of S4 class \code{antsImage} to get values from.}

\item{index}{image index}
}
\value{
array of pixel values
}
\description{
Get spatial point from index of an \code{antsImage}.
}
\examples{
img <- makeImage(c(10, 10), rnorm(100))
pt <- antsTransformIndexToPhysicalPoint(img, c(2, 2))
arr <- as.array(img)
testthat::expect_error(
  antsTransformIndexToPhysicalPoint(arr, c(2, 2)), "antsImage"
)
testthat::expect_error(
  antsTransformIndexToPhysicalPoint(img, c("2", 2)), "index must be"
)
testthat::expect_error(
  antsTransformIndexToPhysicalPoint(img, c(2, 2, 2)), "matrix must be of"
)

}
