% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsTransform_class.R
\name{antsTransformInverse}
\alias{antsTransformInverse}
\title{antsTransformInverse}
\usage{
antsTransformInverse(transform)
}
\arguments{
\item{transform}{trasform to invert}
}
\value{
antsTransform
}
\description{
invert a linear antsTransform
}
\examples{
img <- antsImageRead(getANTsRData("r16"))
tx = new("antsTransform", precision="float", type="AffineTransform", dimension=2 )
antsTransformSetParameters(tx, c(0,-1,1,0,dim(img)[1],0) )
img2 = antsApplyTransformToImage(tx, img, img)
}

