% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsTransformPoints.R
\name{antsTransformPhysicalPointToIndex}
\alias{antsTransformPhysicalPointToIndex}
\title{Get Index from Spatial Point}
\usage{
antsTransformPhysicalPointToIndex(x, point)
}
\arguments{
\item{x}{Image object of S4 class 'antsImage' to get values from.}

\item{point}{image physical point}
}
\value{
array of pixel values
}
\description{
Get index from spatial point of an 'antsImage'.
}
\examples{
img<-makeImage(c(10,10),rnorm(100))
pt<-antsTransformPhysicalPointToIndex(img,c(2,2))
arr = as.array(img)
testthat::expect_error(
antsTransformPhysicalPointToIndex(arr,c(2,2)), "antsImage")
testthat::expect_error(
antsTransformPhysicalPointToIndex(img,c("2",2)), "point must be")
testthat::expect_error(
antsTransformPhysicalPointToIndex(img,c(2,2,2)), "matrix must be of")

}
