% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsTransform_class.R
\name{antsTransformRead}
\alias{antsTransformRead}
\title{antsTransformRead}
\usage{
antsTransformRead(filename, dimension = 3, precision = "float")
}
\arguments{
\item{filename}{filename of transform}
}
\value{
antsTransform
}
\description{
read a transform from file
}
\examples{
img <- antsImageRead(getANTsRData("r16"))
tx = new("antsTransform", precision="float", type="AffineTransform", dimension=2 )
antsTransformSetParameters(tx, c(0,-1,1,0,dim(img)[1],0) )
img2 = antsApplyTransformToImage(tx, img, img)
}

