% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsTransform_class.R
\name{antsTransformSetParameters}
\alias{antsTransformSetParameters}
\title{antsTransformSetParameters}
\usage{
antsTransformSetParameters(tx, parameters)
}
\arguments{
\item{tx}{antsTransform}

\item{parameters}{array of parameters'}
}
\value{
TRUE
}
\description{
Set parameters of transform
}
\examples{
tx = new("antsTransform")
params = antsTransformGetParameters(tx)
antsTransformSetParameters(tx, params*2)
}

