% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply.antsImage.R
\name{apply.antsImage}
\alias{apply.antsImage}
\title{Extension of apply to allow input of antsImage}
\usage{
\method{apply}{antsImage}(X, MARGIN, FUN, ...)
}
\arguments{
\item{X}{an antsImage}

\item{MARGIN}{a vector giving the subscripts which the function will be applied over.}

\item{FUN}{the function to be applied..}

\item{...}{optional arguments to FUN}
}
\value{
output is antsImage of lower dimnesion than input
}
\description{
Extension of apply to allow input of antsImage
}
\examples{

img <- makeImage(c(4,4,4), rnorm(4*4*4))
img2 <- apply.antsImage(img, c(1,2), mean)

}
\author{
Duda JT
}

