% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply.antsImage.R
\name{apply}
\alias{apply}
\alias{apply.default}
\alias{apply.antsImage}
\title{Extension of apply to allow input of antsImage}
\usage{
apply(X, MARGIN, FUN, ...)

\method{apply}{default}(X, MARGIN, FUN, ...)

\method{apply}{antsImage}(X, MARGIN, FUN, ...)
}
\arguments{
\item{X}{an antsImage}

\item{MARGIN}{a vector giving the subscripts which the function will be applied over.}

\item{FUN}{the function to be applied.}

\item{...}{optional arguments to FUN}
}
\value{
output is antsImage of lower dimnesion than input
}
\description{
Returns an antsImage after applying function.
}
\examples{

img <- makeImage(c(4, 4, 4), rnorm(4 * 4 * 4))
img2 <- apply(img, c(1, 2), mean)
is.antsImage(img2)

}
\author{
Duda JT
}
