% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsrTransform_class.R
\name{applyAntsrTransform}
\alias{applyAntsrTransform}
\title{applyAntsrTransform}
\usage{
applyAntsrTransform(transform, data, dataType = "point", reference = NA,
  ...)
}
\arguments{
\item{transform}{antsrTransform}

\item{data}{data to transform}

\item{dataType}{data type for non-antsImage data. Either "point" or "vector"}

\item{reference}{target space for transforming an antsImage}

\item{...}{options passed to `applyAntsrTransformToImage`}
}
\value{
transformed data
}
\description{
Apply transform to point, vector or antsImage data
}
\examples{
tx = createAntsrTransform(dimension=2, precision="float", type="AffineTransform")
params = getAntsrTransformParameters(tx)
setAntsrTransformParameters(tx, params*2)
pt2 = applyAntsrTransform(tx, c(1,2,3))
}
