% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsrTransform_class.R
\name{applyAntsrTransformToPoint}
\alias{applyAntsrTransformToPoint}
\title{applyAntsrTransformToPoint}
\usage{
applyAntsrTransformToPoint(transform, points)
}
\arguments{
\item{transform}{antsrTransform}

\item{points}{a matrix which each row is a spatial point}
}
\value{
array of coordinates
}
\description{
Apply transform to spatial point
}
\examples{
tx <- new("antsrTransform")
params <- getAntsrTransformParameters(tx)
setAntsrTransformParameters(tx, params * 2)
pt2 <- applyAntsrTransformToPoint(tx, c(1, 2, 3))
}
