% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsrTransform_class.R
\name{applyAntsrTransformToVector}
\alias{applyAntsrTransformToVector}
\title{applyAntsrTransformToVector}
\usage{
applyAntsrTransformToVector(transform, vector)
}
\arguments{
\item{transform}{antsrTransform}

\item{vector}{array of vector coordinates}
}
\value{
array of coordinates
}
\description{
Apply transform to spatial vector
}
\examples{
\dontrun{
vec2 = applyAntsrTransformToVector(transform, c(1,2,3))
}
}
