\name{as.antsImage-methods}
\docType{methods}
\alias{as.antsImage-methods}
\alias{as.antsImage,array,character,numeric,numeric-method}
\alias{as.antsImage,matrix,character,numeric,numeric-method}
\title{ Methods for Function \code{as.antsImage} in Package \pkg{ANTsR} }
\description{
Create a an object of class 'antsImage' from an \code{array} or \code{matrix}. See \code{?as.antsImage} for details.
}
\section{Methods}{
\describe{

\item{\code{signature(object = "array", pixeltype = "character", spacing = "numeric", origin = "numeric")}}{
Create a an object of class 'antsImage' from an \code{array}. See \code{?as.antsImage} for details.
}

\item{\code{signature(object = "matrix", pixeltype = "character", spacing = "numeric", origin = "numeric")}}{
Create a an object of class 'antsImage' from an \code{matrix}. See \code{?as.antsImage} for details.
}
}
}

\keyword{methods}
