\name{as.antsMatrix-methods}
\docType{methods}
\alias{as.antsMatrix-methods}
\alias{as.antsMatrix,data.frame,character-method}
\alias{as.antsMatrix,list,character-method}
\title{ Methods for Function \code{as.antsMatrix} in Package \pkg{ANTsR} }
\description{
Create an object of class \code{antsMatrix} from a \code{data.frame} or \code{list}. See \code{?as.antsMatrix} for details.
}
\section{Methods}{
\describe{

\item{\code{signature(object = "data.frame", elementtype = "character")}}{
Create an object of class \code{antsMatrix} from a \code{data.frame}. See \code{?as.antsMatrix} for details.
}

\item{\code{signature(object = "list", elementtype = "character")}}{
Create an object of class \code{antsMatrix} from a \code{list}. See \code{?as.antsMatrix} for details.
}
}
}

\keyword{methods}
