% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsImage_class.R, R/antsMatrix.R, R/c.R
\docType{methods}
\name{dim,antsImage-method}
\alias{dim,antsImage-method}
\alias{is.na,antsImage-method}
\alias{.}
\alias{as.numeric,antsImage-method}
\alias{as.matrix,antsImage-method}
\alias{as.matrix.antsImage}
\alias{as.array,antsImage-method}
\alias{as.array.antsImage}
\alias{as.data.frame,antsMatrix-method}
\alias{as.data.frame.antsMatrix}
\alias{as.matrix,antsMatrix-method}
\alias{as.list,antsMatrix-method}
\alias{c,antsImage-method}
\alias{c,antsImage,NULL,ANY,ANY-method}
\title{Coerce antsImage objects to array}
\usage{
\S4method{dim}{antsImage}(x)

\S4method{is.na}{antsImage}(x)

\S4method{as.numeric}{antsImage}(x, mask = logical(),
  region = new("antsRegion", index = integer(), size = integer()))

\S4method{as.matrix}{antsImage}(x, mask = logical(),
  region = new("antsRegion", index = integer(), size = integer()))

as.matrix.antsImage(x, mask = logical(), region = new("antsRegion", index =
  integer(), size = integer()))

\S4method{as.array}{antsImage}(x, mask = logical(),
  region = new("antsRegion", index = integer(), size = integer()))

as.array.antsImage(x, mask = logical(), region = new("antsRegion", index =
  integer(), size = integer()))

\S4method{as.data.frame}{antsMatrix}(x)

as.data.frame.antsMatrix(x)

\S4method{as.matrix}{antsMatrix}(x)

\S4method{as.list}{antsMatrix}(x)

\S4method{c}{antsImage}(x, ..., recursive = FALSE)
}
\arguments{
\item{x}{object of class \code{\link{antsImage}} or \code{\link{antsMatrix}}}

\item{...}{additional objects}

\item{recursive}{if \code{TRUE}, the function recursively descends 
through lists (and pairlists) combining all their elements into a vector.}

\item{x}{antsImage object}
}
\description{
Converts antsImage, antsImage object to different data types

Extracts the array of an \code{antsImage} object then
converts it into a vector
}
