% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/antsMatrix.R
\name{as.data.frame.antsMatrix}
\alias{as.data.frame.antsMatrix}
\title{convert antsMatrix to data frame}
\usage{
as.data.frame.antsMatrix(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{input matrix}

\item{row.names}{}

\item{optional}{}
}
\value{
data frame is output
}
\description{
convert antsMatrix to data frame
}
\author{
Avants BB
}

