% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/antsMatrix.R
\name{as.list.antsMatrix}
\alias{as.list.antsMatrix}
\title{convert antsMatrix to list}
\usage{
as.list.antsMatrix(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{input antsmatrix}

\item{row.names}{}

\item{optional}{}
}
\value{
list is output
}
\description{
convert antsMatrix to list
}
\author{
Avants BB
}

