% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/aslAveraging.R
\name{aslAveraging}
\alias{aslAveraging}
\title{Average ASL tag-control pairs to estimate perfusion}
\usage{
aslAveraging(asl, mask = NA, nuisance = NA, method = "regression")
}
\arguments{
\item{asl}{input asl image}

\item{mask}{in which to calculate perfusion}

\item{nuisance}{nuisance covariates to include in regression}

\item{method}{method to use for computing average.  One of \code{sincSubtract},
 \code{simpleSubtract}, \code{regression},
 or \code{bayesian}. See \code{Details}.}
}
\description{
This function averages arterial spin labeling (ASL) functional MRI
tag-control image pairs to estimate perfusion.
}
\examples{
nvox <- 5 * 5 * 5 * 10
dims <- c(5, 5, 5, 10)
voxvals <- array(rnorm(nvox) + 500, dim=dims)
voxvals[, , , 5] <- voxvals[, , , 5] + 600
asl <- makeImage(dims, voxvals) \%>\% iMath("PadImage", 2)
censored <- aslCensoring(asl)
avg <- aslAveraging(censored$asl.inlier)
}
\author{
Kandel BM, Avants BB
}

