% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/aslCensoring.R
\name{aslCensoring}
\alias{aslCensoring}
\title{Censor bad volumes from ASL data.}
\usage{
aslCensoring(asl, mask = NA, nuis = NA, method = "outlier")
}
\arguments{
\item{asl}{input asl image}

\item{mask}{mask for calculating perfusion}

\item{method}{one of 'outlier', 'robust', or 'scor'.  See \code{Details}.}
}
\value{
vector of the same length as number of timepoints in \code{asl}, with
 1 indicating the corresponding timepoint is included and 0 indicating exclusion.
}
\description{
Censor bad volumes from ASL data.
}
\examples{
nvox <- 5 * 5 * 5 * 10
dims <- c(5, 5, 5, 10)
asl <- makeImage(dims, rnorm(nvox) + 500) \%>\% iMath("PadImage", 2)
censored <- aslCensoring(asl)
}
\author{
Kandel BM
}

