% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/aslOutlierRejection.R
\name{aslOutlierRejection}
\alias{aslOutlierRejection}
\title{Pair-wise subtraction based outlier rejection.}
\usage{
aslOutlierRejection(asl, mask=NA, centralTendency=median, sigma.mean=2.5, sigma.sd=2.0)
}
\arguments{
\item{asl}{ASL time-series image.}

\item{mask}{Mask to compute average over. If not supplied, estimated from data.}

\item{centralTendency}{Function to compute central tendency.  Defaults to median, but mean may also work.}

\item{sigma.mean}{Scalar defining how many standard deviations away from the mean a volume must be to be rejected.}

\item{sigma.sd}{Scalar defining how many standard deviations away from mean of standard deviations the volume must be to be rejected.}
}
\value{
List of ASL time-series image including only inlier pairs and vector of outlier pair numbers (numbers correspond to *input* ASL time-series).
}
\description{
Performs outlier rejection based on divergence from mean difference.
}
\details{
Computes pair-wise differences in ASL time-series data and removes outliers.  One of two criteria is sufficient to classify an image as an outlier: 1) A difference image for which the mean is higher than the mean of all the images plus \code{sigma.mean} times the standard deviation of the whole time-series; 2) A difference image which has a standard deviation greater than sigam.sd times the mean standard deviation of each difference image.
}
\examples{
\dontrun{ #WIP }
}
\author{
Benjamin M. Kandel
}
\references{
Tan H. et al., ``A Fast, Effective Filtering Method for Improving Clinical Pulsed Arterial Spin Labeling MRI,'' JMRI 2009.
}

