% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/aslPerfusion.R
\name{aslPerfusion}
\alias{aslPerfusion}
\title{ASL-based Perfusion from PASL, CASL or pCASL.}
\usage{
aslPerfusion(asl, maskThresh = 0.75, moreaccurate = 1, dorobust = 0.92,
  m0 = NA, skip = 20, mask = NA, interpolation = "linear",
  checkmeansignal = 100, moco_results = NULL, regweights = NULL,
  useDenoiser = NA, useBayesian = 0, verbose = FALSE, ncompcor = 0,
  N3 = FALSE)
}
\arguments{
\item{asl_antsr_image_or_filename}{The filename to an antsr image or pointer
to an antsr image}

\item{maskThresh=<value>}{A thresholding value for the mask as a fraction of
the mean value of the image ... may need to be adjusted up or down for your
data although the mask does not need to be perfect.}

\item{dorobust=<valuein0to1>}{Controls whether you use a robust regression
when estimating perfusion values}
}
\value{
output is an antsImage containing perfusion values

or

1 -- Failure
}
\description{
This function will estimate perfusion from an ASL time series using a
(robust) regression approach. It will do motion correction, compcorr, factor
out nuisance variables and use regression to estimate the perfusion itself.
It will mask the image too based on a simple procedure ( should fix this in
the future by allowing the user to optionally pass a mask in from the
outside ).  WARNING: the function will estimate the m0 image from the mean
of the control tags assuming that the data is acquired T C T C as is most of
JJ's data.  Quantitative CBF can be obtained by mutiplying the output of
this function by a scalar.
}
\examples{
\dontrun{
asl<-antsImageRead("PEDS012_20131101_pcasl_1.nii.gz",4)
# image available at http://files.figshare.com/1701182/PEDS012_20131101.zip
pcasl.processing <- aslPerfusion( asl, moreaccurate=1, interpolation="linear",
  dorobust=0.95 )
}
}
\author{
Avants BB
}

