% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simlr.R
\name{backup_simlr}
\alias{backup_simlr}
\title{backup simlr implementation}
\usage{
backup_simlr(
  data_matrices,
  smoothingMatrices,
  iterations = 500,
  sparsenessQuantiles = NULL,
  positivities = NULL,
  initialUMatrix = NULL,
  mixAlg = c("svd", "ica", "avg", "rrpca-l", "rrpca-s", "pca", "stochastic",
    "newton-schulz", "ica-newton"),
  repeatedMeasures = NA,
  lineSearchRange = c(-500, 500),
  lineSearchTolerance = 1e-12,
  randomSeed = 0,
  constraint = c("nsaflowx0.5x10", "orthox0x1", "Grassmannx0", "Stiefelx0", "none"),
  energyType = c("cca", "regression", "normalized", "ucca", "lowRank",
    "lowRankRegression", "normalized_correlation", "acc", "nc", "dat", "lrr", "reconorm",
    "logcosh", "exp", "kurtosis", "gauss"),
  vmats = NULL,
  connectors = NULL,
  optimizationStyle = c("bidirectional_lookahead", "armijo_gradient", "lookahead",
    "bidirectional_armijo_gradient"),
  scale = c("center", "eigenvalue"),
  expBeta = 0.9,
  jointInitialization = TRUE,
  sparsenessAlg = "soft",
  orthogonalizeU = FALSE,
  domainMatrices = NULL,
  domainLambdas = NULL,
  sparse_gradient = TRUE,
  verbose = FALSE
)
}
\description{
backup simlr implementation
}
