% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_data.R
\docType{data}
\name{basal_forebrain}
\alias{basal_forebrain}
\title{basal_forebrain}
\format{
A data frame with columns: Label, Description, Side.
}
\usage{
basal_forebrain
}
\description{
A data frame label numbers and names for basal forebrain regions.
}
\keyword{datasets}
