% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigLMStats.R
\name{bigLMStats2}
\alias{bigLMStats2}
\title{Efficiently compute a voxel-wise varying linear regression model}
\usage{
bigLMStats2(dataFrame, voxmat, myFormula)
}
\arguments{
\item{dataFrame}{This data frame contains all relevant predictors except for
the matrix associated with the changing variable, heretofore named \code{vox}.}

\item{voxmat}{The matrix that contains the changing predictor named \code{vox}.}

\item{myFormula}{This is a character string that defines a valid regression formula.}
}
\value{
A list of different matrices that contain names derived from the
formula and the coefficients of the regression model.
}
\description{
This function simplifies calculating p-values from linear models in which
there is a similar formula that is applied many times with a change in only
one predictor.  The outcome variable is constant.  The changing variable
should be named \code{vox} in the input formula.
}
\examples{

set.seed(1500)
nsub = 100
outcome = rnorm( nsub )
covar = rnorm( nsub )
mat = replicate( nsub, rnorm( nsub ) )
myform = " outcome ~ covar + vox "
df = data.frame( outcome = outcome, covar = covar )
result = bigLMStats2( df, mat, myform)
print( names( result ) )
print( rownames( result$pValue ) )

}
\author{
BB Avants.
}

