\name{bold_correlation_matrix}
\alias{bold_correlation_matrix}
\docType{data}
\title{bold_correlation_matrix}
\description{
A data frame with an example square BOLD correlation matrix.
}
\usage{data(bold_correlation_matrix)}
\format{
  A data frame square correlation matrix listed by AAL anatomical column names.
}
\references{
\url{http://en.wikipedia.org/wiki/Automated_Anatomical_Labeling}
}
\examples{
data(bold_correlation_matrix)
}
\keyword{datasets}
