% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist.antsImage.R
\name{boxplot.antsImage}
\alias{boxplot.antsImage}
\title{Boxplot of Values in an Image}
\usage{
\method{boxplot}{antsImage}(x, ..., mask)
}
\arguments{
\item{x}{Object of class \code{antsImage}}

\item{...}{Arguments passed to \code{\link{boxplot.default}}}

\item{mask}{object to subset the image.  If missing, then all
values of the image are plotted.}
}
\value{
Output of \code{\link{boxplot}}
}
\description{
Computes the boxplot of values
of an image with the option for a mask.
}
\examples{
img <- makeImage(c(10, 10), rnorm(100))
mask <- img > 0
boxplot(img, mask = mask)
}
