% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bracket_assign.R, R/bracket_subset.R
\docType{methods}
\name{[<-,antsImage,NULL,ANY,ANY-method}
\alias{[<-,antsImage,NULL,ANY,ANY-method}
\alias{[<-,antsImage,NULL-method}
\alias{[<-,antsImage,logical,ANY,ANY-method}
\alias{[<-,antsImage,logical-method}
\alias{[<-,antsImage,array,ANY,ANY-method}
\alias{[<-,antsImage,array-method}
\alias{[<-,antsImage,matrix,ANY,ANY-method}
\alias{[<-,antsImage,matrix-method}
\alias{[<-,antsImage,list,ANY,ANY-method}
\alias{[<-,antsImage,list-method}
\alias{[<-,antsImage,antsImage,ANY,ANY-method}
\alias{[<-,antsImage,antsImage-method}
\alias{[<-,antsImage,NULL,antsRegion,ANY-method}
\alias{[<-,antsImage,NULL,antsRegion-method}
\alias{[<-,antsImage,logical,antsRegion,ANY-method}
\alias{[<-,antsImage,logical,antsRegion-method}
\alias{[<-,antsImage,array,antsRegion,ANY-method}
\alias{[<-,antsImage,array,antsRegion-method}
\alias{[<-,antsImage,matrix,antsRegion,ANY-method}
\alias{[<-,antsImage,matrix,antsRegion-method}
\alias{[<-,antsImage,NULL,NULL,numeric-method}
\alias{[<-,antsImage,numeric,numeric,numeric-method}
\alias{[<-,antsImage,numeric,NULL,numeric-method}
\alias{[<-,antsImage,NULL,numeric,numeric-method}
\alias{[,antsImage,NULL,ANY,ANY-method}
\alias{[,antsImage,antsImage,ANY,ANY-method}
\alias{[,antsImage,logical,ANY,ANY-method}
\alias{[,antsImage,logical,ANY,ANY-method}
\alias{[,antsImage,logical,ANY-method}
\alias{[,antsImage,ANY,ANY,ANY-method}
\alias{[,antsImage,ANY,ANY,ANY-method}
\alias{[,antsImage,ANY,ANY-method}
\alias{[,antsImage,NULL,NULL,ANY-method}
\alias{[,antsImage,NULL,NULL,ANY-method}
\alias{[,antsImage,NULL,NULL-method}
\alias{[,antsImage,numeric,numeric,ANY-method}
\alias{[,antsImage,numeric,numeric,ANY-method}
\alias{[,antsImage,numeric,numeric-method}
\alias{[,antsImage,numeric,NULL,ANY-method}
\alias{[,antsImage,numeric,NULL,ANY-method}
\alias{[,antsImage,numeric,NULL-method}
\alias{[,antsImage,NULL,numeric,ANY-method}
\alias{[,antsImage,NULL,numeric,ANY-method}
\alias{[,antsImage,NULL,numeric-method}
\alias{[,antsImage,missing,numeric,ANY-method}
\alias{[,antsImage,missing,numeric-method}
\alias{[,antsImage,numeric,missing,ANY-method}
\alias{[,antsImage,numeric,missing-method}
\alias{[,antsImage,missing,missing,ANY-method}
\alias{[,antsImage,missing,missing-method}
\title{Extract or Replace Parts of antsImage Object}
\usage{
\S4method{[}{antsImage,`NULL`,ANY,ANY}(x, i, j, ...) <- value

\S4method{[}{antsImage,logical,ANY,ANY}(x, i, j, ...) <- value

\S4method{[}{antsImage,array,ANY,ANY}(x, i, j, ...) <- value

\S4method{[}{antsImage,matrix,ANY,ANY}(x, i, j, ...) <- value

\S4method{[}{antsImage,list,ANY,ANY}(x, i, j, ...) <- value

\S4method{[}{antsImage,antsImage,ANY,ANY}(x, i, j, ...) <- value

\S4method{[}{antsImage,`NULL`,antsRegion,ANY}(x, i, j, ...) <- value

\S4method{[}{antsImage,logical,antsRegion,ANY}(x, i, j, ...) <- value

\S4method{[}{antsImage,array,antsRegion,ANY}(x, i, j, ...) <- value

\S4method{[}{antsImage,matrix,antsRegion,ANY}(x, i, j, ...) <- value

\S4method{[}{antsImage,`NULL`,`NULL`,numeric}(x, i, j, ...) <- value

\S4method{[}{antsImage,numeric,numeric,numeric}(x, i, j, ...) <- value

\S4method{[}{antsImage,numeric,`NULL`,numeric}(x, i, j, ...) <- value

\S4method{[}{antsImage,`NULL`,numeric,numeric}(x, i, j, ...) <- value

\S4method{[}{antsImage,`NULL`,ANY,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{antsImage,antsImage,ANY,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{antsImage,logical,ANY,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{antsImage,logical,ANY,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{antsImage,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{antsImage,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{antsImage,`NULL`,`NULL`,ANY}(x, i, j, k = NA, l = NA, ...,
  drop)

\S4method{[}{antsImage,`NULL`,`NULL`,ANY}(x, i, j, k = NA, l = NA, ...,
  drop)

\S4method{[}{antsImage,numeric,numeric,ANY}(x, i, j, k = NA, l = NA, ...,
  drop)

\S4method{[}{antsImage,numeric,numeric,ANY}(x, i, j, k = NA, l = NA, ...,
  drop)

\S4method{[}{antsImage,numeric,`NULL`,ANY}(x, i, j, k = NA, l = NA, ...,
  drop)

\S4method{[}{antsImage,numeric,`NULL`,ANY}(x, i, j, k = NA, l = NA, ...,
  drop)

\S4method{[}{antsImage,`NULL`,numeric,ANY}(x, i, j, k = NA, l = NA, ...,
  drop)

\S4method{[}{antsImage,`NULL`,numeric,ANY}(x, i, j, k = NA, l = NA, ...,
  drop)

\S4method{[}{antsImage,missing,numeric,ANY}(x, i, j, k = NA, l = NA, ...,
  drop)

\S4method{[}{antsImage,numeric,missing,ANY}(x, i, j, k = NA, l = NA, ...,
  drop)

\S4method{[}{antsImage,missing,missing,ANY}(x, i, j, k = NA, l = NA, ...,
  drop)
}
\arguments{
\item{x}{antsImage}

\item{i}{logical or i-th dimension}

\item{j}{not used or j-th dimension}

\item{...}{not used}

\item{value}{value to assign to the subset in assignment operator}

\item{drop}{method for missing data}

\item{k}{not used or k-th dimension}

\item{l}{not used or l-th dimension}
}
\description{
Operators acting on antsImage objects 
to extract or replace parts.
}
