% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c.R
\docType{methods}
\name{c,antsImage-method}
\alias{c,antsImage-method}
\title{Create antsImage into a Vector}
\usage{
\S4method{c}{antsImage}(x, ..., recursive = FALSE)
}
\arguments{
\item{x}{antsImage object}

\item{...}{additional objects}

\item{recursive}{if \code{TRUE}, the function recursively descends 
through lists (and pairlists) combining all their elements into a vector.}
}
\description{
Extracts the array of an \code{antsImage} object then
converts it into a vector
}
