% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_data.R
\docType{data}
\name{cerebellum_labels}
\alias{cerebellum_labels}
\title{cerebellum_labels}
\format{
A data frame with columns: Label, Description.
}
\usage{
cerebellum_labels
}
\description{
A data frame label numbers and names for cerebellum regions.
}
\keyword{datasets}
