% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ants.R
\docType{methods}
\name{check_ants-methods}
\alias{check_ants-methods}
\alias{check_ants}
\alias{check_ants,antsImage-method}
\alias{check_ants,ANY-method}
\alias{check_ants,character-method}
\alias{check_ants,list-method}
\title{Check if antsImage or read in}
\usage{
check_ants(x, ...)

\S4method{check_ants}{antsImage}(x, ...)

\S4method{check_ants}{ANY}(x, ...)

\S4method{check_ants}{character}(x, ...)

\S4method{check_ants}{list}(x, ...)
}
\arguments{
\item{x}{character path of image or 
an object of class antsImage}

\item{...}{arguments passed to other methods, 
namely \code{\link{antsImageRead}}}
}
\value{
antsImage object
}
\description{
Simple check to see if input is character, list, \code{nifti},
or class \code{antsImage}
}
\seealso{
\code{\link{antsImageRead}}
}
\author{
John Muschelli \email{muschellij2@gmail.com}
}
