% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscaleSVDxpts.R
\name{check_zero_variance}
\alias{check_zero_variance}
\title{Check a matrix for zero-variance rows and columns}
\usage{
check_zero_variance(mat)
}
\arguments{
\item{mat}{A numeric matrix.}
}
\value{
Boolean checking if any zero-variance rows or columns found.
}
\description{
This function checks a numeric matrix for rows or columns
with zero variance. It reports how many and which ones are affected,
using either their names (if present) or indices.
}
\details{
Zero variance means all values in a row or column are identical
(after removing \code{NA}s).
}
\examples{
m <- matrix(c(1, 1, 1,
              2, 3, 4,
              5, 6, 7),
            nrow = 3, byrow = TRUE)
rownames(m) <- c("rowA", "rowB", "rowC")
colnames(m) <- c("colX", "colY", "colZ")
check_zero_variance(m)

}
