% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_data.R
\docType{data}
\name{cit168_brainstem}
\alias{cit168_brainstem}
\title{cit168_brainstem}
\format{
A data frame with columns: Label, Description, Side.
}
\usage{
cit168_brainstem
}
\description{
A data frame label numbers and names for the CIT168 brainstem atlas.
}
\keyword{datasets}
