% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_data.R
\docType{data}
\name{cit168_reinf_learn}
\alias{cit168_reinf_learn}
\title{cit168_reinf_learn}
\format{
A data frame with columns: Label, Description.
}
\usage{
cit168_reinf_learn
}
\description{
A data frame label numbers and names for the CIT168 reinforcement learning atlas.
}
\keyword{datasets}
