% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsrTransform_class.R
\name{composeAntsrTransforms}
\alias{composeAntsrTransforms}
\title{composeAntsrTransforms}
\usage{
composeAntsrTransforms(transformList)
}
\arguments{
\item{transformList}{a list of antsrTransforms in the reverse order they should be applied}
}
\value{
antsrTransform of type "CompositeTransform"
}
\description{
compose multiple transforms
}
\examples{
tx = new("antsrTransform", precision="float", type="AffineTransform", dimension=2 )
setAntsrTransformParameters(tx, c(0,-1,1,0,0,0) )
tx2 = new("antsrTransform", precision="float", type="AffineTransform", dimension=2 )
setAntsrTransformParameters(tx2, c(0,-1,1,0,0,0) )
tx3 = composeAntsrTransforms( list(tx, tx2) )
}
