% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composeDisplacementFields.R
\name{composeDisplacementFields}
\alias{composeDisplacementFields}
\title{composeDisplacementFields}
\usage{
composeDisplacementFields(displacementField, warpingField)
}
\arguments{
\item{displacementField}{displacement field.}

\item{warpingField}{warping field.}
}
\value{
composite displacement field
}
\description{
Compose displacement fields.
}
\author{
NJ Tustison
}
