% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composeTransformsToField.R
\name{composeTransformsToField}
\alias{composeTransformsToField}
\title{Compose arbitrary transform list into a deformation field.}
\usage{
composeTransformsToField(image, transforms)
}
\arguments{
\item{image}{input image defines the transformation domain}

\item{transforms}{list of transform filenames or antrTransforms}
}
\value{
field deformation object is output
}
\description{
A deformation field can capture any domain to domain correspondence that we
can represent in ANTsR.  This function will map a list of transforms into a
single deformation field object.
}
\examples{

fi <- antsImageRead(getANTsRData("r16"))
mi <- antsImageRead(getANTsRData("r64"))
fi <- resampleImage(fi, c(60, 60), 1, 0)
mi <- resampleImage(mi, c(60, 60), 1, 0) # speed up
mytx <- antsRegistration(fixed = fi, moving = mi, typeofTransform = c("SyN"))
compfield <- composeTransformsToField(fi, mytx$fwd)
atx <- antsrTransformFromDisplacementField(compfield)
wrped <- applyAntsrTransformToImage(atx, mi, fi)

}
\author{
Avants BB, Duda JT
}
