% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessRsBOLD.R, R/preprocessfMRI.R
\name{computeDVARS}
\alias{computeDVARS}
\title{computeDVARS}
\usage{
computeDVARS(boldMatrix)

computeDVARS(boldMatrix)
}
\arguments{
\item{boldMatrix}{matrix of bold signal}

\item{boldMatrix}{matrix of bold signal}
}
\value{
DVARS vector.

DVARS vector.
}
\description{
compute the DVARS quality control metric

compute the DVARS quality control metric
}
\examples{

 mat <- matrix(c(0,1,2,0,0,1,2,2,2),ncol=3)
 dv<-computeDVARS(mat)


 mat <- matrix(c(0,1,2,0,0,1,2,2,2),ncol=3)
 dv<-computeDVARS(mat)

}
\author{
Tustison NJ, Avants BB

Tustison NJ, Avants BB
}

