% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growImage.R
\name{confidenceConnected}
\alias{confidenceConnected}
\title{Grow image using confidence connected filter}
\usage{
confidenceConnected(inimg, seed, radius = 3, multiplier = 1.5,
  iterations = 0L)
}
\arguments{
\item{inimg}{Image to grow}

\item{seed}{xyz components to seed the image}

\item{radius}{the radius of the neighborhood over which the statistics are evaluated}

\item{multiplier}{multiplier to define the confidence interval. 
Multiplier can be anything greater than zero. A typical value is 2.5}

\item{iterations}{the number of iterations to repeat the process}
}
\value{
antsImage region grow image
}
\description{
Grow image using confidence connected filter
}
\note{
After this initial segmentation is calculated, the mean and variance are 
re-calculated. All the pixels in the previous segmentation are 
used to calculate the mean the standard deviation (as opposed to using 
the pixels in the neighborhood of the seed point). 
The segmentation is then recalculated using these refined estimates 
for the mean and variance of the pixel values. This process is 
repeated for the specified number of iterations. Setting the 
`iterations` to zero stops the algorithm after the initial 
segmentation from the seed point.
}
\examples{
library(ANTsR)
img <- makeImage(c(20, 20, 20), 0)
img[1:5, 1:5, 1:2] = 2
img[8:10, 9:14, 3:5] = 5
seed = xyz(as.array(img) == 2)
simg <- confidenceConnected(img, seed = seed, iterations = 20L)
}
