% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectedThresholdImage.R
\name{connectedThreshold}
\alias{connectedThreshold}
\title{Grow image using connected threshold filter}
\usage{
connectedThreshold(inimg, seed, upper = 1, lower = 0)
}
\arguments{
\item{inimg}{Image to grow}

\item{seed}{xyz components to seed the image}

\item{upper}{upper threshold value}

\item{lower}{lower threshold value}
}
\value{
antsImage region grow image
}
\description{
Grow image using connected threshold  filter
}
\examples{
library(ANTsR)
img <- makeImage(c(20, 20, 20), rnorm(20^3))
img[1:5, 1:5, 1:2] = 2
img[8:10, 9:14, 3:5] = 5
seed = xyz(as.array(img) == 2)
simg <- connectedThreshold(img, seed = seed)
}
