% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectedThresholdImage.R
\name{connectedThresholdImage}
\alias{connectedThresholdImage}
\title{Grow image using connected threshold filter}
\usage{
connectedThresholdImage(inimg, seed, upper = 1, lower = 0)
}
\arguments{
\item{inimg}{Image to grow}

\item{seed}{xyz components to seed the image}

\item{upper}{upper threshold value}

\item{lower}{lower threshold value}
}
\value{
antsImage region grow image
}
\description{
Grow image using connected threshold  filter
}
\examples{
library(ANTsR)
img <- makeImage(c(20, 20, 20), rnorm(20^3))
img[2:5, 3:5, 4:8] = 2
seed = floor(colMeans(which(as.array(img) == 2, arr.ind = TRUE)))
simg <- connectedThresholdImage(img, seed = seed,
upper = 2.5, lower = 1.5)
# ortho2(simg, xyz=xyz(as.array(simg) ==1))
}
