% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convolveImage.R
\name{convolveImage}
\alias{convolveImage}
\title{convolve one image with another}
\usage{
convolveImage(image, kernelImage, crop = TRUE)
}
\arguments{
\item{image}{antsImage to convolve}

\item{kernelImage}{antsImage acting as kernel}

\item{crop}{boolean automatically crops kernelImage}
}
\value{
convimage
}
\description{
convolves images together
}
\examples{

fi<-antsImageRead( getANTsRData("r16") ,2)
convimg<-makeImage( c(3,3) , c(1,0,1,0,-4,0,1,0,1) )
convout<-convolveImage( fi, convimg )
convimg2<-makeImage( c(3,3) , c(0,1,0,1,0,-1,0,-1,0) )
convout2<-convolveImage( fi, convimg2 )

}
\author{
Brian B. Avants
}
\keyword{convolution}
\keyword{convolve,}
