% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/corw.R
\name{corw}
\alias{corw}
\title{produces a correlation matrix via weighted correlation.}
\usage{
corw(mat, weights)
}
\arguments{
\item{mat}{input matrix}
}
\value{
matrix is output
}
\description{
Uses weighted regression to compute pairwise correlation matrix on input
matrix - by columns.
}
\examples{
mat <- matrix(c(rep(1,100),rep(0,20)),ncol=10)
wcmat<-corw( mat , weights = rep(1,ncol(mat) ) )
}
\author{
Avants BB
}

