% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/CreateJacobianDeterminantImage.R
\name{CreateJacobianDeterminantImage}
\alias{CreateJacobianDeterminantImage}
\title{CreateJacobianDeterminantImage}
\usage{
CreateJacobianDeterminantImage(...)
}
\arguments{
\item{imageDimension2|3}{Number of dimensions of the input tx}

\item{deformationField}{transformation file name}

\item{outputImage}{result image, cloned from a fixed image}

\item{doLog}{return the log jacobian}
}
\value{
0 -- Success\cr 1 -- Failure
}
\description{
Compute the jacobian determinant from a transformation file
}
\examples{
fi<-antsImageRead( getANTsRData('r16') ,2)
mi<-antsImageRead( getANTsRData('r64') ,2)
mytx<-antsRegistration(fixed=fi , moving=mi, typeofTransform = c("SyN") )
jac<-antsImageClone(fi)
CreateJacobianDeterminantImage(2,mytx$fwdtransforms[[1]],jac,1)
plot(jac)
}
\author{
BB Avants
}

