% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createWarpedGrid.R
\name{createWarpedGrid}
\alias{createWarpedGrid}
\title{Create a warped grid for visualization.}
\usage{
createWarpedGrid(img, gridStep = 10, gridWidth = 2,
  gridDirections = c(TRUE, TRUE), fixedReferenceImage = NULL,
  transform = NA, foreground = 1, background = 0)
}
\arguments{
\item{img}{input image}

\item{gridStep}{width of grid blocks}

\item{gridWidth}{width of grid lines}

\item{gridDirections}{directions in which to draw grid lines, boolean vector}

\item{fixedReferenceImage}{reference image space}

\item{transform}{vector of transforms}

\item{foreground}{intensity value for grid blocks}

\item{background}{intensity value for grid lines}
}
\value{
image is output
}
\description{
Deforming a grid is a helpful way to visualize a deformation field. This
function enables a user to define the grid parameters and apply a deformable
map to that grid.
}
\examples{

fi = antsImageRead( getANTsRData( 'r16' ) )
mi = antsImageRead( getANTsRData( 'r64' ) )
mygr = createWarpedGrid( mi )
\dontrun{
mytx <- antsRegistration(fixed=fi, moving=mi, typeofTransform = c('SyN') )
mywarpedgrid = createWarpedGrid( mi, gridDirections=c(F,T),
  transform=mytx$fwdtransforms, fixedReferenceImage=fi )
plot( fi, mywarpedgrid, alpha=0.75, color.overlay='blue' )
}

}
\author{
Avants BB
}
