% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cropImage.R
\name{cropImage}
\alias{cropImage}
\title{crop a sub-image via a mask}
\usage{
cropImage(image, labelImage, label = 1)
}
\arguments{
\item{image}{antsImage to crop}

\item{labelImage}{antsImage with label values.
If not supplied, estimated from data.}

\item{label}{the label value to use}
}
\value{
subimage
}
\description{
uses a label image to crop a smaller image from within a larger image
}
\examples{

fi <- antsImageRead(getANTsRData("r16"))
cropped <- cropImage(fi)
cropped <- cropImage(fi, fi, 100)

}
\author{
Brian B. Avants, Nicholas J. Tustison
}
\keyword{crop,}
\keyword{extract}
\keyword{sub-image}
