% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/croppingAndPaddingUtilities.R
\name{cropImageFromCenterPoint}
\alias{cropImageFromCenterPoint}
\title{Crop a patch from an image around a center point}
\usage{
cropImageFromCenterPoint(image, centerPoint, patchSize)
}
\arguments{
\item{image}{Input ANTs image}

\item{centerPoint}{Physical space coordinates of center point (vector).}

\item{patchSize}{Vector defining patch size.}
}
\value{
A cropped resampled patch image
}
\description{
Crop a patch from an image around a center point
}
\examples{

library( ANTsR )
image <- antsImageRead( getANTsRData( "r16" ) )
center <- c( 128, 128 )
patch <- cropImageFromCenterPoint( image, center, c( 64, 64 ) )

}
\author{
Tustison NJ
}
