% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cropImage.R
\name{cropIndices}
\alias{cropIndices}
\title{crop a sub-image by image indices}
\usage{
cropIndices(image, lowerind, upperind)
}
\arguments{
\item{image}{antsImage to crop}

\item{lowerind}{vector of lower index, should be length image dimensionality}

\item{upperind}{vector of upper index, should be length image dimensionality}
}
\value{
subimage
}
\description{
create a proper antsImage sub-image by indexing the image with indices.
this is similar to but different from array sub-setting in that the resulting
sub-image can be decropped back into its place without having to store its
original index locations explicitly.
}
\examples{
fi <- antsImageRead( getANTsRData("r16") ,2)
cropped <- cropIndices( fi, c(2,10), c(5,15) )
cropped<-smoothImage( cropped, 5 )
decropped<-decropImage( cropped, fi )
}
\author{
Brian B. Avants, Nicholas J. Tustison
}
\keyword{crop,}
\keyword{extract}
\keyword{sub-image}

