% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist.antsImage.R
\name{cut.antsImage}
\alias{cut.antsImage}
\title{Perform Cut on an image}
\usage{
\method{cut}{antsImage}(x, breaks, ..., mask)
}
\arguments{
\item{x}{Object of class \code{antsImage}}

\item{breaks}{either a numeric vector of two or more unique cut points 
or a single number (greater than or equal to 2) 
giving the number of intervals into which x is to be cut. 
Passed to \code{\link{cut}})}

\item{...}{Arguments passed to \code{\link{cut}}}

\item{mask}{object to subset the image.  If missing, then all 
values of the image are used}
}
\value{
Object of class \code{antsImage} with an \code{attribute} of levels
}
\description{
Cuts a numeric image into an integer factor,
 with the option of a mask.
}
\examples{
img = makeImage(c(10,10),rnorm(100))
mask = img > 0
cut(img, mask = mask, breaks = 4)
}
