% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cvEigenanatomy.R
\name{cvEigenanatomy}
\alias{cvEigenanatomy}
\title{Cross-validation method for eigenanatomy decompositions.}
\usage{
cvEigenanatomy(demog, images, outcome, ratio = 10, mask = NA,
  sparseness = 0.01, nvecs = 50, its = 5, cthresh = 250, ...)
}
\arguments{
\item{demog}{Demographics information that includes outcome and (optional)
covariates.}

\item{images}{n by p input image matrix, where n is the number of subjects
and p is the number of voxels.}

\item{outcome}{Name of outcome variable. Must be present in \code{demog}.}

\item{ratio}{If greater than 1, number of folds for cross-validation.  If
less than 1, one testing-training step will be performed, using \code{ratio}
of the data for training and the rest for testing.}

\item{mask}{Mask image of type \code{antsImage}.}

\item{sparseness}{Desired level of sparsity in decomposition.}

\item{nvecs}{Number of eigenvectors to use in decomposition.}

\item{its}{Number of iterations for decomposition.}

\item{cthresh}{Cluster threshold for decomposition.}

\item{...}{Additional options passed to \code{regressProjections}.}
}
\value{
A result, or (if ratio > 1) list of results, from
\code{regressProjection}.
}
\description{
Perform cross-validation on an image set using eigencomponents to predict an
outcome variable.
}
\examples{
\dontrun{
# generate simulated outcome
nsubjects <- 100
x1 <- seq(1, 10, length.out=nsubjects) + rnorm(nsubjects, sd=2)
x2 <- seq(25, 15, length.out=nsubjects) + rnorm(nsubjects, sd=2)
outcome <- 3 * x1 + 4 * x2 + rnorm(nsubjects, sd=1)
# generate simulated images with outcome predicted
# by sparse subset of voxels
voxel.1 <- 3 * x1 + rnorm(nsubjects, sd=2)
voxel.2 <- rnorm(nsubjects, sd=2)
voxel.3 <- 2 * x2 + rnorm(nsubjects, sd=2)
voxel.4 <- rnorm(nsubjects, sd=3)
input   <- cbind(voxel.1, voxel.2, voxel.3, voxel.4)
mask    <- as.antsImage(matrix(c(1,1,1,1), nrow=2))
# generate sample demographics that do not explain outcome
age <- runif(nsubjects, 50, 75)
demog <- data.frame(outcome=outcome, age=age)
result <- cvEigenanatomy(demog, input, "outcome", ratio=5, mask,
            sparseness=0.25, nvecs=4)
}
}
\author{
Kandel BM and Avants B
}

