% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simlr.R
\name{decode_pymm_names}
\alias{decode_pymm_names}
\title{Decode ANTsPyMM technical shorthand to legible anatomical names}
\usage{
decode_pymm_names(x)
}
\arguments{
\item{x}{A character vector of shorthand names to decode.}
}
\value{
A character vector of legible anatomical names.
}
\description{
This function uses consolidated atlas data (AAL, DKT, CIT168, JHU, etc.) to
translate technical shorthand strings into their original, legible anatomical
descriptions. It performs normalization to handle variations in dots, underscores,
and case.
}
\examples{
decode_pymm_names("cit168.bn.str.pu.left")
decode_pymm_names("dti.fa.thalamus.left.jhu.icbm.labels.1mm")
}
