% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bloodPerfusionSVD.R
\name{deconvolutionSVD}
\alias{deconvolutionSVD}
\title{Regularize SVD (deconvolution) solution of cerebral blood flow.}
\usage{
deconvolutionSVD(arterialInputFunction, thresholdSVD = 0.2)
}
\arguments{
\item{arterialInputFunction}{vector specifying the arterial input function
over time.}

\item{thresholdSVD}{is used to threshold the smaller elements of the
diagonal matrix during the SVD regularization.  0.2 is a common choice
(cf. page 571, end of column 2 in
http://www.ncbi.nlm.nih.gov/pubmed/16971140).}
}
\value{
regularized residue function, i.e. the right-hand side of
        CBF * dT * R = ( V * L^-1 * U^T ) * Cvoi.
}
\description{
Ostergaard's regularization approach for a model independent solution of
cerebral blood flow using a blood pool contrast agent
(http://www.ncbi.nlm.nih.gov/pubmed/8916023).
}
\examples{

S0aif <- 17.62
Saif <- c( 16.25, 16.37, 20.22, 78.96, 230.5, 249.79, 198.58, 147.76,
110.39, 111.64, 129.43 )
Caif <- -log( Saif / S0aif )

dSVD <- deconvolutionSVD( Caif, 0.2 )

}
\author{
Tustison NJ
}
