% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cropImage.R
\name{decropImage}
\alias{decropImage}
\title{decrop a sub-image back into the full image}
\usage{
decropImage(croppedImage, fullImage)
}
\arguments{
\item{croppedImage}{cropped antsImage}

\item{fullImage}{antsImage to put back into}
}
\value{
decroppedImage
}
\description{
the inverse function for \code{cropImage}
}
\examples{
fi <- antsImageRead( getANTsRData("r16"))
mask <- getMask( fi )
cropped <- cropImage( fi, mask, 1 )
cropped <- smoothImage( cropped, 1)
decropped <- decropImage( cropped , fi )
}
\author{
Brian B. Avants, Nicholas J. Tustison
}
\keyword{decrop,}
\keyword{extract}
\keyword{sub-image}

