% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deformationFieldPCA.R
\name{deformationFieldPCA}
\alias{deformationFieldPCA}
\title{Shape analysis of deformation fields based on PCA}
\usage{
deformationFieldPCA(x, mask, k = NA, pcaOption = "PCA")
}
\arguments{
\item{x}{list containing deformation fields}

\item{mask}{mask to apply to the deformation field}

\item{k}{rank to use}

\item{pcaOption}{currently only PCA and randPCA,
the latter being much faster.}
}
\value{
matrix is output
}
\description{
Converts a set of deformation fields to a matrix in order to
enable various forms of PCA.  Returns the components of shape
variability and variance explained.  Can optionally employ
a few different types of decomposition.
}
\examples{

img1 = antsImageRead( getANTsRData( "r16" ) ) \%>\%
  resampleImage( c(4,4) )
img2 = antsImageRead( getANTsRData( "r64" ) ) \%>\%
  resampleImage( c(4,4) )
img3 = antsImageRead( getANTsRData( "r27" ) ) \%>\%
  resampleImage( c(4,4) )
img4 = antsImageRead( getANTsRData( "r30" ) ) \%>\%
  resampleImage( c(4,4) )
reg1 = antsRegistration( img1, img2, 'SyN' )
reg2 = antsRegistration( img1, img3, 'SyN' )
reg3 = antsRegistration( img1, img4, 'SyN' )
w1 = antsImageRead( reg1$fwdtransforms[1] )
w2 = antsImageRead( reg2$fwdtransforms[1] )
w3 = antsImageRead( reg3$fwdtransforms[1] )
mask = getMask( img1 )
dpca = deformationFieldPCA( x, mask )
warpTx = antsrTransformFromDisplacementField( dpca$pcaWarps[[1]] )
warped = applyAntsrTransform( warpTx, data = img1, reference = img1)

}
\author{
Avants BB
}

