% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_data.R
\docType{data}
\name{dkt_cit_labels}
\alias{dkt_cit_labels}
\title{dkt_cit_labels}
\format{
A data frame with columns: Label, Description.
}
\usage{
dkt_cit_labels
}
\description{
A data frame label numbers and names for consolidated DKT and CIT labels.
}
\keyword{datasets}
