% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_data.R
\docType{data}
\name{dkt_labels}
\alias{dkt_labels}
\title{dkt_labels}
\format{
A data frame with columns: Label, Description, type.
}
\usage{
dkt_labels
}
\description{
A data frame label numbers and names for DKT cortical labels.
}
\keyword{datasets}
