% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscaleSVDxpts.R
\name{.calculate_abs_canonical_covariance}
\alias{.calculate_abs_canonical_covariance}
\title{Calculate Absolute Canonical Covariance}
\usage{
.calculate_abs_canonical_covariance(X, U, V)
}
\value{
A single numeric value for the objective.
}
\description{
The objective J = sum(abs(diag(U'XV))) / (||U||*||XV||). This is a
  **maximization** objective.
}
\keyword{internal}
