% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscaleSVDxpts.R
\name{.calculate_abs_canonical_covariance_gradient}
\alias{.calculate_abs_canonical_covariance_gradient}
\title{Calculate Gradient for Absolute Canonical Covariance}
\usage{
.calculate_abs_canonical_covariance_gradient(X, U, V)
}
\value{
A matrix [p x k] representing an **ascent direction** for the objective J.
}
\description{
Calculate Gradient for Absolute Canonical Covariance
}
\keyword{internal}
