% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscaleSVDxpts.R
\name{.calculate_angular_distance}
\alias{.calculate_angular_distance}
\title{Calculate Squared Angular Distance}
\usage{
.calculate_angular_distance(X, U, V)
}
\arguments{
\item{X}{A data matrix [n x p].}

\item{U}{A target basis matrix [n x k].}

\item{V}{A loading matrix [p x k].}
}
\value{
A single numeric value for the energy.
}
\description{
The energy E = || U/||U|| - XV/||XV|| ||_F^2. This is a
  **minimization** objective.
}
\keyword{internal}
