% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscaleSVDxpts.R
\name{.calculate_angular_distance_gradient}
\alias{.calculate_angular_distance_gradient}
\title{Calculate Gradient for Squared Angular Distance}
\usage{
.calculate_angular_distance_gradient(X, U, V)
}
\value{
A matrix [p x k] representing a **descent direction** for the energy.
}
\description{
Computes the analytical gradient of the squared angular distance.
}
\keyword{internal}
