% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscaleSVDxpts.R
\name{.calculate_normed_regression_error}
\alias{.calculate_normed_regression_error}
\title{Calculate normed Regression Error}
\usage{
.calculate_normed_regression_error(X, U, V, center_prediction = TRUE)
}
\arguments{
\item{center_prediction}{Logical, controls if prediction is centered relative to X.}
}
\value{
A single numeric value for the energy.
}
\description{
The energy E = ||X - UV'||^2/||X + epsilon||. This is a **minimization** objective.
}
\keyword{internal}
