% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscaleSVDxpts.R
\name{.calculate_normed_regression_error_gradient}
\alias{.calculate_normed_regression_error_gradient}
\title{Calculate Gradient for normed Centered and Normalized Regression Error}
\usage{
.calculate_normed_regression_error_gradient(X, U, V, center_prediction = TRUE)
}
\arguments{
\item{X}{A data matrix [n x p].}

\item{U}{A target basis matrix [n x k].}

\item{V}{The current loading matrix [p x k].}

\item{center_prediction}{Logical. Should be consistent with the energy function.}
}
\value{
A matrix [p x k] representing a DESCENT direction for the energy E.
}
\description{
This function computes the mathematically precise analytical gradient for the
energy function defined in `.calculate_normed_regression_error`.
}
\keyword{internal}
