% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscaleSVDxpts.R
\name{.calculate_procrustes_correlation}
\alias{.calculate_procrustes_correlation}
\title{Calculate Procrustes Correlation}
\usage{
.calculate_procrustes_correlation(X, U, V)
}
\value{
A single numeric value for the objective.
}
\description{
The objective J = tr(U'XV) / ||U'XV||_F. This is a
  **maximization** objective.
}
\keyword{internal}
