% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eanatDef.R
\name{eanatDef}
\alias{eanatDef}
\title{Eigenanatomy with deflation}
\usage{
eanatDef(inmat, nvecs = 0, mask = NA, smoother = 0, cthresh = 0,
  its = 5, eps = 0.1, positivity = FALSE, priors = NA,
  priorWeight = 0, verbose = FALSE)
}
\arguments{
\item{inmat}{input matrix}

\item{nvecs}{number of eigenanatomy vectors to compute. see
\code{eanatSelect} for a method to compute an optimal \code{nvecs} value.}

\item{mask}{input mask, must match matrix}

\item{smoother}{regularization parameter, typically 0 or 0.5, in voxels}

\item{cthresh}{remove isolated voxel islands of size below this value}

\item{its}{number of iterations}

\item{eps}{gradient descent parameter}

\item{positivity}{return unsigned eigenanatomy vectors}

\item{priors}{external initialization matrix.}

\item{priorWeight}{weight on priors in range 0 to 1.}

\item{verbose}{controls whether computation is silent or not.}
}
\value{
matrix is output, analogous to \code{svd(mat,nu=0,nv=nvecs)}
}
\description{
Simplified, low-parameter eigenanatomy implemented with deflation. The
algorithm is able to automatically select hidden \code{sparseness}
parameters, given the key parameter \code{nvecs}.  The user should select the
\code{cthresh} and  \code{smoother} regularization parameters for a given
application and also based on observing algorithm behavior when
\code{verbose=TRUE}.
}
\examples{
mat <- matrix(rnorm(2000),ncol=50)
nv <- eanatSelect( mat, selectorScale = 1.2 )
esol <- eanatDef( mat, nvecs=nv )
es2 <- sparseDecom( mat, nvecs = nv )
print( paste( "selected", nrow(esol),'pseudo-eigenvectors') )
print( mean( abs( cor( mat \%*\% t(esol)) ) ) ) # what we use to select nvecs
\dontrun{
networkPriors = getANTsRData("fmrinetworks")
ilist = networkPriors$images
mni = antsImageRead( getANTsRData("mni") )
mnireg = antsRegistration( meanbold*mask, mni, typeofTransform = 'Affine')
for ( i in 1:length(ilist) )
  ilist[[i]] = antsApplyTransforms( meanbold,ilist[[i]],mnireg$fwdtransform )
pr = imageListToMatrix( ilist, cortMask )
esol <- eanatDef( boldMat,
  nvecs = length(ilist), cortMask, verbose=FALSE,
  cthresh = 25, smoother = 0, positivity = TRUE, its=10, priors=pr,
  priorWeight=0.15, eps=0.1 )
}

}
\author{
Avants BB, Tustison NJ
}
\references{
Kandel, B. M.; Wang, D. J. J.; Gee, J. C. & Avants, B. B.
Eigenanatomy: sparse dimensionality reduction for multi-modal medical
image analysis. Methods,  2015, 73, 43-53.
PS Dhillon, DA Wolk, SR Das, LH Ungar, JC Gee, BB Avants
Subject-specific functional parcellation via Prior Based Eigenanatomy
NeuroImage, 2014, 99, 14-27.
}
\seealso{
\code{\link{eanatSelect}}
}

