% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eanatDef.R
\name{eanatSelect}
\alias{eanatSelect}
\title{Selection of n eigenvectors and sparsity for eigenanatomy}
\usage{
eanatSelect(inmat, mask = NA, cthresh = 0, maxNEvec = 0,
  selectorScale = 1.1, verbose = FALSE)
}
\arguments{
\item{inmat}{input matrix}

\item{mask}{input mask, must match matrix}

\item{cthresh}{remove isolated voxel islands of size below this value}

\item{maxNEvec}{integer that, if set greater than zero, indicates that we use
a low-rank approximation to the input matrix before proceeding to eanat.
this value should be greater than \code{nvecs}}

\item{selectorScale}{influences automatic selection of \code{nvecs} and tries
to find the knee in the correlation plot. This parameter produces fewer,
less sparse eigenanatomy pseudo-eigenvectors as its value increases.  Its
minimum value is 1 and a reasonable range is between 1 and 2.  The user
should look at the plot produced when verbosity is turned on.}

\item{verbose}{controls whether computation is silent or not.}
}
\value{
nvecs is output, analogous to \code{nvecs} in
\code{svd(mat,nu=0,nv=nvecs)}
}
\description{
The algorithm automatically selects the key \code{nvecs} and hidden
\code{sparseness} parameters.  The user should select the \code{cthresh}
regularization parameters for his or her application. The principle used
here is that we want few but sparse pseudo-eigenvectors that are minimally
correlated in row-space. true left and right eigenvectors are uncorrelated
in both row and column (left and right eigenvector) spaces, but this is not
the case when we impose sparsity.
}
\examples{
mat <- matrix(rnorm(2000),ncol=50)
nvecsSel<-eanatSelect( mat, selectorScale = 1.2 )
esol <- sparseDecom( mat, nvecs = nv )
print(paste("selected", nvecsSel,'pseudo-eigenvectors'))
}
\author{
Avants BB, Tustison NJ
}

