% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/eigSeg.R
\name{eigSeg}
\alias{eigSeg}
\title{Segmentation for eigenanatomy.}
\usage{
eigSeg(mask = NA, imgList = NA, applySegmentationToImages = TRUE)
}
\arguments{
\item{mask}{D-dimensional mask > 0 defining segmentation region.}

\item{applySegmentationToImages}{boolean determines if original image list
is modified by the segmentation.}

\item{imageList}{list containing antsImages or filenames pointing to
antsImages.}
}
\value{
segmentation image.
}
\description{
Segment a mask into regions based on the max value in an image list.  At a
given voxel the segmentation label will contain the index to the image that
has the largest value.  If the 3rd image has the greatest value, the
segmentation label will be 3 at that voxel.
}
\examples{
\dontrun{
mylist<-list(image1,image2)
# both approaches below are ok
myseg<-eigSeg( mask, mylist )
myseg<-eigSeg( mask, c('a.nii.gz','b.nii.gz')  )
}
}
\author{
Avants BB
}

