% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_not_antsImage.R
\name{error_not_antsImage}
\alias{error_not_antsImage}
\title{Stop if not an antsImage object}
\usage{
error_not_antsImage(x, argname = "")
}
\arguments{
\item{x}{object to test}

\item{argname}{Name of argument to say was not an \code{antsImage}}
}
\value{
Invisible \code{NULL}
}
\description{
Stop if not an antsImage object
}
\examples{
img = antsImageRead(getANTsRData('r16'), 2)
error_not_antsImage(img) 
testthat::expect_error(error_not_antsImage(as.array(img)))
}
