% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estSmooth.R
\name{estSmooth}
\alias{estSmooth}
\title{Estimate image smoothness}
\usage{
estSmooth(x, mask, rdf, scaleResid = TRUE, sample = NULL, verbose = TRUE)
}
\arguments{
\item{mask}{input mask, must match matrix}

\item{rdf}{residual degrees of freedom}

\item{scaleResid}{logical. if TRUE residuals are scaled}

\item{sample}{number of images to use for estimating smoothing (default uses all images)}

\item{verbose}{enables verbose output}

\item{object}{object may be an image of class "antsImage" or an image matrix}
}
\value{
Outputs the estimated FWHM and RPV image
}
\description{
The partial derivatives of an image in x, y, and z directions are used to
create a covariance matrix which in turn is used to calculate the 
full-widths at half maxima (FWHM). The FWHM is equivalent to the estimated
image smoothness.

The resels per voxel image (\code{RPVImg}) represents the estimated resel at
each individual voxel throughout the search region. This may be used in 
place of volumetric measurements (or sum voxel measurements) when estimating
the p-value of a cluster using \code{rftPval}. The intent behind using the 
RPV image to estimate cluster level statistics is to offset the natural
probability of obtaining significant clusters solely by chance in very 
smooth regions at low thresholds.

It is possible to use a single statistical field image to estimate the FWHM. 
However, it's recommended that FWHM estimates are obtained from the scaled 
residuals of statistical models (Stefan J.K et al., 1999). Therefore, this 
function is optimized to estimate the pooled smoothness of the residual 
images from a fitted model. By default residuals are scaled
(\code{scaleResid = TRUE}).

A scaling factor is used to correct for differences when using the 
\code{sample} option. Scaling isn't effective when the number of images is 
very low and typically results in an overestimation of the the FWHM. If only
one image or numeric vector is entered then the scaling factor is not used. 
If a numeric vector is entered the \code{imageMake} function is used to 
prepare it for smoothness estimation (see Worsley et al., 1999).
}
\details{
Estimates smoothness of a single image or image matrix
}
\note{
function currently in beta phase
}
\examples{

# estimatation of a single images smoothness
outimg1 <- makeImage(c(10, 10, 10), rnorm(1000))
maskimg <- getMask(outimg1)
myfwhm1 <- estSmooth(outimg1, maskimg)

# estimation of smoothness of overall sample images in a statistical model
outimg2 <- makeImage(c(10,10,10), rnorm(1000))
imat <- imageListToMatrix(list(outimg1, outimg2), maskimg)
variable <- rnorm(2)
fit <- lm(imat ~ variable)
myfwhm2 <- estSmooth(residuals(fit), maskimg)

}
\author{
Zachary P. Christensen
}
\references{
Hayasaka (2004) Nonstationary cluster-size inference with random field and permutation methods.

Worsley K.J. (1992) A Three-Dimensional Statistical Analysis for CBF Activation Studies in Human Brain.

Worsley K.J. (1996) A Unified Statistical Approach for Determining Significant Signals in Images of Cerebral Activation.

Worsley K.J. (1999) Detecting Changes in Nonisotropic Images

Stefan J.K. (1999) Robust Smoothness Estimation in Statistical Parametric Maps Using Standardized Residual from the General Linear Model
}
\seealso{
resels
}

