% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euler.R
\name{euler}
\alias{euler}
\title{Euler Character}
\usage{
euler(u, df, fieldType)
}
\arguments{
\item{u}{statistical value (typically the maxima of a cluster or SPM)}

\item{df}{degrees of freedom expressed as df = c(degrees of interest, degrees of error)}

\item{fieldType:}{\itemize{
\item{T: } {T-field} 
\item{F: } {F-field} 
\item{X: } {Chi-square field'} 
\item{Z: } {Gaussian field}
}}
}
\value{
A vector of estimated euler characteristics for dimensions 0:D
}
\description{
Calculates the euler characteristic at a threshold level
}
\note{
function currently in beta phase
}
\examples{
# generate some data as if we just fitted a linear regression
outimg1 <-makeImage(c(10,10,10), rt(1000))
maskimg <-getMask(outimg1)
fwhm <-estSmooth(outimg1, maskimg)
resels <-resels(maskimg, fwhm$fwhm)
ec <-euler(max(outimg1), c(1,10), fieldType='T')
pvox <-sum(ec*resels)

}
\author{
Zachary P. Christensen
}
\references{
Worlsey K.J., (1996) A Unified Statistical Approach for Determining Significant Signals in Images of Cerebral Activation.
}
\seealso{
rftPval, resels
}

