% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{algorithm}
\name{exemplarInpainting}
\alias{exemplarInpainting}
\title{Uses example images to inpaint or approximate an existing image.}
\usage{
approximg<-exemplarInpainting( img, paintMask, list(img1,img2))
}
\arguments{
\item{img}{- antsImage to be approximated / painted}

\item{featureRadius}{- radius of image neighborhood e.g. 2}

\item{putbackoriginaldata}{- put values in label 1 in approximated image}

\item{sharpen}{- sharpen the approximated image}

\item{painting}{mask with values 1 or values 1 and 2 - if there is a 2 then it will learn from label 1 to paint label 2}

\item{a}{list containing antsImages}
}
\value{
inpainted image
}
\description{
Employs a robust regression approach to learn the relationship between a sample image and a list of images that are mapped to the same space as the sample image.  The regression uses data from an image neighborhood.
}
\examples{
fi<-replicate(100, rnorm(100))
fi[1:10,]<-fi[,1:10]<-fi[91:100,]<-fi[,91:100]<-0
fi<-as.antsImage( fi , 'float' )
SmoothImage(2,fi,1.5,fi)
mo<-as.antsImage( replicate(100, rnorm(100))  )
mo2<-as.antsImage( replicate(100, rnorm(100))  )
mask<-getMask(fi,mean(fi),Inf,TRUE)
ilist<-list(mo,mo2)
painted<-exemplarInpainting(fi,mask,ilist)
}
\author{
Brian B. Avants
}
\keyword{inpainting}
\keyword{template}

