% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscaleSVDxpts.R
\name{exploratory_visualization}
\alias{exploratory_visualization}
\title{Exploratory Clustering and Visualization}
\usage{
exploratory_visualization(data, dotsne = FALSE, verbose = FALSE)
}
\arguments{
\item{data}{A data frame with numeric columns.}

\item{dotsne}{boolean}

\item{verbose}{boolean}
}
\value{
A list containing the combined plot and the optimal k value.
}
\description{
This function performs automated clustering using PAM and silhouette method,
and visualizes the results using ggplot2 and ggdendro.
}
