% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractImagePatches.R
\name{extractImagePatchCoordinates}
\alias{extractImagePatchCoordinates}
\title{Extract 2-D or 3-D image patch coordinates.}
\usage{
extractImagePatchCoordinates(
  image,
  patchSize,
  maxNumberOfPatches = "all",
  strideLength = 1,
  maskImage = NULL,
  physicalCoordinates = TRUE,
  cornerCoordinates = TRUE,
  randomSeed,
  randomize = TRUE
)
}
\arguments{
\item{image}{Input ANTs image with one or more components.}

\item{patchSize}{Width, height, and depth (if 3-D) of patches.}

\item{maxNumberOfPatches}{Maximum number of patches returned.  If
"all" is specified, then all patches in sequence (defined by the
\code{strideLength} are extracted.}

\item{strideLength}{Defines the sequential patch overlap for
\code{maxNumberOfPatches = "all"}.  Can be a image-dimensional vector or a scalar.}

\item{maskImage}{optional image specifying the sampling region for
the patches when \code{maximumNumberOfPatches} does not equal "all".
The way we constrain patch selection using a mask is by forcing
each returned patch to have a masked voxel at its center.}

\item{physicalCoordinates}{boolean to determine whether indices or spatial
coordinates are returned.}

\item{cornerCoordinates}{boolean to determine whether indices or spatial
coordinates of the corner or the center are returned.}

\item{randomSeed}{integer seed that allows reproducible patch extraction
across runs.}

\item{randomize}{boolean controlling whether we randomize indices when masking.}
}
\value{
a matrix of image patch spatial or index coordinates.
}
\description{
Extract 2-D or 3-D image patch coordinates.
}
\examples{

library( ANTsR )

image <- antsImageRead( getANTsRData( "r16" ) )
maskImage <- getMask( image, 1, 1000 )
patchCoordsP <- extractImagePatchCoordinates( image, c( 32, 32 ), 10, c( 32, 32 ),
  maskImage, randomSeed = 0, physicalCoordinates = TRUE )
patchCoordsI <- extractImagePatchCoordinates( image, c( 32, 32 ), 10, c( 32, 32 ),
  maskImage, randomSeed = 0, physicalCoordinates = FALSE )

}
\author{
Tustison NJ, Avants B
}
