% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractImagePatches.R
\name{extractImagePatches}
\alias{extractImagePatches}
\title{Extract 2-D or 3-D image patches.}
\usage{
extractImagePatches(
  image,
  patchSize,
  maxNumberOfPatches = "all",
  strideLength = 1,
  maskImage = NULL,
  randomSeed,
  returnAsArray = FALSE,
  randomize = TRUE
)
}
\arguments{
\item{image}{Input ANTs image with one or more components.}

\item{patchSize}{Width, height, and depth (if 3-D) of patches.}

\item{maxNumberOfPatches}{Maximum number of patches returned.  If
"all" is specified, then all patches in sequence (defined by the
\code{strideLength} are extracted.}

\item{strideLength}{Defines the sequential patch overlap for
\code{maxNumberOfPatches = "all"}.  Can be a image-dimensional vector or a scalar.}

\item{maskImage}{optional image specifying the sampling region for
the patches when \code{maximumNumberOfPatches} does not equal "all".
The way we constrain patch selection using a mask is by forcing
each returned patch to have a masked voxel at its center.}

\item{randomSeed}{integer seed that allows reproducible patch extraction
across runs.}

\item{returnAsArray}{specifies the return type of the function.  If
\code{FALSE} (default) the return type is a list where each element is
a single patch.  Otherwise the return type is an array of size
\code{dim( numberOfPatches, patchSize )}.}

\item{randomize}{boolean controlling whether we randomize indices when masking.}
}
\value{
a list (or array) of image patches.
}
\description{
Extract 2-D or 3-D image patches.
}
\examples{

library( ANTsR )

image <- antsImageRead( getANTsRData( "r16" ) )
maskImage <- getMask( image, 1, 1000 )
patchSet1 <- extractImagePatches( image, c( 32, 32 ), 10, c( 32, 32 ), randomSeed = 0 )
patchSet2 <- extractImagePatches( image, c( 32, 32 ), 10, c( 32, 32 ), randomSeed = 1 )
patchSet3 <- extractImagePatches( image, c( 32, 32 ), 10, c( 32, 32 ), maskImage, randomSeed = 0 )

}
\author{
Tustison NJ
}
