% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmarkTransforms.R
\name{fitTransformToPairedPoints}
\alias{fitTransformToPairedPoints}
\title{fitTransformToPairedPoints}
\usage{
fitTransformToPairedPoints(
  movingPoints,
  fixedPoints,
  transformType = "affine",
  regularization = 0.000001,
  domainImage = NULL,
  numberOfFittingLevels = 4,
  meshSize = 1,
  splineOrder = 3,
  enforceStationaryBoundary = TRUE,
  displacementWeights = NULL,
  numberOfCompositions = 10,
  compositionStepSize = 0.5,
  sigma = 0,
  convergenceThreshold = 0,
  numberOfIntegrationPoints = 2,
  rasterizePoints = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{movingPoints}{Moving points specified in physical space as a \code{n x d}
matrix where \code{n} is the number of points and \code{d} is the dimensionality.}

\item{fixedPoints}{Fixed points specified in physical space as a \code{n x d}
matrix where \code{n} is the number of points and \code{d} is the dimensionality.}

\item{transformType}{'rigid', 'similarity', "affine', 'bspline', 'diffeo', 'syn',
'tv', or 'time-varying'.}

\item{regularization}{Ridge penalty in [0,1] for linear transforms.}

\item{domainImage}{Defines physical domain of the B-spline transform.  Must be defined
for nonlinear transforms.}

\item{numberOfFittingLevels}{Integer specifying the number of fitting levels for the
B-spline interpolation of the displacement field.}

\item{meshSize}{Defines the mesh size at the initial fitting level for the B-spline
interpolation of the displacement field.}

\item{splineOrder}{Spline order of the B-spline displacement field.}

\item{enforceStationaryBoundary}{Ensure no displacements on the image boundary
(B-spline only).}

\item{displacementWeights}{Defines the individual weighting of the corresponding scattered
data value.  Default = None meaning all displacements are weighted the same.}

\item{numberOfCompositions}{Total number of compositions for the diffeomorphic transform.}

\item{compositionStepSize}{Scalar multiplication factor of the weighting of the update field
for the diffeomorphic transforms.}

\item{sigma}{Gaussian smoothing standard deviation of the update field (in mm).}

\item{convergenceThreshold}{Composition-based convergence parameter for the diffeomorphic
transforms using a window size of 10 values.}

\item{numberOfIntegrationPoints}{Time-varying velocity field parameter.}

\item{rasterizePoints}{Use nearest neighbor rasterization of points for estimating update
field (potential speed-up).}

\item{verbose}{Print progress to the screen.}
}
\value{
object containing ANTsR transform, error, and scale (or displacement field)
}
\description{
Estimate a transform using paired point sets.  An ANTsR transform is returned.
}
\examples{
fixed <- matrix( c( 50, 50, 200, 50, 50, 200 ), ncol = 2, byrow = TRUE )
moving <- matrix( c( 50, 50, 50, 200, 200, 200 ), ncol = 2, byrow = TRUE )

# Affine transform
xfrm <- fitTransformToPairedPoints( moving, fixed, transformType = "Affine", regularization = 0 )
params <- getAntsrTransformParameters( xfrm )

# Rigid transform
xfrm <- fitTransformToPairedPoints( moving, fixed, transformType = "Rigid", regularization = 0 )
params <- getAntsrTransformParameters( xfrm )

# Similarity transform
xfrm <- fitTransformToPairedPoints( moving, fixed, transformType = "Similarity", regularization = 0 )
params <- getAntsrTransformParameters( xfrm )

# B-spline transform
domainImage <- antsImageRead( getANTsRData( "r16" ) )
xfrm <- fitTransformToPairedPoints( moving, fixed, transformType = "Bspline", domainImage = domainImage, numberOfFittingLevels = 5 )

# Diffeo transform
domainImage <- antsImageRead( getANTsRData( "r16" ) )
xfrm <- fitTransformToPairedPoints( moving, fixed, transformType = "Diffeo", domainImage = domainImage, numberOfFittingLevels = 6 )

# SyN transform
domainImage <- antsImageRead( getANTsRData( "r16" ) )
xfrm <- fitTransformToPairedPoints( moving, fixed, transformType = "SyN", domainImage = domainImage, numberOfFittingLevels = 6, numberOfCompositions = 10, compositionStepSize = 0.01 )
}
\author{
B Avants, N Tustison
}
