% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscaleSVDxpts.R
\name{frob}
\alias{frob}
\title{Frobenius Norm of a Matrix}
\usage{
frob(A)
}
\arguments{
\item{A}{Numeric matrix.}
}
\value{
Numeric scalar, the Frobenius norm ||A||_F.
}
\description{
Computes the Frobenius norm of a matrix, defined as the square root of the sum of squared elements.
}
\details{
The Frobenius norm is \eqn{\sqrt{\sum_{i,j} A_{ij}^2}}, equivalent to the Euclidean norm of the vectorized matrix.
Used for scaling and computing residuals in NSA-Flow.
}
\examples{
A <- matrix(1:4, 2, 2)
frob(A)
}
