% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsl2antsrTransform.R
\name{fsl2antsrTransform}
\alias{fsl2antsrTransform}
\title{fsl2antsrTransform}
\usage{
fsl2antsrTransform(matrix, reference, moving)
}
\arguments{
\item{matrix}{4x4 matrix of parameters}

\item{reference, }{target image read in as antsImage}

\item{moving}{image, read in as antsImage}
}
\description{
Convert an FSL linear transform to an antsrTransform
}
\examples{
fslMat = matrix(0,4,4)
diag(fslMat) = rep(1,4)
fslMat[1:3,4] = c(2,3,4)
img = antsImageRead(getANTsRData("ch2"))
tx = fsl2antsrTransform(fslMat, img, img)

}
\author{
JT Duda
}
