% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionalLungSegmentation.R
\name{functionalLungSegmentation}
\alias{functionalLungSegmentation}
\title{Ventilation-based segmentation of hyperpolarized gas lung MRI.}
\usage{
functionalLungSegmentation(
  image,
  mask,
  numberOfIterations = 1,
  numberOfAtroposIterations = 0,
  mrfParameters = "[0.3,2x2x2]",
  numberOfClusters = 4,
  clusterCenters = NA,
  verbose = TRUE
)
}
\arguments{
\item{image}{input proton-weighted MRI.}

\item{mask}{mask image designating the region to segment.
0/1 = background/foreground.}

\item{numberOfIterations}{number of Atropos <--> N4 iterations (outer loop).}

\item{numberOfAtroposIterations}{number of Atropos iterations (inner loop).
If \code{numberOfAtroposIterations = 0}, this is equivalent to K-means with
no MRF priors.}

\item{mrfParameters}{parameters for MRF in Atropos.}

\item{numberOfClusters}{number of tissue classes (default = 4)}

\item{clusterCenters}{initialization centers for k-means}

\item{verbose}{print progress to the screen.}
}
\value{
segmentation image, probability images, and processed input
image.
}
\description{
Lung segmentation into four classes based on ventilation as described in
this paper:
}
\details{
\url{https://pubmed.ncbi.nlm.nih.gov/21837781/}
}
\examples{
\dontrun{

library( ANTsR )

image <- antsImageRead( "lung.nii.gz" )
mask <- antsImageRead( "mask.nii.gz" )
output <- functionalLungSegmentation( image, mask )
antsImageWrite( output$segmentationImage, "outputSegmentation.nii.gz" )

}
}
\author{
Tustison NJ
}
