% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bloodPerfusionSVD.R
\name{generateAifMaskImage}
\alias{generateAifMaskImage}
\title{Automatically generate an arterial input function mask.}
\usage{
generateAifMaskImage(
  perfusionImage,
  voiMaskImage,
  maxNumberOfVoxelsToPlot = 0,
  plotDirectory = tempdir(),
  index = NA
)
}
\arguments{
\item{perfusionImage}{time series (n-D + time) perfusion acquisition.}

\item{voiMaskImage}{n-D mask image indicating where the cerebral blood
flow parameter images are calculated.}

\item{maxNumberOfVoxelsToPlot}{Given that several voxels might survive
the criteria for selection, we permit the user to write the fitting
plots and masks to a directory for manual selection.}

\item{plotDirectory}{Directory for writing the manual inspectionfiles.}

\item{index}{If the user prefers a specific voxel if looking at the
plots, they can simply specify the index (gleaned from the file names)
and the corresponding AIF mask image is created.}
}
\value{
list( mask image, fitting results )
}
\description{
Automated approach for generating a mask to be used for modeling the
arterial input function using the method described in
https://www.ncbi.nlm.nih.gov/pubmed/15956519.
}
\examples{
\dontrun{

}
}
\author{
Tustison NJ
}
