% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/geoSeg.R
\name{geoSeg}
\alias{geoSeg}
\title{brain segmentation based on geometric priors}
\usage{
geoSeg(img, brainmask, priors, vesselopt = "none", vesselk = 2,
  gradStep = 1)
}
\arguments{
\item{img}{input image}

\item{brainmask}{binary image}

\item{priors}{spatial priors}

\item{vesselopt}{one of bright, dark or none}

\item{vesselk}{integer for kmeans vessel-based processing}

\item{gradStep}{scalar for registration}
}
\value{
list of segmentation result images
}
\description{
uses topological constraints to enhance accuracy of brain segmentation
}
\examples{
img = antsImageRead( getANTsRData("simple") ,2)
img = n3BiasFieldCorrection( img , 4 )
img = n3BiasFieldCorrection( img , 2 )
bmk = getMask( img )
segs <- kmeansSegmentation( img, 3, bmk )
priors = segs$probabilityimages
seg = geoSeg( img, bmk, priors, 'none' )
}
\author{
Brian B. Avants
}

