% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getANTsRData.R
\name{getANTsRData}
\alias{getANTsRData}
\title{getANTsRData}
\usage{
getANTsRData(fileid, usefixedlocation = FALSE)
}
\arguments{
\item{fileid}{one of the permitted file ids}

\item{usefixedlocation}{directory to which you download files}
}
\value{
filename string
}
\description{
Downloads antsr test data
}
\examples{
fi <- getANTsRData( "r16" )
}
\author{
Avants BB
}

