% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorientImage.R
\name{getCenterOfMass}
\alias{getCenterOfMass}
\title{center of mass}
\usage{
getCenterOfMass(img)
}
\arguments{
\item{img}{antsImage}
}
\value{
vector center of mass
}
\description{
compute an image center of mass in physical space which is defined as the
mean of the intensity weighted voxel coordinate system.
}
\examples{

fi<-antsImageRead( getANTsRData("r16"))
com1<-getCenterOfMass( fi )
fi<-antsImageRead( getANTsRData("r64"))
com2<-getCenterOfMass( fi )

}
\author{
Brian B. Avants
}
\keyword{geometry}
\keyword{image}
