% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMask.R
\name{getMask}
\alias{getMask}
\title{Get Mask}
\usage{
getMask(img, lowThresh, highThresh, cleanup = 2)
}
\arguments{
\item{img}{Input image. Can be an \code{antsImage} of 2, 3 or 4 dimensions.
If \code{img} is NULL, a file chooser dialog will appear.}

\item{lowThresh}{An inclusive lower threshold for voxels to be included in
the mask.}

\item{highThresh}{An inclusive upper threshold for voxels to be included in
the mask.}

\item{cleanup}{If \code{>0}, morphological operations will be applied to
clean up the mask by eroding away small or weakly-connected areas, and
closing holes.}
}
\value{
Object of type \code{antsImage} containing the mask image. The
voxel intensities will be binarized, 1 for voxels in the mask and 0 outside.
}
\description{
Get a binary mask image from the given image after thresholding.
}
\details{
If \code{cleanup} is \code{>0}, the following steps are applied
\enumerate{ \item Erosion with radius 2 voxels \item Retain largest
component \item Dilation with radius 1 voxel \item Morphological closing }

These functions are available in \link{iMath}, see the operations
\dQuote{ME}, \dQuote{GetLargestComponent}, \dQuote{MD}, \dQuote{FillHoles}.
}
\examples{

img<-antsImageRead( getANTsRData("r16"))
mask<-getMask( img )

}
\author{
Shrinidhi KL, Cook PA, Avants BB
}
\keyword{mask}

