% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiResRandomForestRegression.R
\name{getMultiResFeatureMatrix}
\alias{getMultiResFeatureMatrix}
\title{build multiple resolution predictor matrix from feature images}
\usage{
getMultiResFeatureMatrix(x, labelmask, rad = NA, multiResSchedule = c(0),
  nsamples = 10)
}
\arguments{
\item{x}{a list of feature images}

\item{labelmask}{the mask defines the image space
for the associated feature list and the number of parallel predictors.
more labels means more predictors.}

\item{rad}{vector of dimensionality d define nhood radius}

\item{multiResSchedule}{a vector of smoothing values}

\item{nsamples}{defines the number of samples to take for each label.}
}
\value{
mat a matrix of predictors with n samples rows
}
\description{
Represents feature images as a neighborhood across scales.
each subject gets a mask and feature list.  these mask/features should be
the same for all subjects.  e.g each subject has a k-label image
where the labels cover the same anatomy and the feature images are the same.
for each label in a mask, produce a multi-resolution neighborhood
sampling from the data within the label, for a given feature. do this for
each feature.
}
\examples{

img <- antsImageRead( getANTsRData("r16"))
seg <- kmeansSegmentation( img, 3 )$segmentation
flist = list( img, img \%>\% iMath("Grad") )
featMat <- getMultiResFeatureMatrix( flist, seg, rad=c(1,1),
  multiResSchedule = c( 2, 1), nsamples=10 )
}
\author{
Avants BB, Tustison NJ
}

