% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getMultivariateTemplateCoordinates.R
\name{getMultivariateTemplateCoordinates}
\alias{getMultivariateTemplateCoordinates}
\title{Label multivariate components by an anatomical coordinate system.}
\usage{
getMultivariateTemplateCoordinates(imageSetToBeLabeledIn, templateWithLabels,
  labelnames = NA, outprefix = NA, convertToTal = FALSE, pvals = NA,
  threshparam = 1, clustparam = 250, identifier)
}
\arguments{
\item{imageSetToBeLabeledIn}{a template paired with (most likely) the output
of a multivariate sparse decomposition or (alternatively) could be just a
statistical map with zeroes in non-interesting areas}

\item{pvals}{the already computed pvalue for each component}
}
\value{
The output point coordinates are in approximate Talairach / MNI (or
whatever) template space.
}
\description{
This function will provide a mapping that labels the list of input images
and each of their blobs.
}
\details{
Uses getTemplateCoordinates as a sub-routine.

TBN
}
\examples{
\dontrun{
 tem<-antsImageRead('templates/template_brain.nii.gz',3)
 temlab<-antsImageRead('temp.nii.gz',3)
 temlab2<-antsImageRead('temp2.nii.gz',3)
 # try getANTsRData if you have www access
 mymni<-list( antsImageRead(getANTsRData('mni'),3),
       antsImageRead(getANTsRData('mnib'),3),
       antsImageRead(getANTsRData('mnia'),3) )
 mytem<-list(tem,temlab,temlab2)
 mynetworkdescriptor<-getMultivariateTemplateCoordinates(
   mytem, mymni , convertToTal = TRUE , pvals=c(0.01,0.05) )
}
}
\author{
Avants, BB
}
\keyword{Coordinates}
\keyword{Talairach,}
\keyword{Template,}

