% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscaleSVDxpts.R
\name{getNamesFromDataframe}
\alias{getNamesFromDataframe}
\title{Extract column names with concatenated search parameters}
\usage{
getNamesFromDataframe(x, demogIn, exclusions)
}
\arguments{
\item{x}{vector of strings}

\item{demogIn}{the dataframe with column names to search.}

\item{exclusions}{the strings to exclude}
}
\value{
vector of string column names
}
\description{
Extract column names with concatenated search parameters
}
\examples{

# nms = getNamesFromDataframe( c("sp","ed"), cars )

}
\author{
Avants BB
}
