% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNeighborhood.R
\name{getNeighborhoodAtVoxel}
\alias{getNeighborhoodAtVoxel}
\title{Get a hypercube neighborhood at a voxel}
\usage{
getNeighborhoodAtVoxel(image, center, kernel, physical.coordinates = FALSE)
}
\arguments{
\item{image}{Image object of S4 class \code{antsImage} to get values from.}

\item{center}{array of indices for neighborhood center}

\item{kernel}{either an array of values for neighborhood radius (in voxels) or a binary array of the same dimension as the image, specifying the shape of the neighborhood to extract}

\item{physical.coordinates}{a logical indicating if voxel indices and
offsets should be in voxel or physical coordinates}
}
\value{
a list
\itemize{
  \item{values}{numeric vector of values}
  \item{indices}{matrix providing the coordinates for each value}
}
}
\description{
Get the values in a local neighborhood of an \code{antsImage}.
}
\examples{
img<-makeImage(c(10,10),rnorm(100))
center <- dim(img)/2
radius <- rep(3,2)
nhlist <- getNeighborhoodAtVoxel(img,center,radius)
kernel <- 1*(rnorm(49)>0)
dim(kernel) <- c(7,7)
randlist <- getNeighborhoodAtVoxel(img,center,kernel)
randlist <- getNeighborhoodAtVoxel(img,center,kernel, 
physical.coordinates = TRUE)
arr = as.array(img)
testthat::expect_error(getNeighborhoodAtVoxel(arr,center,kernel), "class")
testthat::expect_error(getNeighborhoodAtVoxel(img,as.character(center),kernel), 
"center must be")
testthat::expect_error(getNeighborhoodAtVoxel(img,center,c(radius, 3)),
"kernel must have same") 

}
\author{
Duda JT
}
