% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscaleSVDxpts.R
\name{glm_impute}
\alias{glm_impute}
\title{Impute missing data using GLM models}
\usage{
glm_impute(
  dataframe,
  columns_to_impute,
  predictor_columns,
  family = "gaussian"
)
}
\arguments{
\item{dataframe}{A data frame containing the data to impute.}

\item{columns_to_impute}{A vector of column names to impute.}

\item{predictor_columns}{A vector of column names to use as predictors.}

\item{family}{A string specifying the GLM family (default is 'gaussian').}
}
\value{
A data frame with imputed values.
}
\description{
Impute missing data using GLM models
}
\examples{
set.seed(123)
df <- data.frame(
  age = c(25, 30, 35, NA, 45, 50, NA, 40, 35, NA),
  income = c(50000, 60000, 70000, 80000, 90000, 100000, 110000, NA, 120000, 130000),
  education = c(12, 16, 14, 12, NA, 18, 20, 16, 14, 12)
)
columns_to_impute <- c("age")
predictor_columns <- c( "income", "education")
imputed_data <- glm_impute(df, columns_to_impute, predictor_columns, family = 'gaussian')
print(imputed_data)
}
