% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscaleSVDxpts.R
\name{gradient_invariant_orthogonality_defect_diag_zero}
\alias{gradient_invariant_orthogonality_defect_diag_zero}
\title{Compute the gradient of the invariant orthogonality defect}
\usage{
gradient_invariant_orthogonality_defect_diag_zero(A)
}
\arguments{
\item{A}{Input matrix (n x p, where n >> p)}
}
\value{
Gradient of the invariant orthogonality defect with respect to A
}
\description{
Compute the gradient of the invariant orthogonality defect
}
