% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscaleSVDxpts.R
\name{gradient_measure_orthogonality_norm}
\alias{gradient_measure_orthogonality_norm}
\title{Gradient of the Orthogonality Measure with Respect to the Matrix}
\usage{
gradient_measure_orthogonality_norm(mat)
}
\arguments{
\item{mat}{A numeric matrix.}
}
\value{
A numeric matrix representing the gradient of the orthogonality measure with respect to the input matrix.
}
\description{
This function calculates the gradient of the orthogonality measure with respect to the input matrix.
The gradient indicates the direction to adjust the matrix in order to minimize the orthogonality measure.
}
\examples{
mat <- matrix(runif(9), nrow = 3)
gradient_measure_orthogonality_norm(mat)
}
