% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iMath.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe an object forward}
\usage{
lhs \%>\% rhs
}
\arguments{
\item{lhs}{input from left side}

\item{rhs}{additional params}
}
\description{
Chain operators together
}
\details{
The \code{\%>>\%} operator pipes the object on the left-hand side to the
right-hand side according to the syntax.
}

