% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hausdorffDistance.R
\name{hausdorffDistance}
\alias{hausdorffDistance}
\title{hausdorffDistance}
\usage{
hausdorffDistance(inputImage1, inputImage2)
}
\arguments{
\item{inputImage1}{image 1.}

\item{inputImage2}{image 2.}
}
\value{
data frame with overlap measures
}
\description{
Wrapper for the ITK class HausdorffDistanceImageFilter.  More documentaiton
available here:
}
\details{
\url{https://itk.org/Doxygen/html/classitk_1_1HausdorffDistanceImageFilter.html}
}
\examples{

sourceImage <- antsImageRead(getANTsRData("r16"), 2)
sourceSegmentation <- kmeansSegmentation(sourceImage, 3)$segmentation
referenceImage <- antsImageRead(getANTsRData("r16"), 2)
referenceSegmentation <- kmeansSegmentation(referenceImage, 3)$segmentation
hausdorff <- hausdorffDistance(sourceImage, referenceImage)

}
\author{
Tustison NJ
}
