% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hessianObjectness.R
\name{hessianObjectness}
\alias{hessianObjectness}
\title{hessianObjectness}
\usage{
hessianObjectness(
  image,
  objectDimension = 1,
  isBrightObject = TRUE,
  sigmaMin = 0.1,
  sigmaMax = 10,
  numberOfSigmaSteps = 10,
  useSigmaLogarithmicSpacing = TRUE,
  alpha = 0.5,
  beta = 0.5,
  gamma = 5,
  setScaleObjectnessMeasure = TRUE
)
}
\arguments{
\item{image}{input image}

\item{objectDimension}{0: 'sphere', 1: 'line', or 2: 'plane'.}

\item{isBrightObject}{Set 'true' for enhancing bright objects and 'false' for dark objects.}

\item{sigmaMin}{Define scale domain for feature extraction.}

\item{sigmaMax}{Define scale domain for feature extraction.}

\item{numberOfSigmaSteps}{Define number of samples for scale space.}

\item{useSigmaLogarithmicSpacing}{Define sample spacing the for scale space.}

\item{alpha}{Hessian filter parameter.}

\item{beta}{Hessian filter parameter.}

\item{gamma}{Hessian filter parameter.}

\item{setScaleObjectnessMeasure}{...}
}
\value{
hessian objectness image.
}
\description{
Interface to ITK filter.  Based on the paper by Westin et al., 
"Geometrical Diffusion Measures for MRI from Tensor Basis Analysis" and 
Luca Antiga's Insight Journal paper http://hdl.handle.net/1926/576.
}
\examples{
image <- antsImageRead(getANTsRData("r16"))
hessianObjectnessImage <- hessianObjectness(image)

}
\author{
NJ Tustison
}
