% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogramEqualizeImage.R
\name{histogramEqualizeImage}
\alias{histogramEqualizeImage}
\title{histogramEqualizeImage}
\usage{
histogramEqualizeImage(image, numberOfHistogramBins = 256)
}
\arguments{
\item{image}{image to undergo intensity transformation.}

\item{numberOfHistogramBins}{number of histogram levels.}
}
\value{
equalized image.
}
\description{
Histogram equalize image
}
\examples{
image <- antsImageRead(getANTsRData("r16"), 2)
eqImage <- histogramEqualizeImage(image)

}
\author{
NJ Tustison
}
