% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogramMatchImage.R
\name{histogramMatchImage}
\alias{histogramMatchImage}
\title{histogramMatchImage}
\usage{
histogramMatchImage(
  sourceImage,
  referenceImage,
  numberOfHistogramBins = 255,
  numberOfMatchPoints = 64,
  useThresholdAtMeanIntensity = FALSE
)
}
\arguments{
\item{sourceImage}{image to undergo intensity transformation.}

\item{referenceImage}{image providing reference intensity profile.}

\item{numberOfHistogramBins}{number of histogram levels.}

\item{numberOfMatchPoints}{number of histogram match points.}

\item{useThresholdAtMeanIntensity}{use a simple background exclusion criterion.}
}
\value{
source image intensity matched to reference image.
}
\description{
Match intensity profile with a reference image.
}
\examples{
sourceImage <- antsImageRead(getANTsRData("r16"), 2)
referenceImage <- antsImageRead(getANTsRData("r64"), 2)
matchedImage <- histogramMatchImage(sourceImage, referenceImage)

}
\author{
NJ Tustison
}
