% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogramMatchImage.R
\name{histogramMatchImage2}
\alias{histogramMatchImage2}
\title{Transform image intensities based on histogram mapping.}
\usage{
histogramMatchImage2(
  sourceImage,
  referenceImage,
  sourceMask = NULL,
  referenceMask = NULL,
  matchPoints = 64,
  transformDomainSize = 255
)
}
\arguments{
\item{sourceImage}{source image.}

\item{referenceImage}{reference image.}

\item{sourceMask}{source mask.}

\item{referenceMask}{reference mask.}

\item{matchPoints}{parametric points at which the intensity transform
displacements are specified between [0, 1].  Alternatively, a single
number can be given and the sequence is linearly spaced in [0, 1].}

\item{transformDomainSize}{Defines the sampling resolution of the B-spline warping.}
}
\value{
warped intensity image
}
\description{
Apply B-spline 1-D maps to an input image for intensity warping.
}
\examples{

library( ANTsR )
sourceImage <- antsImageRead(getANTsRData("r16"), 2)
referenceImage <- antsImageRead(getANTsRData("r64"), 2)
matchedImage <- histogramMatchImage2(sourceImage, referenceImage)
}
\author{
Tustison NJ
}
