% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogramWarpImageIntensities.R
\name{histogramWarpImageIntensities}
\alias{histogramWarpImageIntensities}
\title{Transform image intensities based on histogram mapping.}
\usage{
histogramWarpImageIntensities(
  image,
  breakPoints = c(0.25, 0.5, 0.75),
  displacements = NULL,
  clampEndPoints = c(FALSE, FALSE),
  sdDisplacements = 0.05,
  transformDomainSize = 20
)
}
\arguments{
\item{image}{input image.}

\item{breakPoints}{parametric points at which the intensity transform
displacements are specified between [0, 1].  Alternatively, a single
number can be given and the sequence is linearly spaced in [0, 1].}

\item{displacements}{displacements to define intensity warping.  Length
must be equal to the \code{breakPoints}.  Alternatively, if \code{NULL}
random displacements are chosen (random normal:  mean = 0, sd = \code{sdDisplacements}).}

\item{clampEndPoints}{specify non-zero intensity change at the ends of the histogram.}

\item{sdDisplacements}{characterize the randomness of the intensity displacement.}

\item{transformDomainSize}{Defines the sampling resolution of the B-spline warping.}
}
\value{
warped intensity image
}
\description{
Apply B-spline 1-D maps to an input image for intensity warping.
}
\examples{

library( ANTsR )
image <- antsImageRead( getANTsRData( "r16" ) )
transformedImage <- histogramWarpImageIntensities( image, transformDomainSize = 10 )
rm(image); gc()
rm(transformedImage); gc()
}
\author{
Tustison NJ
}
