% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simlr.R
\name{humanize_anatomical_name}
\alias{humanize_anatomical_name}
\title{Humanize anatomical names by expanding abbreviations and reordering}
\usage{
humanize_anatomical_name(x)
}
\arguments{
\item{x}{character string}
}
\value{
humanized character string
}
\description{
Humanize anatomical names by expanding abbreviations and reordering
}
