% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iBind.R
\name{iBind}
\alias{iBind}
\title{iBind}
\usage{
iBind(img1, img2, along = NA)
}
\arguments{
\item{img1}{input object, an antsImage}

\item{img2}{second antsImage, same size as first}

\item{along}{dimension to bind along}
}
\description{
bind two images along their edge
}
\examples{
fi<-antsImageRead( getANTsRData("r16") , 2 )
mi<-antsImageRead( getANTsRData("r62") , 2 )
bi<-iBind( fi, mi , 1 )
multismoo<- fi \%>\% iBind( smoothImage(fi,2) ) 
multismoo = multismoo \%>\% iBind( smoothImage(fi,4) )

}
\author{
BB Avants
}
