% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigLMStats.R
\name{ilr}
\alias{ilr}
\title{Efficiently compute an image-based linear regression model (ilr)}
\usage{
ilr(dataFrame, voxmats, myFormula)
}
\arguments{
\item{dataFrame}{This data frame contains all relevant predictors except for
the matrices associated with the image variables.}

\item{voxmats}{The named list of matrices that contains the changing predictors.}

\item{myFormula}{This is a character string that defines a valid regression formula.}
}
\value{
A list of different matrices that contain names derived from the
formula and the coefficients of the regression model.
}
\description{
This function simplifies calculating p-values from linear models in which
there is a similar formula that is applied many times with a change in
image-based predictors.  Image-based variables are stored in the input
matrix list. They should be named consistently in the input formula and
in the image list.  If they are not, an error will be thrown.  All input
matrices should have the same number of rows and columns.
}
\examples{

set.seed(1500)
nsub = 100
outcome = rnorm( nsub )
covar = rnorm( nsub )
mat = replicate( nsub, rnorm( nsub ) )
mat2 = replicate( nsub, rnorm( nsub ) )
myform = " outcome ~ covar + vox "
df = data.frame( outcome = outcome, covar = covar )
result = ilr( df, list( vox = mat ), myform)
print( names( result ) )
print( rownames( result$pValue ) )
myform = " vox2 ~ covar + vox "
df = data.frame( outcome = outcome, covar = covar )
result = ilr( df, list( vox = mat, vox2=mat2 ), myform)
print( names( result ) )
print( rownames( result$pValue ) )

}
\author{
BB Avants.
}
