% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composeTransformsToField.R
\name{imageDomainToSpatialMatrix}
\alias{imageDomainToSpatialMatrix}
\title{Image domain to spatial position}
\usage{
imageDomainToSpatialMatrix(image, mask)
}
\arguments{
\item{image}{input image}

\item{mask}{optional input mask}
}
\value{
matrix is output
}
\description{
Transform a d-dimensional image into a d by p matrix that encodes the spatial
coordinates of the image domain.
}
\examples{

img = antsImageRead( getANTsRData( 'r16' ) )
spmat = imageDomainToSpatialMatrix( img )

}
\author{
Avants BB
}

