% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/imageListToMatrix.R
\name{imageListToMatrix}
\alias{imageListToMatrix}
\title{Read Images into a Matrix}
\usage{
imageListToMatrix(imageList, mask)
}
\arguments{
\item{imageList}{A character vector containing a list of image files to
read, in order - these are image objects, not file names.}

\item{mask}{An \code{antsImage} containing a binary mask, voxels in the mask
are placed in the matrix. If not provided, estimated from first image in list.}
}
\value{
A matrix containing the masked data, the result of calling
\code{as.numeric(image, mask)} on each input image.
}
\description{
Read images into rows of a matrix.
}
\examples{
img <- antsImageRead(getANTsRData('r16'), 2)
 imglist <- list()
 nvox <- dim(img)[1] * dim(img)[2]
 nsubj <- 50
 for(ii in 1:nsubj){
   imglist[ii] <- img + rnorm(nvox, sd=mean(img[img!=0]))
 }
 mask <- getMask(img)
 imgmat <- imageListToMatrix(imglist, mask)
}
\author{
Cook PA, Avants B, Kandel BM
}
\seealso{
\code{\link{matrixToImages}, \link{getMask}}
}

