% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/imageMath.R
\name{imageMath}
\alias{imageMath}
\title{imageMath}
\usage{
imageMath(...)
}
\arguments{
\item{...}{all parameters}
}
\description{
Perform math-operations on the given image.  The first argument should
be dimensionality, the 2nd the output image, the third a string
describing the operation.  The fourth should be in the input image.
Additional parameters should be specific for each function.  See the
the imageMath help in ANTs.
}
\examples{
fi<-antsImageRead( getANTsRData("r16") ,2)
mask<-getMask(fi)
imageMath( 2 , fi , "GD", fi , 1 )  # gray matter dilation
imageMath( 2 , mask , "Neg", mask )  # negate
imageMath( 2 , mask , "D", mask )  # distance transform
plot(mask)
}
\author{
Shrinidhi KL
}

