% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageSimilarity.R
\name{imageSimilarity}
\alias{imageSimilarity}
\title{imageSimilarity}
\usage{
imageSimilarity(
  fixed,
  moving,
  type = c("MeanSquares", "MattesMutualInformation", "ANTSNeighborhoodCorrelation",
    "Correlation", "Demons", "JointHistogramMutualInformation"),
  fixed.mask = NULL,
  moving.mask = NULL,
  sampling.strategy = "none",
  sampling.percentage = 1,
  nBins = 32,
  radius = 3
)
}
\arguments{
\item{fixed}{the fixed antsImage}

\item{moving}{the moving antsImage}

\item{type}{image metric to calculate
\itemize{
  \item{MeanSquares}{}
  \item{Correlation}{}
  \item{ANTSNeighborhoodCorrelation}{}
  \item{MattesMutualInformation}{}
  \item{JointHistogramMutualInformation}{}
  \item{Demons}{}
}}

\item{fixed.mask}{mask for the fixed image}

\item{moving.mask}{mask for the moving image}

\item{sampling.strategy}{sampling strategy, default if full sampling
\itemize{
  \item{none}{Full sampling}
  \item{random}{}
  \item{regular}{}
}}

\item{sampling.percentage}{percentage of data to sample when calculating metric}

\item{nBins}{the number of bins for histogram metrics}

\item{radius}{neighborhood radius of ANTSNeighborhoodCorrelation}
}
\value{
value of image to image metric
}
\description{
measure similarity between two images
}
\examples{
x <- antsImageRead(getANTsRData("r16"))
y <- antsImageRead(getANTsRData("r30"))
metric <- imageSimilarity(x, y, type = "MeanSquares")
}
