% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/iMath.R
\name{iMath}
\alias{iMath}
\title{R access to the ANTs program iMath}
\usage{
iMath(...)
}
\arguments{
\item{...}{all parameters}
}
\description{
Perform math-operations on the given image.  The first input is the
input (usually an 'antsImage' object), the second is the operation.
Additional parameters should be specific for each function.  See the
the iMath help in ANTs.
}
\examples{
fi<-antsImageRead( getANTsRData("r16") ,2)
mask<-getMask(fi)
mask[1:256,100:150] = 0
mask = iMath( mask, "GetLargestComponent" )
plot(mask)
}
\author{
Duda JT
}

