% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscaleSVDxpts.R
\name{indicator_opt_both_ways}
\alias{indicator_opt_both_ways}
\title{Helper Function to Optimize Indicator Matrix with Best Sum}
\usage{
indicator_opt_both_ways(m, verbose = FALSE)
}
\arguments{
\item{m}{A numeric matrix for which the optimization will be performed.}

\item{verbose}{Logical. If TRUE, reports the objective value and convergence progress at each iteration.}
}
\value{
m*I
}
\description{
This function runs the optimization on both the input matrix `m` and 
its negative counterpart `m * (-1)`, returning the result that maximizes the sum `sum(m * I)`.
}
\examples{
set.seed(123)
m <- matrix(rnorm(500), nrow = 5)
indicator_opt_both_ways(m)
}
