\name{int_antsExtractXptrAsString-methods}
\docType{methods}
\alias{int_antsExtractXptrAsString-methods}
\alias{int_antsExtractXptrAsString,antsImage-method}
\alias{int_antsExtractXptrAsString,antsImageList-method}
\alias{int_antsExtractXptrAsString,antsMatrix-method}
\alias{int_antsExtractXptrAsString,character-method}
\alias{int_antsExtractXptrAsString,numeric-method}
\title{ Methods for Function \code{int_antsExtractXptrAsString} in Package \pkg{ANTsR} }
\description{
Extract the external pointer in an object with a slot named 'pointer' as an R string. This generic and its methods are for internal use only.
}
\section{Methods}{
\describe{

\item{\code{signature(object = "antsImage")}}{
}

\item{\code{signature(object = "antsImageList")}}{
}

\item{\code{signature(object = "antsMatrix")}}{
}

\item{\code{signature(object = "character")}}{
}

\item{\code{signature(object = "numeric")}}{
}
}}

\keyword{methods}
\keyword{internal}
