% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresholdImage.R
\name{integrateVectorField}
\alias{integrateVectorField}
\title{Integrate vector field}
\usage{
integrateVectorField(
  referenceImage,
  vectorFieldFileName,
  deformationFieldFileName,
  lowerTime = 0,
  upperTime = 1,
  deltaTime = 0.01
)
}
\arguments{
\item{referenceImage}{defines the image domain}

\item{deformationFieldFileName}{the deformation field output file name.}

\item{lowerTime}{the starting time, usually zero for forward transformation
and one for the inverse transformation.}

\item{upperTime}{the ending time, usually one for forward transformation
and zero for the inverse transformation.}

\item{deltaTime}{the integration time step}

\item{velocityFieldFileName}{the velocity field exists on disk.}
}
\description{
Utility function to integrate a vector field and create a deformation field.
}
\examples{
\dontrun{
set.seed(1234)
fi <- ( ri(1) )
mi <- ( ri(2) )
mytx2 <- antsRegistration(fixed=fi, mi, typeofTransform = "SyN" )
integrateVectorField( fi, mytx2$velocityfield,  "/tmp/def.nii.gz" )
}
}
\author{
Avants BB
}
