% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrateVelocityField.R
\name{integrateVelocityField}
\alias{integrateVelocityField}
\title{integrateVelocityField}
\usage{
integrateVelocityField(
  velocityField,
  lowerIntegrationBound = 0,
  upperIntegrationBound = 1,
  numberOfIntegrationSteps = 10
)
}
\arguments{
\item{velocityField}{time-varying displacement field}

\item{lowerIntegrationBound}{Lower time bound for integration in [0, 1]}

\item{upperIntegrationBound}{Upper time bound for integration in [0, 1]}

\item{numberOfIntegrationSteps}{Number of integration steps used in the Runge-Kutta solution}
}
\value{
integrated field
}
\description{
Integrate velocity field
}
\examples{
fi <- antsImageRead( getANTsRData( "r16" ) )
mi <- antsImageRead( getANTsRData( "r27" ) )
reg <- antsRegistration( fi, mi, "TV[2]" )
velocityField <- antsImageRead( reg$velocityfield )
field <- integrateVelocityField( velocityField, 0.0, 1.0, 10 )

}
\author{
NJ Tustison
}
