% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresholdImage.R
\name{integrateVelocityField2}
\alias{integrateVelocityField2}
\title{Integrate velocity field}
\usage{
integrateVelocityField2(
  referenceImage,
  velocityFieldFileName,
  deformationFieldFileName,
  lowerTime = 0,
  upperTime = 1,
  deltaTime = 0.01
)
}
\arguments{
\item{referenceImage}{defines the image domain}

\item{velocityFieldFileName}{the velocity field exists on disk.}

\item{deformationFieldFileName}{the deformation field output file name.}

\item{lowerTime}{the starting time, usually zero for forward transformation
and one for the inverse transformation.}

\item{upperTime}{the ending time, usually one for forward transformation
and zero for the inverse transformation.}

\item{deltaTime}{the integration time step}
}
\description{
Utility function to integrate a velocity field and create a deformation field.
}
\examples{
\dontrun{
set.seed(1234)
fi <- ( ri(1) )
mi <- ( ri(2) )
mytx2 <- antsRegistration(fixed=fi, mi, typeofTransform = "TV[4]" )
integrateVelocityField2( fi, mytx2$velocityfield,  "/tmp/def.nii.gz" )
qq=antsApplyTransforms( fi, mi, mytx2$fwdtransforms )
pp=antsApplyTransforms( fi, mi, "/tmp/def.nii.gz" )
antsImageMutualInformation( fi, mi )
antsImageMutualInformation( fi, qq )
antsImageMutualInformation( fi, pp )
}
}
\author{
Avants BB
}
