% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simlr.R
\name{inv_sqrt_sym_newton}
\alias{inv_sqrt_sym_newton}
\title{Inverse square root of a symmetric matrix via Newton-Schulz}
\usage{
inv_sqrt_sym_newton(
  A,
  epsilon = 1e-10,
  max_iter = 10L,
  tol = 1e-06,
  verbose = FALSE
)
}
\description{
Inverse square root of a symmetric matrix via Newton-Schulz
}
