% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simlr.R
\name{invariant_orthogonality_defect}
\alias{invariant_orthogonality_defect}
\title{Calculate the invariant orthogonality defect}
\usage{
invariant_orthogonality_defect(A)
}
\arguments{
\item{A}{input matrix}
}
\description{
Calculate the invariant orthogonality defect
}
