% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invariantImageSimilarity.R
\name{invariantImageSimilarity}
\alias{invariantImageSimilarity}
\title{similarity metrics between two images as a function of geometry}
\usage{
invariantImageSimilarity(in_image1, in_image2, localSearchIterations = 0,
  metric = "MI", thetas, thetas2, thetas3, scaleImage = 1,
  doReflection = 0, txfn = "")
}
\arguments{
\item{in_image1}{reference image}

\item{in_image2}{moving image}

\item{localSearchIterations}{integer controlling local search in multistart}

\item{metric}{which metric MI or GC (string)}

\item{thetas}{numeric vector of search angles in degrees}

\item{thetas2}{numeric vector of search angles in degrees around principal axis 2 (3D)}

\item{thetas3}{numeric vector of search angles in degrees around principal axis 3 (3D)}

\item{scaleImage}{global scale}

\item{doReflection}{reflect image about principal axis}

\item{txfn}{if present, write optimal tx to .mat file}
}
\value{
dataframe with metric values and transformation parameters
}
\description{
compute similarity metric between two images as image is rotated about its
center w/or w/o optimization
}
\examples{
fi<-antsImageRead( getANTsRData("r16") )
mi<-antsImageRead( getANTsRData("r64") )
mival<-invariantImageSimilarity( fi, mi, thetas = c(0,10,20) )

}
\author{
Brian B. Avants
}
\keyword{image}
\keyword{similarity}

