% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsrTransform_class.R
\name{invertAntsrTransform}
\alias{invertAntsrTransform}
\title{invertAntsrTransform}
\usage{
invertAntsrTransform(transform)
}
\arguments{
\item{transform}{trasform to invert}
}
\value{
antsrTransform
}
\description{
invert a linear antsrTransform
}
\examples{
img <- antsImageRead(getANTsRData("r16"))
tx = new("antsrTransform", precision="float", type="AffineTransform", dimension=2 )
setAntsrTransformParameters(tx, c(0,-1,1,0,dim(img)[1],0) )
txinv =  invertAntsrTransform(tx)
}

