% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invertDisplacementField.R
\name{invertDisplacementField}
\alias{invertDisplacementField}
\title{invertDisplacementField}
\usage{
invertDisplacementField(
  displacementField,
  inverseFieldInitialEstimate,
  maximumNumberOfIterations = 20,
  meanErrorToleranceThreshold = 0.001,
  maxErrorToleranceThreshold = 0.1,
  enforceBoundaryCondition = TRUE
)
}
\arguments{
\item{displacementField}{displacement field.}

\item{inverseFieldInitialEstimate}{initial guess.}

\item{maximumNumberOfIterations}{number of iterations.}

\item{meanErrorToleranceThreshold}{mean error tolerance threshold.}

\item{maxErrorToleranceThreshold}{max error tolerance threshold.}

\item{enforceBoundaryCondition}{enforce stationary boundary condition.}
}
\value{
inverse displacement field
}
\description{
Invert displacement field.
}
\author{
NJ Tustison
}
