% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsImage_class.R
\name{is.antsImage}
\alias{is.antsImage}
\title{is.antsImage}
\usage{
is.antsImage(x)
}
\arguments{
\item{x}{An object}
}
\value{
TRUE if \code{object} is antsImage; FALSE otherwise.
}
\description{
Tests if object is antsImage
}
\examples{
is.antsImage(antsImageRead(getANTsRData("r16"), 2))
}
