% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_data.R
\docType{data}
\name{jhu_md_labels}
\alias{jhu_md_labels}
\title{jhu_md_labels}
\format{
A data frame with columns: Label, Description.
}
\usage{
jhu_md_labels
}
\description{
A data frame label numbers and names for the JHU MD white matter labels.
}
\keyword{datasets}
