% Generated by roxygen2 (4.0.2): do not edit by hand
\name{jointIntensityFusion}
\alias{jointIntensityFusion}
\title{Estimates an image from another set of images}
\usage{
outlist<-jointIntensityFusion(targetI, targetIMask,
     atlasList, beta = 1, rad = NA,
     labelList = NA, doscale = FALSE)
}
\arguments{
\item{targetI}{antsImage to be approximated}

\item{targetIMask}{mask with value 1}

\item{atlasList}{list containing antsImages}

\item{beta}{weight sharpness}

\item{rad}{neighborhood radius}

\item{labelList}{list containing antsImages - not yet used}

\item{doscale}{scale neighborhood intensities}
}
\value{
approximated image
}
\description{
intensity generalization of joint label fusion.
}
\examples{
set.seed(123)
fi<-antsImageRead( getANTsRData('r16') ,2)
mi<-antsImageRead( getANTsRData('r64') ,2)
mi2<-antsImageRead( getANTsRData('r64') ,2)
mi3<-antsImageRead( getANTsRData('r64') ,2)
refmask<-getMask(fi)
ilist<-list(mi,mi2,mi3)
pp<-jointIntensityFusion(ref,refmask,ilist)
}
\author{
Brian B. Avants, Hongzhi Wang, Paul Yushkevich
}
\keyword{fusion,}
\keyword{template}

