% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/jointIntensityFusion3D.R
\name{jointIntensityFusion3D}
\alias{jointIntensityFusion3D}
\title{jointIntensityFusion3D}
\usage{
jointIntensityFusion3D(targetI, targetIMask, atlasList, beta = 1, rad = NA,
  labelList = NA, doscale = TRUE, doNormalize = TRUE,
  maxAtlasAtVoxel = c(1, Inf), rho = 0.1, useSaferComputation = FALSE,
  usecor = FALSE)
}
\arguments{
\item{targetI}{antsImage to be approximated}

\item{targetIMask}{mask with value 1}

\item{atlasList}{list containing antsImages}

\item{beta}{weight sharpness, default to 2}

\item{rad}{neighborhood radius, default to 4}

\item{labelList}{list containing antsImages}

\item{doscale}{scale neighborhood intensities}

\item{doNormalize}{normalize each image range to 0, 1}

\item{maxAtlasAtVoxel}{min/max n atlases to use at each voxel}

\item{rho}{ridge penalty increases robustness to outliers}

\item{useSaferComputation}{slower but more error checking}

\item{usecor}{employ correlation as local similarity}
}
\value{
approximated image, segmentation and probabilities
}
\description{
Estimates an image/labelset from another set of 3D images
}
\details{
intensity generalization of joint label fusion, still supports segmentation.
this version is more efficient, memory-wise, for 3D images. it is a thin
wrapper.
}
\examples{
# see jointIntensityFusion
}
\author{
Brian B. Avants, Hongzhi Wang, Paul Yushkevich
}
\keyword{fusion,}
\keyword{template}

