% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscaleSVDxpts.R
\name{jointSmoothMatrixReconstruction}
\alias{jointSmoothMatrixReconstruction}
\title{joint smooth matrix prediction}
\usage{
jointSmoothMatrixReconstruction(x, nvecs, parameters, iterations = 10,
  gamma = 1e-06, sparsenessQuantile = 0.5, positivity = FALSE,
  smoothingMatrix = NA, smoothingWeight = 0.5, rowWeights = NA,
  repeatedMeasures = NA, verbose = FALSE)
}
\arguments{
\item{x}{input list of matrices to be jointly predicted.}

\item{nvecs}{number of basis vectors to compute}

\item{parameters}{should be a ncomparisons by 3 matrix where the first two
columns define the pair to be matched and the last column defines the weight
in the objective function.}

\item{iterations}{number of gradient descent iterations}

\item{gamma}{step size for gradient descent}

\item{sparsenessQuantile}{quantile to control sparseness - higher is sparser}

\item{positivity}{restrict to positive solution (beta) weights}

\item{smoothingMatrix}{a list containing smoothing matrices of the same
length as x.}

\item{smoothingWeight}{between zero and one, increases smoothing.}

\item{rowWeights}{vectors of weights with size n (assumes diagonal covariance)}

\item{repeatedMeasures}{list of repeated measurement identifiers. this will
allow estimates of per identifier intercept.}

\item{verbose}{boolean option}
}
\value{
matrix list each of size p by k is output
}
\description{
Joints reconstruct a n by p, q, etc matrices or predictors.
The reconstruction can be regularized.
# norm( x - u_x v_x^t ) + norm( y - u_y v_y^t) + .... etc
}
\examples{

\dontrun{
mat<-replicate(100, rnorm(20))
mat2<-replicate(100, rnorm(20))
mat<-scale(mat)
mat2<-scale(mat2)
wt<-0.666
mat3<-mat*wt+mat2*(1-wt)
params = matrix( nrow = 2, ncol = 3 )
params[1,] = c(1,2,1)
params[2,] = c(2,1,1)
x = list( (mat), (mat3 ))
jj = jointSmoothMatrixReconstruction( x, 2, params,
 gamma = 1e-4, sparsenessQuantile=0.5, iterations=10,
 smoothingMatrix = list(NA,NA), verbose=TRUE )
}
}
\author{
Avants BB
}
