% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscaleSVDxpts.R
\name{knnSmoothImage}
\alias{knnSmoothImage}
\title{k-nearest neighbors constrained image smoothing}
\usage{
knnSmoothImage(img, mask, radius, intensitySigma = 50000, spatialSigma = 20,
  iterations = 1)
}
\arguments{
\item{img}{input image to smooth}

\item{mask}{input image mask}

\item{radius}{number of neighbors, higher causes more smoothing}

\item{intensitySigma}{sigma for gaussian defining intensity differences}

\item{spatialSigma}{for gaussian defining spatial distances}

\item{iterations}{number of iterations over which to apply smoothing kernel}
}
\value{
antsImage is output
}
\description{
Compute a smoothing matrix based on an input matrix of point coordinates as
well as neighborhood intensity patterns.  this performs a form of edge
preserving smoothing.
}
\examples{

\dontrun{
img = antsImageRead( getANTsRData( 'r16' ) )
mask = getMask( img )
simg = knnSmoothImage( img=img, mask=mask, radius=2, intensitySigma=5000,
  spatialSigma=1.5, iterations=1 )
}
}
\author{
Avants BB
}
