% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/LabelGeometryMeasures.R
\name{LabelGeometryMeasures}
\alias{LabelGeometryMeasures}
\title{LabelGeometryMeasures}
\usage{
LabelGeometryMeasures(...)
}
\arguments{
\item{...}{see ants function}
}
\value{
none
}
\description{
Wrapper for the ANTs funtion LabelGeometryMeasures
}
\author{
Avants BB
}

