% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/labelImageCentroids.R
\name{labelImageCentroids}
\alias{labelImageCentroids}
\title{labelImageCentroids}
\usage{
labelImageCentroids(img, physical = FALSE, convex = TRUE)
}
\arguments{
\item{img}{ants image of labels (ints)}

\item{physical}{boolean if you want physical space coordinates or not}

\item{convex}{- if TRUE, return centroid, if FALSE return point
with min average distance to other  points with same label}
}
\value{
list with labels , array of label values, and
centroids <- coordinates of label centroids
}
\description{
Converts a label image to coordinates summarizing their positions
}
\examples{
labelImage<-makeImage( c(2,2,2) , 0:7 )
labelImageCentroids(labelImage)
}
\author{
Avants BB, Duda JT
}

