% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelOverlapMeasures.R
\name{labelOverlapMeasures}
\alias{labelOverlapMeasures}
\title{labelOverlapMeasures}
\usage{
labelOverlapMeasures(sourceLabelImage, targetLabelImage)
}
\arguments{
\item{sourceLabelImage}{label image for source image.}

\item{targetLabelImage}{label image for target/reference image.}
}
\value{
data frame with overlap measures
}
\description{
Wrapper for the ANTs funtion LabelOverlapMeasures
}
\author{
Avants BB, Tustison NJ
}
