% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/labelStats.R
\name{labelStats}
\alias{labelStats}
\title{labelStats}
\usage{
labelStats(image, labelImage)
}
\arguments{
\item{image}{Image to calculate statistics from.}

\item{labelImage}{Label image.}
}
\value{
Data frame with columns :
  LabelValues, Mean, Min, Max, Variance, Count, Volume
}
\description{
Get label statistics from image.
}
\examples{
img <- antsImageRead( getANTsRData("r16") , 2 )
img <- resampleImage( img, c(64,64), 1, 0 )
mask <- getMask(img)
segs1 <- atropos( d = 2, a = img, m = '[0.2,1x1]',
 c = '[2,0]',  i = 'kmeans[3]', x = mask )
labelStats(img, segs1$segmentation)
}

