% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/landmarkTransform.R
\name{landmarkTransform}
\alias{landmarkTransform}
\title{Create an itk transform from landmarks (rigid, similarity and affine)}
\usage{
landmarkTransform(lmFix, lmMoving, type = c("rigid", "similarity", "affine"),
  RAS2IJK = NULL, ...)
}
\arguments{
\item{lmFix}{fix landmarks}

\item{lmMoving}{moving landmarks}

\item{type}{of transform (can be "rigid","similarity" or "affine")}

\item{RAS2IJK}{4x4 (3D) or 3x3 (2D) matrix with transform from RAS to IJK space (e.g. when landmarks are placed with Slicer).}

\item{...}{additional parameters (such as landmark weights) to be passed to \code{\link{rotonto}}, in case type != affine.}
}
\value{
writes a transform to a tempfile and returns the path to that transform
}
\description{
Create an itk transform from landmarks (rigid, similarity and affine) to transform an image
}
\details{
lmMoving are landmarks placed on the image to be transformed to the position of lmFix
}
\examples{
fi <- antsImageRead(getANTsRData("r16") ,2)
mi <- antsImageRead(getANTsRData("r64") ,2)
load(system.file("extdata/landmarks.RData",package="ANTsR"))
trafo <- landmarkTransform(miLM,fiLM,RAS2IJK = diag(c(-1,-1,1)))
fiTransformed <- antsApplyTransforms(mi,fi,trafo)
plot(fiTransformed)
}
\author{
Schlager S
}

