% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.ANTsR.R
\name{load.ANTsR}
\alias{load.ANTsR}
\title{Load R sessions with ANTsR objects}
\usage{
load.ANTsR(filename=file.path('.','.ANTsRsession'), env=as.environment(1))
}
\arguments{
\item{filename}{path of the saved ANTsR session. If not
specified, the default behavior is to search for the 
folder \".ANTsRsesssion\" in the current working directory.}

\item{env}{the environment level. Don't change unless you know
what are you doing.}
}
\description{
Loads a previously saved ANTsR session or object. Simply
point the function to the folder where you previously
saved the data with save.ANTsR.
}
\author{
Dorian Pustina
}
