% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_data.R
\docType{data}
\name{lobes}
\alias{lobes}
\title{lobes}
\format{
A data frame with columns: Label, Description, Side, SideNumber.
}
\usage{
lobes
}
\description{
A data frame label numbers and names for brain lobes.
}
\keyword{datasets}
