% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_data.R
\docType{data}
\name{lobes_brainstem}
\alias{lobes_brainstem}
\title{lobes_brainstem}
\format{
A data frame with columns: Label, Description, Side.
}
\usage{
lobes_brainstem
}
\description{
A data frame label numbers and names for lobes and brainstem regions.
}
\keyword{datasets}
