% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whiten.R
\name{localGyrificationIndex}
\alias{localGyrificationIndex}
\title{Local gyrification index.}
\usage{
localGyrificationIndex(segmentation, sigma = 3, k = 25, ksigma = 3)
}
\arguments{
\item{segmentation}{- potentially from antsCorticalThickness}

\item{sigma}{- scale parameter}

\item{k}{- size of neighborhood}

\item{ksigma}{- sigma for k-neighborhood}
}
\value{
lGI image
}
\description{
Estimate the gyrification index at a specific scale.  This will produce images of
local sulcification and gyrification as as well as the difference of these
two measurements.  The most gyral values are near 1 and most sulcal negative 1.
The function is not limited to brain data and should work on any 3D shape.
See the paper entitled the shape operator for differential image analysis.
}
\examples{
\dontrun{
seg = antsImageRead( 'segmentation.nii.gz' )
wm = thresholdImage( seg, 3, 3 ) \%>\% morphology( "dilate", 1 )
lgi = localGyrificationIndex( wm )
}

}
\author{
Avants BB
}
