% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jointLabelFusion.R
\name{localJointLabelFusion}
\alias{localJointLabelFusion}
\title{local joint label and intensity fusion}
\usage{
localJointLabelFusion(
  targetI,
  whichLabels,
  targetMask,
  initialLabel,
  atlasList,
  labelList,
  submaskDilation = 10,
  typeofTransform = "SyN",
  affMetric = "meansquares",
  synMetric = "mattes",
  synSampling = 32,
  regIterations = c(40, 20, 0),
  affIterations,
  localMaskTransform,
  maxLabelPlusOne = FALSE,
  noZeroes = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{targetI}{antsImage to be labeled}

\item{whichLabels}{label number(s) from the library on which to focus}

\item{targetMask}{a mask for the target image (optional), passed to joint fusion}

\item{initialLabel}{the initial approximate label(s) for the target region.}

\item{atlasList}{list containing antsImages with intensity images}

\item{labelList}{list containing antsImages with segmentation labels}

\item{submaskDilation}{amount to dilate initial mask to define region on which
we perform focused registration}

\item{typeofTransform}{passed to \code{antsRegistration}.}

\item{affMetric}{the metric for the affine part (GC, mattes, meansquares)}

\item{synMetric}{the metric for the syn part (CC, mattes, meansquares, demons)}

\item{synSampling}{the nbins or radius parameter for the syn metric}

\item{regIterations}{vector of iterations for syn.  we will set the smoothing
and multi-resolution parameters based on the length of this vector.
passed to \code{antsRegistration}.}

\item{affIterations}{vector of iterations for low-dimensional transforms.}

\item{localMaskTransform}{type of transform for local mask initialization;
would usually set to Rigid, Similarity or Affine}

\item{maxLabelPlusOne}{boolean
this will add max label plus one to the non-zero parts of each label where the target mask
is greater than one.  NOTE: this will have a side effect of adding to the original label
images that are passed to the program.  It also guarantees that every position in the
labels have some label, rather than none.  Ie it guarantees to explicitly parcellate the
input data.}

\item{noZeroes}{boolean will zero out target mask regions that have any zero label.
this prevents JLF from computing a solution in regions not covered by the initial library.}

\item{verbose}{boolean}

\item{...}{extra parameters passed to JLF}
}
\value{
label probabilities and segmentations
}
\description{
A local version of joint label fusion that focuses on one or more specific labels.
This is primarily different from standard JLF because it performs registration
on a per label basis and focuses JLF on the label(s) alone.  It requires an
initial segmentation of the target region which can be provided either by
a manual or automated initialization.  Registration by SyN is a good choice
for the latter approach.
}
\author{
Brian B. Avants
}
\keyword{fusion,}
\keyword{template}
