% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/networkEiganat.R
\name{lowrankRowMatrix}
\alias{lowrankRowMatrix}
\title{Produces a low rank version of the input matrix}
\usage{
lowrankRowMatrix(A, k = 2)
}
\arguments{
\item{A}{input matrix}

\item{k}{rank to use}
}
\value{
matrix is output
}
\description{
Produces a low rank version of the input matrix
}
\examples{
mat <- matrix(rnorm(300),ncol=50)
lrmat <- lowrankRowMatrix( mat , 2 )
}
\author{
Avants BB
}

