% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePowersPointsImage.R
\name{makePowersPointsImage}
\alias{makePowersPointsImage}
\title{Label image from Powers points}
\usage{
makePowersPointsImage(pts, mask, radius = 5)
}
\arguments{
\item{pts}{input powers points}

\item{mask}{antsImage mask defining target space}

\item{radius}{for the points}
}
\value{
antsImage is output
}
\description{
Creates spherical points in the coordinate space of the target image.
}
\examples{
\dontrun{
mni <- antsImageRead(getANTsRData("mni")) \%>\% getMask()
data("powers_areal_mni_itk", package = "ANTsR", envir = environment())
powersLabels <- makePowersPointsImage(powers_areal_mni_itk, mni)
# if you have a different target space
# pts = antsApplyTransformsToPoints( 3, powers_areal_mni_itk,
#          transformlist = concatenatedMaps$toTemplate,
#          whichtoinvert = concatenatedMaps$toTemplateInversion )
}

}
\author{
Avants BB, Duda JT
}
