% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscaleSVDxpts.R
\name{mapAsymVar}
\alias{mapAsymVar}
\title{Convert left/right variables to a measure of asymmetry}
\usage{
mapAsymVar(
  mydataframe,
  leftvar,
  leftname = "left",
  rightname = "right",
  replacer = "Asym"
)
}
\arguments{
\item{mydataframe}{dataframe containing relevant variables}

\item{leftvar}{left side variable names ie the full names of the variables to asym}

\item{leftname}{the variable substring indicating left side}

\item{rightname}{the variable substring indicating right side}

\item{replacer}{string to replace left with in column names of output}
}
\value{
fixed x
}
\description{
Convert left/right variables to a measure of asymmetry
}
\author{
Avants BB
}
