% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist.antsImage.R
\name{mask_values}
\alias{mask_values}
\title{Simple Value extractor for antsImage Values}
\usage{
mask_values(x, mask)
}
\arguments{
\item{x}{Object of class \code{antsImage}}

\item{mask}{object to subset the image.  If missing, then all
values of the image are plotted.}
}
\value{
Vector of values
}
\description{
Takes in a mask and an image and then returns the values
}
\examples{
img = makeImage(c(10,10),rnorm(100))
mask = img > 0
mask_values(img, mask)
mask_values(img)
}
