% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/matrix2timeseries.R
\name{matrix2timeseries}
\alias{matrix2timeseries}
\title{Simple matrix2timeseries function.}
\usage{
matrix2timeseries(referenceImage, maskImage, timeSeriesMatrix)
}
\arguments{
\item{referenceImage}{reference 4D image}

\item{maskImage}{mask image defining voxels of interest}

\item{timeSeriesMatrix}{matrix to convert to image}
}
\value{
antsImage in 4D is output
}
\description{
matrix2timeseriess converts a matrix to a 4D image.
}
\author{
Avants BB
}

