% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_Summary.R
\name{mean.antsImage}
\alias{mean.antsImage}
\title{Mean for antsImage Objects}
\usage{
\method{mean}{antsImage}(x, ..., mask = NULL, na.rm = FALSE)
}
\arguments{
\item{x}{is an object of class \code{antsImage}.}

\item{...}{additional arguments to pass to \code{\link{mean}}}

\item{mask}{binary mask of values to subset}

\item{na.rm}{Remove missing values}
}
\description{
Overloaded Mean for antsImage objects
}
\details{
#' @rdname antsImagemath
#' @aliases xor,antsImage-method
setMethod(f = "xor", signature(x = "antsImage",
                             y = "antsImage"), definition = function(x) {
  xx = as.array(x)
  yy = as.array(y)
  xor(xx, yy)
  !a2
})
}
\examples{
img =  makeImage(c(10, 10, 10), rnorm(1000))
mean(img)
mean(img, na.rm = TRUE)
mean(img > 0)
mean(img > 0, na.rm = TRUE)
mean(img, mask = img > 0)
mean(img, mask = img > 0,  na.rm = TRUE)
arr = as.array(img)
arr[1,1,1] = NA
img = as.antsImage(arr, reference = img)
m = mean(img)
stopifnot(is.nan(m))
mean(img, na.rm = TRUE)
m = mean(0 < img)
stopifnot(is.nan(m))
m = mean(img > 0)
# stopifnot(is.nan(m))
mean(img > 0, na.rm = TRUE)
m = mean(img, mask = img > 0)
stopifnot(is.nan(m))
mean(img, mask = img > 0, na.rm = TRUE)
}
