% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscaleSVDxpts.R
\name{measure_orthogonality}
\alias{measure_orthogonality}
\title{Measure the Orthogonality of a Matrix}
\usage{
measure_orthogonality(mat)
}
\arguments{
\item{mat}{A square numeric matrix.}
}
\value{
A numeric value indicating the deviation from orthogonality.
}
\description{
Computes a measure of orthogonality for a square matrix.
}
\examples{
A <- matrix(c(1, 0, 0, 1), nrow = 2)
measure_orthogonality(A) # Should be 0 for an orthogonal matrix
B <- matrix(c(1, 1, 0, 1), nrow = 2)
measure_orthogonality(B) # Should be greater than 0 for a non-orthogonal matrix
}
