% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscaleSVDxpts.R
\name{measure_orthogonality_norm}
\alias{measure_orthogonality_norm}
\title{Measure the Orthogonality of a Matrix}
\usage{
measure_orthogonality_norm(mat)
}
\arguments{
\item{mat}{A numeric matrix.}
}
\value{
A numeric value representing the orthogonality measure of the matrix.
}
\description{
This function calculates a measure of how far a given matrix is from being orthogonal.
The measure is invariant to the Frobenius norm of the input matrix and is zero if the matrix is diagonal.
}
\examples{
mat <- matrix(c(1, 2, 3, 4, 5, 6, 7, 8, 9), nrow = 3)
measure_orthogonality_norm(mat)
}
