% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measureImageSimilarity.R
\name{measureImageSimilarity}
\alias{measureImageSimilarity}
\title{Wrapper for the ANTs function measureImageSimilarity}
\usage{
measureImageSimilarity(fixedImage, movingImage, fixedMask = NA,
  movingMask = NA, metric = "cc", radiusOrNumberOfBins = 4,
  samplingStrategy = "none", samplingPercentage = 1)
}
\arguments{
\item{fixedImage}{Fixed image.}

\item{movingImage}{Moving image.}

\item{fixedMask}{Optional mask limiting the voxels considered by the metric.}

\item{movingMask}{Optional mask limiting the voxels considered by the metric.}

\item{metric}{Possible options include: ANTs neighborhoo cross-correlation
('cc'), joint mutual information ('mi'), Mattes mutual information ('mattes'),
mean squares ('meansquares'), demons ('demons'), global cross-correlation
('gc').}

\item{radiusOrNumberOfBins}{Parameter which depends on the metric.  For 'cc',
this value determines the radius of the correlation window.  For either of
the mutual information metrics, this determines the number of bins.}

\item{samplingStrategy}{Sampling strategy employed by the metric.  Either
'none', 'regular', or 'random'.}

\item{samplingPercentage}{Metric sampling percentage (in the range [0,1]).}
}
\value{
scalar value
}
\description{
Calculate the similarity between a pair of images using metrics available
via ITK and optional fixed and/or moving image masks.
}
\examples{

fixedImage <- antsImageRead( getANTsRData( "r16" ), 2 )
movingImage <- antsImageRead( getANTsRData( "r64" ), 2 )
value <- measureImageSimilarity( fixedImage, movingImage )

}
\author{
Tustison NJ, Avants BB
}
