% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_Summary.R
\name{median.antsImage}
\alias{median.antsImage}
\title{Median for antsImage Objects}
\usage{
\method{median}{antsImage}(x, na.rm = FALSE, ..., mask = NULL)
}
\arguments{
\item{x}{is an object of class \code{antsImage}.}

\item{na.rm}{a logical value indicating whether NA should be removed}

\item{...}{additional arguments to send to \code{median}}

\item{mask}{is an object of class \code{antsImage}}
}
\description{
Overloaded Median for antsImage objects
}
