% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeChannels.R
\name{mergeChannels}
\alias{mergeChannels}
\title{merge images into a multiChannel antsImage}
\usage{
mergeChannels(imageList)
}
\arguments{
\item{imageList}{a list of antsImage objects to merge}
}
\value{
A multiChannel antsImage object
}
\description{
merge images into a multiChannel antsImage
}
\examples{
 r <- floor( seq(1:(64*64)) / (64*64) * 255 )
 dim(r) <- c(64,64)
 r <- as.antsImage(r)
 g <- r*0
 b <- r*0
 rgbImage = mergeChannels( list(r,g,b) )

}
\author{
Duda, JT
}
