% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitNDImageToList.R
\name{mergeListToNDImage}
\alias{mergeListToNDImage}
\title{mergeListToNDImage}
\usage{
mergeListToNDImage(img, imgList)
}
\arguments{
\item{img}{input image of class antsImage, last dimension will be split}

\item{imgList}{input list to be merged back into dimension d space}
}
\value{
output list filled with d-1 volumes of class antsImage
}
\description{
This function will copy a list of d-1 images into a d-dimension target image.
This function reverses the operation performed by \code{splitNDImageToList}.
}
\examples{

my4Dimage = makeImage( c(5,5,5,4), rnorm( 5*5*5*4 ) )
my3Dlist = splitNDImageToList( my4Dimage )
my4DimageR = mergeListToNDImage( my4Dimage, my3Dlist )

}
\author{
Pustina D, Avants B
}
