% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphology.R
\name{morphology}
\alias{morphology}
\title{Morphological image operations}
\usage{
morphology(input, operation, radius, type = "binary", value = 1,
  shape = "ball", radiusIsParametric = FALSE, thickness = 1, lines = 3,
  includeCenter = FALSE)
}
\arguments{
\item{input}{input image}

\item{operation}{operation to apply
\itemize{
\item{"close"}{ Morpholgical closing}
\item{"dilate"}{ Morpholgical dilation}
\item{"erode"}{ Morpholgical erosion}
\item{"open"}{ Morpholgical opening}
}}

\item{radius}{radius of structuring element}

\item{type}{type of morphology
\itemize{
\item{"binary"}{ Binary operation on a single value}
\item{"grayscale"}{ Grayscale operations}
}}

\item{value}{value to operation on (type='binary' only)}

\item{shape}{shape of the structuring element ( type='binary' only )
\itemize{
\item{"ball"}{ spherical structuring element}
\item{"box"}{ box shaped structuring element}
\item{"cross"}{ cross shaped structuring element}
\item{"annulus"}{ annulus shaped structuring element}
\item{"polygon"}{ polygon structuring element}
}}

\item{radiusIsParametric}{used parametric radius boolean (shape='ball' and shape='annulus' only)}

\item{thickness}{thickness (shape='annulus' only)}

\item{lines}{number of lines in polygon (shape='polygon' only)}

\item{includeCenter}{include center of annulus boolean (shape='annulus' only)}
}
\value{
antsImage is output
}
\description{
Apply imaage morphological operations
}
\examples{

fi<-antsImageRead( getANTsRData("r16") , 2 )
mask<-getMask( fi )
dilatedBall = morphology( mask, operation="dilate", radius=3, type="binary", shape="ball")
erodedBox = morphology( mask, operation="erode", radius=3, type="binary", shape="box")
openedAnnulus = morphology( mask, operation="open", radius=5, type="binary", shape="annulus", thickness=2)

}
\author{
Avants BB
}

