% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ants_motion_estimation.R
\name{.motion_correction}
\alias{.motion_correction}
\title{Simple motion_correction function.}
\usage{
.motion_correction(
  img,
  fixed = NULL,
  moreaccurate = 1,
  txtype = "Affine",
  verbose = FALSE
)
}
\arguments{
\item{img}{4D antsImage}

\item{fixed}{target fixed image}

\item{moreaccurate}{0, 1 or 2 with higher values being more accurate
(use 2 for real applications, 0 for testing)}

\item{txtype}{Transformation type}

\item{verbose}{verbosity boolean}
}
\value{
list of outputs
}
\description{
motion corrects 4D time series imaging data
}
\note{
This function may give different results on multiple runs.
}
\examples{
testimg <- makeImage(c(10, 10, 10, 5), rnorm(5000))
testimg <- iMath(testimg, "PadImage", 5)
mocorr <- .motion_correction(testimg)
aimg_to_array <- function(x) {
  if (is.antsImage(x)) {
    x <- as.array(x)
  }
  x
}
amocorr <- lapply(mocorr, aimg_to_array)
mocorr2 <- .motion_correction(testimg)
amocorr2 <- lapply(mocorr2, aimg_to_array)

# This function may give different results on multiple runs
# without setting the seed
}
\author{
Avants BB
}
