% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mrvnrfs.R
\name{mrvnrfs.predict}
\alias{mrvnrfs.predict}
\title{multi-res voxelwise neighborhood random forest segmentation}
\usage{
mrvnrfs.predict(rflist, x, labelmask, rad = NA, multiResSchedule = c(4, 2,
  1), asFactors = TRUE)
}
\arguments{
\item{rflist}{a list of random forest models from mrvnrfs}

\item{x}{a list of lists where each list contains feature images}

\item{labelmask}{a mask for the features (all in the same image space)}

\item{rad}{vector of dimensionality d define nhood radius}

\item{multiResSchedule}{an integer vector defining multi-res levels}

\item{asFactors}{boolean - treat the y entries as factors}
}
\value{
list a 4-list with the rf model, training vector, feature matrix
and the random mask
}
\description{
Represents multiscale feature images as a neighborhood and uses the features
to apply a random forest segmentation model to a new image
}
\author{
Avants BB, Tustison NJ
}

