% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_data.R
\docType{data}
\name{mtl_labels}
\alias{mtl_labels}
\title{mtl_labels}
\format{
A data frame with columns: Label, Description.
}
\usage{
mtl_labels
}
\description{
A data frame label numbers and names for medial temporal lobe regions.
}
\keyword{datasets}
