% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multichannelPCA.R
\name{multichannelToVector}
\alias{multichannelToVector}
\title{Convert multichannel \code{antsImage} to a vector.}
\usage{
multichannelToVector(multichannelimage, mask)
}
\arguments{
\item{multichannelimage}{input multichannel image}

\item{mask}{mask of same dimensionality as multichannelimage}
}
\value{
vector is output
}
\description{
Employs a mask to flatten each channel of a multi-channel image into a
contiguous piece of a vector.
}
\examples{

# see vectorToMultichannel

}
\author{
Avants BB
}
