% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscaleSVDxpts.R
\name{multigrep}
\alias{multigrep}
\title{Grep entries with a vector search parameters}
\usage{
multigrep(x, desc, intersect = FALSE)
}
\arguments{
\item{x}{a vector of search terms}

\item{desc}{target vector of items to be searched}

\item{intersect}{boolean whether to use intersection or union otherwise}
}
\value{
result of grep (indices of desc that match x)
}
\description{
Grep entries with a vector search parameters
}
\author{
Avants BB
}
