% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscaleSVDxpts.R
\name{multiview_pca}
\alias{multiview_pca}
\title{Multiview PCA}
\usage{
multiview_pca(
  views,
  n_components,
  sparse = 0.5,
  max_iter = 100,
  sparsenessAlg = "basic",
  verbose = FALSE
)
}
\arguments{
\item{views}{A list of data matrices for each view.}

\item{n_components}{Number of principal components to compute.}

\item{sparse}{vector of length views with values between zero and one}

\item{max_iter}{Maximum number of iterations for the optimization.}

\item{sparsenessAlg}{NA is default otherwise basic, spmp or orthorank}

\item{verbose}{Logical, whether to print information about energy and sparsity.}
}
\value{
A list containing the common representation Z and transformation matrices W.
}
\description{
Perform Multiview PCA on multiple datasets with an option for sparse PCA.
}
\examples{
set.seed(123)
n_samples <- 100
n_features_1 <- 50
n_features_2 <- 60
n_features_3 <- 70
n_components <- 5
view1 <- matrix(rnorm(n_samples * n_features_1), nrow = n_samples, ncol = n_features_1)
view2 <- matrix(rnorm(n_samples * n_features_2), nrow = n_samples, ncol = n_features_2)
view3 <- matrix(rnorm(n_samples * n_features_3), nrow = n_samples, ncol = n_features_3)
result <- multiview_pca(list(view1, view2, view3), n_components, sparse = rep(0.5,3), 
 verbose = TRUE)
print(result)
}
