% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_data.R
\docType{data}
\name{nbm3_labels}
\alias{nbm3_labels}
\title{nbm3_labels}
\format{
A data frame with columns: Label, Description, Side.
}
\usage{
nbm3_labels
}
\description{
A data frame label numbers and names for NBM3 regions.
}
\keyword{datasets}
