% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscaleSVDxpts.R
\name{NNHEmbed_update_residuals}
\alias{NNHEmbed_update_residuals}
\title{Update Residuals for SiMLR}
\usage{
NNHEmbed_update_residuals(
  mats,
  x,
  covariate,
  blaster2,
  allnna,
  n.comp,
  opt = NULL
)
}
\arguments{
\item{mats}{A list of matrices.}

\item{x}{An index to select a matrix from the list.}

\item{covariate}{A character string specifying the covariate transformation to apply.}

\item{blaster2}{A data frame containing additional covariate data.}

\item{allnna}{A logical vector indicating non-missing data points.}

\item{n.comp}{An integer specifying the number of components.}

\item{opt}{A character string. If set to "opt", the function will print available covariate options.}
}
\value{
The residual matrix after applying the specified covariate transformation.
If `opt` is set to "opt", the function will print available covariate options and return NULL.
}
\description{
This function updates residuals for SiMLR by applying different covariate transformations.
}
\examples{
\dontrun{
NNHEmbed_update_residuals(mats, 1, "whiten", blaster2, allnna, 5)
NNHEmbed_update_residuals(opt = "opt")
}
}
