% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/croppingAndPaddingUtilities.R
\name{padImageByFactor}
\alias{padImageByFactor}
\title{Pad an image based on a factor.}
\usage{
padImageByFactor(image, factor)
}
\arguments{
\item{image}{Input ANTs image}

\item{factor}{padding factor.  Can be an integer or vector of size
equal to the image dimensionality.}
}
\value{
a padded image
}
\description{
Pad image of size \code{(x, y, z)} to \code{(x', y', z')} where
\code{(x', y', z')} is a divisible by a user-specified factor.
}
\examples{

library( ANTsR )
image <- antsImageRead( getANTsRData( "r16" ) )
image <- cropImage( image )
paddedImage <- padImageByFactor( image, 4 )

}
\author{
Tustison NJ, Avants BB
}
