% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscaleSVDxpts.R
\name{pairwise_matrix_similarity}
\alias{pairwise_matrix_similarity}
\title{pairwise application of matrix similarity to matrix List projected onto a feature list}
\usage{
pairwise_matrix_similarity(mat_list, feat_list, FUN = adjusted_rvcoef)
}
\arguments{
\item{mat_list}{List of matrices}

\item{feat_list}{List of feature lists corresponding to each matrix}

\item{FUN}{function to apply}
}
\value{
Matrix of RV coefficients, where each entry [i, j] represents the similarity between the low-rank projections of matrices i and j
}
\description{
Computes a measure such as the RV coefficient between low-rank projections of all pairs of matrices in a list.
}
\examples{
mat_list <- list(matrix(rnorm(60), nrow = 10), matrix(rnorm(60), nrow = 10))
feat_list <- list(matrix(rnorm(60), nrow = 10), matrix(rnorm(60), nrow = 10))
pairwise_matrix_similarity(mat_list, feat_list )

}
