% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partialVolumeCorrection.R
\name{partialVolumeCorrection}
\alias{partialVolumeCorrection}
\title{Perform partial volume correction for ASL images.}
\usage{
partialVolumeCorrection(img, img.gm, img.wm, mask = NULL, proportion = 0.4)
}
\arguments{
\item{img}{Low-resolution image to be corrected.  All input images can be
either of type antsImage or numeric vectors (if numeric vectors, the mask is
ignored).}

\item{img.gm}{Gray matter probability image for partial volume correction.}

\item{img.wm}{White matter probability image for partial volume
correction.}

\item{mask}{Brain mask for image.}

\item{proportion}{Ratio of activity for white matter to gray matter.
Assumed to be 0.4.}
}
\value{
Returns partial volume corrected antsImage.
}
\description{
This function performs a partial volume correction for ASL images by
dividing the observed CBF by the gray matter and white matter probabilities,
as described in Johnson et al., Radiology 2005: CBF_corrected = CBF_observed
/ (GM_prob + 0.4 * WM_prob)
}
\examples{

activity.gm <- 10
activity.wm <- activity.gm * 0.4
percent.gm <- matrix(seq(0.1, 1, by=0.1), nrow=2)
percent.wm <- -percent.gm + 1
activity.obs <- percent.gm * rnorm(n=length(percent.gm), mean=activity.gm, sd=5) +
                  rnorm(n=length(percent.wm), mean=activity.wm, sd=5)
activity.corrected <- partialVolumeCorrection(activity.obs, percent.gm, percent.wm)

}
\references{
Method: Pattern of cerebral hypoperfusion in Alzheimer disease
and mild cognitive impairment measured with arterial spin-labeling MR
imaging: initial experience.  Johnson NA, Jahng GH, Weiner MW, Miller BL,
Chui HC, Jagust WJ, Gorno-Tempini ML, Schuff N. Radiology 2005.

Ratio of GM to WM activity: Quantitative magnetic resonance imaging of human
brain perfusion at 1.5 T using steady-state inversion of arterial water.
Roberts DA, Detre JA, Bolinger L, Insko EK, Leigh JS Jr. PNAS 1994.
}
\author{
Kandel BM %% ~~who you are~~
}
