% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscaleSVDxpts.R
\name{plot_features}
\alias{plot_features}
\title{Plot Features}
\usage{
plot_features(data_list, take_abs = TRUE, n_limit = 12)
}
\arguments{
\item{data_list}{A list of data frames.}

\item{take_abs}{boolean}

\item{n_limit}{Integer, limit features to top n_limit with highest value}
}
\value{
A list of ggplot objects.
}
\description{
Create bar plots for each column in each data frame, showing only non-zero values.  Normalize each feature s.t. max is one.
}
\examples{
\dontrun{
# Simulate data
set.seed(123)
feature_names <- paste("Feature", 1:10)
data_list <- list(
  df1 = data.frame(matrix(ifelse(runif(100) < 0.5, 0, runif(100)), nrow = 10)),
  df2 = data.frame(matrix(ifelse(runif(100) < 0.5, 0, runif(100)), nrow = 10))
)

# Set rownames
rownames(data_list$df1) <- feature_names
rownames(data_list$df2) <- feature_names

# Use the function
plots <- plot_features(data_list)

# Display the plots
for (i in 1:length(plots)) {
  print(plots[[i]])
}
}
}
