% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/points2LabelImage.R
\name{points2LabelImage}
\alias{points2LabelImage}
\title{create a label image from landmark coordinates}
\usage{
points2LabelImage(pts, neighbours = 16, spacing = rep(1, ncol(pts)),
  margin = 0.1, labelcolors = seq_len(nrow(pts)), RAS2IJK = NULL,
  refimage = NULL)
}
\arguments{
\item{pts}{k x 2 or k x 3 matrix containing landmark coordinates}

\item{neighbours}{integer: number of closest points in the image grid to label}

\item{spacing}{spacing of the output image}

\item{margin}{margin around the landmarks (determined by the extend of the bounding box) to be part of the image}

\item{labelcolors}{vector of length nrow(pts) with a color for each landmark label}

\item{RAS2IJK}{RAS2IJK 4x4 (3D) or 3x3 (2D) matrix with transform from RAS to IJK space (e.g. when landmarks are placed with Slicer).}

\item{refimage}{optional: an object of class antsImage to project the point label image into the image domain defined by refimage.}
}
\value{
returns an antsImage
}
\description{
create a label image from landmark coordinates
}
\examples{
load(system.file("extdata/landmarks.RData",package="ANTsR"))
fi <- antsImageRead(getANTsRData("r16") ,2)
li <- points2LabelImage(fiLM,refimage=fi,RAS2IJK=diag(c(-1,-1,1)))
plot(fi,list(li),alpha=0.3,add=T)
}
\author{
Schlager S
}

