% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{powers_areal_mni_itk}
\alias{powers_areal_mni_itk}
\title{powers_areal_mni_itk}
\format{
A data frame listing the following variables.
\describe{
  \item{\code{x}}{ x coordinate }
  \item{\code{y}}{ y coordinate }
  \item{\code{z}}{ z coordinate }
  \item{\code{ROI}}{ unique integer for each node }
  \item{\code{SystemLabel}}{ unique integer for each system }
  \item{\code{SystemName}}{ unique name for each system }
  \item{\code{Color}}{ name of color for system }
  \item{\code{r}}{ red value for system }
  \item{\code{g}}{ green value for system }
  \item{\code{b}}{ blue value for system }
  \item{\code{Anatomy}}{ anatomical location as 
  determined by OASIS labeling }
  \item{\code{Lobe}}{ lobe location of node }
  \item{\code{Brodmann}}{ Broadmann area number }
  \item{\code{AAL}}{ AAL region label }
}
#' @references \url{http://www.nil.wustl.edu/labs/petersen/Resources.html}
}
\usage{
powers_areal_mni_itk
}
\description{
A data frame providing coordinates for the nodes 
identified by Powers at al. Coordinates have been modified from nifti 
space to ITK space
}
\examples{
data(powers_areal_mni_itk)
\dontrun{
  fixed <- antsImageRead( getANTsRData("ch2") )
  moving <- antsImageRead( getANTsRData("mni") )
  mytx <- antsRegistration( fixed=fixed, moving=moving, 
  typeofTransform = c("SyN") )
  data( "powers_areal_mni_itk", package = "ANTsR", envir = environment() )
  coords = powers_areal_mni_itk[,1:3]
  ch2reg   = antsRegistration( fixed, moving, typeofTransform = "SyN" )
  coordsw <- antsApplyTransformsToPoints( 
  dim=3, points=coords,
  transformlist=ch2reg$fwdtransforms,
  whichtoinvert=c(FALSE,FALSE) )
  ptrd = 3
  powersLabels = makePointsImage( coordsw, moving, radius = pard )
  plot( moving, powersLabels, axis=3, nslices=30 )
}
}
\keyword{datasets}
