% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powersACTrsfMRIprocessing.R
\name{powersACTrsfMRIprocessing}
\alias{powersACTrsfMRIprocessing}
\title{Powers and ants cortical thickness processing for fMRI.}
\usage{
powersACTrsfMRIprocessing(img, fdthresh = 0.2, repeatMotionEst = 1,
  freqLimits = c(0.008, 0.09), nCompCor = 4, structuralImage = NA,
  structuralSeg = NA, structuralNodes = NA, templateMap = NA,
  smoothingSigmas = NA, extraRuns = NA, verbose = FALSE)
}
\arguments{
\item{img}{input time series antsImage}

\item{fdthresh}{threshold for framewise displacement.  determines what time
frames should be interpolated. Set typically between 0.1 and 0.5.}

\item{repeatMotionEst}{number of times to repeat motion estimation.}

\item{freqLimits}{pair defining bandwidth of interest from low to high.}

\item{nCompCor}{number of compcor components to use.}

\item{structuralImage}{the structural antsImage of the brain.}

\item{structuralSeg}{a 3 or greater class tissue segmentation of the structural image.}

\item{structuralNodes}{regions of interest for network analysis, in the structural image space.}

\item{templateMap}{antsRegistration output mapping template space (as moving) to struturalImage (fixed).}

\item{smoothingSigmas}{4-vector defining amount of smoothing in FWHM units}

\item{extraRuns}{a list containing additional BOLD images (runs) to be merged with the first image}

\item{verbose}{enables visualization as well as commentary.}
}
\value{
outputs a list containing:
\itemize{
  \item{boldMat: }{Matrix of filtered BOLD data.}
  \item{boldMask: }{BOLD mask.}
  \item{motionCorr: }{Motion corrected data.}
  \item{nuisance: }{Nuisance variables.}
  \item{connMatNodes: }{User provided nodal system connectivity matrix.}
  \item{connMatNodesPartialCorr: }{TUser provided nodal system partial correlation matrix.}
  \item{FD: }{mean framewise displacement.}
  \item{badtimes: }{time frames that are above FD threshold.}
  \item{seg2bold: }{strutural segmentation in BOLD space.}
  \item{nodes2bold: }{strutural nodes in BOLD space.}
  \item{mapsToTemplate: }{invertible maps from BOLD to template space.}
  \item{runID: }{Identifies which run over time series.}
  \item{dmnBetas: }{Default mode network beta map.}
  \item{networkPriors2Bold: }{WIP: standard network priors in BOLD space.}
  \item{connMatPowers: }{WIP: Powers nodes connectivity matrix. Assumes template maps are to MNI space.}
}
}
\description{
This function leverages structural image processing based on ANTs
cortical thickness to implement Powers' functional image processing
recommendations.  The function will perform motion correction and
produce a variety of nuisance regressors.  It will also perform
spatial and temporal filtering.  In sum, these regressors mitigate
the wide variety of confounds that impact functional MRI.
}
\examples{
# this example is long-running ( perhaps 10 minutes on an OSX laptop 2016 )
\dontrun{
exrun = powersACTrsfMRIprocessing( verbose = TRUE ) # will download ex data
myid = "BBAvants" # some MRI data
pre = paste("~/rsfTest/",sep='')
fn = list.files( pre, full.names = TRUE, recursive = TRUE,
  pattern = glob2rx( paste( myid, "*rsfMRI0.nii.gz", sep='')  ) )
img  = antsImageRead( fn  )
sfn = list.files( pre, full.names = TRUE, recursive = TRUE,
  pattern = glob2rx( paste( myid, "*BrainSegmentation.nii.gz", sep='')  ) )
seg = antsImageRead( sfn )
t1fn = list.files( pre, full.names = TRUE, recursive = TRUE,
  pattern = glob2rx( paste( myid, "*BrainSegmentation0N4.nii.gz", sep='')  ) )
t1 = antsImageRead( t1fn   )
tt = powersACTrsfMRIprocessing( img, fdthresh=0.2, repeatMotionEst=1,
  structuralImage=t1, structuralSeg=seg, verbose= TRUE )
# bold to template
antsApplyTransforms( mni, getAverageOfTimeSeries( img ),
   transformlist=tt$mapsToTemplate$toTemplate,
   whichtoinvert=tt$mapsToTemplate$toTemplateInversion )
# template to bold
antsApplyTransforms( getAverageOfTimeSeries( img ), mni,
   transformlist=tt$mapsToTemplate$toBold,
   whichtoinvert=tt$mapsToTemplate$toBoldInversion )
}

}
\author{
Avants BB, Duda JT
}

