% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_data.R
\docType{data}
\name{ppmi_template_500Parcels_Yeo2011_17Networks_2023_homotopic}
\alias{ppmi_template_500Parcels_Yeo2011_17Networks_2023_homotopic}
\title{ppmi template 500Parcels Yeo2011 17Networks 2023 homotopic}
\description{
ppmi template 500Parcels Yeo2011 17Networks 2023 homotopic
}
