% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_data.R
\docType{data}
\name{ppmi_yeo_labels}
\alias{ppmi_yeo_labels}
\title{ppmi_yeo_labels}
\format{
A data frame with columns: x, y, z, ROI, SystemName, AAL.
}
\usage{
ppmi_yeo_labels
}
\description{
A data frame providing coordinates and names for the PPMI 500 parcels Yeo atlas.
}
\keyword{datasets}
