% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscaleSVDxpts.R
\name{predictSimlr}
\alias{predictSimlr}
\title{predict output matrices from simlr solution}
\usage{
predictSimlr(x, simsol)
}
\arguments{
\item{x}{A list that contains the named matrices.}

\item{simsol}{the simlr solution}
}
\value{
A list of variance explained, predicted matrices and error metrics
}
\description{
predict output matrices from simlr solution
}
\examples{
# see simlr examples
}
\author{
BB Avants.
}
