% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/preprocessfMRI.R
\name{preprocessfMRI}
\alias{preprocessfMRI}
\title{Preprocess BOLD fMRI image data.}
\usage{
preprocessfMRI(boldImage, maskImage = NA, maskingMeanRatioThreshold = 0.75,
  initialNuisanceVariables = NA, numberOfCompCorComponents = 6,
  doMotionCorrection = TRUE, useMotionCorrectedImage = FALSE,
  motionCorrectionAccuracyLevel = 1,
  meanBoldFixedImageForMotionCorrection = NA, frequencyLowThreshold = NA,
  frequencyHighThreshold = NA, spatialSmoothingType = "none",
  spatialSmoothingParameters = 0, residualizeMatrix = TRUE)
}
\arguments{
\item{boldImage}{4-D ANTs image fMRI data.}

\item{maskImage}{3-D ANTs image defining the region of interest.}

\item{maskingMeanRatioThreshold}{If mask image is not specified,
a mask image is
created using the specified threshold which is in terms of the mean of the
average image ie 0.8 means threshold at 0.8 of the mean.}

\item{initialNuisanceVariables}{Optional initial nuisance variables.}

\item{numberOfCompCorComponents}{Numer of CompCor nuisance components.}

\item{doMotionCorrection}{Boolean indicating whether motion correction
should be performed and used in nuisance regression.}

\item{useMotionCorrectedImage}{Boolean indicating whether or not the motion
corrected image should be used in the rest of the pipeline.  This is off by
default to avoid additional interpolation.}

\item{motionCorrectionAccuracyLevel}{Accuracy for the motion correcting
registration: 0 = fast/debug parameters, 1 = intrasession parameters, or 2 =
intersession/intersubject parameters.}

\item{meanBoldFixedImageForMotionCorrection}{Optional target fixed image for
motion correction.}

\item{frequencyLowThreshold}{Lower threshold for bandpass filtering.}

\item{frequencyHighThreshold}{Upper threshold for bandpass filtering.}

\item{spatialSmoothingType}{Either \code{none}, \code{gaussian} (isotropic) or
\code{perona-malik} (anisotropic) smoothing.}

\item{spatialSmoothingParameters}{For gaussian smoothing, this is a single
scalar designating the smoothing sigma (in mm).  For perona-malik, a vector
needs to be specified with the conductance parameter and the number of
iterations, e.g. \code{c(0.25, 5)}.}

\item{residualizeMatrix}{boolean}
}
\value{
List of:
\itemize{
  \item{cleanBOLDImage: }{Cleaned BOLD image.}
  \item{maskImage: }{mask image.}
  \item{DVARS: }{Framewise change in BOLD signal, as in Powers et al.}
  \item{DVARSPostCleaning: }{DVARS after cleaning image.}
  \item{FD: }{Framewise displacement.}
  \item{globalSignal: }{Global signal.}
  \item{nuisanceVariables: }{Nuisance variables used in denoising.}
}
}
\description{
Preprocess fMRI data by performing compcor/motion correction, nuisance
regression, band-pass filtering, and spatial smoothing.
}
\examples{
set.seed(123)
nvox <- 10*10*10*20
dims <- c(10,10,10,20)
boldImage <- makeImage(dims, rnorm(nvox) + 500) \%>\% iMath("PadImage", 2)
# for real data: boldImage <- antsImageRead(getANTsRData('pcasl'))
cleanfMRI <- preprocessfMRI(boldImage)
}
\author{
Tustison NJ, Avants BB
}
\references{
Power et al. 2012, "Spurious but systematic correlations
in functional connectivity MRI networks arise from subject motion."
NeuroImage 59, 2142-2154.
}

