% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessRestingBOLD.R
\name{preprocessRestingBOLD}
\alias{preprocessRestingBOLD}
\title{Preprocess resting BOLD fMRI image data.}
\usage{
preprocessRestingBOLD(
  boldImage,
  maskImage = NULL,
  maskingMeanRatioThreshold = 0.75,
  initialNuisanceVariables,
  denseFramewise = FALSE,
  numberOfCompCorComponents = 6,
  doMotionCorrection = TRUE,
  useMotionCorrectedImage = FALSE,
  motionCorrectionAccuracyLevel = 1,
  motionCorrectionIterations = 1,
  meanBoldFixedImageForMotionCorrection = NULL,
  frequencyLowThreshold = NA,
  frequencyHighThreshold = NA,
  spatialSmoothingType = "none",
  spatialSmoothingParameters = 0,
  residualizeMatrix = TRUE,
  num_threads = 1,
  seed = NULL
)
}
\arguments{
\item{boldImage}{4-D ANTs image fMRI data.}

\item{maskImage}{3-D ANTs image defining the region of interest.}

\item{maskingMeanRatioThreshold}{If mask image is not specified,
a mask image is
created using the specified threshold which is in terms of the mean of the
average image ie 0.8 means threshold at 0.8 of the mean.}

\item{initialNuisanceVariables}{Optional initial nuisance variables.}

\item{denseFramewise}{boolean for dense framewise stats}

\item{numberOfCompCorComponents}{Numer of CompCor nuisance components.}

\item{doMotionCorrection}{Boolean indicating whether motion correction
should be performed and used in nuisance regression.}

\item{useMotionCorrectedImage}{Boolean indicating whether or not the motion
corrected image should be used in the rest of the pipeline.  This is off by
default to avoid additional interpolation.}

\item{motionCorrectionAccuracyLevel}{Accuracy for the motion correcting
registration: 0 = fast/debug parameters, 1 = intrasession parameters, or 2 =
intersession/intersubject parameters.}

\item{motionCorrectionIterations}{number of iteration for motion correction}

\item{meanBoldFixedImageForMotionCorrection}{Optional target fixed image for
motion correction.}

\item{frequencyLowThreshold}{Lower threshold for bandpass filtering.}

\item{frequencyHighThreshold}{Upper threshold for bandpass filtering.}

\item{spatialSmoothingType}{Either \code{none}, \code{gaussian} (isotropic) or
\code{perona-malik} (anisotropic) smoothing.}

\item{spatialSmoothingParameters}{For gaussian smoothing, this is a single
scalar designating the smoothing sigma (in mm).  For perona-malik, a vector
needs to be specified with the conductance parameter and the number of
iterations, e.g. \code{c(0.25, 5)}.}

\item{residualizeMatrix}{boolean}

\item{num_threads}{will execute
\code{Sys.setenv(ITK_GLOBAL_DEFAULT_NUMBER_OF_THREADS = num_threads)} before
running to attempt a more reproducible result.  See
\url{https://github.com/ANTsX/ANTs/wiki/antsRegistration-reproducibility-issues}
for discussion.  If \code{NULL}, will not set anything.}

\item{seed}{will execute
\code{Sys.setenv(ANTS_RANDOM_SEED = seed)} before
running to attempt a more reproducible result.  See
\url{https://github.com/ANTsX/ANTs/wiki/antsRegistration-reproducibility-issues}
for discussion.  If \code{NULL}, will not set anything.}
}
\value{
List of:
\itemize{
  \item{cleanBOLDImage: }{Cleaned BOLD image.}
  \item{maskImage: }{mask image.}
  \item{DVARS: }{Framewise change in BOLD signal, as in Powers et al.}
  \item{DVARSPostCleaning: }{DVARS after cleaning image.}
  \item{FD: }{Framewise displacement.}
  \item{globalSignal: }{Global signal.}
  \item{nuisanceVariables: }{Nuisance variables used in denoising.}
}
}
\description{
Preprocess resting fMRI by performing compcor/motion correction, nuisance
regression, band-pass filtering, and spatial smoothing.
}
\examples{

set.seed(123)
n <- 8
nvox <- n * n * n * 6
dims <- c(n, n, n, 6)
boldImage <- makeImage(dims, rnorm(nvox) + 500) \%>\% iMath("PadImage", 2)
# for real data: boldImage <- antsImageRead(getANTsRData('pcasl'))
cleanfMRI <- preprocessRestingBOLD(boldImage)

}
\references{
Power et al. 2012, "Spurious but systematic correlations
in functional connectivity MRI networks arise from subject motion."
NeuroImage 59, 2142-2154.
}
\author{
Tustison NJ, Avants BB
}
