% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priorBasedSegmentation.R
\name{priorBasedSegmentation}
\alias{priorBasedSegmentation}
\title{spatial prior-based image segmentation.}
\usage{
priorBasedSegmentation(
  img,
  priors,
  mask,
  priorWeight = 0.25,
  mrf = 0.1,
  iterations = 25
)
}
\arguments{
\item{img}{input image or image list for multivariate segmentation}

\item{priors}{list of priors that cover the number of classes}

\item{mask}{segment inside this mask}

\item{priorWeight}{usually 0 (priors used for initialization only), 0.25 or 0.5.}

\item{mrf}{regularization, higher is smoother, a numerical value in range 0.0 to 0.2}

\item{iterations}{maximum number of iterations.  could be a large value eg 25.}
}
\value{
segmentation and probability images
}
\description{
markov random field regularized, prior-based image segmentation that is a
wrapper around atropos (see ANTs and related publications).
}
\examples{

fi = antsImageRead( getANTsRData( "r16" ) )
seg = kmeansSegmentation( fi, 3 )
msk = thresholdImage( seg$segmentation, 1, Inf )
pseg = priorBasedSegmentation( fi, seg$probabilityimages, msk, 0.25, 0.1, 3 )

}
\author{
Brian B. Avants
}
