% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscaleSVDxpts.R
\name{project_to_nonneg_orthogonal_optim}
\alias{project_to_nonneg_orthogonal_optim}
\title{Project a matrix to the nearest non-negative orthogonal matrix}
\usage{
project_to_nonneg_orthogonal_optim(X, tol = 1e-06, max_iter = 10)
}
\arguments{
\item{X}{a matrix}

\item{tol}{tolerance for convergence (default: 1e-6)}

\item{max_iter}{maximum number of iterations (default: 1000)}
}
\value{
a non-negative orthogonal matrix
}
\description{
Project a matrix to the nearest non-negative orthogonal matrix
}
