% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscaleSVDxpts.R
\name{project_to_orthonormal_nonnegative}
\alias{project_to_orthonormal_nonnegative}
\title{Project Non-Negative Matrix to Orthonormal Columns Matrix}
\usage{
project_to_orthonormal_nonnegative(
  X,
  max_iter = 100,
  tol = 1e-04,
  constraint = "positive"
)
}
\arguments{
\item{X}{A square numeric matrix with non-negative entries.}

\item{max_iter}{Maximum number of iterations for the alternating projection.}

\item{tol}{Tolerance for convergence. The projection stops if the change
in the matrix norm is below this tolerance, or if max_iter is reached.}

\item{constraint}{either or positive}
}
\value{
A matrix Y that is non-negative and whose columns are orthonormal,
        closest to X in a sense defined by the alternating projections.
}
\description{
Iteratively projects a non-negative matrix onto the intersection
             of non-negative matrices and matrices with orthonormal columns.
}
