% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/projectImageAlongAxis.R
\name{projectImageAlongAxis}
\alias{projectImageAlongAxis}
\title{Simple projectImageAlongAxis function.}
\usage{
projectImageAlongAxis(imageND, referenceImageNDminus1, projtype = 0,
  axis = NA)
}
\arguments{
\item{imageND}{input image}

\item{referenceImageNDminus1}{down-dimensional image to help define physical #' space}

\item{projtype}{projection type 0, 1 or 2}

\item{axis}{should be less than image dimension and greater than or equal to
0}
}
\value{
image of n-dimensions-1 is output
}
\description{
Will turn an N-D image into an N-1-D image by summation (0), max-intensity
(1) or min-intensity (2) projection along an orthogonal axis.
}
\examples{
img<-makeImage(c(5,5,5,5),rnorm(5^4))
img3d<-makeImage(c(5,5,5),rnorm(5^3))
mask4dproj<-projectImageAlongAxis( img, img3d, axis=3, projtype=0 )
}
\author{
Avants BB
}

