% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/quantifySNPs.R
\name{quantifySNPs}
\alias{quantifySNPs}
\title{Simple quantifySNPs function.}
\usage{
quantifySNPs(snps, freqthresh = 0.1, shiftit = FALSE, replaceWithF = T,
  traitvecin = NA, trainvec = NA)
}
\arguments{
\item{snps}{input matrix}

\item{freqthresh}{remove snps below this frequency}

\item{shiftit}{shift the snps to smooth the estimate}

\item{replaceWithF}{replaces snps with frequency values}

\item{traitvecin}{map snps to trait vector}

\item{trainvec}{defines training data}
}
\value{
matrix is output
}
\description{
quantifySNPs converts trinary snps to frequency data
}
\examples{
mat <- matrix(rnorm(300),ncol=50)
wmat<-quantifySNPs( mat )
wmat2<-quantifySNPs( mat, 2, TRUE )
}
\author{
Avants BB
}

