% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomMask.R
\name{randomMask}
\alias{randomMask}
\title{Get a random mask}
\usage{
randomMask(img, nsamples, perLabel = F)
}
\arguments{
\item{img}{an antsImage, either continuous or mask-like. This will
define the space where to sample voxels from. If this image is
continous (i.e., a T1), voxels will be sampled from all nonzero
voxels.}

\item{nsamples}{the number of random voxels to select from}

\item{perLabel}{logical, if true the input image must be multi-label
and the the output will have nsamples from each label (i.e., 200
samples from each label)}
}
\value{
binary antsImage with random voxels set to 1
}
\description{
Get a mask with a specific number of voxels randomly
distributed on the input image/mask
}
\examples{

img<-antsImageRead( getANTsRData("r16"))
randmask<-randomMask(img,200)

}
\author{
Pustina D
}

