% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randsccan.R
\name{randsvd}
\alias{randsvd}
\title{randomized svd function.}
\usage{
randsvd(A, k, seed = NA)
}
\arguments{
\item{A}{input matrix}

\item{k}{rank to use}

\item{seed}{for testing}
}
\value{
randomized svd is output
}
\description{
uses random matrix to estimate svd results
}
\examples{

A <- matrix(rnorm(3000), ncol = 50)
k <- 10
dr <- randsvd(A, k)$d[1:k]
dt <- svd(A)$d[1:k]
cor(dr, dt)
}
\references{
N. Halko, P.G. Martinsson, J. Tropp "Finding structure with randomness: Stochastic algorithms for constructing approximate matrix decompositions" arXiv 0909.4061
}
\author{
Avants BB
}
\keyword{svd}
