% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscaleSVDxpts.R
\name{rankBasedMatrixSegmentation}
\alias{rankBasedMatrixSegmentation}
\title{rank-based segmentation of a matrix}
\usage{
rankBasedMatrixSegmentation(
  v,
  sparsenessQuantile,
  basic = FALSE,
  positivity = "positive"
)
}
\arguments{
\item{v}{input matrix}

\item{sparsenessQuantile}{a value in zero to one}

\item{basic}{simplest keep top k-entries in each row}

\item{positivity}{one of positive, negative, either}
}
\value{
matrix
}
\description{
rank-based segmentation of a matrix
}
\examples{

mat=replicate(100, rnorm(20))
matr=rankBasedMatrixSegmentation( mat, 0.9, basic=FALSE, positivity='positive' )

}
\author{
Avants BB
}
