% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankIntensity.R
\name{rankIntensity}
\alias{rankIntensity}
\title{rank intensity transformation for an image}
\usage{
rankIntensity(x, mask, getMask = TRUE, method = "max")
}
\arguments{
\item{x}{input image}

\item{mask}{optional mask}

\item{getMask}{boolean only used if mask is not present}

\item{method}{optional method passed to \code{ties.method} in \code{rank}}
}
\value{
image is output
}
\description{
Rank transform the intensity of an image to reduce the impact of outliers;
useful when using correlation-based metrics for registration and possibly
for segmentation tasks. best used with a mask.
}
\examples{

fi <- makeImage(c(10, 10, 10), rnorm(1000))
rfi <- rankIntensity(fi)

}
\author{
Avants BB
}
