% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rapidlyInspectImageData.R
\name{rapidlyInspectImageData}
\alias{rapidlyInspectImageData}
\title{Simple rapidlyInspectImageData function.}
\usage{
rapidlyInspectImageData(myfiles)
}
\arguments{
\item{myfiles}{input list of filenames}
}
\value{
matrix is output
}
\description{
rapidlyInspectImageData collects basic statistics over a dataset and returns them
in a dataframe.
}
\examples{
\dontrun{
 fnl<-c( getANTsRData('r16'),
 getANTsRData('r27'),
 getANTsRData('r62'),
 getANTsRData('r64'),
 getANTsRData('r85') )
mm<-rapidlyInspectImageData( fnl )
if ( !usePkg("DMwR") | ! usePkg('fpc') )
   { print("Need DMwR and fpc packages") } else {
  pamres <- pamk(mm,1:4)
  outlier.scores <- lofactor( mm, k=3 )
  outliers <- order(outlier.scores)
  }
}
}
\author{
Avants BB
}

