% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscaleSVDxpts.R
\name{read_simlr}
\alias{read_simlr}
\title{Read all SiMLR outputs from disk}
\usage{
read_simlr(dir)
}
\arguments{
\item{dir}{A character string specifying the directory with SiMLR outputs.}
}
\value{
A named list reconstructing the original SiMLR object.
}
\description{
Reconstructs SiMLR outputs written by \code{\link{write_simlr}} using the manifest
with relative paths for portability. Restores row names for matrices and data frames
from the CSV files.
}
