% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsrTransform_class.R
\name{readAntsrTransform}
\alias{readAntsrTransform}
\title{readAntsrTransform}
\usage{
readAntsrTransform(filename, dimension = NA, precision = "float")
}
\arguments{
\item{filename}{filename of transform}

\item{dimension}{spatial dimension of transform}

\item{precision}{numerical precision of transform}
}
\value{
antsrTransform
}
\description{
read a transform from file
}
\examples{
trans <- c(3, 4, 5)
tx <- createAntsrTransform(type = "Euler3DTransform", translation = trans)
txfile <- tempfile(fileext = ".mat")
writeAntsrTransform(tx, txfile)
tx2 <- readAntsrTransform(txfile)
testthat::expect_error(readAntsrTransform(txfile, 2), "space dim")
}
