% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeNormalizedPopulationData.R
\name{readNormalizedPopulationData}
\alias{readNormalizedPopulationData}
\title{Read normalized population from h5 file}
\usage{
readNormalizedPopulationData(filename)
}
\arguments{
\item{filename}{input filename for hdf5 file.}
}
\value{
list containing demographics, imageMat, imageMask and imageBoolean
}
\description{
This function reads a file created by \code{writeNormalizedPopulationData}.
}
\examples{
\dontrun{
ilist = getANTsRData( "population" )
mask = getMask( ilist[[ 1 ]] )
imat = imageListToMatrix( ilist, mask )
demog = data.frame( age = c( 11, 7, 18, 22 ), sex = c("M","M","F","F") )
ibool = c( TRUE, TRUE, TRUE, FALSE )
tfn = tempfile(fileext=".h5")
if ( usePkg( "h5" ) ) writeNormalizedPopulationData( demog, imat, mask, ibool, tfn )
if ( usePkg( "h5" ) ) dlist = readNormalizedPopulationData( tfn )
}
}
\author{
Avants BB
}
