% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstructImageFromPatches.R
\name{reconstructImageFromPatches}
\alias{reconstructImageFromPatches}
\title{Reconstruct image from a list of patches.}
\usage{
reconstructImageFromPatches(
  patches,
  domainImage,
  strideLength = 1,
  domainImageIsMask = FALSE
)
}
\arguments{
\item{patches}{List or array of patches defining an image.  Patches are assumed
to have the same format as returned by \code{extractImagePatches}.}

\item{domainImage}{Image or mask to define the geometric information of the
reconstructed image.  If this is a mask image, the reconstruction will only
use patches in the mask.}

\item{strideLength}{Defines the sequential patch overlap for
\code{maxNumberOfPatches = all}.  Can be a image-dimensional vector or a scalar.}

\item{domainImageIsMask}{boolean specifying whether the domain image is a
mask used to limit the region of reconstruction from the patches.}
}
\value{
an ANTs image.
}
\description{
Reconstruct image from a list of patches.
}
\examples{

library( ANTsR )

image <- antsImageRead( getANTsRData( "r16" ) )
patchSet <- extractImagePatches( image, c( 64, 64 ), "all", c( 8, 8 ) )
imageReconstructed <-
  reconstructImageFromPatches( patchSet, image, c( 8, 8 ) )
testthat::expect_equal(as.array(image), as.array(imageReconstructed))
rm(image); gc()
rm(patchSet); gc()
rm(imageReconstructed); gc()
}
\author{
Tustison NJ
}
