% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCentroids_import.R, R/maskImage_import.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{getCentroids}
\alias{maskImage}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{ANTsRCore}{\code{\link[ANTsRCore]{getCentroids}}, \code{\link[ANTsRCore]{maskImage}}}
}}

