% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/iMath.R
\name{reflectImage}
\alias{reflectImage}
\title{reflectImage}
\usage{
reflectImage(img1, axis = NA, tx = NA)
}
\arguments{
\item{axis}{which dimension to reflect across}

\item{tx}{transformation type to estimate after reflection}

\item{img}{input object, an antsImage}
}
\description{
reflects an image along its axis
}
\examples{
fi<-antsImageRead( getANTsRData("r16") , 2 )
fir<-reflectImage( fi, 1, "Affine" )
}
\author{
BB Avants
}

