% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reflectImage.R
\name{reflectionMatrix}
\alias{reflectionMatrix}
\title{reflectionMatrix}
\usage{
reflectionMatrix(img1, axis = NA, reflectionMatrixFilename)
}
\arguments{
\item{img1}{input object, an antsImage}

\item{axis}{which dimension to reflect across, numbered from 0 to imageDimension-1}

\item{reflectionMatrixFilename}{a filename ending in .mat}
}
\description{
return a filename pointing to a reflection matrix
}
\author{
BB Avants
}
