% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressionMatchImage.R
\name{regressionMatchImage}
\alias{regressionMatchImage}
\title{Image intensity normalization using linear regression.}
\usage{
regressionMatchImage(
  sourceImage,
  referenceImage,
  mask = NULL,
  polyOrder = 1,
  truncate = TRUE
)
}
\arguments{
\item{sourceImage}{image whose intensities we will match to the
\code{referenceImage} intensities.}

\item{referenceImage}{defines the reference intensity function.}

\item{mask}{Defines voxels for regression modeling.}

\item{polyOrder}{of polynomial fit.  Default is 1 (linear fit).}

\item{truncate}{boolean which turns on/off the clipping of intensities.}
}
\value{
the \code{sourceImage} matched to the \code{referenceImage}.
}
\description{
Image intensity normalization by regressing the image
intensities of the reference image with the source image.
}
\examples{
library(ANTsRCore)
sourceImage <- antsImageRead( getANTsRData( "r16" ) )
referenceImage <- antsImageRead( getANTsRData( "r64" ) )
matchedImage <- regressionMatchImage( sourceImage, referenceImage )
bad_source = sourceImage[1:200, 1:200]
testthat::expect_error(regressionMatchImage( bad_source, referenceImage ))
}
\author{
Avants BB
}
