\name{regressProjections}
\alias{regressProjections}
\title{Regression on image set projection}
\description{
  Perform regression on a training set of images, projected onto (provided) eigenvectors, and test on testing images.
}

\usage{
  regressProjections <- function(input.train, input.test, demog.train, demog.test, vector.names, mask, outcome, covariates="1", dim=3, method="optimal")
}

\arguments{
  \item{input.train}{Masked imaging data from training set of type \code{antsImage} or \code{matrix}.  This will typically be read from a \code{train.mha} file generated by, e.g., \code{sccan --imageset-to-matrix}.}
  \item{input.test}{Masked imaging data from testing set of type \code{antsImage} or \code{matrix}.  This will typically be read from a \code{test.mha} file generated by, e.g., \code{sccan --imageset-to-matrix}.}
  \item{demog.train}{Data frame of demographics information for training images.}
  \item{demog.test}{Data frame of demographics information for testing images.}
  \item{vector.names}{List of eigenvector image names to read for dimensionality reduction.  Only the filenames are required, not the actual images.}
  \item{mask}{Mask image of type \code{antsImage}.}
  \item{outcome}{Name of outcome variable to be predicted.  Must be present in \code{demog.train} and \code{demog.test}.}
  \item{covariates}{List of names of covariates to be used for the prediction.  All names must be present in \code{demog.train} and \code{demog.test}.}
  \item{dim}{Dimensionality of images.}
  \item{method}{Method for selecting eigenvectors.  Can be either \code{"all"} (uses all eigenvectors in \code{vector.names}) or \code{"optimal"} (uses
  BIC stepwise model selection to select eigenvectors.}
}

\details{
\code{regressProjections} is a convenient way to perform training and testing of predictions of demographic information from imaging data.  It takes as
input demographics information, imaging data, and eigenvectors, and performs prediction of the outcome variable from the projection of the imaging data on
the eigenvectors. 
}

\value{
A list of diagnostic and statistical information generated from the prediction, including: 
  \code{stats}: Mean squared error, correlation coefficients, and p-value of prediction for training and testing data.
  \code{outcome.comparison}: Data frame comparing real vs. predicted values for testing data. 
  \code{vectors.used}: Names of eigenvectors retained in model building. 
}

\seealso{
  \code{\link{sparseDecom2}}
}

\author{
Kandel BM
}

\examples{
\dontrun{
fileNames = c("eigvec1.nii.gz", "eigvec2.nii.gz") 
input.train <- antsImageRead('train.mha', dim=2, 'float')
input.test <- antsImageRead('test.mha', dim=2, 'float')
demog.train <- read.csv('demogtrain.csv')
demog.test <- read.csv('demogtest.csv')
mask <- antsImageRead('mask.nii.gz', dim=3, 'float')
outcome <- "age"
result <- regressProjections( input.train, input.test, demog.train, demog.test,
             vector.names=fileNames, mask, outcome, method='all')
}
}
