% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reho.R
\name{reho}
\alias{reho}
\title{REHO}
\usage{
reho(rsfmri, mask, shifter = 1, method = "kendall",
  maxDistance = 1.1)
}
\arguments{
\item{rsfmri}{antsImage usually 4D}

\item{mask}{antsImage binary mask}

\item{shifter}{size of neighborhood}

\item{method}{one of kendall, spearman or pearson}

\item{maxDistance}{maximum distance from center voxel}
}
\value{
antsImage is output
}
\description{
regional homogeneity for a time series image
}
\examples{
\dontrun{
rehoimg = reho( rsfmri )
}
}
\references{
Zang, Y., Jiang, T., Lu, Y., He, Y., Tian, L., 2004. Regional
homogeneity approach to fMRI data analysis. Neuroimage 22, 394-400.
}
\author{
Tustison, N.
}
