% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renormalizeProbabilityImages.R
\name{renormalizeProbabilityImages}
\alias{renormalizeProbabilityImages}
\title{Normalize (modified) probability images}
\usage{
renormalizeProbabilityImages(probs, mask, k = NA)
}
\arguments{
\item{probs}{list of images}

\item{mask}{image}

\item{k}{which image, if any, to prioritize}
}
\value{
list is output
}
\description{
Probability image lists should sum to one at every voxel. Smoothing or other
common operations may break this constraint.  This function allows one to
easily ensure a list of proability images sums to one everywhere.  It also
allows one prioritize specific images in the probability set such that its
probability is maintained while others are normalized with respect to the
prioritized value.  Will also set values less than zero, to zero.
}
\examples{

mat <- t( matrix( rnorm(4000),ncol=10) )
mask = makeImage( c(20,20), 1 )
plist = matrixToImages( mat, mask )
rplist<-renormalizeProbabilityImages( plist, mask )
rplist2<-renormalizeProbabilityImages( plist, mask , 2 )

}
\author{
Avants BB
}

