% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/reorientImage.R
\name{reorientImage}
\alias{reorientImage}
\title{reorient image by its principal axis}
\usage{
reo<-reorientImage( i1 , axis )
}
\arguments{
\item{img}{antsImage}

\item{axis2}{vector of size dim for 3D}

\item{doreflection}{vector of size dim for 3D}

\item{txfn}{file name for transformation}

\item{axis}{vector of size dim, might need to play w/axis sign}
}
\value{
reoriented image
}
\description{
align along a specific axis
}
\examples{
fi<-antsImageRead( getANTsRData('r16') ,2)
reofi<-reorientImage(fi,c(1,0))
}
\author{
Brian B. Avants
}
\keyword{geometry}
\keyword{image}

