% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/resampleImage.R
\name{resampleImage}
\alias{resampleImage}
\title{resampleImage}
\usage{
resampleImage(image, resampleParams, useVoxels = 1, interpType = 1)
}
\arguments{
\item{image}{input antsImage matrix}

\item{resampleParams}{vector of size dimension with numeric values}

\item{useVoxels}{true means interpret resample params as voxel counts}

\item{interpType}{one of 0, 1, 2, 3, 4}
}
\value{
output antsImage
}
\description{
Resample image by spacing or other parameters
}
\details{
Change an image by its spacing or number of voxels
}
\examples{
fi<-antsImageRead( getANTsRData('r16') ,2)
finn<-resampleImage(fi,c(50,60),1,0)
filin<-resampleImage(fi,c(1.5,1.5),0,1)
}
\author{
Avants BB
}

