% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampleImageToTarget.R
\name{resampleImageToTarget}
\alias{resampleImageToTarget}
\title{resampleImageToTarget}
\usage{
resampleImageToTarget(image, target, interpType = 1)
}
\arguments{
\item{image}{input antsImage matrix}

\item{target}{reference image to use for voxel size and header info}

\item{interpType}{one of 0 (linear), 1 (nearest neighbor),
2 (gaussian), 3 (windowed sinc), 4 (bspline)}
}
\value{
output antsImage
}
\description{
Resample image by using another image as target reference.
}
\examples{

fi<-antsImageRead( getANTsRData("r16"))
fi.ref<-antsImageRead( getANTsRData("r64"))
fi <- resampleImage(fi, c(2.2, 2.2, 2.2), useVoxels = 0, interpType = 1)
finew <- resampleImageToTarget(fi, fi.ref)

}
\author{
Pustina D
}

