% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rftPval.R
\name{rftPval}
\alias{rftPval}
\title{RFT p-values}
\usage{
rftPval(D, c, k, u, n, resels, df, fieldType)
}
\arguments{
\item{D}{image dimensions}

\item{c}{Threshold}

\item{k}{spatial extent in resels (minimum cluster size in resels)}

\item{u}{Number of clusters}

\item{n}{number of statistical field in conjunction}

\item{resels}{resel measurements of the search region}

\item{df}{degrees of freedom expressed as df = c(degrees of interest, degrees of error)}

\item{fieldType}{\itemize{
\item{T: } {T-field} 
\item{F: } {F-field} 
\item{X: } {Chi-square field'} 
\item{Z: } {Gaussian field}
}}
}
\value{
The probability of obtaining the specified cluster
\itemize{
{Pcor: } {corrected p-value}
{Pu: } {uncorrected p-value}
{Ec: } {expected number of clusters}
{ek: } {expected number of resels per cluster}
}
}
\description{
Calculates p-values of a statistical field using random field theory
}
\details{
This function calculates p-values of a thresholded statistical field at various levels:

set-level
rft.pval(D, c, k, u, n, resels, df, fieldType)

cluster-level
rft.pval(D, 1, k, u, n, resels, df, fieldType)

peak-level
rft.pval(D, 1, 0, u, n, resels, df, fieldType)

Where set-level refers to obtaining the set of clusters, cluster-level refers to a specific 
cluster, and peak-level refers to the maximum (or peak) of a single cluster.
}
\examples{
\dontrun{
# using rftPval for hypothetical 3D t-statistical image
# assume resels have been calculated and df = c(dfi, dfe)

# peak RFT p-value (peak = the maximum of a specific cluster)
peakP <- rftPval(3, 1, 0, peak, 1, resels, df, fieldType = "T")$Pcor 

# cluster RFT p-value (u = the value the statistical field was threshold at
# and k = the size of the cluster in resels)
clusterP <- rftPval(3, 1, k, u, 1, resels, df, fieldType = "T")$Pcor 

# set RFT p-value 
setP <- rftPval(3, numberOfClusters, minimumClusterSize, u, 1, resels, df,
                fieldType = "T")$Pcor
}
}
\references{
Friston K.J., (1994) Assessing the Significance of Focal Activations Using Their Spatial Extent.

Friston K.J., (1996) Detecting Activations in PET and fMRI: Levels of Inference and Power.

Worlsey K.J., (1996) A Unified Statistical Approach for Determining Significant Signals in Images of Cerebral Activation.
}
\seealso{
rftResults, resels
}
\author{
Zachary P. Christensen
}
