% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rimgs.R
\name{ri}
\alias{ri}
\title{Minimal function for creating r images}
\usage{
ri(x)
}
\arguments{
\item{x}{one of 16, 27, 30, 62, 64, 85 or 'all'. values 1 through 6 are also acceptable.}
}
\value{
image or image list is output
}
\description{
r16, r27, r30, r62, r64, r85 are t1-weighted MRI slices of the
human brain that have been around for a long time.  We use them
for quick registration, segmentation and other processing tests.
This function will return these images to the user.
}
\examples{

r16 <- ri(16)

}
\author{
Avants BB
}
