% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ripmmarc.R
\name{ripmmarcBasisImage}
\alias{ripmmarcBasisImage}
\title{Generate an antsImage from ripmmarc basis data}
\usage{
ripmmarcBasisImage(canonicalFrame, patchBasisVector, eps = 1e-12)
}
\arguments{
\item{canonicalFrame}{canonical frame image generated by ripmmarc}

\item{patchBasisVector}{the basis vector to convert to image space}

\item{eps}{small value threshold to help generate mask from canonicalFrame}
}
\value{
antsImage
}
\description{
This function converts a vectorized image patch back into an
antsImage such that it may be displayed in its extrinsic dimensionality.
}
\examples{

img <- antsImageRead( getANTsRData( "r16" ) )
msk <- thresholdImage( img, quantile( img[ img > 0 ] )[1], max( img ) )
ripped <- ripmmarc( img, msk, patchRadius=3, patchSamples=2000, patchVarEx=5)
bimg = ripmmarcBasisImage( ripped$canonicalFrame, ripped$basisMat[4,] )

}
\author{
Kandel BM, Avants BB
}
