% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustMatrixTransform.R
\name{robustMatrixTransform}
\alias{robustMatrixTransform}
\title{Simple robustMatrixTransform function.}
\usage{
robustMatrixTransform(mat)
}
\arguments{
\item{mat}{input matrix}
}
\value{
matrix is output
}
\description{
Transform a matrix such that each column has rank-ordered entries where rank #' is determined by the column values.
}
\examples{
mat<-replicate(100, rnorm(20))
rmat<-robustMatrixTransform( mat )

}
\author{
Avants BB
}

