% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscaleSVDxpts.R
\name{rvcoef}
\alias{rvcoef}
\title{RV Coefficient of Two Matrices}
\usage{
rvcoef(X, Y)
}
\arguments{
\item{X}{First matrix}

\item{Y}{Second matrix}
}
\value{
RV coefficient (a value between 0 and 1)
}
\description{
Computes the RV coefficient, a measure of similarity between two matrices.
}
\examples{
X <- matrix(rnorm(100), nrow = 10)
Y <- matrix(rnorm(120), nrow = 10)
rvcoef(X, Y)

}
