% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscaleSVDxpts.R
\name{rvcoef_components}
\alias{rvcoef_components}
\title{Computes the RV-Coefficient (Internal)}
\usage{
rvcoef_components(X, Y)
}
\arguments{
\item{X}{A numeric matrix (n observations, p variables).}

\item{Y}{A numeric matrix (n observations, q variables).}
}
\value{
values.
}
\description{
This function automatically selects the fastest algorithm based on data dimensions.
Internal Dispatcher: Returns all components from the fastest method.
}
