% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscaleSVDxpts.R
\name{rvcoef_trace_impl}
\alias{rvcoef_trace_impl}
\title{Computes the RV-Coefficient (Internal) Trace Method for Wide data}
\usage{
rvcoef_trace_impl(X_centered, Y_centered)
}
\arguments{
\item{X_centered}{A numeric matrix (n observations, p variables).}

\item{Y_centered}{A numeric matrix (n observations, q variables).}
}
\value{
values.
}
\description{
Computes the RV-Coefficient (Internal) Trace Method for Wide data
}
