% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentationRefinement.R
\name{segmentationRefinement.predict}
\alias{segmentationRefinement.predict}
\title{Segmentation refinement using corrective learning (prediction)}
\usage{
segmentationRefinement.predict(segmentationImage, labelSet, labelModelsStage1,
  featureImagesStage1, featureImageNamesStage1, labelModelsStage2,
  featureImagesStage2, featureImageNamesStage2, dilationRadius = 2,
  neighborhoodRadiusStage1 = 0, neighborhoodRadiusStage2 = 0,
  useLabelDistances = TRUE, normalizeSamplesPerLabel = TRUE)
}
\arguments{
\item{segmentationImage}{image to refine via corrective learning.}

\item{labelSet}{a vector specifying the labels of interest.  Must be specified.}

\item{labelModelsStage1}{a list of models for stage 1 (prediction of misclassifed voxels).
Each element of the labelSet requires a model.}

\item{featureImagesStage1}{a list of lists of feature images for Stage 1.  Each list of
feature images corresponds to a single subject.  Possibilities are outlined in
the above-cited paper.}

\item{featureImageNamesStage1}{a vector of character strings naming the set of features.
Must be specified.}

\item{featureImagesStage2}{a list of lists of feature images for Stage 2.  Each list of
feature images corresponds to a single subject.  Possibilities are outlined in
the above-cited paper.}

\item{featureImageNamesStage2}{a vector of character strings naming the set of features.
Must be specified.}

\item{dilationRadius}{specifies the dilation radius for determining the ROI for
each label.}

\item{neighborhoodRadiusStage1}{specifies which voxel neighbors should be included in
prediction for Stage 1.  The user can specify a scalar or vector but it must match
with what was used for training.}

\item{neighborhoodRadiusStage2}{specifies which voxel neighbors should be included in
prediction for Stage 2.  The user can specify a scalar or vector but it must match
with what was used for training.}

\item{useLabelDistances}{if TRUE, Euclidean distance maps for each label are
created and added to the list of featureImages.  This feature image type
was recommended in Wang et al.}

\item{normalizeSamplesPerLabel}{if TRUE, the samples from each ROI are normalized
by the mean of the voxels in that ROI}

\item{labelModelsStage1}{a list of models for stage 2 (prediction of background/foreground).
Each element of the labelSet requires a model.}
}
\value{
a list consisting of the refined segmentation estimate (RefinedSegmentationImage) 
        and a list of the foreground probability images (ForegroundProbabilityImages).
}
\description{
A random forest implementation of the corrective learning wrapper introduced
in Wang, et al., Neuroimage 2011 (http://www.ncbi.nlm.nih.gov/pubmed/21237273).
The prediction process involves using the two-stage training models (per label)
to refine an initial segmentation.  The two stages of refinement are:
    Stage 1:  Predict potentially misclassified voxels per label
    Stage 2:  Predict background/foreground voxels per label
}
\author{
Tustison NJ
}

