% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{setANTsDeterministic}
\alias{setANTsDeterministic}
\title{Set ANTs deterministic behavior}
\usage{
setANTsDeterministic(on = TRUE, seedValue = 123)
}
\arguments{
\item{on}{Set deterministic behavior.}

\item{seedValue}{Assign random seed value.}
}
\description{
Register a pair of images either through the full or simplified interface
to the ANTs registration method.
}
