% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsrTransform_class.R
\name{setAntsrTransformFixedParameters}
\alias{setAntsrTransformFixedParameters}
\title{setAntsrTransformFixedParameters}
\usage{
setAntsrTransformFixedParameters(transform, parameters)
}
\arguments{
\item{transform}{antsrTransform}

\item{parameters}{array of parameters'}
}
\value{
TRUE
}
\description{
Set parameters of transform
}
\examples{
tx <- new("antsrTransform")
params <- getAntsrTransformFixedParameters(tx)
setAntsrTransformFixedParameters(tx, params * 2)
}
