% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simlr.R
\name{shorten_pymm_names}
\alias{shorten_pymm_names}
\title{shorter antspymm names}
\usage{
shorten_pymm_names(x)
}
\arguments{
\item{x}{string or strings}
}
\value{
string
}
\description{
shorter antspymm names
}
\author{
Avants BB
}
