% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscaleSVDxpts.R
\name{simlr.parameters}
\alias{simlr.parameters}
\title{Generate Parameter Grid for SIMLR Search}
\usage{
simlr.parameters(
  nsimlr_options,
  prescaling_options,
  objectiver_options,
  mixer_options,
  sparval_options,
  expBeta_options,
  positivities_options,
  optimus_options,
  constraint_options = list("none"),
  sparsenessAlg = list("soft"),
  num_samples = 10,
  search_type = c("random", "deterministic", "full")
)
}
\arguments{
\item{nsimlr_options}{A list of options for the `nsimlr` parameter.}

\item{prescaling_options}{A list of options for the `prescaling` parameter.}

\item{objectiver_options}{A list of options for the `objectiver` parameter.}

\item{mixer_options}{A list of options for the `mixer` parameter.}

\item{sparval_options}{A list of options for the `sparval` parameter.}

\item{expBeta_options}{A list of options for the `ebber` parameter.}

\item{positivities_options}{A list of options for the `pizzer` parameter.}

\item{optimus_options}{A list of options for the `optimus` parameter.}

\item{constraint_options}{A list of options for the `constraint` parameter, default is `list("none")`.}

\item{sparsenessAlg}{A list of options for the `sparsenessAlg` parameter, default is `list(NA)`.}

\item{num_samples}{if not full, then the size of the subset space}

\item{search_type}{string vector either full random or deterministic}
}
\value{
A list containing all (or a subset of) combinations of the provided parameters. Each row in the data frame represents a unique combination of the parameters.

The columns of the returned data frame include:
\itemize{
  \item \code{nsimlr}: Values corresponding to the `nsimlr` parameter.
  \item \code{prescaling}: Values corresponding to the `prescaling` parameter.
  \item \code{objectiver}: Values corresponding to the `objectiver` parameter.
  \item \code{mixer}: Values corresponding to the `mixer` parameter.
  \item \code{constraint}: Values corresponding to the `constraint` parameter.
  \item \code{sparval}: Values corresponding to the `sparval` parameter.
  \item \code{ebber}: Values corresponding to the `ebber` parameter.
  \item \code{pizzer}: Values corresponding to the `pizzer` parameter.
  \item \code{optimus}: Values corresponding to the `optimus` parameter.
}

Each row represents a specific set of parameters that can be passed to the `simlr.search` function for evaluation. The returned data frame is intended for use in parameter tuning and optimization.
}
\description{
This function creates and optionally subsets all possible combinations of input parameters for use in a SIMLR grid search.
}
