% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscaleSVDxpts.R
\name{simlr.perm}
\alias{simlr.perm}
\title{Assess Significance of SiMLR Call}
\usage{
simlr.perm(
  voxmats,
  smoothingMatrices,
  iterations = 10,
  sparsenessQuantiles,
  positivities,
  initialUMatrix,
  mixAlg = c("svd", "ica", "avg", "rrpca-l", "rrpca-s", "pca", "stochastic"),
  orthogonalize = FALSE,
  repeatedMeasures = NA,
  lineSearchRange = c(-1e+10, 1e+10),
  lineSearchTolerance = 1e-08,
  randomSeed,
  constraint = c("none", "Grassmann", "Stiefel"),
  energyType = c("cca", "regression", "normalized", "ucca", "lowRank",
    "lowRankRegression", "normalized_correlation"),
  vmats,
  connectors = NULL,
  optimizationStyle = c("lineSearch", "mixed", "greedy"),
  scale = c("centerAndScale", "sqrtnp", "np", "center", "norm", "none", "impute",
    "eigenvalue", "robust"),
  expBeta = 0,
  jointInitialization = TRUE,
  sparsenessAlg = NA,
  verbose = FALSE,
  nperms = 50,
  FUN = "mean"
)
}
\arguments{
\item{voxmats}{A list of voxel matrices.}

\item{smoothingMatrices}{A list of smoothing matrices.}

\item{iterations}{Number of iterations. Default is 10.}

\item{sparsenessQuantiles}{A vector of sparseness quantiles.}

\item{positivities}{A vector of positivity constraints.}

\item{initialUMatrix}{Initial U matrix for the algorithm.}

\item{mixAlg}{The mixing algorithm to use. Default is 'svd'.}

\item{orthogonalize}{Logical indicating whether to orthogonalize. Default is FALSE.}

\item{repeatedMeasures}{Repeated measures data. Default is NA.}

\item{lineSearchRange}{Range for line search. Default is c(-1e+10, 1e+10).}

\item{lineSearchTolerance}{Tolerance for line search. Default is 1e-08.}

\item{randomSeed}{Seed for random number generation.}

\item{constraint}{The constraint type. Default is 'none'.}

\item{energyType}{The energy type. Default is 'cca'.}

\item{vmats}{List of V matrices - optional initialization matrices}

\item{connectors}{List of connectors. Default is NULL.}

\item{optimizationStyle}{The optimization style. Default is 'lineSearch'.}

\item{scale}{Scaling method. Default is 'centerAndScale'.}

\item{expBeta}{Exponential beta value. Default is 0.}

\item{jointInitialization}{Logical indicating joint initialization. Default is TRUE.}

\item{sparsenessAlg}{Sparseness algorithm. Default is NA.}

\item{verbose}{Logical indicating whether to print verbose output. Default is FALSE. values > 1 lead to more verbosity}

\item{nperms}{Number of permutations for significance testing. Default is 50.}

\item{FUN}{function for summarizing variance explained}
}
\value{
A data frame containing p-values for each permutation.
}
\description{
This function performs permutation tests to assess the significance of a SiMLR analysis.
For more detail on input parameters, see the original function.
}
