% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscaleSVDxpts.R
\name{simlr.search}
\alias{simlr.search}
\title{Perform Parallel SIMLR Grid Search (with Progress Bar)}
\usage{
simlr.search(
  mats,
  regs,
  options_df,
  maxits = 100,
  connectors = NULL,
  nperms = 1,
  verbose = 0,
  FUN = rvcoef,
  cores = 1
)
}
\arguments{
\item{mats}{List of input matrices for SIMLR.}

\item{regs}{List of regularization parameters for SIMLR.}

\item{options_df}{Data frame of parameter combinations.}

\item{maxits}{Maximum iterations for SIMLR.}

\item{connectors}{List of connectors for modalities.}

\item{nperms}{Number of permutations for significance test.}

\item{verbose}{Verbosity level.}

\item{FUN}{Evaluation function.}

\item{cores}{Number of CPU cores to use (default: detectCores() - 1). \code{parallel::detectCores() - 1}}
}
\value{
List with best SIMLR result, significance, best parameters, and search table.
}
\description{
This version runs parameter sets in parallel using all available CPU cores (minus one by default),
and shows a progress bar with ETA.
}
