% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscaleSVDxpts.R
\name{simlr_impute}
\alias{simlr_impute}
\title{Impute missing SiMLR data in a specified column based on other columns}
\usage{
simlr_impute(
  dataframe,
  nms,
  vecnum,
  toimpute,
  separator = "PC",
  family = "gaussian"
)
}
\arguments{
\item{dataframe}{A data frame containing the data to impute.}

\item{nms}{A vector of base column names.}

\item{vecnum}{A numeric value to append to the column names.}

\item{toimpute}{The base name of the target column to be imputed.}

\item{separator}{A string specifying the separator between the column name and feature.}

\item{family}{A string specifying the GLM family (default is 'gaussian').}
}
\value{
A data frame with imputed values.
}
\description{
Impute missing SiMLR data in a specified column based on other columns
}
\examples{
set.seed(123)
n=50
df <- data.frame(
  t1PC1 = rnorm(n),
  t1aPC1 = rnorm(n),
  dtPC1 = rnorm(n),
  dtaPC1 = rnorm(n),
  rsfPC1 = rnorm(n),
  perfPC1 = rnorm(n)
)
df[ sample(1:nrow(df),20),6 ]=NA
nms <- c("t1", "t1a", "dt", "dta", "rsf", "perf")
vecnum <- 1
toimpute <- "perf"
df = simlr_impute(df, nms, vecnum, toimpute, family = 'gaussian')
}
