% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscaleSVDxpts.R
\name{simlr_path_models}
\alias{simlr_path_models}
\title{SiMLR Path Models helper}
\usage{
simlr_path_models(n, type = 0)
}
\arguments{
\item{n}{The length of the output list.}

\item{type}{The type of modeling. Can be one of the following:
- 0: "Exclude self" - each entry contains all integers except the one corresponding to its position.
- Integer greater than 0: "focus on a target" - each entry contains the target integer, except for the entry at the target position, which contains all integers except the target.
- 'pca': "Identity" - each entry contains its own position number.}
}
\value{
A list of length n with contents based on the input type.
}
\description{
Creates the list that should be passed to the connectors argument
to simlr.   the different options are default (0) which connects
each to all others but not itself, a single integer > 0 which forces
modeling to focus on a given target and then a pca like option.
}
