% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscaleSVDxpts.R
\name{simlrU}
\alias{simlrU}
\title{Compute the low-dimensional u matrix for simlr}
\usage{
simlrU(
  projections,
  mixingAlgorithm,
  initialW,
  orthogonalize = FALSE,
  connectors = NULL
)
}
\arguments{
\item{projections}{A list that contains the low-dimensional projections.}

\item{mixingAlgorithm}{the elected mixing algorithm.  see \code{simlr}.  can
be 'svd', 'ica', 'rrpca-l', 'rrpca-s', 'pca', 'stochastic' or 'avg'.}

\item{initialW}{initialization matrix size \code{n} by \code{k} for fastICA.}

\item{orthogonalize}{boolean}

\item{connectors}{a list ( length of projections or number of modalities )
that indicates which modalities should be paired with current modality}
}
\value{
u matrix for modality i
}
\description{
simlr minimizes reconstruction error across related modalities.  One crucial
component of the reconstruction is the low-dimensional cross-modality basis.
This function computes that basis, given a mixing algorithm.
}
\examples{

set.seed(1500)
nsub <- 25
npix <- c(100, 200, 133)
nk <- 5
outcome <- matrix(rnorm(nsub * nk), ncol = nk)
outcome1 <- matrix(rnorm(nsub * nk), ncol = nk)
outcome2 <- matrix(rnorm(nsub * nk), ncol = nk)
u <- simlrU(list(outcome, outcome1, outcome2), 2, "avg")

}
\seealso{
\code{\link{simlr}}
}
\author{
BB Avants.
}
