% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateBiasField.R
\name{simulateBiasField}
\alias{simulateBiasField}
\title{Simulate random bias field}
\usage{
simulateBiasField(
  domainImage,
  numberOfPoints = 10,
  sdBiasField = 1,
  numberOfFittingLevels = 4,
  meshSize = 1
)
}
\arguments{
\item{domainImage}{image to define the spatial domain of the bias field.}

\item{numberOfPoints}{number of randomly defined points to define the bias
field (default = 10).}

\item{sdBiasField}{characterize the standard deviation of the amplitude
(default = 1).}

\item{numberOfFittingLevels}{B-spline fitting parameter.}

\item{meshSize}{B-spline fitting parameter (scalar or vector of size image
dimension).}
}
\value{
simulated bias field
}
\description{
Low frequency, spatial varying simulated random bias field using
random points and B-spline fitting.
}
\examples{

library( ANTsR )
image <- antsImageRead( getANTsRData( "r16" ) )
logField <- simulateBiasField(image, numberOfPoints = 10, sdBiasField = 1.0,
   numberOfFittingLevels = 2, meshSize = 10 ) \%>\% iMath( "Normalize" )
logField <- ( exp( logField ) )^4
image <- image * logField
rm(image); gc()
rm(logField); gc()
}
\author{
Tustison NJ
}
