% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateDisplacementField.R
\name{simulateDisplacementField}
\alias{simulateDisplacementField}
\title{simulateDisplacementField}
\usage{
simulateDisplacementField(
  domainImage,
  fieldType = c("bspline", "exponential"),
  numberOfRandomPoints = 1000,
  sdNoise = 10,
  enforceStationaryBoundary = TRUE,
  numberOfFittingLevels = 4,
  meshSize = 1,
  sdSmoothing = 4
)
}
\arguments{
\item{domainImage}{image to define the domain of the field.}

\item{fieldType}{either "bspline" or "exponential".}

\item{numberOfRandomPoints}{number of displacement points.
Default = 1000.}

\item{sdNoise}{standard deviation of the displacement field
noise (in mm).  Default = 10.0.}

\item{enforceStationaryBoundary}{boolean determining fixed boundary
conditions.  Default = TRUE.}

\item{numberOfFittingLevels}{(bspline only) number of fitting levels.
Default = 4.}

\item{meshSize}{(bspline only) scalar or n-D vector determining fitting
resolution.  Default = 1.}

\item{sdSmoothing}{(exponential only) standard deviation of the
Gaussian smoothing in mm.  Default = 4.0.}
}
\value{
ANTsR displacement field.
}
\description{
Simulate a displacement field using random points with B-spline
smoothing
}
\examples{
domainImage <- antsImageRead( getANTsRData( "r16" ), 2 )
expField <- simulateDisplacementField( domainImage, fieldType = "exponential" )
bsplineField <- simulateDisplacementField( domainImage, fieldType = "bspline" )
warp <- antsrTransformFromDisplacementField( bsplineField * 3 )
wimg <- applyAntsrTransformToImage( warp, domainImage, domainImage )

}
\author{
NJ Tustison
}
