% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skeletonize.R
\name{skeletonize}
\alias{skeletonize}
\title{Skeletonize a binary segmentation.}
\usage{
skeletonize(x, laplacianThreshold = 0.25, propagationOption = 1)
}
\arguments{
\item{x}{input binary image}

\item{laplacianThreshold}{between zero and one, controls the speed image}

\item{propagationOption}{either 0 (none), 1 (well-composed) or 2 (topological) constraint}
}
\value{
output binary image
}
\description{
Employ a topologically constrained label propagation tool in order to thin
an existing segmentation into a skeleton.  The function will perform better
if the initial segmentation is well-composed.
}
\examples{

img = antsImageRead( getANTsRData( 'r16') ) \%>\% thresholdImage( "Otsu", 3 )
skel = skeletonize( thresholdImage( img, 3, 3 ) )

}
\author{
Avants BB
}

