% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sliceTimingCorrection.R
\name{sliceTimingCorrection}
\alias{sliceTimingCorrection}
\title{slice timing correction for fMRI.}
\usage{
sliceTimingCorrection(fmri, sliceTime = NA, interpolation = "sinc",
  sincRadius = 4, bsplineOrder = 3)
}
\arguments{
\item{fmri}{input bold image}

\item{sliceTime}{the TR}

\item{interpolation}{sinc recommended but linear is much faster}

\item{sincRadius}{recommend 4 for interpolation}

\item{bsplineOrder}{recommend 3 for interpolation}
}
\value{
matrix is output
}
\description{
Alters BOLD signal according to acquisition parameters (TR).
}
\author{
Avants BB
}

