% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/smoothImage.R
\name{smoothImage}
\alias{smoothImage}
\title{Smooth image}
\usage{
smoothImage( inimg, sigma )
}
\arguments{
\item{inimg}{Input image to operate on}

\item{sigma}{Smoothing factor.  Either scalar, or vector of length \code{dim} for \code{dim}-dimensional image.}
}
\value{
antsImage smoothed
}
\description{
Perform smoothing on the given image with a given sigma, defined in physical space units
}
\examples{
img<-makeImage(c(5,5),rnorm(25))
simg<-smoothImage( img ,c(1.2,1.5) )
}
\author{
Shrinidhi KL, Avants BB
}

