% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscaleSVDxpts.R
\name{sparseDistanceMatrix}
\alias{sparseDistanceMatrix}
\title{Create sparse distance, covariance or correlation matrix}
\usage{
sparseDistanceMatrix(x, k = 3, r = Inf, kmetric = c("euclidean",
  "correlation", "covariance"))
}
\arguments{
\item{x}{input matrix, should be n (samples) by p (measurements)}

\item{k}{number of neighbors}

\item{r}{radius of epsilon-ball}

\item{kmetric}{similarity or distance metric determining k nearest neighbors}
}
\value{
matrix sparse matrix is output
}
\description{
Exploit k-nearest neighbor algorithms to estimate a sparse similarity matrix.
See relevant publications by Mauro Maggioni and other authors.
}
\examples{

mat = matrix(rnorm(300),ncol=50)
smat = sparseDistanceMatrix( mat )

}
\author{
Avants BB
}
\references{
\url{http://www.math.jhu.edu/~mauro/multiscaledatageometry.html}
}

