% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscaleSVDxpts.R
\name{sparseDistanceMatrixXY}
\alias{sparseDistanceMatrixXY}
\title{Create sparse distance, covariance or correlation matrix from x, y}
\usage{
sparseDistanceMatrixXY(
  x,
  y,
  k = 3,
  r = Inf,
  sigma = NA,
  kmetric = c("euclidean", "correlation", "covariance", "gaussian"),
  eps = 1e-06,
  kPackage = "RcppHNSW",
  ncores = NA,
  verbose = FALSE
)
}
\arguments{
\item{x}{input matrix, should be n (samples) by p (measurements)}

\item{y}{input matrix second view, should be n (samples) by q (measurements)}

\item{k}{number of neighbors}

\item{r}{radius of epsilon-ball}

\item{sigma}{parameter for kernel PCA.}

\item{kmetric}{similarity or distance metric determining k nearest neighbors}

\item{eps}{epsilon error for rapid knn}

\item{kPackage}{name of package to use for knn}

\item{ncores}{number of cores to use}

\item{verbose}{verbose output}
}
\value{
matrix sparse p by q matrix is output with p by k nonzero entries
}
\description{
Exploit k-nearest neighbor algorithms to estimate a sparse matrix measuring
the distance, correlation or covariance between two matched datasets.
Critical to the validity of this function is the basic mathematical
relationships between euclidean distance and correlation and between
correlation and covariance.  For applications of such matrices, one may see
relevant publications by Mauro Maggioni and other authors.
}
\examples{
\dontrun{
set.seed(120)
mat = matrix( rnorm(60), nrow=6 )
mat2 = matrix( rnorm(120), nrow=6 )
smat = sparseDistanceMatrixXY( mat, mat2, 3 )
smat2 = sparseDistanceMatrixXY( mat2, mat, 3 )
testthat::expect_is(smat, "Matrix")
testthat::expect_is(smat, "dgCMatrix")
testthat::expect_is(smat2, "Matrix")
testthat::expect_is(smat2, "dgCMatrix")
testthat::expect_equal(sum(smat), 154.628961265087)
testthat::expect_equal(sum(smat2), 63.7344262003899)
}
}
\references{
\url{http://www.math.jhu.edu/~mauro/multiscaledatageometry.html}
}
\author{
Avants BB
}
