% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mrvnrfs.R
\name{splitMask}
\alias{splitMask}
\title{split a mask into n labeled sub-masks}
\usage{
splitMask(mask, n)
}
\arguments{
\item{mask}{antsImage mask}

\item{n}{number of mask chunks}
}
\value{
relabeledMask
}
\description{
split a mask into n labeled sub-masks
}
\examples{
mask = getMask( antsImageRead( getANTsRData("r16" ) ) )
smask = splitMask( mask, 10 )
}
\author{
Avants BB, Tustison NJ
}

