% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitNDImageToList.R
\name{splitNDImageToList}
\alias{splitNDImageToList}
\title{splitNDImageToList}
\usage{
splitNDImageToList(img)
}
\arguments{
\item{img}{input image of class antsImage, last dimension will be split}
}
\value{
output list filled with d-1 volumes of class antsImage
}
\description{
This function splits the dimensionality d input image into a list of images
with dimensonality d-1 by slicing the input image along its last dimension.
}
\examples{

my4Dimage <- makeImage(c(5, 5, 5, 4))
my3Dlist <- splitNDImageToList(my4Dimage)

}
\author{
Pustina D, Avants B
}
