% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bracket_subset.R
\name{subset.antsImage}
\alias{subset.antsImage}
\title{Subsetting antsImage Objects}
\usage{
\method{subset}{antsImage}(x, subset, ...)
}
\arguments{
\item{x}{antsImage object}

\item{subset}{logical expression of subset}

\item{...}{further arguments to be passed to or from other methods.}
}
\description{
Returns subsets of antsImage objects as vectors
}
