% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscaleSVDxpts.R
\name{sum_preserving_matrix_partition}
\alias{sum_preserving_matrix_partition}
\title{Sum preserving matrix partition}
\usage{
sum_preserving_matrix_partition(X, option = 2, tol = 0.001)
}
\arguments{
\item{X}{A matrix of size p x k}

\item{option}{An integer indicating whether to allow +/- values (option 1)
or only + values (option 2). Default is 1.}

\item{tol}{A numeric value indicating the tolerance for the row sums.
Default is 0.1.}
}
\value{
A matrix with segmented rows
}
\description{
This function takes a matrix as input and partitions each column such that
only one entry in each row is non-zero, and the sums of each row are roughly
equivalent.
}
\details{
The function uses a greedy algorithm to select the entry with the
         maximum absolute value (or maximum positive value) in each column.
         The rows are then normalized to ensure that the sums are roughly
         equivalent.
}
\examples{
X <- matrix(rnorm(3000), nrow = 1000)
Y <- segment_matrix(X, option = 1)
Y <- segment_matrix(X, option = 2)

}
