% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscaleSVDxpts.R
\name{symilrU}
\alias{symilrU}
\title{Compute the low-dimensional u matrix for symilr}
\usage{
symilrU(projections, i, mixingAlgorithm, initialW)
}
\arguments{
\item{projections}{A list that contains the low-dimensional projections.}

\item{i}{which modality to predict from the others.}

\item{mixingAlgorithm}{the elected mixing algorithm.  see \code{symilr}.  can
be 'svd', 'ica', 'rrpca-l', 'rrpca-s' or 'avg'.}

\item{initialW}{initialization matrix size \code{n} by \code{k} for fastICA.}
}
\value{
u matrix for modality i
}
\description{
SyMILR minimizes reconstruction error across related modalities.  One crucial
component of the reconstruction is the low-dimensional cross-modality basis.
This function computes that basis, given a mixing algorithm.
}
\examples{

set.seed(1500)
nsub = 25
npix = c(100,200,133)
nk = 5
outcome = matrix(rnorm( nsub * nk ),ncol=nk)
outcome1 = matrix(rnorm( nsub * nk ),ncol=nk)
outcome2 = matrix(rnorm( nsub * nk ),ncol=nk)
u = symilrU( list( outcome, outcome1, outcome2 ), 2, 'avg' )

}
\seealso{
\code{\link{symilr}}
}
\author{
BB Avants.
}
