% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscaleSVDxpts.R,
%   R/multiscaleSVDxpts_test.R, R/multiscaleSVDxpts_test2.R
\name{symlrU}
\alias{symlrU}
\title{Compute the low-dimensional u matrix for symlr}
\usage{
symlrU(projections, i, mixingAlgorithm, initialW, orthogonalize = FALSE)

symlrU(projections, i, mixingAlgorithm, initialW, orthogonalize = FALSE)

symlrU(projections, i, mixingAlgorithm, initialW, orthogonalize = FALSE)
}
\arguments{
\item{projections}{A list that contains the low-dimensional projections.}

\item{i}{which modality to predict from the others.}

\item{mixingAlgorithm}{the elected mixing algorithm.  see \code{symlr}.  can
be 'svd', 'ica', 'rrpca-l', 'rrpca-s', 'pca', 'stochastic' or 'avg'.}

\item{initialW}{initialization matrix size \code{n} by \code{k} for fastICA.}

\item{orthogonalize}{boolean}

\item{projections}{A list that contains the low-dimensional projections.}

\item{mixingAlgorithm}{the elected mixing algorithm.  see \code{symlr}.  can
be 'svd', 'ica', 'rrpca-l', 'rrpca-s', 'pca', 'stochastic' or 'avg'.}

\item{initialW}{initialization matrix size \code{n} by \code{k} for fastICA.}

\item{orthogonalize}{boolean}

\item{projections}{A list that contains the low-dimensional projections.}

\item{i}{which modality to predict from the others.}

\item{mixingAlgorithm}{the elected mixing algorithm.  see \code{symlr}.  can
be 'svd', 'ica', 'rrpca-l', 'rrpca-s', 'pca', 'stochastic' or 'avg'.}

\item{initialW}{initialization matrix size \code{n} by \code{k} for fastICA.}

\item{orthogonalize}{boolean}
}
\value{
u matrix for modality i

u matrix for modality i

u matrix for modality i
}
\description{
symlr minimizes reconstruction error across related modalities.  One crucial
component of the reconstruction is the low-dimensional cross-modality basis.
This function computes that basis, given a mixing algorithm.

symlr minimizes reconstruction error across related modalities.  One crucial
component of the reconstruction is the low-dimensional cross-modality basis.
This function computes that basis, given a mixing algorithm.

symlr minimizes reconstruction error across related modalities.  One crucial
component of the reconstruction is the low-dimensional cross-modality basis.
This function computes that basis, given a mixing algorithm.
}
\examples{

set.seed(1500)
nsub = 25
npix = c(100,200,133)
nk = 5
outcome = matrix(rnorm( nsub * nk ),ncol=nk)
outcome1 = matrix(rnorm( nsub * nk ),ncol=nk)
outcome2 = matrix(rnorm( nsub * nk ),ncol=nk)
u = symlrU( list( outcome, outcome1, outcome2 ), 2, 'avg' )


set.seed(1500)
nsub = 25
npix = c(100,200,133)
nk = 5
outcome = matrix(rnorm( nsub * nk ),ncol=nk)
outcome1 = matrix(rnorm( nsub * nk ),ncol=nk)
outcome2 = matrix(rnorm( nsub * nk ),ncol=nk)
u = symlrU( list( outcome, outcome1, outcome2 ), 2, 'avg' )


set.seed(1500)
nsub = 25
npix = c(100,200,133)
nk = 5
outcome = matrix(rnorm( nsub * nk ),ncol=nk)
outcome1 = matrix(rnorm( nsub * nk ),ncol=nk)
outcome2 = matrix(rnorm( nsub * nk ),ncol=nk)
u = symlrU( list( outcome, outcome1, outcome2 ), 2, 'avg' )

}
\seealso{
\code{\link{symlr}}

\code{\link{symlr}}

\code{\link{symlr}}
}
\author{
BB Avants.

BB Avants.

BB Avants.
}
