% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresholdImage.R
\name{thresholdImage}
\alias{thresholdImage}
\title{Threshold Image}
\usage{
thresholdImage(inimg, lothresh, hithresh, inval=1, outval=0)
}
\arguments{
\item{inimg}{Input image to operate on}

\item{lothresh}{Lower edge of threshold window}

\item{hithresh}{Higher edge of threshold window}

\item{inval}{Output value for image voxels in between \code{lothresh} and \code{hithresh}}

\item{outval}{Output value for image voxels lower than \code{lothresh} or higher than \code{hithresh}}
}
\value{
antsImage
}
\description{
converts a scalar image into a binary image by thresholding operations
}
\examples{
set.seed(1234)
img <- makeImage(c(5,5), rnorm(25)+0.5)
imgt<-thresholdImage( img, 0.5, Inf )
testthat::expect_equal(sum(imgt), 9)
imgt<-thresholdImage( img > 0.5, 0.1, Inf )
testthat::expect_equal(sum(imgt), 9)
}
\author{
Shrinidhi KL
}
