% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/timeseries2matrix.R
\name{timeseries2matrix}
\alias{timeseries2matrix}
\title{Time-series image to matrix}
\usage{
timeseries2matrix(img, mask)
}
\arguments{
\item{img}{Input image of type 'antsImage' or R array.}

\item{mask}{Input mask of type 'antsImage' or R array. In either case, the
number of voxels in the mask may be either equal to that of input image
'img' or equal to number of voxels in one time unit of input image 'img'. In
the second case, the mask is reused for every time unit of the 'img'.  A
mask of n-labels will cause the function to return a matrix containing the
mean time series within each label.}
}
\value{
Success -- an R matrix of dimensions ( dim(img)[length(dim(img))] ,
sum(mask==1) )\cr
}
\description{
Extract a matrix from a time-series image after applying a mask.
}
\examples{
img <- makeImage( c(10,10,10,5) , 0 )
# or use antsImageRead ...
mask <- array( 1 , dim(img)[1:3] )
mat <- timeseries2matrix( img , mask )
}
\author{
Shrinidhi KL
}

