% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_data.R
\docType{data}
\name{tissues}
\alias{tissues}
\title{tissues}
\format{
A data frame with columns: Label, Description, ProbabilityNumber.
}
\usage{
tissues
}
\description{
A data frame label numbers and names for tissue types.
}
\keyword{datasets}
