\name{tracts}
\alias{tracts}
\docType{data}
\title{tracts}
\description{
A data frame label numbers and names for the white matter tracts labels.
}
\usage{data(tracts)}
\format{
  A data frame listing the following variables.
  \describe{
    \item{\code{label_num}}{ Numerical label value. }
    \item{\code{label_name}}{ Shorthand anatomical value. }
  }
}
\references{
\url{http://www.ncbi.nlm.nih.gov/pubmed/18407524}
}
\examples{
data(tracts)
}
\keyword{datasets}
