% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/landmarkTransform.R
\name{transform2mat}
\alias{transform2mat}
\title{write an affine transform matrix to ITK mat format}
\usage{
transform2mat(affinemat, file, RAS2IJK = NULL)
}
\arguments{
\item{affinemat}{a 4x4 (3D case) or 3x3 (2D case) homogenous transform matrix}

\item{file}{filename}

\item{RAS2IJK}{transform from RAS to IJK space (e.g. when landmarks are placed with Slicer).}
}
\value{
returns the character of the filename
}
\description{
write an affine transform matrix to ITK mat format
}
\examples{
require(Morpho)
fi <- antsImageRead(getANTsRData("r16") ,2)
## project landmarks into IJK space
fiLM2IJK <- applyTransform(fiLM,diag(c(-1,-1,1)))
load(system.file("extdata/landmarks.RData",package="ANTsR"))
## now compute a rigid transform to permuted landmarks
rotmat <- computeTransform(fiLM2IJK,fiLM2IJK[4:1,])
trafo <- transform2mat(rotmat,"rotmat2D.mat")
## write transfrom to disk
fiRotated <- antsApplyTransforms(fi,fi,trafo)
plot(fiRotated)
}
\author{
Schlager S
}

