% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bloodPerfusionSVD.R
\name{trapz}
\alias{trapz}
\title{Calculate the area under a curve (or set of curves) using the trapezoidal rule.}
\usage{
trapz(x, y)
}
\arguments{
\item{x}{vector of samples for the dependent variable.}

\item{y}{vector or matrix of samples for the independent variable.  In the case of the
latter, curves are organized column-wise.}
}
\value{
area (areas) under the sampled curve (curves).
}
\description{
Implementation of the deconvolution technique of Ostergaard et al.
(http://www.ncbi.nlm.nih.gov/pubmed/8916023).  Given an arterial input function,
}
\examples{

x <- seq( 0, 1, by = 0.0001 )
y <- exp( x )

Compare with true area of exp( 1 ) - 1 = 1.718282...
areaEstimate <- trapz( x, y )

}
\author{
Tustison NJ
}

