% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscaleSVDxpts.R
\name{undo_transform_matrix}
\alias{undo_transform_matrix}
\title{Undo a previously applied matrix transform}
\usage{
undo_transform_matrix(Xs, params)
}
\arguments{
\item{Xs}{Transformed matrix.}

\item{params}{List of parameters returned by \code{transform_matrix()}.}
}
\value{
The untransformed (original scale) matrix.
}
\description{
Undo a previously applied matrix transform
}
