% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_Operators.R
\docType{methods}
\name{var}
\alias{var}
\alias{var,antsImage-method}
\title{Variance of Image}
\usage{
var(x, y = NULL, ...)

\S4method{var}{antsImage}(x, y = NULL, ...)
}
\arguments{
\item{x}{antsImage object}

\item{y}{antsImage object, but likely null}

\item{...}{additional arguments passed to \code{\link{var}}}
}
\description{
Calculates the variance of an image
}
