% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscaleSVDxpts.R
\name{vector_to_df}
\alias{vector_to_df}
\title{Convert a Vector to a Data Frame}
\usage{
vector_to_df(vector, column_name)
}
\arguments{
\item{vector}{The input vector}

\item{column_name}{The base name for the columns}
}
\value{
A data frame with one row and columns determined by the length of the vector
}
\description{
Convert a Vector to a Data Frame
}
\examples{
vector_to_df(1:10, "nm")
}
