% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deformationFieldPCA.R
\name{vectortomultichannel}
\alias{vectortomultichannel}
\title{Convert vector to multichannel \code{antsImage}.}
\usage{
vectortomultichannel(v, mask)
}
\arguments{
\item{v}{input multichannel vector}

\item{mask}{mask of same dimensionality as multichannelimage}
}
\value{
multichannelimage is output
}
\description{
Converts a vector of size n-entries in mask times number of channels to
a multi-channel image with the same dimensionality as the mask.
}
\examples{

fi <- antsImageRead(getANTsRData("r16") ) \%>\%
  resampleImage(c(60,60),1,0)
mi <- antsImageRead(getANTsRData("r64") ) \%>\%
  resampleImage(c(60,60),1,0)
mytx <- antsRegistration(fixed=fi, moving=mi, typeofTransform = c('SyN') )
mcimg = antsImageRead( mytx$fwd[1] )
msk = getMask( fi )
vv=multichanneltovector(mcimg,msk)
mcimg2=vectortomultichannel( vv, msk )
vv2=multichanneltovector(mcimg2,msk)
cor.test(vv2,vv)
stopifnot( all( mcimg2[30,30] == mcimg[30,30] ) )

}
\author{
Avants BB
}

