% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simlr.R
\name{visualize_lowrank_relationships}
\alias{visualize_lowrank_relationships}
\title{Visualize Low-Rank Relationships}
\usage{
visualize_lowrank_relationships(
  X1,
  X2,
  V1,
  V2,
  plot_title = NULL,
  nm1 = "X1",
  nm2 = "X2",
  max_pairs = 5
)
}
\arguments{
\item{X1}{First matrix (Subjects x Features)}

\item{X2}{Second matrix (Subjects x Features)}

\item{V1}{First feature loading matrix (Features x Components)}

\item{V2}{Second feature loading matrix (Features x Components)}

\item{plot_title}{Title of the plot}

\item{nm1}{Label for view 1}

\item{nm2}{Label for view 2}

\item{max_pairs}{Maximum components to include in the pairs plot (Safety cap)}
}
\description{
Visualize Low-Rank Relationships
}
