% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/vwnrfs.R
\name{vwnrfs}
\alias{vwnrfs}
\title{voxelwise neighborhood random forest segmentation and prediction}
\usage{
vwnrfs(y, x, labelmask, rad = NA, nsamples = 1, ntrees = 500,
  asFactors = TRUE)
}
\arguments{
\item{y}{list of training label images, can be a factor or numeric vector
this can also be a regular old vector}

\item{x}{a list of lists where each list contains feature images}

\item{labelmask}{a mask for the features (all in the same image space)
the labelmask defines the number of parallel samples that will be used
per subject sample. two labels will double the number of predictors
contributed from each feature image.}

\item{rad}{vector of dimensionality d define nhood radius}

\item{nsamples}{(per subject to enter training)}

\item{ntrees}{(for the random forest model)}

\item{asFactors}{boolean - treat the y entries as factors}
}
\value{
list a 3-list with the rf model, training vector and feature matrix
}
\description{
Represents feature images as a neighborhood and uses the features
to build a random forest prediction from an image population
}
\examples{
mask<-makeImage( c(10,10), 0 )
mask[ 3:6, 3:6 ]<-1
mask[ 5, 5 ]<-2
ilist<-list()
lablist<-list()
inds<-1:50
scl<-0.33 # a noise parameter
for ( i in inds ) {
  img<-antsImageClone(mask)
  imgb<-antsImageClone(mask)
  img[ 3:6, 3:6 ]<-rnorm(16)*scl+(i \%\% 2)+scl*mean(rnorm(1))
  imgb[ 3:6, 3:6 ]<-rnorm(16)*scl+(i \%\% 2)+scl*mean(rnorm(1))
  limg<-antsImageClone(mask)
  limg[ 3:6, 3:6 ]<-(i \%\% 2)  # the label image is constant
  ilist[[i]]<-list(img,imgb)  # two features
  lablist[[i]]<-limg
}
rfm<-vwnrfs( lablist , ilist, mask, rad=c(0,0) )
sum(abs(as.numeric(rfm$tv)-as.numeric(predict(rfm$rfm))))
rfm<-vwnrfs( lablist , ilist, mask, rad=c(2,2) )
sum(abs(as.numeric(rfm$tv)-as.numeric(predict(rfm$rfm))))
}
\author{
Avants BB
}

