% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_data.R
\docType{data}
\name{wm_major_tracts}
\alias{wm_major_tracts}
\title{wm_major_tracts}
\format{
A data frame with columns: Label, Description, side.
}
\usage{
wm_major_tracts
}
\description{
A data frame label numbers and names for major white matter tracts.
}
\keyword{datasets}
