% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_data.R
\docType{data}
\name{wmh_evidence}
\alias{wmh_evidence}
\title{wmh_evidence}
\format{
A data frame with columns: Value, Description.
}
\usage{
wmh_evidence
}
\description{
A data frame describing WMH evidence features.
}
\keyword{datasets}
