% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsrTransform_class.R
\name{writeAntsrTransform}
\alias{writeAntsrTransform}
\title{writeAntsrTransform}
\usage{
writeAntsrTransform(transform, filename)
}
\arguments{
\item{transform}{antsrTransform}

\item{filename}{filename of transform}
}
\value{
TRUE
}
\description{
read a transform from file
}

