% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_data.R
\docType{data}
\name{aal}
\alias{aal}
\title{AAL Labels}
\format{
A data frame listing the following variables.
\describe{
  \item{\code{label_num}}{ Numerical label value. }
  \item{\code{label_name}}{ Shorthand anatomical value. }
  \item{\code{cerebellum}}{binary value}
  \item{\code{cortex}}{binary value}
  \item{\code{deepgrey}}{binary value}
  \item{\code{frontal}}{binary value}
  \item{\code{isdmn}}{binary value}
  \item{\code{issalience}}{binary value}
  \item{\code{left}}{binary value}
  \item{\code{limbic}}{binary value}
  \item{\code{occipital}}{binary value}
  \item{\code{parietal}}{binary value}
  \item{\code{right}}{binary value}
  \item{\code{temporal}}{binary value}
  \item{\code{unknown}}{binary value}
}
}
\usage{
aal
}
\description{
A data frame label numbers and names 
for the neuroanatomical AAL labels.
}
\examples{
data(aal)
}
\references{
\url{http://en.wikipedia.org/wiki/Automated_Anatomical_Labeling}
}
\keyword{datasets}
