% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/abpBrainExtraction.R
\name{abpBrainExtraction}
\alias{abpBrainExtraction}
\title{An ants-based brain extraction script.}
\usage{
abpBrainExtraction(img = NA, tem = NA, temmask = NA, tempriors = NA,
  tdir = NA)
}
\arguments{
\item{img}{image to which we map a brain mask}

\item{tem}{Template image which has an associated label mask.}

\item{temmask}{Template's antsImage brain mask.}

\item{tempriors}{e.g. c(img1,img2,...,imgN) Vector of spatial priors - may
not be implemented yet}

\item{tdir}{temporary directory}
}
\value{
outputs a brain image and brain mask.
}
\description{
Brain extraction based on mapping a template image and its mask to the input
image.  Should be preceded by abpN4.
}
\examples{
# short example first
fn<-getANTsRData('r16')
img<-antsImageRead(fn,2)
tf<-getANTsRData('r64')
tem<-antsImageRead(tf,2)
temmask<-antsImageClone( tem )
temmask[ tem  > 20 ]<-1
temmask[ tem  <= 20 ]<-0
bm<-abpBrainExtraction(img=img,tem=tem,temmask=temmask)
}
\author{
Tustison N, Avants BB
}

