% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abpN4.R
\name{abpN4}
\alias{abpN4}
\title{MR image bias correction based on the N4 algorithm.}
\usage{
abpN4(img, intensityTruncation = c(0.025, 0.975, 256), mask = NA,
  usen3 = FALSE)
}
\arguments{
\item{img}{image to be bias corrected}

\item{intensityTruncation}{quantiles for image truncation.}

\item{mask}{optional antsImage mask}

\item{usen3}{Use N3 algorithm instead of N4}
}
\value{
outputs a bias corrected image. 1 indicates failure.
}
\description{
Truncate outlier intensities and bias correct with the N4 algorithm.
}
\examples{

img <- antsImageRead(getANTsRData("r16"))
img2 <- abpN4( img )

}
\author{
Tustison N, Avants BB
}

