% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscaleSVDxpts.R
\name{adjusted_rvcoef}
\alias{adjusted_rvcoef}
\title{Computes the Adjusted RV-Coefficient using the Fast RV-Coefficient Function}
\usage{
adjusted_rvcoef(X, Y)
}
\arguments{
\item{X}{A numeric matrix (n observations, p variables).}

\item{Y}{A numeric matrix (n observations, q variables).}
}
\value{
A single scalar value for the adjusted RV-coefficient.
}
\description{
This function calculates the adjusted RV-coefficient, correcting for the
expected value under the null hypothesis of independence.
}
