% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscaleSVDxpts.R
\name{adjusted_rvcoef}
\alias{adjusted_rvcoef}
\title{Adjusted RV Coefficient}
\usage{
adjusted_rvcoef(X, Y, lambda = 1e-06)
}
\arguments{
\item{X}{First matrix}

\item{Y}{Second matrix}

\item{lambda}{The ridge penalty parameter (default is 1e-6).}
}
\value{
Adjusted RV coefficient (a value between 0 and 1)
}
\description{
Computes the adjusted RV coefficient between two matrices, as proposed by Mordant and Segers.
}
\examples{
X <- matrix(rnorm(100), nrow = 10)
Y <- matrix(rnorm(100), nrow = 10)
adjusted_rvcoef(X, Y)

}
\references{
Mordant, G., & Segers, J. (2006). A note on the RV coefficient. Journal of Multivariate Analysis, 97(10), 2155-2164.
}
