% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsAffineInitializer.R
\name{affineInitializer}
\alias{affineInitializer}
\title{a multi-start optimizer for affine registration}
\usage{
affineInitializer(fixedImage, movingImage, searchFactor = 20,
  radianFraction = 0.1, usePrincipalAxis = FALSE,
  localSearchIterations = 10, mask, txfn, reproducible = TRUE)
}
\arguments{
\item{fixedImage}{the fixed reference image}

\item{movingImage}{the moving image to be mapped to the fixed space}

\item{searchFactor}{degree of increments on the sphere to search}

\item{radianFraction}{between zero and one, defines the arc to search over}

\item{usePrincipalAxis}{boolean to initialize by principal axis}

\item{localSearchIterations}{gradient descent iterations}

\item{mask}{optional mask to restrict registration}

\item{txfn}{filename for the transformation}

\item{reproducible}{if \code{TRUE}, will execute 
\code{Sys.setenv(ITK_GLOBAL_DEFAULT_NUMBER_OF_THREADS = 1)} before
running to attempt a more reproducible result.  See
\url{https://github.com/ANTsX/ANTs/wiki/antsRegistration-reproducibility-issues}
for discussion.}
}
\value{
transformationMatrix
}
\description{
Searches over the sphere to find a good initialization for further
registration refinement, if needed.  This is a wrapper for the ANTs
function antsAffineInitializer.
}
\note{
See 
\url{https://github.com/ANTsX/ANTs/issues/444} for reproducibility discussion
}
\examples{

fi <- antsImageRead(getANTsRData("r16"))
mi <- antsImageRead(getANTsRData("r27"))
tx <- affineInitializer( fi, mi )


}
\author{
Avants BB
}
