% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alff.R
\name{alffmap}
\alias{alffmap}
\title{alffmap}
\usage{
alffmap(x, flo = 0.01, fhi = 0.1, tr = 1, detrend = TRUE,
  takesqrt = FALSE, kernel)
}
\arguments{
\item{x}{input vector for the time series of interest}

\item{flo}{low frequency, typically 0.01}

\item{fhi}{high frequency, typically 0.1}

\item{tr}{the period associated with the vector x (inverse of frequency)}

\item{detrend}{detrend the input time series}

\item{takesqrt}{take the sqrt of the computed values}

\item{kernel}{smoothing kernel in estimate, see \code{spec.pgram}}
}
\value{
vector is output showing ALFF and fALFF values
}
\description{
Amplitude of Low Frequency Fluctuations (ALFF; Zang et al., 2007) and
fractional Amplitude of Low Frequency Fluctuations (f/ALFF; Zou et al., 2008)
are related measures that quantify the amplitude of low frequency
oscillations (LFOs).  This function outputs ALFF and fALFF for the input.
}
\examples{


set.seed(1)
mat <- matrix(rnorm(3000),ncol=50)
fallf = apply( mat, FUN=alffmap, MARGIN=2 )
k = kernel("daniell", rep( 2, 3 ) )
fallf2 = apply( mat, FUN=alffmap, MARGIN=2, kernel=k )
stopifnot(abs(mean(fallf2) - 3.00367017972746) < 1e-13)


}
\author{
Avants BB
}
