% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/antsApplyTransforms.R
\name{antsApplyTransforms}
\alias{antsApplyTransforms}
\title{A simplified (or full) interface to antsApplyTransforms.}
\usage{
antsApplyTransforms(fixed = NA, moving = NA, transformlist = "",
  interpolator = "Linear", imagetype = 0, whichtoinvert = NA, ...)
}
\arguments{
\item{fixed}{fixed image defining domain into which the moving image is
transformed.}

\item{transformlist}{list of transforms generated by antsRegistration where
each transform is a filename.}

\item{interpolator}{choose the Linear, NearestNeighbor, MultiLabel or any
other interpolator available in antsApplyTransforms.}

\item{imagetype}{choose 0/1/2/3 mapping to scalar/vector/tensor/time-series
.}

\item{movingImage}{moving image to be mapped to fixed space.}
}
\value{
an antsImage is output. 1 -- Failure
}
\description{
Apply a transform list to map an image from one domain to another.
}
\examples{
# will give the full form of help
antsApplyTransforms( "-h")
# example 1 - simplified
mywarpedimage<-antsApplyTransforms(fixed=fixed,moving=moving,
  transformlist=mytx$fwdtransforms)
mywarpedimage<-antsApplyTransforms(fixed=moving,moving=fixed,
  transformlist=mytx$invtransforms)
# full access via listing the inputs in standard ANTs format
antsApplyTransforms( list( d=as.character(2),
 i=moving,o=outimg,r=fixed,t=mytx$inv[1],sep='') ) )
antsApplyTransforms( list( d=as.character(2),
 i=fixed,o=outimg,r=moving,t=paste("[",mytx$inv[1],",1]",sep='') ) )
}
\author{
Shrinidhi KL, Avants BB
}

