% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsApplyTransforms.R
\name{antsApplyTransforms}
\alias{antsApplyTransforms}
\title{Apply transforms to images.}
\usage{
antsApplyTransforms(fixed, moving, transformlist = "",
  interpolator = "Linear", imagetype = 0, whichtoinvert = NA, ...)
}
\arguments{
\item{fixed}{fixed image defining domain into which the moving image is
transformed.}

\item{moving}{moving image to be mapped to fixed space.}

\item{transformlist}{list of transforms generated by antsRegistration where
each transform is a filename.}

\item{interpolator}{Choice of interpolator.  One of: Linear,
NearestNeighbor, MultiLabel, CosineWindowedSinc, WelchWindowedSinc,
HammingWindowedSinc, LanczosWindowedSinc, Gaussian, or BSpline.
other interpolator available in antsApplyTransforms.}

\item{imagetype}{choose 0/1/2/3 mapping to scalar/vector/tensor/time-series}

\item{whichtoinvert}{optional list of booleans, same length as transforms. 
whichtoinvert[i] is TRUE if transformlist[i] is a matrix, and the matrix should
be inverted. If transformlist[i] is a warp field, whichtoinvert[i] must be FALSE.

If the transform list is a matrix followed by a warp field, whichtoinvert
defaults to c(TRUE,FALSE). Otherwise it defaults to rep(FALSE, length(transformlist)).}

\item{...}{other parameters to pass to antsApplyTransforms.}
}
\value{
an antsImage is output. 1 -- Failure
}
\description{
Apply a transform list to map an image from one domain to another. In image
registration, one computes mappings between (usually) pairs of images.
These transforms are often a sequence of increasingly complex maps, e.g.
from translation, to rigid, to affine to deformation.  The list of such
transforms is passed to this function to interpolate one image domain
into the next image domain, as below.  The order matters strongly and the
user is advised to familiarize with the standards established in examples.
}
\examples{

# will give the full form of help
antsApplyTransforms("-h")
# see antsRegistration
# example 1 - simplified
fixed <- antsImageRead( getANTsRData("r16") ,2)
moving <- antsImageRead( getANTsRData("r64") ,2)
fixed <- resampleImage(fixed,c(64,64),1,0)
moving <- resampleImage(moving,c(64,64),1,0)
mytx <- antsRegistration(fixed=fixed , moving=moving ,
  typeofTransform = c("SyN") )
mywarpedimage <- antsApplyTransforms( fixed=fixed,moving=moving,
  transformlist=mytx$fwdtransforms )
mywarpedimage <- antsApplyTransforms( fixed=moving,moving=fixed,
  transformlist=mytx$invtransforms )
# full access via listing the inputs in standard ANTs format

}
\author{
Shrinidhi KL, Avants BB
}
\seealso{
\code{\link{antsRegistration}}
}

