% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/antsAverageImages.R
\name{antsAverageImages}
\alias{antsAverageImages}
\title{antsAverageImages}
\usage{
antsAverageImages(imageList, reference, normalize = FALSE)
}
\arguments{
\item{imageList}{must be an antsImage list}

\item{reference}{must be an antsImage of the same size as those in the list}

\item{normalize}{true or false}
}
\description{
calculate the mean of a list of antsImages imageList: can contain anything that be cast via as.array()
}
\author{
Avants BB

Avants BB
}

