% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsAverageImages.R
\name{antsAverageImages}
\alias{antsAverageImages}
\title{Computes average of image list}
\usage{
antsAverageImages(imageList, normalize = FALSE)
}
\arguments{
\item{imageList}{list of antsImages}

\item{normalize}{boolean determines if image is divided by mean before
averaging}
}
\description{
Calculate the mean of a list of antsImages
}
\examples{
r16 <- antsImageRead(getANTsRData('r16'))
r64 <- antsImageRead(getANTsRData('r64'))
mylist <- list(r16, r64)
antsAverageImages(mylist)
}
\author{
Avants BB, Pustina D
}

