% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/antsCopyImageInfo.R
\name{antsCopyImageInfo}
\alias{antsCopyImageInfo}
\title{Copy image information.}
\usage{
antsCopyImageInfo(reference, target)
}
\arguments{
\item{reference}{Reference image to take header information from.}

\item{target}{Target image, which will have its header replaced.}
}
\value{
antsImage.
}
\description{
Copy image header from reference image to target.
}
\examples{
img1 <- makeImage(c(4,5), matrix(rnorm(20), nrow=4))
img2 <- antsImageClone(img1)
antsSetSpacing(img1, c(2,2))
antsCopyImageInfo(img1, img2)
}

