% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/antsCopyImageInfo.R
\name{antsCopyImageInfo}
\alias{antsCopyImageInfo}
\title{Copy header info}
\usage{
antsCopyImageInfo(reference, target)
}
\arguments{
\item{reference}{image object of S4 class \code{antsImage} to get values from.}

\item{target}{image object of S4 class \code{antsImage} to copy values to.}
}
\value{
Target image with reference header information.
}
\description{
Copy origin, direction, and spacing from one \code{antsImage} to another
}
\examples{
img <- makeImage(c(10,10),rnorm(100))
img2 <- makeImage(c(10,10), rnorm(100))
img2 <- antsCopyImageInfo(img, img2)
}

