% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/antsImage_class.R
\name{antsGetNeighborhoodMatrix}
\alias{antsGetNeighborhoodMatrix}
\title{antsGetNeighborhoodMatrix for the masked image}
\usage{
antsGetNeighborhoodMatrix(image, mask, radius, physical.coordinates = FALSE,
  boundary.condition = "NA", spatial.info = FALSE)
}
\arguments{
\item{image}{Image object of S4 class 'antsImage' to get values from.}

\item{mask}{Image object of S4 class 'antsImage' indicating which voxels to
examine. Each voxel > 0 will be used as the center of a neighborhood}

\item{radius}{array of values for neighborhood radius (in voxels)}

\item{physical.coordinates}{a logical indicating if voxel indices and
offsets should be in voxel or physical coordinates}

\item{boundary.condition}{a string indicating how to handle voxels in a
neighborhood, but not in the mask 'NA' - fill value with NA 'image' - use
image value even if not in mask 'mean' - use mean of all non-NA values for
that neighborhood}

\item{spatial.info}{a boolean indicating of voxel locations and neighborhood
offsets should be returned along with pixel values.}
}
\value{
if spatial.info is false: a matrix of pixel values where the number of rows
is the size of the neighborhood and there is a column for each voxel

if spatial.info is true, a list containing three matrices: values: matrix of
pixel values where the number of rows is the size of the neighborhood and
there is a column for each voxel indices: matrix providing the center
coordinates for each neighborhood offsets: matrix providing the offsets from
center for each voxel in a neighborhood
}
\description{
Summarize neighborhoods for voxels in an antsImage.
}
\examples{
r16<-getANTsRData("r16")
r16<-antsImageRead(r16,2)
mask<-getMask(r16,lowThresh=mean(r16),cleanup=1)
radius <- rep(2,2)
mat<-antsGetNeighborhoodMatrix(r16,mask,radius)
}
\author{
Duda JT
}

