% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/antsImage_class.R
\name{antsGetPixels}
\alias{antsGetPixels}
\title{Get Pixels}
\usage{
antsGetPixels(x, i = NA, j = NA, k = NA, l = NA)
}
\arguments{
\item{x}{Image object of S4 class 'antsImage' to get values from.}

\item{i}{indices in first dimension}

\item{j}{indices in second dimension}

\item{k}{indices in third dimension}

\item{l}{indices in forth dimension}
}
\value{
array of pixel values
}
\description{
Get pixel values from an 'antsImage'.
}
\examples{
img<-makeImage(c(10,10),rnorm(100))
pixel<-antsGetPixels(img,i=c(1,2),j=1)
}

