% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsImage_class.R
\docType{methods}
\name{mean,antsImage-method}
\alias{*.antsImage}
\alias{+.antsImage}
\alias{-.antsImage}
\alias{/.antsImage}
\alias{\%\%.antsImage}
\alias{^.antsImage}
\alias{abs.antsImage}
\alias{antsImageArith}
\alias{exp.antsImage}
\alias{log.antsImage}
\alias{mean,antsImage-method}
\alias{sum,antsImage-method}
\title{arith.antsImage}
\usage{
\S4method{mean}{antsImage}(x, mask = logical())

\S4method{sum}{antsImage}(x, mask = logical())

\method{+}{antsImage}(x, y)

\method{-}{antsImage}(x, y)

\method{/}{antsImage}(x, y)

\method{*}{antsImage}(x, y)

\method{^}{antsImage}(x, y)

\method{\%\%}{antsImage}(x, y)

\method{log}{antsImage}(x, ...)

\method{exp}{antsImage}(x)

\method{abs}{antsImage}(x)
}
\arguments{
\item{x}{antsImage}

\item{mask}{antsImage logical mask (optional)}

\item{y}{antsImage or numeric}

\item{...}{Additional arguments passed to underlying R operator}
}
\description{
Atomic arithmetic operators for antsImages
}
\examples{
r16 <- antsImageRead(getANTsRData('r16'), 2)
r64 <- antsImageRead(getANTsRData('r64'), 2)
r16 + r64
r16 + 5
r16 / 10
log(r16, base=10)
}

