% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ants_set_get.R
\name{antsImageGetSet}
\alias{antsImageGetSet}
\alias{antsGetSpacing}
\alias{antsSetSpacing}
\alias{antsGetOrigin}
\alias{antsSetOrigin}
\alias{antsGetDirection}
\alias{antsSetDirection}
\title{antsImageGetSet}
\usage{
antsGetSpacing(x)

antsSetSpacing(x, spacing)

antsGetOrigin(x)

antsSetOrigin(x, origin)

antsGetDirection(x)

antsSetDirection(x, direction)
}
\arguments{
\item{x}{antsImage to access, of dimensionality \code{d}.}

\item{spacing}{numeric vector of length \code{d}.}

\item{origin}{numeric vector of length \code{d}.}

\item{direction}{matrix of size \code{d * d}.}
}
\value{
For \code{get} methods, vector of length \code{d} (origin, spacing) or matrix of size \code{d * d} (direction).
For \code{set} methods, 0 to indicate success.
}
\description{
Get and set methods for image header information
}
\examples{
img <- makeImage(c(5, 5), rnorm(25))
antsGetSpacing(img)
antsSetSpacing(img, c(2.0, 2.0))
antsGetOrigin(img)
antsSetOrigin(img, c(0.5, 0.5))
testthat::expect_error(antsGetSpacing(as.array(img)))
testthat::expect_error(antsSetSpacing(as.array(img), c(2, 2)), "class")
testthat::expect_error(antsSetSpacing(img, c("2", 2)), "numeric")
testthat::expect_error(antsSetSpacing(img, c(3, 3, 3)), "dimensions")

testthat::expect_error(antsGetOrigin(as.array(img)))
testthat::expect_error(antsSetOrigin(as.array(img), c(0.5, 0.5)))
testthat::expect_error(antsSetOrigin(img, c("0.5", 0.5)))
testthat::expect_error(antsSetOrigin(img, c(0.5, 0.5, 0.5)))
img <- makeImage(c(5, 5), rnorm(25))
antsGetDirection(img)
testthat::expect_error(antsGetDirection(as.array(img)))
outimg <- makeImage(c(5, 5), rnorm(25))
antsGetDirection(outimg)
direct <- antsGetDirection(outimg)
antsSetDirection(outimg, direct)
testthat::expect_error(antsSetDirection(as.array(outimg), direct))
testthat::expect_error(antsSetDirection(outimg, as.numeric(direct)))
testthat::expect_error(antsSetDirection(outimg, diag(length(dim(outimg)) + 1)))
}
