% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsImage_class.R
\name{antsImageGetSet}
\alias{antsImageGetSet}
\alias{antsGetSpacing}
\alias{antsSetSpacing}
\alias{antsGetOrigin}
\alias{antsSetOrigin}
\alias{antsGetDirection}
\alias{antsSetDirection}
\title{antsImageGetSet}
\usage{
antsGetSpacing(x)

antsSetSpacing(x, spacing)

antsGetOrigin(x)

antsSetOrigin(x, origin)

antsGetDirection(x)

antsSetDirection(x, direction)
}
\arguments{
\item{x}{antsImage to access, of dimensionality \code{d}.}

\item{spacing}{numeric vector of length \code{d}.}

\item{origin}{numeric vector of length \code{d}.}

\item{direction}{matrix of size \code{d * d}.}
}
\value{
For \code{get} methods, vector of length \code{d} (origin, spacing) or matrix of size \code{d * d} (direction).
For \code{set} methods, 0 to indicate success.
}
\description{
Get and set methods for image header information
}
\examples{
img <- makeImage(c(5,5), rnorm(25))
antsSetSpacing(img, c(2.0, 2.0))
antsSetOrigin(img, c(0.5, 0.5))
}
