% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/antsImage_class.R
\docType{class}
\name{antsImageList-class}
\alias{antsImageList-class}
\title{An S4 class to hold a list of antsImages}
\description{
An S4 class to hold a list of antsImages
}
\section{Slots}{

\describe{
\item{\code{pixeltype}}{}

\item{\code{dimension}}{}

\item{\code{pointer}}{the memory location}
}}

