% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_Operators.R
\docType{methods}
\name{antsImage-math}
\alias{antsImage-math}
\alias{Math,antsImage-method}
\alias{!,antsImage-method}
\title{Math for antsImage Objects}
\usage{
\S4method{Math}{antsImage}(x)

\S4method{!}{antsImage}(x)
}
\arguments{
\item{x}{is an object of class \code{antsImage}.}
}
\description{
Overloaded math for antsImage objects
}
