% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsImageMutualInformation.R
\name{antsImageMutualInformation}
\alias{antsImageMutualInformation}
\title{mutual information between two images}
\usage{
antsImageMutualInformation(in_image1, in_image2)
}
\arguments{
\item{in_image1}{antsImage}

\item{in_image2}{antsImage}
}
\value{
mutual information value
}
\description{
compute mutual information between two images
}
\examples{

fi<-antsImageRead( getANTsRData("r16") ,2)
mi<-antsImageRead( getANTsRData("r64") ,2)
mival<-antsImageMutualInformation(fi,mi)

}
\author{
Brian B. Avants
}
\keyword{image}
\keyword{information}
\keyword{mutual}
\keyword{similarity,}
