% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_Operators.R
\docType{methods}
\name{antsImage-operators}
\alias{antsImage-operators}
\alias{Ops,antsImage,antsImage-method}
\alias{Ops,antsImage,numeric-method}
\alias{Ops,antsImage,missing-method}
\alias{Ops,antsImage,numeric-method}
\alias{Ops,numeric,antsImage-method}
\alias{Ops,antsImage,logical-method}
\alias{Ops,logical,antsImage-method}
\alias{Ops,antsImage,array-method}
\alias{Ops,array,antsImage-method}
\alias{Ops,list,antsImage-method}
\alias{Ops,antsImage,list-method}
\title{Operations for antsImage Objects}
\usage{
\S4method{Ops}{antsImage,antsImage}(e1, e2)

\S4method{Ops}{antsImage,numeric}(e1, e2)

\S4method{Ops}{antsImage,missing}(e1, e2)

\S4method{Ops}{numeric,antsImage}(e1, e2)

\S4method{Ops}{antsImage,logical}(e1, e2)

\S4method{Ops}{logical,antsImage}(e1, e2)

\S4method{Ops}{antsImage,array}(e1, e2)

\S4method{Ops}{array,antsImage}(e1, e2)

\S4method{Ops}{list,antsImage}(e1, e2)

\S4method{Ops}{antsImage,list}(e1, e2)
}
\arguments{
\item{e1}{is an object of class \code{antsImage}.}

\item{e2}{is an object of class \code{antsImage}.}
}
\description{
Overloaded operators for antsImage objects
}
\examples{

img01 <- as.antsImage(array(1:64, c(4,4,4,1)))
img02 <- as.antsImage(array(64:1, c(4,4,4,1)))
is.antsImage(img01 + img02)
is.antsImage(sqrt(2) * img01)
is.antsImage(img02 / pi)
}
\author{
John Muschellli \email{muschellij2@gmail.com}
}
