% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsImagePhysicalSpaceConsistency.R
\name{antsImagePhysicalSpaceConsistency}
\alias{antsImagePhysicalSpaceConsistency}
\title{Check for physical space consistency}
\usage{
antsImagePhysicalSpaceConsistency(img1, img2, coordinate.tolerance = 0.01,
  direction.tolerance = 0.01, data.type = FALSE)
}
\arguments{
\item{img1}{Image object of S4 class \code{antsImage}.}

\item{img2}{Image object of S4 class \code{antsImage}.}

\item{coordinate.tolerance}{floating point error tolerance in origin}

\item{direction.tolerance}{floating point error tolerance in direction matrix}

\item{data.type}{boolean, if TRUE check pixeltype and components, default is FALSE}
}
\value{
Boolean indicating consistency of physical space
}
\description{
Check if two \code{antsImage} objects occupy the same physical space
}
\examples{

img1 <- makeImage(c(10,10), rnorm(100))
img2 <- makeImage(c(10,10), rnorm(100))
check <- antsImagePhysicalSpaceConsistency(img1, img2)


}

