% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsImageRead.R
\name{antsImageRead}
\alias{antsImageRead}
\title{Image Read}
\usage{
antsImageRead(filename, dimension = NULL, pixeltype = "float")
}
\arguments{
\item{filename}{Name of the file to read the image from.}

\item{dimension}{Number of dimensions of the image read. This need not be
the same as the dimensions of the image in the file. Allowed values: 2, 3,
4. If not provided, the dimension is obtained from the image file}

\item{pixeltype}{C++ datatype to be used to represent the pixels read. This
datatype need not be the same as the datatype used in the file. Allowed
values: 'double', 'float' , 'unsigned int' , 'unsigned char'.}
}
\value{
S4 object of Class 'antsImage' -- Success\cr 1 -- Failure
}
\description{
Read an image file into an S4 object of class 'antsImage'.
}
\examples{

fn <- getANTsRData( "r16" )
fi <- antsImageRead( fn )
img <- antsImageRead( fn , dimension = 2  )
img <- antsImageRead( fn , dimension = 2 , 'double' )

}
\seealso{
\code{\link{antsImageWrite}}
}
\author{
Shrinidhi KL
}
