% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_Operators.R
\docType{methods}
\name{antsImage-summary}
\alias{antsImage-summary}
\alias{Summary,antsImage-method}
\title{Summary for antsImage Objects}
\usage{
\S4method{Summary}{antsImage}(x, ..., na.rm = FALSE)
}
\arguments{
\item{x}{is an object of class \code{antsImage}.}
}
\description{
Overloaded Summary for antsImage objects
}
