% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsImageWrite.R
\name{antsImageWrite}
\alias{antsImageWrite}
\title{Image Write}
\usage{
antsImageWrite(image, filename)
}
\arguments{
\item{image}{Image object of S4 class \code{antsImage} to be written.}

\item{filename}{Name of the file to write the image to.}
}
\value{
0 -- Success\cr 1 -- Failure
}
\description{
Write an image object of S4 class \code{antsImage} to a file.
}
\examples{

fn <- getANTsRData( "r16" )
fi <- antsImageRead( fn )
antsImageWrite( fi , tempfile( fileext = ".nii.gz" ) )
antsImageWrite( fi , tempfile( fileext = ".mha" ) )
antsImageWrite( fi , tempfile( fileext = ".nrrd" ) )
antsImageWrite( antsImageClone( fi, "unsigned int" ) ,
  tempfile( fileext = ".jpg" )  )
antsImageWrite( antsImageClone( fi, "float" ) ,
  tempfile( fileext = ".tif" )  )
antsImageWrite( fi, tempfile( fileext = ".mrc" )  )
antsImageWrite( fi, tempfile( fileext = ".hd5" )  )

}
\seealso{
\code{\link{antsImageRead}}
}
\author{
Shrinidhi KL
}
