% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/antsMatrix.R
\docType{class}
\name{antsMatrix-class}
\alias{antsMatrix-class}
\alias{as.data.frame,antsMatrix-method}
\alias{as.list,antsMatrix-method}
\alias{initialize,antsMatrix-method}
\title{An S4 class to hold an antsMatrix imported from ITK types
 C++ type used to represent an element of the matrix pointer
 to the actual image
 C++ type 'itk::image< pixeltype , dimension >::Pointer'}
\usage{
\S4method{initialize}{antsMatrix}(.Object, elementtype)

\S4method{as.data.frame}{antsMatrix}(x)

\S4method{as.list}{antsMatrix}(x)
}
\arguments{
\item{.Object}{input object to convert}

\item{elementtype}{string e.g. "float"}

\item{x}{input object to convert}
}
\description{
An S4 class to hold an antsMatrix imported from ITK types
 C++ type used to represent an element of the matrix pointer
 to the actual image
 C++ type 'itk::image< pixeltype , dimension >::Pointer'
}
\section{Methods (by generic)}{
\itemize{
\item \code{initialize}: 

\item \code{as.data.frame}: 

\item \code{as.list}: 
}}
\section{Slots}{

\describe{
\item{\code{elementtype}}{}
}}

