% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/antsMotionCalculation.R
\name{antsMotionCalculation}
\alias{antsMotionCalculation}
\title{Correct ASL data for motion.}
\usage{
antsMotionCalculation(img, mask=NA, fixed=NA, moreaccurate=1, framewise=1)
}
\arguments{
\item{img}{Name of image file, e.g. 'img.nii.gz'.}

\item{mask}{Name of mask file, e.g. 'mask.nii.gz'.  If not provided, estimated from data.}

\item{fixed}{Fixed image to register all timepoints to.  If not provided, mean image is used.}

\item{moreaccurate}{Level of accuracy desired for motion correction.  Higher is more accurate.}

\item{framewise}{Calculate framewise displacement?}
}
\value{
List containing:
\itemize{
 \item{moco_img}{Motion corrected time-series image.}
 \item{moco_params}{Data frame of translation parameters.}
 \item{moco_avg_img}{Average motion-corrected image.}
 \item{moco_mask}{Mask used to calculate framewise displacement.}
 \item{tsDisplacement}{Time-series displacement image.}
 \item{dvars}{DVARS, derivative of frame-wise intensity changes.}
}
}
\description{
Correct ASL data for motion.
}
\examples{
antsMotionCalculation(getANTsRData('pcasl'))
}
\author{
Benjamin M. Kandel
}

