% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/antsMotionCalculation.R
\name{antsMotionCalculation}
\alias{antsMotionCalculation}
\title{antsMotionCalculation}
\usage{
antsMotionCalculation(img, mask = NA, fixed = NA, moreaccurate = 1,
  txtype = "Affine", framewise = 1)
}
\arguments{
\item{img}{antsImage, usually 4D.}

\item{mask}{mask for image (3D).  If not provided, estimated from data.}

\item{fixed}{Fixed image to register all timepoints to.  If not provided, mean image is used.}

\item{moreaccurate}{Level of accuracy desired for motion correction.  Higher is more accurate.}

\item{txtype}{Type of transform.  One of \code{"Affine"}, \code{"Rigid"}, or
\code{"SyN"}.}

\item{framewise}{Calculate framewise displacement?}
}
\value{
List containing:
\itemize{
 \item{moco_img}{ Motion corrected time-series image.}
 \item{moco_params}{ Data frame of translation parameters.}
 \item{moco_avg_img}{ Average motion-corrected image.}
 \item{moco_mask}{ Mask used to calculate framewise displacement.}
 \item{tsDisplacement}{ Time-series displacement image.}
 \item{fd}{ Time-series mean and max displacements.}
 \item{dvars}{ DVARS, derivative of frame-wise intensity changes.}
}
}
\description{
Correct 4D time-series data for motion.
}
\examples{
set.seed(120)
simimg<-makeImage(rep(5,4), rnorm(5^4))
# for real data, use simimg <- antsImageRead(getANTsRData('pcasl'), 4)
antsMotionCalculation(simimg,moreaccurate=0)
}
\author{
Benjamin M. Kandel
}

