% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/antsRegistration.R
\name{antsRegistration}
\alias{antsRegistration}
\title{Perform registration between two images.}
\usage{
antsRegistration(fixed = NA, moving = NA, typeofTransform = "SyN",
  initialTransform = NA, outprefix = "", mask = NA, ...)
}
\arguments{
\item{fixed}{fixed image to which we register the moving image.}

\item{moving}{moving image to be mapped to fixed space.}

\item{typeofTransform}{Either a one stage rigid/affine mapping or a 2-stage
affine+syn mapping.  Mutual information metric by default. See \code{Details.}
One of \code{Rigid}, \code{Affine}, \code{AffineFast}, \code{SyN}, \code{SyNCC},}

\item{initialTransform}{transforms to prepend}

\item{outprefix}{output will be named with this prefix.}

\item{mask}{mask the registration.}

\item{gradStep}{gradient step size (not for all tx)}

\item{...}{additional options see antsRegistration in ANTs}
}
\value{
outputs a list containing:
\itemize{
  \item{warpedmovout: }{Moving image warped to space of fixed image.}
  \item{warpedfixout: }{Fixed image warped to space of moving image.}
  \item{fwdtransforms: }{Transforms to move from moving to fixed image.}
  \item{invtransforms: }{Transforms to move from fixed to moving image.}
}
Ouptut of 1 indicates failure
}
\description{
Register a pair of images either through the full or simplified interface
to the ANTs registration method.
}
\details{
typeofTransform can be one of:
\itemize{
  \item{"Rigid": }{Rigid transformation: Only rotation and translation.}
  \item{"Affine": }{Affine transformation: Rigid + scaling.}
  \item{"AffineFast": }{Fast version of \code{Affine}.}
  \item{"SyN": }{Symmetric normalization: Affine + deformable transformation,
    with mutual information as optimization metric.}
  \item{"SyNCC": }{SyN, but with cross-correlation as the metric.}
  \item{"SynBold": }{SyN, but optimized for registrations between
    BOLD and T1 images.}
  \item{"SyNAggro": }{SyN, but with more aggressive registration
    (fine-scale matching and more deformation).  Takes more time than \code{SyN}.}
  \item{"TVMSQ": }{time-varying diffeomorphism with mean square metric}
}
}
\examples{
fi <- antsImageRead(getANTsRData("r16") ,2)
mi <- antsImageRead(getANTsRData("r64") ,2)
fi<-resampleImage(fi,c(60,60),1,0)
mi<-resampleImage(mi,c(60,60),1,0) # speed up
mytx <- antsRegistration(fixed=fi, moving=mi, typeofTransform = c('SyN') )
mywarpedimage <- antsApplyTransforms( fixed=fi, moving=mi,
  transformlist=mytx$fwdtransforms )
}
\author{
Shrinidhi KL, Avants BB
}

