% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsrMotionCalculation.R
\name{antsrMotionCalculation}
\alias{antsrMotionCalculation}
\title{antsrMotionCalculation}
\usage{
antsrMotionCalculation(
  img,
  fixed,
  mask,
  typeofTransform = c("Rigid", "QuickRigid", "BOLDRigid", "Affine", "AffineFast",
    "BOLDAffine", "SyN", "SyNOnly"),
  getMotionDescriptors = TRUE,
  verbose = FALSE,
  num_threads = 1,
  seed = NULL,
  ...
)
}
\arguments{
\item{img}{antsImage, usually ND where D=4.}

\item{fixed}{Fixed image to register all timepoints to.  If not provided, mean image is used.}

\item{mask}{mask for image (ND-1).  If not provided, estimated from data.}

\item{typeofTransform}{One of \code{"Affine"}, \code{"Rigid"},
'\code{"BOLDAffine"}, \code{"BOLDRigid"}, \code{"QuickRigid"}.}

\item{getMotionDescriptors}{computes dvars and framewise displacement.  May
take additional memory.}

\item{verbose}{enables verbose output.}

\item{num_threads}{will execute
\code{Sys.setenv(ITK_GLOBAL_DEFAULT_NUMBER_OF_THREADS = num_threads)} before
running to attempt a more reproducible result.  See
\url{https://github.com/ANTsX/ANTs/wiki/antsRegistration-reproducibility-issues}
for discussion.  If \code{NULL}, will not set anything.}

\item{seed}{will execute
\code{Sys.setenv(ANTS_RANDOM_SEED = seed)} before
running to attempt a more reproducible result.  See
\url{https://github.com/ANTsX/ANTs/wiki/antsRegistration-reproducibility-issues}
for discussion.  If \code{NULL}, will not set anything.}

\item{...}{extra parameters passed to antsRegistration}
}
\value{
List containing:
\itemize{
 \item{moco_img}{ Motion corrected time-series image.}
 \item{moco_params}{ Data frame of translation parameters.}
 \item{moco_avg_img}{ Average motion-corrected image.}
 \item{moco_mask}{ Mask used to calculate framewise displacement.}
 \item{fd}{ Time-series mean and max displacements.}
 \item{dvars}{ DVARS, derivative of frame-wise intensity changes.}
}
}
\description{
Correct time-series data for motion.
}
\note{
For reproducible results, you should run
\code{Sys.setenv(ITK_GLOBAL_DEFAULT_NUMBER_OF_THREADS = 1)},
which is what the \code{num_threads = 1} flag will do.
See \url{https://github.com/ANTsX/ANTs/wiki/antsRegistration-reproducibility-issues}
and \url{https://github.com/ANTsX/ANTsR/issues/210#issuecomment-377511054}
for discussion
}
\examples{
Sys.setenv(ITK_GLOBAL_DEFAULT_NUMBER_OF_THREADS = 1)
Sys.setenv(ANTS_RANDOM_SEED = 1)
set.seed(120)
simimg <- makeImage(rep(5, 4), rnorm(5^4))
testthat::expect_equal(mean(simimg), 0.0427369860965759)
res <- antsrMotionCalculation(simimg, seed = 1234)
res2 <- antsrMotionCalculation(simimg, seed = 1234)
res3 <- antsrMotionCalculation(simimg, num_threads = 1, seed = 1)
testthat::expect_equal(res, res2)
# testthat::expect_failure(testthat::expect_equal(res, res3))
print(res$fd)
print(res3$fd)
print(res$moco_params)
print(res3$moco_params)

}
\author{
BB Avants, Benjamin M. Kandel, JT Duda, Jeffrey S. Phillips
}
