% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_data.R
\docType{data}
\name{antsrVersions}
\alias{antsrVersions}
\title{antsrVersions}
\format{
A data frame listing the following variables.
\describe{
  \item{\code{Dependency}}{ Name of software dependency. }
  \item{\code{GitTag}}{ The git tag.  This can also be used to
  trace other dependencies, e.g. the ITK version used by the
  current ANTs version. }
}
}
\description{
A data frame defining the git tag
for the current ANTsR version.  One can also see
antsrVersions from ANTsRCore to get the git tag for
those dependencies.
}
\examples{
data(antsrVersions)
}
\references{
\url{https://github.com/stnava/ANTs}
}
\keyword{datasets}
