% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/antsImage_class.R
\name{antsSetPixels}
\alias{antsSetPixels}
\title{Set a pixel value at an index}
\usage{
antsSetPixels(x, i = NA, j = NA, k = NA, l = NA, value)
}
\arguments{
\item{x}{Image object of S4 class 'antsImage'.}

\item{i}{the slowest moving index to the image}

\item{j}{the next slowest moving index to the image, similar for k ( 2d )}

\item{k}{the next slowest moving index to the image ( 3d )}

\item{l}{the next slowest moving index to the image ( 4d )}

\item{value}{the value to place at this location}
}
\value{
array of pixel values
}
\description{
Set a pixel value at an index in an 'antsImage'.
}
\examples{
img<-makeImage(c(10,10),rnorm(100))
antsSetPixels(img,2,3,value=Inf)
}

