% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/antsSpatialICAfMRI.R
\name{antsSpatialICAfMRI}
\alias{antsSpatialICAfMRI}
\title{Perform spatial ICA on fMRI bold data.}
\usage{
antsSpatialICAfMRI(boldImages, maskImage = NA, numberOfICAComponents = 20,
  normalizeComponentImages = TRUE)
}
\arguments{
\item{boldImages}{a list of 4-D ANTs image fMRI data.}

\item{maskImage}{A 3-D ANTs image defining the region of interest.  This
must be specified.}

\item{numberOfICAComponents}{Number of estimated observers (components).}

\item{normalizeComponentImages}{Boolean to specify whether each component
vector element is normalized to its z-score.}
}
\value{
Output list includes standard ICA matrices from the fastICA
algorithm:

X = pre-processed data matrix

K = pre-whitening matrix that projects data onto the first n.comp principal
components

W = estimated un-mixing matrix (see definition in details)

A = estimated mixing matrix

S = estimated source matrix

and the component images.
}
\description{
Perform spatial ICA on group or individual fMRI data.  Preprocessing should
be performed prior to calling this function (cf preprocessfMRI.R).
}
\examples{
set.seed( 123 )
boldImages <- list()
nvox <- 5*5*5*10
dims <- c(5,5,5,10)
boldImages[[1]] <- makeImage( dims , rnorm( nvox )+500 ) \%>\% iMath("PadImage" , 2 )
boldImages[[2]] <- makeImage( dims , rnorm( nvox )+500 ) \%>\% iMath("PadImage" , 2 )
boldImages[[3]] <- makeImage( dims , rnorm( nvox )+500 ) \%>\% iMath("PadImage" , 2 )
boldImages[[4]] <- makeImage( dims , rnorm( nvox )+500 ) \%>\% iMath("PadImage" , 2 )

cleanBoldImages <- list()
for( i in 1:length( boldImages ) )
  {
  fmri <- preprocessfMRI( boldImages[[i]] )
  if( i == 1 )
    {
    maskImage <- fmri$maskImage
    }
  cleanBoldImages[[i]] <- fmri$cleanBoldImage
  }

icaResults <- antsSpatialICAfMRI( cleanBoldImages, maskImage,
  numberOfICAComponents = 2 )
componentImages <- icaResults$componentImages

# write out the component images
for( i in 1:length( icaResults$componentImages ) )
  {
  componentFileName <- paste0( 'componentImage', i, '.nii.gz' )
  cat( 'Writing ', componentFileName, '.\\n' )
  # antsImageWrite( componentImages[[i]], componentFileName )
 }
}
\author{
Tustison NJ, Avants BB
}

