% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/antsImage_class.R
\name{antsTransformIndexToPhysicalPoint}
\alias{antsTransformIndexToPhysicalPoint}
\title{Get Spatial Point from Index}
\usage{
antsTransformIndexToPhysicalPoint(x, index)
}
\arguments{
\item{x}{Image object of S4 class 'antsImage' to get values from.}

\item{index}{image index}
}
\value{
array of pixel values
}
\description{
Get spatial point from index of an 'antsImage'.
}
\examples{
img<-makeImage(c(10,10),rnorm(100))
pt<-antsTransformIndexToPhysicalPoint(img,c(2,2))
}

