% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsImage_class.R
\name{as.antsImage}
\alias{as.antsImage}
\title{as.antsImage}
\usage{
as.antsImage(object, pixeltype = "float", spacing = as.numeric(seq.int(from
  = 1, by = 0, length.out = length(dim(object)))),
  origin = as.numeric(seq.int(from = 0, by = 0, length.out =
  length(dim(object)))), direction = diag(length(dim(object))),
  components = FALSE, reference = NA, ...)
}
\arguments{
\item{object}{An object}

\item{pixeltype}{a character string e.g. "float"}

\item{spacing}{numeric vector matching image dimensionality e.g. c(1.2,1.2)}

\item{origin}{numeric vector matching image dimensionality e.g. c(0,0)}

\item{direction}{numeric matrix matching image dimensionality eg diag(2)}

\item{components}{number of components per voxel, e.g. 1}

\item{reference}{optional reference antsImage providing all header info}

\item{...}{Extra named arguments passed to FUN}
}
\description{
convert types to antsImage
}
