% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/antsImage_class.R
\docType{methods}
\name{as.antsImage}
\alias{as.antsImage}
\alias{as.antsImage,array-method}
\alias{as.antsImage,matrix-method}
\title{as.antsImage}
\usage{
as.antsImage(object, ...)

\S4method{as.antsImage}{matrix}(object, pixeltype = "float",
  spacing = as.numeric(seq.int(from = 1, by = 0, length.out =
  length(dim(object)))), origin = as.numeric(seq.int(from = 0, by = 0,
  length.out = length(dim(object)))))

\S4method{as.antsImage}{array}(object, pixeltype = "float",
  spacing = as.numeric(seq.int(from = 1, by = 0, length.out =
  length(dim(object)))), origin = as.numeric(seq.int(from = 0, by = 0,
  length.out = length(dim(object)))))
}
\arguments{
\item{object}{An object}

\item{...}{Extra named arguments passed to FUN}

\item{data}{Numeric vector or data.frame}

\item{Fun}{Function. Default function is \code{sum}}
}
\description{
convert types to antsImage
}
\section{Methods (by class)}{
\itemize{
\item \code{matrix}: 

\item \code{array}: 
}}

