% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/antsImage_class.R
\docType{methods}
\name{[,antsImage,NULL,ANY,ANY-method}
\alias{[,antsImage,ANY,ANY,ANY-method}
\alias{[,antsImage,NULL,ANY,ANY-method}
\alias{[,antsImage,NULL,NULL,ANY-method}
\alias{[,antsImage,NULL,numeric,ANY-method}
\alias{[,antsImage,logical,ANY,ANY-method}
\alias{[,antsImage,numeric,NULL,ANY-method}
\alias{[,antsImage,numeric,numeric,ANY-method}
\alias{[<-,antsImage,NULL,ANY,ANY-method}
\alias{[<-,antsImage,NULL,antsRegion,ANY-method}
\alias{[<-,antsImage,array,ANY,ANY-method}
\alias{[<-,antsImage,array,antsRegion,ANY-method}
\alias{[<-,antsImage,list,ANY,ANY-method}
\alias{[<-,antsImage,logical,ANY,ANY-method}
\alias{[<-,antsImage,logical,antsRegion,ANY-method}
\alias{[<-,antsImage,matrix,ANY,ANY-method}
\alias{[<-,antsImage,matrix,antsRegion,ANY-method}
\alias{as.antsImage}
\alias{as.antsImage,array-method}
\alias{as.antsImage,matrix-method}
\title{as.antsImage}
\usage{
\S4method{[}{antsImage,`NULL`,ANY,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{antsImage,logical,ANY,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{antsImage,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{antsImage,`NULL`,`NULL`,ANY}(x, i, j, k = NA, l = NA, ...,
  drop)

\S4method{[}{antsImage,numeric,numeric,ANY}(x, i, j, k = NA, l = NA, ...,
  drop)

\S4method{[}{antsImage,numeric,`NULL`,ANY}(x, i, j, k = NA, l = NA, ...,
  drop)

\S4method{[}{antsImage,`NULL`,numeric,ANY}(x, i, j, k = NA, l = NA, ...,
  drop)

\S4method{[}{antsImage,`NULL`,ANY,ANY}(x, i, j, ...) <- value

\S4method{[}{antsImage,logical,ANY,ANY}(x, i, j, ...) <- value

\S4method{[}{antsImage,array,ANY,ANY}(x, i, j, ...) <- value

\S4method{[}{antsImage,matrix,ANY,ANY}(x, i, j, ...) <- value

\S4method{[}{antsImage,list,ANY,ANY}(x, i, j, ...) <- value

\S4method{[}{antsImage,`NULL`,antsRegion,ANY}(x, i, j, ...) <- value

\S4method{[}{antsImage,logical,antsRegion,ANY}(x, i, j, ...) <- value

\S4method{[}{antsImage,array,antsRegion,ANY}(x, i, j, ...) <- value

\S4method{[}{antsImage,matrix,antsRegion,ANY}(x, i, j, ...) <- value

as.antsImage(object, ...)

\S4method{as.antsImage}{matrix}(object, pixeltype = "float",
  spacing = as.numeric(seq.int(from = 1, by = 0, length.out =
  length(dim(object)))), origin = as.numeric(seq.int(from = 0, by = 0,
  length.out = length(dim(object)))))

\S4method{as.antsImage}{array}(object, pixeltype = "float",
  spacing = as.numeric(seq.int(from = 1, by = 0, length.out =
  length(dim(object)))), origin = as.numeric(seq.int(from = 0, by = 0,
  length.out = length(dim(object)))))
}
\arguments{
\item{...}{Extra named arguments passed to FUN}

\item{object}{An object}

\item{pixeltype}{a character string e.g. "float"}

\item{spacing}{numeric vector matching image dimensionality e.g. c(1.2,1.2)}

\item{origin}{numeric vector matching image dimensionality e.g. c(0,0)}
}
\description{
convert types to antsImage
}
\section{Methods (by class)}{
\itemize{
\item \code{x = antsImage,i = NULL,j = ANY,drop = ANY}: 

\item \code{x = antsImage,i = logical,j = ANY,drop = ANY}: 

\item \code{x = antsImage,i = ANY,j = ANY,drop = ANY}: 

\item \code{x = antsImage,i = NULL,j = NULL,drop = ANY}: 

\item \code{x = antsImage,i = numeric,j = numeric,drop = ANY}: 

\item \code{x = antsImage,i = numeric,j = NULL,drop = ANY}: 

\item \code{x = antsImage,i = NULL,j = numeric,drop = ANY}: 

\item \code{x = antsImage,i = NULL,j = ANY,value = ANY}: 

\item \code{x = antsImage,i = logical,j = ANY,value = ANY}: 

\item \code{x = antsImage,i = array,j = ANY,value = ANY}: 

\item \code{x = antsImage,i = matrix,j = ANY,value = ANY}: 

\item \code{x = antsImage,i = list,j = ANY,value = ANY}: 

\item \code{x = antsImage,i = NULL,j = antsRegion,value = ANY}: 

\item \code{x = antsImage,i = logical,j = antsRegion,value = ANY}: 

\item \code{x = antsImage,i = array,j = antsRegion,value = ANY}: 

\item \code{x = antsImage,i = matrix,j = antsRegion,value = ANY}: 

\item \code{matrix}: 

\item \code{array}: 
}}

