% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsMatrix.R
\name{as.antsMatrix}
\alias{as.antsMatrix}
\title{as.antsMatrix}
\usage{
as.antsMatrix(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{other parameters}

\item{elementtype}{e.g. "float" or "double"}
}
\description{
convert types to an antsMatrix
}
\examples{
as.antsMatrix(matrix(rnorm(10), nrow=2))
}
