% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/antsMatrix.R
\docType{methods}
\name{as.antsMatrix}
\alias{as.antsMatrix}
\alias{as.antsMatrix,data.frame-method}
\alias{as.antsMatrix,list-method}
\title{as.antsMatrix}
\usage{
as.antsMatrix(object, ...)

\S4method{as.antsMatrix}{list}(object, elementtype)

\S4method{as.antsMatrix}{data.frame}(object, elementtype)
}
\arguments{
\item{object}{An object}

\item{...}{Extra named arguments passed to FUN}

\item{elementtype}{e.g. "float" or "double"}

\item{data}{Numeric vector or data.frame}
}
\description{
convert types to an antsMatrix
}
\section{Methods (by class)}{
\itemize{
\item \code{list}: 

\item \code{data.frame}: 
}}

