% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/antsMatrix.R
\docType{methods}
\name{as.antsMatrix}
\alias{as.antsMatrix}
\alias{as.antsMatrix,data.frame-method}
\alias{as.antsMatrix,list-method}
\alias{as.antsMatrix,matrix-method}
\title{as.antsMatrix}
\usage{
as.antsMatrix(object, ...)

\S4method{as.antsMatrix}{list}(object, elementtype = "float")

\S4method{as.antsMatrix}{data.frame}(object, elementtype = "float")

\S4method{as.antsMatrix}{matrix}(object, elementtype = "float")
}
\arguments{
\item{object}{An object}

\item{...}{other parameters}

\item{elementtype}{e.g. "float" or "double"}
}
\description{
convert types to an antsMatrix
}
\section{Methods (by class)}{
\itemize{
\item \code{list}: 

\item \code{data.frame}: 

\item \code{matrix}: 
}}
\examples{
as.antsMatrix(matrix(rnorm(10), nrow=2))
}

