% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/antsMatrix.R
\docType{methods}
\name{as.antsMatrix}
\alias{as.antsMatrix}
\alias{as.antsMatrix,data.frame-method}
\alias{as.antsMatrix,list-method}
\alias{as.list,antsMatrix-method}
\title{as.antsMatrix}
\usage{
as.antsMatrix(object, ...)

\S4method{as.antsMatrix}{list}(object, elementtype)

\S4method{as.antsMatrix}{data.frame}(object, elementtype)

\S4method{as.list}{antsMatrix}(x)
}
\arguments{
\item{object}{An object}

\item{...}{Extra named arguments passed to FUN}

\item{data}{Numeric vector or data.frame}

\item{Fun}{Function. Default function is \code{sum}}
}
\description{
convert types to an antsMatrix
}
\section{Methods (by class)}{
\itemize{
\item \code{list}: 

\item \code{data.frame}: 

\item \code{antsMatrix}: 
}}

