% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsImage_class.R, R/antsMatrix.R
\docType{methods}
\name{dim,antsImage-method}
\alias{dim,antsImage-method}
\alias{is.na,antsImage-method}
\alias{as.numeric,antsImage-method}
\alias{as.matrix,antsImage-method}
\alias{as.array,antsImage-method}
\alias{as.data.frame,antsMatrix-method}
\alias{as.matrix,antsMatrix-method}
\alias{as.list,antsMatrix-method}
\title{Coerce antsImage objects to array}
\usage{
\S4method{dim}{antsImage}(x)

\S4method{is.na}{antsImage}(x)

\S4method{as.numeric}{antsImage}(x, mask = logical(),
  region = new("antsRegion", index = integer(), size = integer()))

\S4method{as.matrix}{antsImage}(x, mask = logical(),
  region = new("antsRegion", index = integer(), size = integer()))

\S4method{as.array}{antsImage}(x, mask = logical(),
  region = new("antsRegion", index = integer(), size = integer()))

\S4method{as.data.frame}{antsMatrix}(x)

\S4method{as.matrix}{antsMatrix}(x)

\S4method{as.list}{antsMatrix}(x)
}
\arguments{
\item{x}{object of class \code{\link{antsImage}} or \code{\link{antsMatrix}}}
}
\description{
Converts antsImage, antsImage object to different data types
}
