% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsImage_class.R, R/antsMatrix.R
\name{as.array}
\alias{as.array}
\alias{dim,antsImage-method}
\alias{as.numeric,antsImage-method}
\alias{as.matrix,antsImage-method}
\alias{as.matrix.antsImage}
\alias{as.array,antsImage-method}
\alias{as.array.antsImage}
\alias{as.character.antsImage}
\alias{as.factor.antsImage}
\alias{factor.antsImage}
\alias{as.data.frame,antsMatrix-method}
\alias{as.data.frame.antsMatrix}
\alias{as.matrix,antsMatrix-method}
\alias{as.list,antsMatrix-method}
\title{Coerce antsImage objects to array}
\usage{
\S4method{dim}{antsImage}(x)

\S4method{as.numeric}{antsImage}(
  x,
  mask = logical(),
  region = new("antsRegion", index = integer(), size = integer())
)

\S4method{as.matrix}{antsImage}(
  x,
  mask = logical(),
  region = new("antsRegion", index = integer(), size = integer())
)

\method{as.matrix}{antsImage}(
  x,
  ...,
  mask = logical(),
  region = new("antsRegion", index = integer(), size = integer())
)

\S4method{as.array}{antsImage}(
  x,
  ...,
  mask = logical(),
  region = new("antsRegion", index = integer(), size = integer())
)

\method{as.array}{antsImage}(
  x,
  ...,
  mask = logical(),
  region = new("antsRegion", index = integer(), size = integer())
)

\method{as.character}{antsImage}(
  x,
  ...,
  mask = logical(),
  region = new("antsRegion", index = integer(), size = integer())
)

\method{as.factor}{antsImage}(
  x,
  ...,
  mask = logical(),
  region = new("antsRegion", index = integer(), size = integer())
)

\method{factor}{antsImage}(
  x,
  ...,
  mask = logical(),
  region = new("antsRegion", index = integer(), size = integer())
)

\S4method{as.data.frame}{antsMatrix}(x)

\method{as.data.frame}{antsMatrix}(x, row.names = NULL, optional = FALSE, ...)

\S4method{as.matrix}{antsMatrix}(x, ...)

\S4method{as.list}{antsMatrix}(x, ...)
}
\arguments{
\item{x}{object of class \code{antsImage} or \code{antsMatrix}}

\item{mask}{a logical vector/array or binary antsImage object}

\item{region}{a \code{antsRegion} object}

\item{...}{additional arguments passed to functions}

\item{row.names}{NULL or a character vector giving the row names for the
data frame.}

\item{optional}{passsed to \code{\link{as.data.frame}}}
}
\description{
Converts antsImage, antsImage object to different data types
}
\examples{
outimg<-makeImage( c(2,10) , rnorm(20))
as.numeric(outimg)
as.numeric(outimg, mask = outimg > 1)
 testthat::expect_error(as.numeric(outimg, mask = outimg))

outimg<-makeImage( c(2,10) , rnorm(20))
as.matrix(outimg)
as.matrix(outimg, mask = outimg > 1)
testthat::expect_error(as.matrix(outimg, mask = outimg) )
outimg<-makeImage( c(2,10,2) , rnorm(40))
testthat::expect_error(as.matrix(outimg) )
outimg<-makeImage( c(2,10) , rnorm(20))
as.matrix(outimg)
as.matrix(outimg, mask = outimg > 1)
outimg<-makeImage( c(2,10,2) , rnorm(40))
testthat::expect_error(as.matrix(outimg) )
img <- antsImageRead( getANTsRData( "r16" ) )
img[img > 5] = 0
sort(unique(as.character(img)))
factor(img)
table(img, img)
}
