% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/aslPerfusion.R
\name{aslPerfusion}
\alias{aslPerfusion}
\title{ASL-based Perfusion from PASL, CASL or pCASL.}
\usage{
aslPerfusion(asl, maskThresh = 0.75, moreaccurate = 1, dorobust = 0.92,
  m0 = NA, skip = 20, mask = NA, interpolation = "linear",
  checkmeansignal = 100, moco_results = NULL, regweights = NULL,
  useDenoiser = NA, useBayesian = 0, verbose = FALSE, ncompcor = 0,
  N3 = FALSE)
}
\arguments{
\item{asl}{input asl image}

\item{maskThresh}{for estimating a brain mask}

\item{moreaccurate}{zero, one or two with the last being the most accurate}

\item{dorobust}{robustness parameter, lower value keeps more data}

\item{m0}{known M0 if any}

\item{skip}{stride to speed up robust regression weight estimates}

\item{mask}{known brain mask}

\item{interpolation}{e.g. 'linear'}

\item{checkmeansignal}{throw out volumes with mean lower than this thresh}

\item{moco_results}{passes prior motion results so moco does not get repeated}

\item{regweights}{known temporal weights on regression solution, if any}

\item{useDenoiser}{use the aslDenoiser if this value is a range gt than zero}

\item{useBayesian}{strength of bayesian prior}

\item{verbose}{bool}

\item{ncompcor}{number of compcor parameters}

\item{N3}{bool correct target image before motion corr}
}
\value{
output a list of relevant objects
}
\description{
This function will estimate perfusion from an ASL time series using a
(robust) regression approach. It will do motion correction, compcorr, factor
out nuisance variables and use regression to estimate the perfusion itself.
It will mask the image too based on a simple procedure ( should fix this in
the future by allowing the user to optionally pass a mask in from the
outside ).  WARNING: the function will estimate the m0 image from the mean
of the control tags assuming that the data is acquired T C T C as is most of
JJ's data.  Quantitative CBF can be obtained by mutiplying the output of
this function by a scalar.
}
\examples{
# image available at http://files.figshare.com/1701182/PEDS012_20131101.zip
# fn<-'PEDS012_20131101_pcasl_1.nii.gz'
# asl<-antsImageRead(fn,4)
set.seed(1)
nvox <- 5*5*5*10
dims <- c(5,5,5,10)
asl <- makeImage( dims , rnorm( nvox )+500 ) \%>\% iMath("PadImage" , 2 )
aslmean <- getAverageOfTimeSeries( asl )
aslmask <- getMask( aslmean , 0.001 , Inf )
aslmat<-timeseries2matrix( asl, aslmask )
for ( i in 1:10 ) aslmat[,i*2]<-aslmat[,i*2]*2
asl<-matrix2timeseries( asl, aslmask, aslmat )
pcasl.processing <- aslPerfusion( asl, moreaccurate=1, dorobust=0 )
pcasl.processing <- aslPerfusion( asl, moreaccurate=1, ncompcor=2 )
# allow some rejection
pcasl.processing <- aslPerfusion( asl, moreaccurate=1, dorobust=0.925 )
}
\author{
Avants BB
}

