% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Atropos.R
\name{Atropos}
\alias{Atropos}
\title{FMM Segmentation}
\usage{
Atropos(d, a, x, i = "kmeans[3]", m = "[0.2,1x1]", c = "[5,0]",
  priorweight = 0.5, ...)
}
\arguments{
\item{d}{This option forces the image to
be treated as a specified-dimensional image. If not specified, Atropos tries
to infer the dimensionality from the first input image. Allowed values: 2,
3, 4.}

\item{a}{One or more scalar images is specified for segmentation using the
-a/--intensity-image option. For segmentation scenarios with no prior
information, the first scalar image encountered on the command line is used
to order labelings such that the class with the smallest intensity signature
is class '1' through class 'N' which represents the voxels with the largest
intensity values. The optional adaptive smoothing weight parameter is
applicable only when using prior label or probability images. This scalar
parameter is to be specified between [0,1] which smooths each labeled region
separately and modulates the intensity measurement at each voxel in each
intensity image between the original intensity and its smoothed counterpart.
The smoothness parameters are governed by the -b/--bspline option.}

\item{x}{mask image}

\item{i}{initialization see Atropos in ANTs for full set of options}

\item{m}{mrf parameters see Atropos for format of these}

\item{c}{convergence parameters see Atropos for format of these}

\item{priorweight}{usually 0, 0.25 or 0.5}

\item{...}{more parameters, see Atropos help in ANTs}
}
\value{
0 -- Success\cr 1 -- Failure
}
\description{
A finite mixture modeling (FMM) segmentation approach with possibilities for
specifying prior constraints. These prior constraints include the
specification of a prior label image, prior probability images (one for each
class), and/or an MRF prior to enforce spatial smoothing of the labels.
Similar algorithms include FAST and SPM.  Atropos can also perform
multivariate segmentation if you pass a list of images in: e.g.
a=c(img1,img2).
}
\examples{
img<-antsImageRead( getANTsRData('r16') ,2)
mask<-getMask(img)
segs1<-Atropos( d = 2, a = img, m = '[0.2,1x1]',c = '[5,0]',  i = 'kmeans[3]', x = mask)
}
\author{
Shrinidhi KL, B Avants
}

