% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atropos.R
\name{atropos}
\alias{atropos}
\title{FMM Segmentation}
\usage{
atropos(a, x, i = "KMeans[3]", m = "[0.2,1x1]", c = "[5,0]",
  priorweight = 0.25, ...)
}
\arguments{
\item{a}{One or more scalar images to segment. If priors are not used, the intensities
of the first image are used to order the classes in the segmentation output, from lowest
to highest intensity. Otherwise the order of the classes is dictated by the order of the prior
images.}

\item{x}{mask image.}

\item{i}{initialization usually \code{KMeans[N]} for N classes or a list of N prior probability images.
See Atropos in ANTs for full set of options.}

\item{m}{mrf parameters as a string, usually \code{"[smoothingFactor,radius]"} where \code{smoothingFactor}
determines the amount of smoothing and radius determines the MRF neighborhood, as an ANTs style neighborhood
vector eg "1x1x1" for a 3D image. The radius must match the dimensionality of the image, eg 1x1 for 2D and
The default in ANTs is \code{smoothingFactor=0.3} and \code{radius=1}. See Atropos for more options.}

\item{c}{convergence parameters, \code{"[numberOfIterations,convergenceThreshold]"}. A threshold of 0 runs
the full \code{numberOfIterations}, otherwise Atropos tests convergence by comparing the mean maximum posterior
probability over the whole region of interest defined by the mask \code{x}.}

\item{priorweight}{usually 0 (priors used for initialization only), 0.25 or 0.5.}

\item{...}{more parameters, see Atropos help in ANTs}
}
\value{
0 -- Success\cr 1 -- Failure
}
\description{
A finite mixture modeling (FMM) segmentation approach with possibilities for
specifying prior constraints. These prior constraints include the
specification of a prior label image, prior probability images (one for each
class), and/or an MRF prior to enforce spatial smoothing of the labels.
Similar algorithms include FAST and SPM.  atropos can also perform
multivariate segmentation if you pass a list of images in: e.g.
\code{a=c(img1,img2)}.
}
\examples{

img <- antsImageRead( getANTsRData("r16") , 2 )
img <- resampleImage( img, c(64,64), 1, 0 )
mask <- getMask(img)
segs1 <- atropos( d = 2, a = img, m = '[0.2,1x1]',
   c = '[2,0]',  i = 'kmeans[3]', x = mask )

# Use probabilities from k-means seg as priors

segs2 <- atropos( d = 2, a = img, m = '[0.2,1x1]',
   c = '[2,0]',  i = segs1$probabilityimages, x = mask )

feats <- list(img, iMath(img,"Laplacian"), iMath(img,"Grad") )
segs3 <- atropos( d = 2, a = feats, m = '[0.2,1x1]',
   c = '[2,0]',  i = segs1$probabilityimages, x = mask )

}
\author{
Shrinidhi KL, B Avants
}

