% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/basicInPaint.R
\name{basicInPaint}
\alias{basicInPaint}
\title{Inpaints missing imaging data from boundary data}
\usage{
basicInPaint(img, paintMask, speedimage = NA, its = 0, gparam = 0.05)
}
\arguments{
\item{img}{antsImage to be approximated / painted}

\item{paintMask}{painting mask with values 1 or
values 1 and 2 - if there is a 2 then it will learn
from label 1 to paint label 2.  should cover the brain.}

\item{speedimage}{- larger means faster/better}

\item{its}{- iterations of graddescent}

\item{gparam}{- graddescent param e.g. 0.05}
}
\value{
inpainted image
}
\description{
Smooths data along the boundary into the missing region.
}
\examples{
set.seed(123)
fi<-abs(replicate(100, rnorm(100)))
fi[1:10,]<-fi[,1:10]<-fi[91:100,]<-fi[,91:100]<-0
mask<-fi
mask[ mask > 0 ]<-1
mask2<-mask
mask2[11:20,11:20]<-2
mask<-as.antsImage( mask2 , 'float' )
fi<-as.antsImage( fi , 'float' )
fi<-smoothImage(fi,3)
painted<-basicInPaint(fi,mask)
\dontrun{
lmask<-antsImageRead('brainmask.nii.gz',2)
limg<-antsImageRead('r16slice_lesion.nii.gz',2)
mm<-basicInPaint(limg,lmask)
plot(mm)
mm2<-basicInPaint(limg,lmask,its=10,gparam=0.05)
plot(mm2)
}
}
\author{
Brian B. Avants
}
\keyword{inpainting}
\keyword{template}

