% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bayesianCBF.R
\name{bayesianCBF}
\alias{bayesianCBF}
\title{Uses probabilistic segmentation to constrain pcasl-based cbf computation.}
\usage{
bayesianCBF(pcasl, seg, tissuelist, myPriorStrength = 30,
  useDataDrivenMask = 3, denoisingComponents = 1:8,
  robustnessvalue = 0.95, localweights = F, priorBetas = NA)
}
\arguments{
\item{pcasl}{img antsImage for cbf}

\item{seg}{segmentation image, should cover the brain.}

\item{tissuelist}{a list containing antsImages eg list(prob1,...,probN)}

\item{myPriorStrength}{- e.g 30}

\item{useDataDrivenMask}{- morphology parameters e.g. 3}

\item{denoisingComponents}{- data-driven denoising parameters}

\item{robustnessvalue}{- value (e.g. 0.95) that throws away time points}

\item{localweights}{Use estimate of voxel-wise reliability to inform prior
weight?}

\item{priorBetas}{prior betas for each tissue and predictor}
}
\value{
estimated cbf image
}
\description{
Employs a robust regression approach to learn the relationship between a
sample image and a list of images that are mapped to the same space as the
sample image.
}
\examples{
\dontrun{
set.seed(123)
# see fMRIANTs github repository for data and I/O suggestions
}
}
\author{
Brian Beaumont Avants and Benjamin M. Kandel
}
\keyword{asl,}
\keyword{bayesian}
\keyword{blood}
\keyword{cerebral}
\keyword{flow,}

