% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bold_correlation_matrix}
\alias{bold_correlation_matrix}
\title{BOLD Correlation Matrix}
\format{
A data frame square correlation matrix listed by AAL 
anatomical column names.
}
\usage{
bold_correlation_matrix
}
\description{
A data frame with an example square BOLD correlation matrix.
}
\examples{
data(bold_correlation_matrix)
}
\references{
\url{http://en.wikipedia.org/wiki/Automated_Anatomical_Labeling}
}
\keyword{datasets}
