% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscaleSVDxpts.R
\name{calculate_simlr_energy}
\alias{calculate_simlr_energy}
\title{Calculate SIMLR Similarity Energy}
\usage{
calculate_simlr_energy(V, X, U, energy_type)
}
\arguments{
\item{V}{A candidate loading matrix [p x k] to evaluate.}

\item{X}{The data matrix for the current modality [n x p].}

\item{U}{The shared basis matrix [n x k].}

\item{energy_type}{A string specifying the similarity objective.}
}
\value{
A single numeric value for the similarity energy. The sign is adjusted
  such that the value should always be minimized.
}
\description{
This dispatcher calculates the similarity/reconstruction part of the
objective function for use in an optimization routine.
}
