% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineNuisancePredictors.R
\name{combineNuisancePredictors}
\alias{combineNuisancePredictors}
\title{Combine and reduce dimensionality of nuisance predictors.}
\usage{
combineNuisancePredictors(
  inmat,
  target,
  globalpredictors = NA,
  maxpreds = 4,
  localpredictors = NA,
  method = "cv",
  k = 5,
  covariates = NA,
  ordered = FALSE
)
}
\arguments{
\item{inmat}{Input predictor matrix.}

\item{target}{Target outcome matrix.}

\item{globalpredictors}{Global predictors of size \code{nrow(inmat)}
by \code{n}, where \code{n} is the number of global predictors.}

\item{maxpreds}{Maximum number of predictors to output.}

\item{localpredictors}{Local predictor array of size \code{nrow(inmat)}
by \code{ncol(inmat)} by \code{m}, where \code{m} is the number of
local predictors.}

\item{method}{Method of selecting noisy voxels.  One of 'svd' or 'cv'.
See \code{Details}.}

\item{k}{Number of cross-validation folds.}

\item{covariates}{Covariates to be considered when assessing prediction
of \code{target}.}

\item{ordered}{Can the predictors be assumed to be ordered from most
important to least important, as in output from PCA?
Computation is much faster if so.}
}
\value{
Array of size \code{nrow(aslmat)} by \code{npreds},
containing a timeseries of all the nuisance predictors.
If \code{localpredictors} is not NA, array is of size \code{nrow(aslmat)}
by \code{ncol(aslmat)} by \code{npreds}.
}
\description{
Combine and select nuisance predictors to maximize
correlation between \code{inmat} and \code{target}.
}
\examples{
set.seed(120)
simimg <- makeImage( c(10,10,10,20) , rnorm( 10*10*10*20)+1 )
moco <- antsMotionCalculation( simimg , moreaccurate=0)
# for real data use below
# moco <- antsMotionCalculation(getANTsRData("pcasl"))
aslmat <- timeseries2matrix(moco$moco_img, moco$moco_mask)
tc <- rep(c(0.5, -0.5), length.out=nrow(aslmat))
noise <- getASLNoisePredictors(aslmat, tc, 0.5 )
noise.sub <- combineNuisancePredictors(aslmat, tc, noise, 2)
}
\author{
Benjamin M. Kandel, Brian B. Avants
}
