% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/compcor.R
\name{compcor}
\alias{compcor}
\title{Simple compcor function.}
\usage{
compcor(fmri, ncompcor = 4, variance_extreme = 0.975, mask = NA,
  fastsvd = FALSE, useimagemath = FALSE, randomSamples = 1,
  returnv = FALSE, returnhighvarmatinds = FALSE, highvarmatinds = NA)
}
\arguments{
\item{returnv}{returns the singular vectors in v-space}

\item{mat}{input fmri image or matrix}

\item{returnhighvarmat}{returns the high variance matrix on which svd is
performed}
}
\value{
dataframe is output
}
\description{
Compcors the input matrix using SVD and returns the result.
}
\examples{
mat <- matrix( rnorm(50000) ,ncol=500)
compcorrdf<-compcor( mat )
}
\author{
Avants BB
}

