% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compcor.R
\name{compcor}
\alias{compcor}
\title{Compcors the input matrix using SVD and returns the result.}
\usage{
compcor(fmri, ncompcor = 4, variance_extreme = 0.975, mask = NULL,
  randomSamples = 1, returnv = FALSE, returnhighvarmat = FALSE,
  returnhighvarmatinds = FALSE, highvarmatinds = NA, scale = TRUE)
}
\arguments{
\item{fmri}{input fmri image or matrix}

\item{ncompcor}{n compcor vectors}

\item{variance_extreme}{high variance threshold e.g 0.95 for 95 percent}

\item{mask}{optional mask for image}

\item{randomSamples}{take this many random samples to speed things up}

\item{returnv}{return the spatial vectors}

\item{returnhighvarmat}{bool to return the high variance matrix}

\item{returnhighvarmatinds}{bool to return the high variance matrix indices}

\item{highvarmatinds}{index list}

\item{scale}{scale the matrix of high variance voxels, default FALSE. note
that you may get slightly different results by scaling the input matrix
before passing into this function.}
}
\value{
dataframe of nuisance predictors is output
}
\description{
Compcors the input matrix using SVD and returns the result.
}
\examples{

mat <- matrix( rnorm(50000) ,ncol=500)
compcorrdf<-compcor( mat )

}
\author{
Avants BB
}
