% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessRestingBOLD.R
\name{computeDVARS}
\alias{computeDVARS}
\title{computeDVARS}
\usage{
computeDVARS(boldMatrix)
}
\arguments{
\item{boldMatrix}{matrix of bold signal}
}
\value{
DVARS vector.
}
\description{
compute the DVARS quality control metric
}
\examples{

mat <- matrix(c(0, 1, 2, 0, 0, 1, 2, 2, 2), ncol = 3)
dv <- computeDVARS(mat)

}
\author{
Tustison NJ, Avants BB
}
