% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/CreateJacobianDeterminantImage.R
\name{CreateJacobianDeterminantImage}
\alias{CreateJacobianDeterminantImage}
\title{CreateJacobianDeterminantImage}
\usage{
CreateJacobianDeterminantImage(dim, tx, doLog = 0)
}
\arguments{
\item{doLog}{return the log jacobian}

\item{imageDimension2|3}{Number of dimensions of the input tx}

\item{deformationField}{transformation file name}
}
\value{
jacobianImage
}
\description{
Compute the jacobian determinant from a transformation file
}
\examples{
fi<-antsImageRead( getANTsRData('r16') ,2)
mi<-antsImageRead( getANTsRData('r64') ,2)
mytx<-antsRegistration(fixed=fi , moving=mi, typeofTransform = c("SyN") )
jac<-CreateJacobianDeterminantImage(2,mytx$fwdtransforms[[1]],1)
plot(jac)
}
\author{
BB Avants
}

