% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createJacobianDeterminantImage.R
\name{createJacobianDeterminantImage}
\alias{createJacobianDeterminantImage}
\title{createJacobianDeterminantImage}
\usage{
createJacobianDeterminantImage(domainImg, tx, doLog = FALSE,
  geom = FALSE)
}
\arguments{
\item{domainImg}{image that defines transformation domain}

\item{tx}{deformation transformation file name}

\item{doLog}{return the log jacobian}

\item{geom}{use the geometric jacobian calculation (boolean)}
}
\value{
jacobianImage
}
\description{
Compute the jacobian determinant from a transformation file
}
\examples{
fi<-antsImageRead( getANTsRData("r16") ,2)
mi<-antsImageRead( getANTsRData("r64") ,2)
fi<-resampleImage(fi,c(128,128),1,0)
mi<-resampleImage(mi,c(128,128),1,0)
mytx<-antsRegistration(fixed=fi , moving=mi, typeofTransform = c("SyN") )
jac<-createJacobianDeterminantImage(fi,mytx$fwdtransforms[[1]],1)
# plot(jac)
}
\author{
BB Avants
}
