% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/CreateJacobianDeterminantImage.R
\name{CreateJacobianDeterminantImage}
\alias{CreateJacobianDeterminantImage}
\title{CreateJacobianDeterminantImage}
\usage{
CreateJacobianDeterminantImage(dim, tx, doLog = 0)
}
\arguments{
\item{dim}{Number of dimensions of the input tx}

\item{tx}{deformation transformation file name}

\item{doLog}{return the log jacobian}
}
\value{
jacobianImage
}
\description{
Compute the jacobian determinant from a transformation file
}
\examples{
fi<-antsImageRead( getANTsRData('r16') ,2)
mi<-antsImageRead( getANTsRData('r64') ,2)
fi<-resampleImage(fi,c(128,128),1,0)
mi<-resampleImage(mi,c(128,128),1,0)
mytx<-antsRegistration(fixed=fi , moving=mi, typeofTransform = c('SyN') )
jac<-CreateJacobianDeterminantImage(2,mytx$fwdtransforms[[1]],1)
# plot(jac)
}
\author{
BB Avants
}

