% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/crossvalidatedR2.R
\name{crossvalidatedR2}
\alias{crossvalidatedR2}
\title{Cross-Validated R^2 value}
\usage{
crossvalidatedR2(x, y, ngroups=5, covariates=NA, fast=F)
}
\arguments{
\item{x}{Input predictor matrix.}

\item{y}{Target dependent variable.}

\item{ngroups}{Number of cross-validation folds to use or the fold labels themselves, equal to the length of y.  e.g. c(1,1,1,2,2,2...)}

\item{covariates}{Covariate predictors.}

\item{fast}{Use low-level \code{lm.fit} instead of \code{lm}.  Much faster, but less error checking.}
}
\value{
Matrix of size \code{ngroups} by \code{ncol(x)}, which each row corresponding to one fold and the columns corresponding to the R2 values for each predictor.
}
\description{
Computes an R^2 value for predicting an outcome measure using a k-fold cross-validation scheme.
}
\examples{
set.seed(300)
ncol <- 30
nrow <- 20
covariate <- sin((1:nrow)*2*pi/nrow)
x <- matrix(rep(NA, nrow*ncol), nrow=nrow)
xsig <- seq(0,1,length.out=nrow)
y <- xsig + covariate + rnorm(nrow, sd=0.5)
for(i in 1:ncol){
  x[, i] <- xsig + rnorm(nrow, sd=i/ncol)
}
r2 <- crossvalidatedR2(x, y, covariates=covariate)
}
\author{
Brian B Avants, Benjamin M. Kandel
}

