\name{DesikanKillianyTourville}
\alias{DesikanKillianyTourville}
\docType{data}
\title{DesikanKillianyTourville}
\description{
A data frame label numbers and names for the neuroanatomical DesikanKillianyTourville labels.
}
\usage{data(DesikanKillianyTourville)}
\format{
  A data frame listing the following variables.
  \describe{
    \item{\code{label_num}}{ Numerical label value. }
    \item{\code{label_name}}{ Shorthand anatomical value. }
    \item{\code{tissue}}{ Shorthand tissue value. }
    \item{\code{contralateral}}{ contralateral anatomical value. }
    \item{\code{lobe}}{ which lobe: temporal, parietal, occipital, frontal. }
    \item{\code{hemisphere}}{ left or right hemisphere. }
  }
}
\references{
\url{http://www.ncbi.nlm.nih.gov/pubmed/23227001}
}
\examples{
data(DesikanKillianyTourville)
}
\keyword{datasets}
