% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscaleSVDxpts.R
\name{divide_by_column_sum}
\alias{divide_by_column_sum}
\title{Divide Each Column by Its Sum}
\usage{
divide_by_column_sum(m)
}
\arguments{
\item{m}{A numeric matrix}
}
\value{
A matrix with each column divided by its sum
}
\description{
Divide Each Column by Its Sum
}
