% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscaleSVDxpts.R
\name{.calculate_regression_gradient}
\alias{.calculate_regression_gradient}
\title{Calculate Gradient for Basic Regression Error}
\usage{
.calculate_regression_gradient(X, U, V)
}
\value{
A matrix [p x k] representing a **descent direction** for the energy E.
}
\description{
Calculate Gradient for Basic Regression Error
}
\keyword{internal}
