% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigSeg.R
\name{eigSeg}
\alias{eigSeg}
\title{Segmentation for eigenanatomy.}
\usage{
eigSeg(mask = NA, imgList = NA, applySegmentationToImages = FALSE,
  cthresh = 0, smooth = 1)
}
\arguments{
\item{mask}{D-dimensional mask > 0 defining segmentation region.}

\item{imgList}{list containing antsImages or a matrix matching the mask.}

\item{applySegmentationToImages}{boolean determines if original image list
is modified by the segmentation.}

\item{cthresh}{throw away isolated clusters smaller than this value}

\item{smooth}{smooth the input data first by this value}
}
\value{
segmentation image.
}
\description{
Segment a mask into regions based on the max value in an image list.  At a
given voxel the segmentation label will contain the index to the image that
has the largest value.  If the 3rd image has the greatest value, the
segmentation label will be 3 at that voxel.
}
\examples{

mylist<-list( antsImageRead( getANTsRData("r16") ),
  antsImageRead( getANTsRData("r27") ),
  antsImageRead( getANTsRData("r85") ) )
myseg<-eigSeg( getMask( mylist[[1]] ) , mylist )
mat=imageListToMatrix( mylist, getMask( mylist[[1]] ) )
myseg<-eigSeg( getMask( mylist[[1]] ) , mat )

}
\author{
Avants BB
}

