% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscaleSVDxpts.R
\name{estimate_rank_by_permutation_rv}
\alias{estimate_rank_by_permutation_rv}
\title{Estimate Joint Rank via Cross-View RV Analysis}
\usage{
estimate_rank_by_permutation_rv(
  mat_list,
  n_permutations = 20,
  var_explained_threshold = 0.99,
  return_max = FALSE
)
}
\arguments{
\item{mat_list}{A list of numeric matrices [subjects x features].}

\item{n_permutations}{The number of permutations to create the null model.
If set to 0, the function will use the elbow detection method instead.}

\item{var_explained_threshold}{The variance threshold to determine the
upper bound on k.}

\item{return_max}{boolean just return the max likely rank from an individual matrix}
}
\value{
A list containing `optimal_k`, a results tibble, and a plot.
}
\description{
Estimates the shared rank of a multi-view dataset. If permutations
  are used, it finds the rank that maximizes the signal-to-noise ratio against
  a permuted null model. If `n_permutations = 0`, it finds the "elbow" of the
  real data's signal curve.
}
