% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/exemplarInpainting.R
\name{exemplarInpainting}
\alias{exemplarInpainting}
\title{Uses example images to inpaint or approximate an existing image.}
\usage{
approximg<-exemplarInpainting( img, paintMask, list(img1,img2),
  featureRadius, sharpen, feather )
}
\arguments{
\item{img}{antsImage to be approximated / painted}

\item{paintMask}{painting mask with values 1 or
values 1 and 2 - if there is a 2 then it will learn
from label 1 to paint label 2.  should cover the brain.}

\item{imageList}{a list containing antsImages}

\item{featureRadius}{- radius of image neighborhood e.g. 2}

\item{scaleInpaintIntensity}{- brighter or darker painted voxels,
default of 0 sets this parameter automatically}

\item{sharpen}{- sharpen the approximated image}

\item{feather}{- value (e.g. 1) that helps feather the mask for smooth blending}

\item{predalgorithm}{- string svm or lm}

\item{debug}{- TRUE or FALSE}
}
\value{
inpainted image
}
\description{
Employs a robust regression approach to learn the relationship between a sample image and a list of images that are mapped to the same space as the sample image.  The regression uses data from an image neighborhood.
}
\examples{
set.seed(123)
fi<-abs(replicate(100, rnorm(100)))
fi[1:10,]<-fi[,1:10]<-fi[91:100,]<-fi[,91:100]<-0
mask<-fi
mask[ mask > 0 ]<-1
mask2<-mask
mask2[11:20,11:20]<-2
mask<-as.antsImage( mask , 'float' )
fi<-as.antsImage( fi , 'float' )
SmoothImage(2,fi,3,fi)
mo<-as.antsImage( replicate(100, rnorm(100)) , 'float' )
mo2<-as.antsImage( replicate(100, rnorm(100)) , 'float' )
ilist<-list(mo,mo2)
painted<-exemplarInpainting(fi,mask,ilist)
mask2<-as.antsImage( mask2 , 'float' )
painted2<-exemplarInpainting(fi,mask2,ilist)
# just use 1 image, so no regression is performed
painted3<-exemplarInpainting(fi,mask2, list(ilist[[1]]))
}
\author{
Brian B. Avants
}
\keyword{inpainting}
\keyword{template}

