% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cropImage.R
\name{extractSlice}
\alias{extractSlice}
\title{extract a slice from an image}
\usage{
extractSlice(image, slice, direction)
}
\arguments{
\item{image}{antsImage to crop}

\item{slice}{which slice, integer}

\item{direction}{which axis, integer}
}
\value{
antsImage of dimension - 1
}
\description{
extract a slice from an image and return an image of dimensionality, d-1
}
\examples{

fi <- makeImage( c(10,10,10), rnorm(1000) )
slice <- extractSlice( fi, 1, 1 )

}
\author{
Brian B. Avants, Nicholas J. Tustison
}
\keyword{extract}

