% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fastMarchingExtension.R
\name{fastMarchingExtension}
\alias{fastMarchingExtension}
\title{Fast Marching Extension filter will extend an intensity or label image
from a known region into the unknown region along a geodesic path}
\usage{
fastMarchingExtension(speedImage, labelImage, valueImage)
}
\arguments{
\item{speedImage}{defines the cost or distance function}

\item{labelImage}{defines the known (value 1) and unknown (value 2) regions}

\item{valueImage}{these values are extended into the unknown regions}
}
\value{
antsImage
}
\description{
This in an implementation of ITK's FastMarchingExtensionImageFilter - From http://www.itk.org/Doxygen/html/classitk_1_1FastMarchingExtensionImageFilter.html - Fast marching can be used to extend auxiliary variables smoothly from the zero level set. Starting from an initial position on the front, this class simultaneously calculate the signed distance and extend a set of auxiliary values. This class is templated over the level set image type, the auxiliary variable type and the number of auxiliary variables to extended. The initial front is specified by two containers: one containing the known points and one containing the trial points. The auxiliary variables on the front are represented by two auxiliary variable containers: one containing the value of the variables at the know points and on containing the value of the variables at the trail points. Implemenation of this class is based on Chapter 11 of "Level Set Methods and Fast Marching Methods", J.A. Sethian, Cambridge Press, Second edition, 1999.
}
\author{
Duda, JT
}

