% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastMarchingExtension.R
\name{fastMarchingExtension}
\alias{fastMarchingExtension}
\title{Fast Marching Extension filter will extend an intensity or label image
from a known region into the unknown region along a geodesic path}
\usage{
fastMarchingExtension(speedImage, labelImage, valueImage)
}
\arguments{
\item{speedImage}{defines the cost or distance function}

\item{labelImage}{defines the known (value 1) and unknown (value 2) regions}

\item{valueImage}{these values are extended into the unknown regions}
}
\value{
antsImage
}
\description{
This in an implementation of ITK's FastMarchingExtensionImageFilter - From http://www.itk.org/Doxygen/html/classitk_1_1FastMarchingExtensionImageFilter.html - Fast marching can be used to extend auxiliary variables smoothly from the zero level set. Starting from an initial position on the front, this class simultaneously calculate the signed distance and extend a set of auxiliary values. Implemenation of this class is based on Chapter 11 of "Level Set Methods and Fast Marching Methods", J.A. Sethian, Cambridge Press, Second edition, 1999.
}
\examples{

img = makeImage( c( 7, 7 ) , 0 )
img[3:5,3:5]=1:9
mask = getMask( img )
mask[ mask == 0 ] = 2
speed = smoothImage( mask, 1 )
extendedImg = fastMarchingExtension( speed, mask, img )

}
\author{
Duda, JT
}

