% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fastMarchingExtension.R
\name{fastMarchingExtension}
\alias{fastMarchingExtension}
\title{Fast Marching Extension}
\usage{
fastMarchingExtension(speedImage, labelImage, valueImage)
}
\arguments{
\item{speedImage}{}

\item{labelImage}{}

\item{valueImage}{}
}
\value{
antsImage
}
\description{
Fast marching extension
}
\author{
Duda, JT
}

