% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fastwhiten.R
\name{fastwhiten}
\alias{fastwhiten}
\title{Simple fastwhitening function.}
\usage{
fastwhiten(x, mynu = NA)
}
\arguments{
\item{mat}{input matrix}
}
\value{
matrix is output
}
\description{
whitens the input matrix using a fast SVD and returns the result.  can
select number of principal components to use.  might want to scale the
matrix before input.
}
\examples{
\dontrun{
mat<-replicate(400, rnorm(600))
print(mean(abs(cor(mat))))
wmat<-fastwhiten( scale( mat ) )
print(mean(abs(cor(wmat))))
wmat<-fastwhiten( scale( mat ), 20 )
print(mean(abs(cor(wmat))))
}
}
\author{
Avants BB
}

