% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitTransformToPairedPoints.R
\name{fitTimeVaryingTransformToPointSets}
\alias{fitTimeVaryingTransformToPointSets}
\title{fitTimeVaryingTransformToPointSets}
\usage{
fitTimeVaryingTransformToPointSets(
  pointSets,
  timePoints = NULL,
  numberOfIntegrationPoints = NULL,
  domainImage = NULL,
  numberOfFittingLevels = 4,
  meshSize = 1,
  splineOrder = 3,
  displacementWeights = NULL,
  numberOfCompositions = 10,
  compositionStepSize = 0.5,
  sigma = 0,
  convergenceThreshold = 0,
  rasterizePoints = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{pointSets}{Corresponding points across sets specified in physical space as a
\code{n x d} matrix where \code{n} is the number of points and \code{d} is the
dimensionality.}

\item{timePoints}{Set of scalar values, one for each point-set designating its time
position in the velocity flow.  If not set, it defaults to equal spacing between 0
and 1.}

\item{numberOfIntegrationPoints}{Time-varying velocity field parameter.  Needs to
be equal to or greater than the number of point sets.  If not specified, it
defaults to the number of point sets.}

\item{domainImage}{defines physical domain of the nonlinear transforms.}

\item{numberOfFittingLevels}{integer specifying the number of fitting levels}

\item{meshSize}{scalar or vector defining the mesh size at the initial fitting level}

\item{splineOrder}{spline order of the B-spline object.  Default = 3.}

\item{displacementWeights}{vector defining the individual weighting of the corresponding
scattered data value.  Default = NULL meaning all displacements are
weighted the same.}

\item{numberOfCompositions}{total number of compositions.}

\item{compositionStepSize}{scalar multiplication factor.}

\item{sigma}{gaussian smoothing sigma (in mm).}

\item{convergenceThreshold}{Composition-based convergence parameter for the diffeomorphic
transforms using a window size of 10 values.}

\item{rasterizePoints}{Use nearest neighbor rasterization of points for estimating update
field (potential speed-up).}

\item{verbose}{Print progress to the screen.}
}
\value{
object containing ANTsR transform, error, and scale (or displacement field)
}
\description{
Estimate a time-varying transform from corresponding point sets (> 2).
}
\examples{
fixed <- matrix( c( 50, 50, 200, 50, 50, 200 ), ncol = 2, byrow = TRUE )
moving <- matrix( c( 50, 50, 50, 200, 200, 200 ), ncol = 2, byrow = TRUE )

# Affine transform
xfrm <- fitTransformToPairedPoints( moving, fixed, transformType = "Affine", regularization = 0 )
params <- getAntsrTransformParameters( xfrm )

# Rigid transform
xfrm <- fitTransformToPairedPoints( moving, fixed, transformType = "Rigid", regularization = 0 )
params <- getAntsrTransformParameters( xfrm )

# Similarity transform
xfrm <- fitTransformToPairedPoints( moving, fixed, transformType = "Similarity", regularization = 0 )
params <- getAntsrTransformParameters( xfrm )

# B-spline transform
domainImage <- antsImageRead( getANTsRData( "r16" ) )
xfrm <- fitTransformToPairedPoints( moving, fixed, transformType = "Bspline", domainImage = domainImage, numberOfFittingLevels = 5 )

# Diffeo transform
domainImage <- antsImageRead( getANTsRData( "r16" ) )
xfrm <- fitTransformToPairedPoints( moving, fixed, transformType = "Diffeo", domainImage = domainImage, numberOfFittingLevels = 6 )

# SyN transform
domainImage <- antsImageRead( getANTsRData( "r16" ) )
xfrm <- fitTransformToPairedPoints( moving, fixed, transformType = "SyN", domainImage = domainImage, numberOfFittingLevels = 6, numberOfCompositions = 10, compositionStepSize = 0.01 )
}
\author{
B Avants, N Tustison
}
