% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequencyFilterfMRI.R
\name{frequencyFilterfMRI}
\alias{frequencyFilterfMRI}
\title{Band pass filtering for BOLD image.}
\usage{
frequencyFilterfMRI(boldmat, tr, freqLo = 0.01, freqHi = 0.1,
  opt = "butt")
}
\arguments{
\item{boldmat}{Time series matrix for bold image}

\item{tr}{The sequence's TR value , typically 3 or 4.}

\item{freqLo}{The lower frequency limit, e.g. 0.01 in band-pass
filter}

\item{freqHi}{The higher frequency limit, e.g. 0.1 in band-pass
filter}

\item{opt}{one of 'trig','butt','stl' Type of filter to use: butterworth,
trigonometric, stl.}
}
\value{
output is the filtered time series.
}
\description{
This function works for a BOLD time-series data
}
\examples{

fmat<-replicate(1000, rnorm(200))
k<-1
tr = 4
for ( ftype in c("butt","stl","trig") ) {
  myres <- frequencyFilterfMRI( fmat, tr = tr, freqLo = 0.01, freqHi = 0.05, opt = ftype )
  comparemat <- cbind( fmat[,k], myres[,k] )
  plot( ts( comparemat ), main=ftype )
  Sys.sleep(0.3)
  temp = spectrum( ts(fmat[,k], frequency=1/tr ) )
  plot( temp$freq, temp$spec, type='l' )
  temp = spectrum( ts(myres[,k], frequency=1/tr ) )
  plot( temp$freq, temp$spec, type='l' )
}

}
\author{
Avants BB
}

