% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/geoSeg.R
\name{geoSeg}
\alias{geoSeg}
\title{brain segmentation based on geometric priors}
\usage{
geoSeg(img, brainmask, priors, seginit, vesselopt = "none", vesselk = 2,
  gradStep = 1.25, mrfval = 0.1, atroposits = 10, jacw = NA,
  beta = 0.9)
}
\arguments{
\item{img}{input image or list of images (multiple features) where 1st image
would typically be the primary constrast}

\item{brainmask}{binary image}

\item{priors}{spatial priors, assume first is csf, second is gm, third is wm}

\item{seginit}{a previously computed segmentation which should have the structure of \code{atropos} or \code{kmeansSegmentation} output}

\item{vesselopt}{one of bright, dark or none}

\item{vesselk}{integer for kmeans vessel-based processing}

\item{gradStep}{scalar for registration}

\item{mrfval}{e.g. 0.05 or 0.1}

\item{atroposits}{e.g. 5 iterations}

\item{jacw}{precomputed diffeo jacobian}

\item{beta}{for sigma transformation ( thksig output variable )}
}
\value{
list of segmentation result images
}
\description{
uses topological constraints to enhance accuracy of brain segmentation
}
\examples{
img = antsImageRead( getANTsRData("simple") ,2)
img = n3BiasFieldCorrection( img , 4 )
img = n3BiasFieldCorrection( img , 2 )
bmk = getMask( img )
segs <- kmeansSegmentation( img, 3, bmk )
priors = segs$probabilityimages
seg = geoSeg( img, bmk, priors )
}
\author{
Brian B. Avants
}

