% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getANTsRData.R
\name{getANTsRData}
\alias{getANTsRData}
\title{getANTsRData}
\usage{
getANTsRData(
  fileid,
  usefixedlocation = FALSE,
  verbose = FALSE,
  method = ifelse(Sys.info()["sysname"] == "Linux", "wget", "auto"),
  quiet = FALSE
)
}
\arguments{
\item{fileid}{one of the permitted file ids or pass "show" to list all
valid possibilities. Note that most require internet access to download.}

\item{usefixedlocation}{directory to which you download files}

\item{verbose}{optional boolean}

\item{method}{Method to be used for downloading files,
passed to \code{\link{download.file}}}

\item{quiet}{If \code{TRUE}, suppress status messages 
(if any), and the progress bar.}
}
\value{
filename string
}
\description{
Downloads antsr test data
}
\examples{

fi <- getANTsRData( "r16" )

}
\author{
Avants BB
}
