% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getANTsRData.R
\name{getANTsRData}
\alias{getANTsRData}
\title{getANTsRData}
\usage{
getANTsRData(fileid, usefixedlocation = FALSE, verbose = FALSE)
}
\arguments{
\item{fileid}{one of the permitted file ids or pass "show" to list all
valid possibilities. Note that most require internet access to download.}

\item{usefixedlocation}{directory to which you download files}

\item{verbose}{optional boolean}
}
\value{
filename string
}
\description{
Downloads antsr test data
}
\examples{

fi <- getANTsRData( "r16" )

}
\author{
Avants BB
}
