% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsrTransform_class.R
\name{getAntsrTransformFixedParameters}
\alias{getAntsrTransformFixedParameters}
\title{getAntsrTransformParameters}
\usage{
getAntsrTransformFixedParameters(transform)
}
\arguments{
\item{transform}{antsrTransform}
}
\value{
array of parameters'
}
\description{
Get parameters of transform
}
\examples{
tx = new("antsrTransform")
params = getAntsrTransformFixedParameters(tx)
}

