% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getAverageOfTimeSeries.R
\name{getAverageOfTimeSeries}
\alias{getAverageOfTimeSeries}
\title{Returns a 3D average of a 4D time series.}
\usage{
getAverageOfTimeSeries(timeseriesimage)
}
\arguments{
\item{img}{input 4D image}
}
\value{
3D ants image is output
}
\description{
Averages the 3D sub-volumes of a 4D input ants image.
}
\examples{
img<-as.antsImage( array(data = rep(0,10^4), dim = c(10,10,10,10) ) )
avg<-getAverageOfTimeSeries( img )
print(dim(avg))
avg<-getAverageOfTimeSeries( avg )
}
\author{
Avants BB
}

