% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getfMRInuisanceVariables.R
\name{getfMRInuisanceVariables}
\alias{getfMRInuisanceVariables}
\title{Extract generic fMRI nuisance variables for ASL or BOLD}
\usage{
getfMRInuisanceVariables(fmri, maskThresh = 500, moreaccurate = 1,
  mask = NULL)
}
\arguments{
\item{fmri}{input antsImage or filename}

\item{maskThresh}{will use this intensity threshold to estimate a mask
otherwise will use the mask passed in}

\item{moreaccurate}{zero, one or two with increasing accuracy/computation}

\item{mask}{binary image masking the intput image, precedent over mask thresh}
}
\value{
outputs list described above.
}
\description{
Will motion correct, run compcorr and estimate global signal. Outputs a list
with the time series data matrix (time by voxels), motion and other nuisance
variables, global signal (for BOLD or ASL), the mask and the average time
series image.  Meant to be used before filterfMRIforNetworkAnalysis or any
other results-oriented processing.
}
\examples{

\dontrun{
if (!exists("fn") ) fn<-getANTsRData("pcasl")
pcasl<-antsImageRead( fn )
aslmean<-getAverageOfTimeSeries( pcasl )
aslmask<-getMask(aslmean)
ee<-getfMRInuisanceVariables( pcasl, mask = aslmask ,
        moreaccurate=F )
}

}
\author{
Avants BB
}
