% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiResRandomForestRegression.R
\name{getMultiResFeatureMatrix}
\alias{getMultiResFeatureMatrix}
\title{build multiple resolution predictor matrix from feature images}
\usage{
getMultiResFeatureMatrix(x, labelmask, rad = NA,
  multiResSchedule = c(0), nsamples = 10)
}
\arguments{
\item{x}{a list of feature images}

\item{labelmask}{the mask defines the image space for the associated feature
and increases the number of predictors.  more labels means more
predictors.  a different mask may be used for each feature, if desired, in
which case this should be a list of masks where the length of the list
is equal to the number of features.}

\item{rad}{vector of dimensionality d define nhood radius}

\item{multiResSchedule}{a vector of smoothing values}

\item{nsamples}{defines the number of samples to take for each label.}
}
\value{
mat a matrix of predictors with n samples rows
}
\description{
Represents feature images as a neighborhood across scales. each subject
gets a label image and feature list.  these labels/features should be
the same type for all subjects.  e.g each subject has a k-label image
where the labels cover the same anatomy and the feature images are the same.
for each label in a mask, produce a multi-resolution neighborhood
sampling from the data within the label, for a given feature. do this for
each feature. so, 2 labels will double the width of the predictor matrix.
an additional scale parameter or feature will increase the width.  more
samples will increase the number of rows in the predictor matrix. the labels
allow one to collect predictors side-by-side from different parts of an
image-based feature set.  future work will allow other covariates.
}
\examples{

img <- antsImageRead( getANTsRData("r16"))
seg <- kmeansSegmentation( img, 3 )$segmentation
flist = list( img, img \%>\% iMath("Grad") )
featMat <- getMultiResFeatureMatrix( flist, seg, rad=c(1,1),
  multiResSchedule = c( 2, 1), nsamples=10 )
}
\author{
Avants BB, Tustison NJ
}
