% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getTemplateCoordinates.R
\name{getTemplateCoordinates}
\alias{getTemplateCoordinates}
\title{Define an anatomical coordinate system in a new image based on a template}
\usage{
getTemplateCoordinates(imagePairToBeLabeled, templatePairWithLabels,
  labelnames = NA, outprefix = NA, convertToTal = FALSE)
}
\arguments{
\item{imagePairToBeLabeled}{e.g. the template image and the activation map}

\item{templatePairWithLabels}{e..g the mni image and brodmann label set}

\item{labelnames}{a list of names for the labels}

\item{outprefix}{if output to a file, provide file prefix}

\item{convertToTal}{bool, return talairach coordinates}
}
\value{
The output point is in approximate template space.
}
\description{
This function will provide a mapping that labels an input image and its
blobs.
}
\details{
Uses Matthew Brett's mni2tal to get the final Talairach coordinates from MNI
space.

This is a standard approach but it's not very accurate.
}
\examples{
\dontrun{
#
# ch2bet is available in chris rorden's mricron
#  but you can do something with any other image
#  e.g. a statistical image
#
  tem<-antsImageRead("ch2bet.nii.gz",3)
  clust <- antsImageClone( tem )
  clust[ tem < 80 ]<- 0
  clust[ tem > 90 ]<- 0
  clust[  tem > 80 & tem < 90 ]<- 1
  ImageMath(3, clust,"ME",clust,1)  # erosion
  clust <- labelClusters( clust , minClusterSize=30, minThresh=1, maxThresh=1)
  if ( ! exists("mymni") ) {
  # try getANTsRData if you have www access
    mymni<-list( antsImageRead(getANTsRData("mni"),3),
                 antsImageRead(getANTsRData("mnib"),3),
                 antsImageRead(getANTsRData("mnia"),3) )
  }
  template_cluster_pair<-list(tem,clust)
  gcoords<-getTemplateCoordinates( template_cluster_pair ,
      mymni , convertToTal = TRUE )
# output will be like
# > gcoords$templatepoints
#     x   y   z t label Brodmann                 AAL
# 1  -12  13  -3 0     1        0           Caudate_R
# 2   13  16   5 0     2        0           Caudate_L
#
# can also use a white matter label set ...
#
}
}
\author{
Avants, BB
}
\keyword{Coordinates}
\keyword{Talairach,}
\keyword{Template,}

