% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscaleSVDxpts.R
\name{gradient_invariant_orthogonality_defect_diag_zero}
\alias{gradient_invariant_orthogonality_defect_diag_zero}
\title{Gradient of the Invariant Orthogonality Defect Measure}
\usage{
gradient_invariant_orthogonality_defect_diag_zero(A)
}
\arguments{
\item{A}{A numeric matrix.}
}
\value{
A numeric matrix representing the gradient of the orthogonality defect measure.
}
\description{
This function computes the gradient of the orthogonality defect measure with respect to the input matrix `A`.
The gradient is useful for optimization techniques that require gradient information. The gradient will be zero
for matrices where `AtA` equals the diagonal matrix `D`.
}
\examples{
# library(salad)
# A <- matrix(runif(20), nrow = 10, ncol = 2)
# gradient_invariant_orthogonality_defect_diag_zero(A)
}
