% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icawhiten.R
\name{icawhiten}
\alias{icawhiten}
\title{Simple icawhitening function.}
\usage{
icawhiten(Xin, n.comp, verbose = FALSE)
}
\arguments{
\item{Xin}{input matrix}

\item{n.comp}{number of components on which to project}

\item{verbose}{bool}
}
\value{
matrix is output
}
\description{
Icawhitens the input matrix using SVD and returns the result.
}
\examples{

mat <- matrix(c(rep(1,100),rep(0,200)),ncol=50)
wmat<-icawhiten( mat, 2 )

}
\author{
Avants BB, fastICA package (see CRAN)
}
