% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/image2ClusterImages.R
\name{image2ClusterImages}
\alias{image2ClusterImages}
\title{Converts an image to several independent images.}
\usage{
image2ClusterImages(x, minClusterSize = 50, minThresh = 1e-06,
  maxThresh = 1)
}
\arguments{
\item{img}{input antsImage e.g. a statistical map}
}
\value{
the original image broken into a list of cluster images is the
output
}
\description{
Produces a unique image for each connected component 1 through N of size >
minClusterSize
}
\examples{
\dontrun{
imageclusterlist<-image2ClusterImages( img )
}
}
\author{
Avants BB
}

