% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/imageFileNames2ImageList.R
\name{imageFileNames2ImageList}
\alias{imageFileNames2ImageList}
\title{Simple imageFileNames2ImageListing function.}
\usage{
imageFileNames2ImageList(x, dim)
}
\arguments{
\item{x}{input file name list}

\item{dim}{img dimensionality (optional)}
}
\value{
a list like this : mylist<-list( img1, img2 , etcetera ) is the
output ... img* are antsImages
}
\description{
ImageFileNames2ImageLists converts the input list of file names to a list
containing antsImages.
}
\examples{
\dontrun{
  # gglb<-paste('gmView1vec*.nii.gz',sep='')
  # gfnl<-imageFileNames2ImageList( list.files(path=statdir,
  #   pattern = glob2rx(gglb),full.names = T,recursive = T) )
}
}
\author{
Avants BB
}

