% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageFileNames2ImageList.R
\name{imageFileNames2ImageList}
\alias{imageFileNames2ImageList}
\title{Simple imageFileNames2ImageListing function.}
\usage{
imageFileNames2ImageList(x)
}
\arguments{
\item{x}{input file name list}
}
\value{
a list containing the images  this : mylist<-list( img1, img2 , etcetera ) is the
}
\description{
ImageFileNames2ImageLists converts the input list of file names to a list
containing antsImages.
}
\examples{
dir  <- paste(tempdir(), .Platform$file.sep, sep='')
img <- makeImage(c(4,4))
for (ii in 1:6) {
  antsImageWrite(img, paste(dir, 'image', ii, '.nii.gz', sep=''))
}
imagenames <- list.files(dir, glob2rx('*.nii.gz'), full.names=TRUE)
images <- imageFileNames2ImageList(imagenames)

}
\author{
Avants BB, Kandel BM
}

