% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageListToMatrix.R
\name{imageListToMatrix}
\alias{imageListToMatrix}
\title{Read Images into a Matrix}
\usage{
imageListToMatrix(imageList, mask, sigma = NA, epsilon = 0)
}
\arguments{
\item{imageList}{A character vector containing a list of image files to
read, in order - these are image objects, not file names.}

\item{mask}{An \code{antsImage} containing a binary mask, voxels in the mask
are placed in the matrix. If not provided, estimated from first image in list.
If the mask is a different size than the image, the images will be downsampled
and smoothed to the size of the mask.}

\item{sigma}{smoothing operation in physical space.  See \code{smoothImage}.}

\item{epsilon}{threshold value determining what is included in the mask}
}
\value{
A matrix containing the masked data, the result of calling
\code{as.numeric(image, mask)} on each input image.
}
\description{
Read images into rows of a matrix.
}
\examples{
 img <- antsImageRead(getANTsRData('r16') )
 imglist <- list()
 nvox <- dim(img)[1] * dim(img)[2]
 nsubj <- 50
 for(ii in 1:nsubj){
   imglist[[ ii ]] <- img + rnorm(nvox, sd=mean(img[img!=0]))
 }
 mask <- getMask(img) \%>\% resampleImage( c( 2,2 ) )
 imgmat <- imageListToMatrix(imglist, mask)

}
\seealso{
\code{\link{matrixToImages}, \link{getMask}}
}
\author{
Cook PA, Avants B, Kandel BM
}
