% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ImageMath.R
\name{ImageMath}
\alias{ImageMath}
\title{ImageMath}
\usage{
ImageMath(...)
}
\arguments{
\item{imageDimension2|3}{Number of dimensions of the input image}

\item{outputImage}{result image}

\item{operator}{Must be one of the operators listed by ImageMath()}

\item{inputImage}{inputImage}

\item{otherParams}{see ImageMath() for details for each option}
}
\value{
0 -- Success\cr 1 -- Failure
}
\description{
Perform math-operations on the given image
}
\examples{
fi<-antsImageRead( getANTsRData('r16') ,2)
mask<-getMask(fi)
ImageMath( 2 , fi , 'GD', fi , 1 )  # gray matter dilation
ImageMath( 2 , mask , 'Neg', mask )  # negate
ImageMath( 2 , mask , 'D', mask )  # distance transform
plot(mask)
}
\author{
Shrinidhi KL
}

