% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagesToMatrix.R
\name{imagesToMatrix}
\alias{imagesToMatrix}
\title{Read Images into a Matrix}
\usage{
imagesToMatrix(imageList, mask)
}
\arguments{
\item{imageList}{A character vector containing a list of image files to
read, in order.}

\item{mask}{An \code{antsImage} containing a binary mask, voxels in the mask
are placed in the matrix.}
}
\value{
A matrix containing the masked data, the result of calling
\code{as.numeric(image, mask)} on each input image.
}
\description{
Read images into rows of a matrix, given a mask - much faster for large
datasets as it is based on C++ implementations.
}
\examples{


# make some simulated images and convert them to a matrix

n <- 2
tdir <- tempdir()
for (i in 1:n) {
  simimg <- as.antsImage(replicate(64, rnorm(64)))
  antsImageWrite(simimg, tempfile(fileext = ".mha"))
}
imageList <- list.files(tdir, pattern = ".mha", full.names = TRUE)
mask <- getMask(antsImageRead(imageList[1]))
mat <- imagesToMatrix(imageList, mask)
print(dim(mat))

}
\seealso{
\code{\link{matrixToImages}, \link{getMask}}
}
\author{
Cook PA, Avants BB (C++ version)
}
