% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/iMath.R
\name{iMath}
\alias{iMath}
\title{iMath}
\usage{
iMath(img, operation, param = NA, ...)
}
\arguments{
\item{img}{input object, usually antsImage}

\item{operation}{a character string e.g. "GetLargestComponent" ... the
special case of "GetOperations" or "GetOperationsFull" will return
a list of operations and brief description.
Some operations may not be valid (WIP), but most are.}

\item{param}{... additional parameters}

\item{...}{further parameter options}
}
\description{
Perform various (often mathematical) operations on the input image/s.
Additional parameters should be specific for each operation.  See the
the full iMath in ANTs, on which this function is based.
}
\examples{
fi<-antsImageRead( getANTsRData("r16") , 2 )
mask<-getMask( fi )
op1<-iMath( fi , "GD" , 1 )  # gray matter dilation by 1 voxel
op2<-iMath( mask , "D" )  # distance transform
op3<-iMath( 0 , "GetOperations" )  # list all ops

if ( usePkg("magrittr") ) { # string ops together
  lapgd <- fi \%>\% iMath("Laplacian",1)  \%>\% iMath("GD",3)
}
}
\author{
JT Duda
}

