% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/iMath.R
\name{iMath}
\alias{iMath}
\title{iMath}
\usage{
iMath(img, operation, ...)
}
\arguments{
\item{img}{input object, usually antsImage}

\item{operation}{a character string e.g. "GetLargestComponent" ... the
special case of "GetOperations" will return a list of operations and brief
description.  Some operations may not be valid (WIP), but most are.}

\item{...}{additional parameters}
}
\description{
Perform various (often mathematical) operations on the input image.
Additional parameters should be specific for each operation.  See the
the full ImageMath in ANTs, on which this function is based.
}
\examples{
fi<-antsImageRead( getANTsRData("r16") , 2 )
mask<-getMask( fi )
op1<-iMath( fi , "GD" , 1 )  # gray matter dilation by 1 voxel
op2<-iMath( mask , "Neg" )  # negate
op3<-iMath( mask , "D" )  # distance transform
ops<-iMath( mask , "GetOperations" )  # list all ops
}
\author{
BB Avants
}

