% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscaleSVDxpts.R
\name{initializeSyMLR}
\alias{initializeSyMLR}
\title{Initialize SyMLR}
\usage{
initializeSyMLR(voxmats, k, jointReduction = FALSE, zeroUpper = FALSE,
  uAlgorithm = "pca")
}
\arguments{
\item{voxmats}{list that contains the named matrices.}

\item{k}{rank of U matrix}

\item{jointReduction}{boolean determining whether one or length of list bases are}

\item{zeroUpper}{boolean determining whether upper triangular part of
initialization is zeroed out}

\item{uAlgorithm}{either \code{"pca"} (default) or \code{"ica"}}
}
\value{
A single matrix or list of matrices
}
\description{
Four initialization approaches for SyMLR.  Returns either a single matrix
derived from dimensionality reduction on all matrices
(\code{jointReduction=TRUE}) or a list of reduced
dimensionality matrices, one for each input.  Either PCA or ICA can be used
as the reduction method.
}
\examples{

set.seed(1500)
nsub = 3
npix = c(10,6,13)
nk = 2
outcome = initializeSyMLR(
  list( matrix(rnorm( nsub * npix[1] ),ncol=npix[1]),
        matrix(rnorm( nsub * npix[2] ),ncol=npix[2]),
        matrix(rnorm( nsub * npix[3] ),ncol=npix[3]) ), 2 )

}
\author{
BB Avants.
}
