% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/inspectImageData3D.R
\name{inspectImageData3D}
\alias{inspectImageData3D}
\title{Simple inspectImageData3D function.}
\usage{
inspectImageData3D(myfiles)
}
\arguments{
\item{fn}{input list of filenames}
}
\value{
matrix is output
}
\description{
InspectImageData3D collects basic statistics over a dataset and returns them
in a dataframe.
}
\examples{
\dontrun{
mm<-inspectImageData3D( fnl<-c('r16slice.nii.gz','r64slice.nii.gz' ) )
usePkg('fpc')
pamres <- pamk(mm)
plot(pamres$pamobject)
outlier.scores <- DMwR::lofactor( mm, k=5)
outliers <- order(outlier.scores, decreasing=T)[1:5]
}
}
\author{
Avants BB
}

