% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/inspectImageData3D.R
\name{inspectImageData3D}
\alias{inspectImageData3D}
\title{Simple inspectImageData3D function.}
\usage{
inspectImageData3D(myfiles)
}
\arguments{
\item{myfiles}{input list of filenames}
}
\value{
matrix is output
}
\description{
InspectImageData3D collects basic statistics over a dataset and returns them
in a dataframe.
}
\examples{
mm<-inspectImageData3D( fnl<-c(getANTsRData('r16'), getANTsRData('r64') ) )
if ( !usePkg("DMwR") | ! usePkg('fpc') )
   { print("Need DMwR and fpc packages") } else {
pamres <- pamk(mm)
plot(pamres$pamobject)
outlier.scores <- lofactor( mm, k=5)
outliers <- order(outlier.scores, decreasing=T)[1:5]
}
}
\author{
Avants BB
}

