% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrateVelocityField.R, R/thresholdImage.R
\name{integrateVelocityField}
\alias{integrateVelocityField}
\title{integrateVelocityField}
\usage{
integrateVelocityField(
  referenceImage,
  velocityFieldFileName,
  deformationFieldFileName,
  lowerTime = 0,
  upperTime = 1,
  deltaTime = 0.01
)

integrateVelocityField(
  referenceImage,
  velocityFieldFileName,
  deformationFieldFileName,
  lowerTime = 0,
  upperTime = 1,
  deltaTime = 0.01
)
}
\arguments{
\item{referenceImage}{defines the image domain}

\item{velocityFieldFileName}{the velocity field exists on disk.}

\item{deformationFieldFileName}{the deformation field output file name.}

\item{lowerTime}{the starting time, usually zero for forward transformation
and one for the inverse transformation.}

\item{upperTime}{the ending time, usually one for forward transformation
and zero for the inverse transformation.}

\item{deltaTime}{the integration time step}

\item{velocityField}{time-varying displacement field}

\item{lowerIntegrationBound}{Lower time bound for integration in [0, 1]}

\item{upperIntegrationBound}{Upper time bound for integration in [0, 1]}

\item{numberOfIntegrationSteps}{Number of integration steps used in the Runge-Kutta solution}
}
\value{
integrated field
}
\description{
Integrate velocity field

Utility function to integrate a velocity field and create a deformation field.
}
\examples{
fi <- antsImageRead( getANTsRData( "r16" ) )
mi <- antsImageRead( getANTsRData( "r27" ) )
reg <- antsRegistration( fi, mi, "TV[2]" )
velocityField <- antsImageRead( reg$velocityfield )
field <- integrateVelocityField( velocityField, 0.0, 1.0, 10 )

\dontrun{
set.seed(1234)
fi <- ( ri(1) )
mi <- ( ri(2) )
mytx2 <- antsRegistration(fixed=fi, mi, typeofTransform = "TV[4]" )
integrateVelocityField( fi, mytx2$velocityfield,  "/tmp/def.nii.gz" )
qq=antsApplyTransforms( fi, mi, mytx2$fwdtransforms )
pp=antsApplyTransforms( fi, mi, "/tmp/def.nii.gz" )
antsImageMutualInformation( fi, mi )
antsImageMutualInformation( fi, qq )
antsImageMutualInformation( fi, pp )
}
}
\author{
NJ Tustison

Avants BB
}
