% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/interleaveMatrixWithItself.R
\name{interleaveMatrixWithItself}
\alias{interleaveMatrixWithItself}
\title{Simple interleaveMatrixWithItself function.}
\usage{
interleaveMatrixWithItself(x, n = 1)
}
\arguments{
\item{x}{input matrix}

\item{n}{number of interleaves}
}
\value{
matrix is output
}
\description{
Replicate columns of input matrix n times as neighbors of columns in
original matrix.
}
\examples{
mat<-replicate(100, rnorm(20))
wmat<-interleaveMatrixWithItself( mat , 5 )
}
\author{
Avants BB
}

