% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscaleSVDxpts.R
\name{invariant_orthogonality_defect}
\alias{invariant_orthogonality_defect}
\title{Calculate the invariant orthogonality defect that is zero for diagonal matrices}
\usage{
invariant_orthogonality_defect(A)
}
\arguments{
\item{A}{Input matrix (n x p, where n >> p)}
}
\value{
The invariant orthogonality defect that is zero for diagonal matrices
}
\description{
Calculate the invariant orthogonality defect that is zero for diagonal matrices
}
