% Generated by roxygen2 (4.0.2): do not edit by hand
\name{invariantImageSimilarity}
\alias{invariantImageSimilarity}
\title{mutual information between two images as a function of geometry}
\usage{
mival<-invariantImageSimilarity( i1, i2, thetas=0:360, 0 )
}
\arguments{
\item{thetas}{numeric vector}

\item{localSearchIterations}{integer controlling local search in multistart}

\item{metric}{which metric MI or GC (string)}

\item{scaleImage}{global scale}

\item{img}{antsImage}

\item{img}{antsImage}
}
\value{
vector of similarity values
}
\description{
compute mutual information between two images as image is rotated about its center w/or w/o optimization
}
\examples{
fi<-antsImageRead( getANTsRData('r16') ,2)
mi<-antsImageRead( getANTsRData('r64') ,2)
mival<-invariantImageSimilarity( fi, mi, c(0,10,20) )
}
\author{
Brian B. Avants
}
\keyword{image}
\keyword{similarity}

