% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/invariantImageSimilarity.R
\name{invariantImageSimilarity}
\alias{invariantImageSimilarity}
\title{mutual information between two images as a function of geometry}
\usage{
mival<-invariantImageSimilarity( i1, i2, thetas=0:360, 0 )
}
\arguments{
\item{thetas}{numeric vector}

\item{localSearchIterations}{integer controlling local search in multistart}

\item{metric}{which metric MI or GC (string)}

\item{scaleImage}{global scale}

\item{doReflection}{reflect image about principal axis}

\item{fixedImg}{antsImage that is fixed ie does not move}

\item{movingImg}{antsImage to be mapped to fixed}

\item{txFilename}{if present, write optimal tx to .mat file}
}
\value{
vector of similarity values
}
\description{
compute mutual information between two images as image is rotated about its center w/or w/o optimization
}
\examples{
fi<-antsImageRead( getANTsRData('r16') ,2)
mi<-antsImageRead( getANTsRData('r64') ,2)
mival<-invariantImageSimilarity( fi, mi, c(0,10,20) )
}
\author{
Brian B. Avants
}
\keyword{image}
\keyword{similarity}

