% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/jointIntensityFusion.R
\name{jointIntensityFusion}
\alias{jointIntensityFusion}
\title{joint intensity fusion}
\usage{
jointIntensityFusion(targetI, targetIMask, atlasList, beta = 4, rad = NA,
  labelList = NA, doscale = TRUE, doNormalize = TRUE,
  maxAtlasAtVoxel = c(1, Inf), rho = 0.01, useSaferComputation = FALSE,
  usecor = FALSE, boundary.condition = "mean", rSearch = 2,
  segvals = NA)
}
\arguments{
\item{targetI}{antsImage to be approximated}

\item{targetIMask}{mask with value 1}

\item{atlasList}{list containing antsImages}

\item{beta}{weight sharpness, default to 2}

\item{rad}{neighborhood radius, default to 4}

\item{labelList}{list containing antsImages}

\item{doscale}{scale neighborhood intensities}

\item{doNormalize}{normalize each image range to 0, 1}

\item{maxAtlasAtVoxel}{min/max n atlases to use at each voxel}

\item{rho}{ridge penalty increases robustness to outliers but also
makes image converge to average}

\item{useSaferComputation}{slower but more error checking}

\item{usecor}{employ correlation as local similarity}

\item{boundary.condition}{one of 'image' 'mean' 'NA'}

\item{rSearch}{radius of search, default is 2}

\item{segvals}{list of labels to expect}
}
\value{
approximated image, segmentation and probabilities
}
\description{
Estimates an image from another set of images - intensity generalization of joint label fusion.  Search radius used only when employing labels - WIP to speed it up.
}
\examples{
set.seed(123)
ref<-antsImageRead( getANTsRData("r16"), 2)
ref<-resampleImage(ref,c(50,50),1,0)
ImageMath(2,ref,"Normalize",ref)
mi<-antsImageRead( getANTsRData("r27"),  2)
mi2<-antsImageRead( getANTsRData("r30") ,2)
mi3<-antsImageRead( getANTsRData("r62") ,2)
mi4<-antsImageRead( getANTsRData("r64") ,2)
mi5<-antsImageRead( getANTsRData("r85") ,2)
refmask<-getMask(ref,mean(ref),Inf,1)
ImageMath(2,refmask,"ME",refmask,10) # just to speed things up
ilist<-list(mi,mi2,mi3,mi4,mi5)
seglist<-list()
for ( i in 1:length(ilist) )
 {
 ImageMath(2,ilist[[i]],"Normalize",ilist[[i]])
 mytx<-antsRegistration(fixed=ref , moving=ilist[[i]] ,
   typeofTransform = c("Affine"), outprefix=tempfile() )
 mywarpedimage<-antsApplyTransforms(fixed=ref,moving=ilist[[i]],
   transformlist=mytx$fwdtransforms)
 ilist[[i]]=mywarpedimage
 seg<-kmeansSegmentation( ilist[[i]], k=3, kmask = refmask)
 seglist[[i]]<-seg$segmentation
 }
r<-2
d<-2
pp<-jointIntensityFusion(ref,refmask,ilist, rSearch=0,
  labelList=seglist, rad=rep(r,d) )
\dontrun{
mm<-imageListToMatrix(ilist,refmask)
avg<-makeImage(refmask,colMeans(mm)) # compare to pp[[1]]
# save memory by separating masks
refmaske<-antsImageClone(refmask)
ImageMath(d,refmaske,"ME",refmask,15)
refmask[refmaske==1]<-0
pp1<-jointIntensityFusion(ref,refmask,ilist,
  beta=2,rad=rep(r,d))
pp2<-jointIntensityFusion(ref,refmaske,ilist,
  beta=2,rad=rep(r,d))
pp1[[1]][refmaske==1]<-pp2[[1]][refmaske==1]
}
}
\author{
Brian B. Avants, Hongzhi Wang, Paul Yushkevich
}
\keyword{fusion,}
\keyword{template}

