% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jointIntensityFusion.R
\name{jointIntensityFusion}
\alias{jointIntensityFusion}
\title{joint intensity fusion}
\usage{
jointIntensityFusion(targetI, targetIMask, atlasList, beta = 4, rad = NA,
  doscale = TRUE, doNormalize = TRUE, maxAtlasAtVoxel = c(1, Inf),
  rho = 0.01, usecor = FALSE, boundary.condition = "image", rSearch = 2,
  jifImage = NA, imputeList = NA, imputedImage = NA)
}
\arguments{
\item{targetI}{antsImage to be approximated}

\item{targetIMask}{mask with value 1}

\item{atlasList}{list containing antsImages}

\item{beta}{weight sharpness, default to 2}

\item{rad}{neighborhood radius, default to 3}

\item{doscale}{scale neighborhood intensities}

\item{doNormalize}{normalize each image range to 0, 1}

\item{maxAtlasAtVoxel}{min/max n atlases to use at each voxel}

\item{rho}{ridge penalty increases robustness to outliers but also
makes image converge to average}

\item{usecor}{employ correlation as local similarity}

\item{boundary.condition}{one of 'image' 'mean' 'NA'}

\item{rSearch}{radius of search, default is 2}

\item{jifImage}{the current estimated jif image (helps speed slice by slice)}

\item{imputeList}{2nd modality to impute}

\item{imputedImage}{current estimate of 2nd modality}
}
\value{
approximated image, segmentation and probabilities
}
\description{
Estimates an image from another set of images - intensity generalization of
joint label fusion.  Search radius used only when employing labels.
}
\examples{

set.seed(123)
ref<-antsImageRead( getANTsRData("r16"))
ref<-resampleImage(ref,c(50,50),1,0)
ref<-iMath(ref,"Normalize")
mi<-antsImageRead( getANTsRData("r27"))
mi2<-antsImageRead( getANTsRData("r30"))
mi3<-antsImageRead( getANTsRData("r62"))
mi4<-antsImageRead( getANTsRData("r64"))
mi5<-antsImageRead( getANTsRData("r85"))
refmask<-getMask(ref)
refmask<-iMath(refmask,"ME",2) # just to speed things up
ilist<-list(mi,mi2,mi3,mi4,mi5)
implist<-list()
for ( i in 1:length(ilist) )
 {
 ilist[[i]]<-iMath(ilist[[i]],"Normalize")
 mytx<-antsRegistration(fixed=ref , moving=ilist[[i]] ,
   typeofTransform = c("Affine") )
 mywarpedimage<-antsApplyTransforms(fixed=ref,moving=ilist[[i]],
   transformlist=mytx$fwdtransforms)
 ilist[[i]]=mywarpedimage
 implist[[i]]=mywarpedimage \%>\% iMath("Laplacian")
 }
r<-2
d<-2
pp<-jointIntensityFusion( ref, refmask, ilist, rSearch=1, rad=rep(r,d) )
pp<-jointIntensityFusion( ref, refmask, ilist, rSearch=1, rad=rep(r,d),
  imputeList=implist )

}
\author{
Brian B. Avants, Hongzhi Wang, Paul Yushkevich
}
\keyword{fusion,}
\keyword{template}

