% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/KellyKapowski.R
\name{KellyKapowski}
\alias{KellyKapowski}
\title{Compute cortical thickness using the DiReCT algorithm.}
\usage{
KellyKapowski(d = NA, s = NA, g = NA, w = NA, c = "[45,0.0,10]",
  r = 0.025, m = 1.5, outimg = outimg, ...)
}
\arguments{
\item{s}{segmentation image}

\item{g}{gray matter probability image}

\item{w}{white matter probability image}

\item{c}{convergence params - first controls iterations}

\item{r}{gradient descent update parameter}

\item{m}{gradient field smoothing parameter}

\item{imageDimension2|3|4}{Number of dimensions of the input image}

\item{inputImage}{Input image to operate on}
}
\value{
0 -- Success\cr 1 -- Failure
}
\description{
Diffeomorphic registration-based cortical thickness based on probabilistic
segmentation of an image.  This is an optimization algorithm.
}
\examples{
\dontrun{
KellyKapowski( d=3, s=simg, g=gimg,w=wimg,c=45,r=0.5,m=1,o=oimg )
}
}
\author{
Shrinidhi KL, Avants BB
}

