% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/KellyKapowski.R
\name{KellyKapowski}
\alias{KellyKapowski}
\title{Compute cortical thickness using the DiReCT algorithm.}
\usage{
KellyKapowski(s = NA, g = NA, w = NA, its = 50, r = 0.025, m = 1.5,
  outimg = outimg, ...)
}
\arguments{
\item{s}{segmentation image}

\item{g}{gray matter probability image}

\item{w}{white matter probability image}

\item{its}{convergence params - controls iterations}

\item{r}{gradient descent update parameter}

\item{m}{gradient field smoothing parameter}

\item{...}{anything else, see KK help in ANTs}
}
\value{
0 -- Success\cr 1 -- Failure
}
\description{
Diffeomorphic registration-based cortical thickness based on probabilistic
segmentation of an image.  This is an optimization algorithm.
}
\examples{
img<-antsImageRead( getANTsRData("r16") ,2)
mask<-getMask( img )
segs<-kmeansSegmentation( img, k=3, kmask = mask)
KellyKapowski( s=segs$segmentation, g=segs$probabilityimages[[2]],
  w=segs$probabilityimages[[3]],its=45,r=0.5,m=1 )
}
\author{
Shrinidhi KL, Avants BB
}

