% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kellyKapowski.R
\name{kellyKapowski}
\alias{kellyKapowski}
\title{Compute cortical thickness using the DiReCT algorithm.}
\usage{
kellyKapowski(s = NA, g = NA, w = NA, its = 50, r = 0.025, m = 1.5,
  ...)
}
\arguments{
\item{s}{segmentation image}

\item{g}{gray matter probability image}

\item{w}{white matter probability image}

\item{its}{convergence params - controls iterations}

\item{r}{gradient descent update parameter}

\item{m}{gradient field smoothing parameter}

\item{...}{anything else, see KK help in ANTs}
}
\value{
thickness antsImage
}
\description{
Diffeomorphic registration-based cortical thickness based on probabilistic
segmentation of an image.  This is an optimization algorithm.
}
\examples{

img<-antsImageRead( getANTsRData("r16") ,2)
img<-resampleImage(img,c(64,64),1,0)
mask<-getMask( img )
segs<-kmeansSegmentation( img, k=3, kmask = mask)
thk<-kellyKapowski( s=segs$segmentation, g=segs$probabilityimages[[2]],
  w=segs$probabilityimages[[3]],its=45,r=0.5,m=1 )

}
\author{
Shrinidhi KL, Avants BB
}
