% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/kmeansSegmentation.R
\name{kmeansSegmentation}
\alias{kmeansSegmentation}
\title{k means image segmentation.}
\usage{
kmeansSegmentation(img, k, kmask = NA, mrf = 0.1)
}
\arguments{
\item{img}{input image}

\item{k}{integer number of classes}

\item{kmask}{segment inside this mask}

\item{mrf}{smoothness, higher is smoother}
}
\value{
segmentation and probability images
}
\description{
k means image segmentation that is a wrapper around Atropos
}
\examples{
fi<-antsImageRead( getANTsRData("r16") ,2)
fi<-n3BiasFieldCorrection(fi,4)
seg<-kmeansSegmentation( fi, 3 )
}
\author{
Brian B. Avants
}

