% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscaleSVDxpts.R
\name{knnSmoothImage}
\alias{knnSmoothImage}
\title{k-nearest neighbors constrained image smoothing}
\usage{
knnSmoothImage(img, mask, radius, intensityWeight = 0.1,
  spatialSigma = 20, iterations = 1, returnMatrix = FALSE)
}
\arguments{
\item{img}{input image to smooth}

\item{mask}{input image mask}

\item{radius}{number of neighbors, higher causes more smoothing}

\item{intensityWeight}{weight for intensity component, value 1 will weight
local voxel intensity roughly equally to spatial component}

\item{spatialSigma}{for gaussian defining spatial distances}

\item{iterations}{number of iterations over which to apply smoothing kernel}

\item{returnMatrix}{boolean, will return smoothing matrix instead of image.}
}
\value{
antsImage is output
}
\description{
Compute a smoothing matrix based on an input matrix of point coordinates as
well as neighborhood intensity patterns.  this performs a form of edge
preserving smoothing.
}
\examples{

\dontrun{
img = antsImageRead( getANTsRData( 'r16' ) )
mask = getMask( img )
simg = knnSmoothImage( img=img, mask=mask, radius=2, intensityWeight=1,
  spatialSigma=1.5, iterations=1 )
}
}
\author{
Avants BB
}
