% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscaleSVDxpts.R
\name{knnSmoothingMatrix}
\alias{knnSmoothingMatrix}
\title{k-nearest neighbors constrained smoothing}
\usage{
knnSmoothingMatrix(x, k, sigma, segmentation)
}
\arguments{
\item{x}{input matrix of point coordinates of dimensions n-spatial
spatial dimensions by p points}

\item{k}{number of neighbors, higher causes more smoothing}

\item{sigma}{sigma for the gaussian function}

\item{segmentation}{optional boolean to restrict specific rows to have minimal respons}
}
\value{
sparse matrix is output
}
\description{
Compute a smoothing matrix based on an input matrix of point coordinates
}
\examples{

\dontrun{
mask = getMask( antsImageRead( getANTsRData( 'r16' ) ) )
spatmat = t( imageDomainToSpatialMatrix( mask, mask ) )
smoothingMatrix = knnSmoothingMatrix( spatmat, k = 25, sigma = 3.0 )
rvec = rnorm( nrow( smoothingMatrix ) )
srvec = smoothingMatrix \%*\% rvec
rvi = makeImage( mask, rvec )
srv = makeImage( mask,  as.numeric( srvec ) )
}
}
\author{
Avants BB
}
