% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/labelClusters.R
\name{labelClusters}
\alias{labelClusters}
\title{Simple labelClustering function.}
\usage{
labelClusters(imagein, minClusterSize = 50, minThresh = 1e-06,
  maxThresh = 1)
}
\arguments{
\item{img}{input antsImage e.g. a statistical map}
}
\value{
labeled cluster image is output
}
\description{
This will give a unique ID to each connected component 1 through N of size >
minClusterSize
}
\examples{
\dontrun{
img<-antsImageRead( getANTsRData('mnib'), 3 )
outimage<-labelClusters( img )
}
}
\author{
Avants BB
}

