% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/labelGeometryMeasures.R
\name{labelGeometryMeasures}
\alias{labelGeometryMeasures}
\title{labelGeometryMeasures}
\usage{
labelGeometryMeasures(...)
}
\arguments{
\item{...}{see ants function}
}
\value{
none
}
\description{
Wrapper for the ANTs funtion labelGeometryMeasures
}
\author{
Avants BB
}

