% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/labelGeometryMeasures.R
\name{labelGeometryMeasures}
\alias{labelGeometryMeasures}
\title{labelGeometryMeasures}
\usage{
labelGeometryMeasures(labelImage, intensityImage = NA)
}
\arguments{
\item{labelImage}{on which to compute geometry}

\item{intensityImage}{optional}
}
\value{
none
}
\description{
Wrapper for the ANTs funtion labelGeometryMeasures
}
\examples{
fi<-antsImageRead( getANTsRData("r16") , 2 )
seg<-kmeansSegmentation( fi, 3 )$segmentation
geom<-labelGeometryMeasures(seg,fi)
}
\author{
Avants BB
}

