% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelOverlapMeasures.R
\name{labelOverlapMeasures}
\alias{labelOverlapMeasures}
\title{labelOverlapMeasures}
\usage{
labelOverlapMeasures(sourceLabelImage, targetLabelImage)
}
\arguments{
\item{sourceLabelImage}{label image for source image.}

\item{targetLabelImage}{label image for target/reference image.}
}
\value{
data frame with overlap measures
}
\description{
Wrapper for the ANTs funtion LabelOverlapMeasures.  More documentaiton
available here:
}
\details{
\url{https://www.insight-journal.org/browse/publication/707}
}
\examples{

sourceImage <- antsImageRead(getANTsRData("r16"), 2)
sourceSegmentation <- kmeansSegmentation(sourceImage, 3)$segmentation
referenceImage <- antsImageRead(getANTsRData("r16"), 2)
referenceSegmentation <- kmeansSegmentation(referenceImage, 3)$segmentation
overlap <- labelOverlapMeasures(sourceSegmentation, referenceSegmentation)

}
\author{
Avants BB, Tustison NJ
}
