% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelStats.R
\name{labelStats}
\alias{labelStats}
\title{labelStats}
\usage{
labelStats(image, labelImage)
}
\arguments{
\item{image}{Image to calculate statistics from.}

\item{labelImage}{Label image.}
}
\value{
Data frame with one row per label (including zero) and columns :
  LabelValues, Mean, Min, Max, Variance, Count, Volume
}
\description{
Get label statistics from image.
}
\examples{

img <- antsImageRead( getANTsRData("r16") , 2 )
img <- resampleImage( img, c(64,64), 1, 0 )
mask <- getMask(img)
segs1 = kmeansSegmentation( img, 3 )
labelStats(img, segs1$segmentation)

}
