% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lappend.R
\name{lappend}
\alias{lappend}
\title{Simple list append tool}
\usage{
lappend(lst, obj)
}
\arguments{
\item{lst}{list to which you will append the 2nd item}

\item{obj}{to be added to the list}
}
\value{
outputs the appended list
}
\description{
Append an item to a list , creating a new list or combine two lists
}
\examples{

\dontrun{
mat<-replicate(100, rnorm(20))
mat2<-replicate(100, rnorm(20))
mylist<-list(mat)
mylist<-lappend( mylist, mat2 )
}

}
\author{
Avants BB, the internet
}

