% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeImage.R
\name{makeImage}
\alias{makeImage}
\title{Simple makeImage function.}
\usage{
makeImage(
  imagesize,
  voxval = 0,
  spacing = c(NA),
  origin = c(NA),
  direction = c(NA),
  components = FALSE,
  pixeltype = "float"
)
}
\arguments{
\item{imagesize}{input image size or mask}

\item{voxval}{input image value or vector, size of mask}

\item{spacing}{image spatial resolution}

\item{origin}{image spatial origin}

\item{direction}{direction matrix to convert from index to physical space}

\item{components}{whether there are components per pixel or not}

\item{pixeltype}{data type of image values}
}
\value{
antsImage is output
}
\description{
Make an image with given size and voxel value or given a mask and vector.
}
\examples{

outimg<-makeImage( c(2,10) , 1)
outimg<-makeImage( outimg ,  c(2,10) )
testthat::expect_error(makeImage(outimg, c(2,10), components = 2))

}
\author{
Avants BB
}
