% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/makeImage.R
\name{makeImage}
\alias{makeImage}
\title{Simple makeImage function.}
\usage{
makeImage(imagesize, voxval = 0, spacing = c(NA), origin = c(NA),
  direction = c(NA), components = 1, pixeltype = "float")
}
\arguments{
\item{imagesize}{input image size or mask}

\item{voxval}{input image value or vector, size of mask}
}
\value{
antsImage is output
}
\description{
Make an image with given size and voxel value or given a mask and vector.
}
\examples{
outimg<-makeImage( c(2,10) , 1)
outimg<-makeImage( outimg ,  c(2,10) )
}
\author{
Avants BB
}

