% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePowersPointsImage.R
\name{makePointsImage}
\alias{makePointsImage}
\title{Create label image from physical space points}
\usage{
makePointsImage(pts, mask, radius = 5)
}
\arguments{
\item{pts}{input powers points}

\item{mask}{antsImage mask defining target space}

\item{radius}{for the points}
}
\value{
antsImage is output
}
\description{
Creates spherical points in the coordinate space of the target image based
on the n-dimensional matrix of points that the user supplies. The image
defines the dimensionality of the data so if the input image is 3D then
the input points should be 2D or 3D.
}
\examples{

\dontrun{
mni <- antsImageRead( getANTsRData( "mni" ) ) \%>\% getMask()
data( "powers_areal_mni_itk", package = "ANTsR", envir = environment() )
powersLabels = makePointsImage( powers_areal_mni_itk[,1:3], mni, radius=3 )
}

}
\author{
Avants BB, Duda JT
}
