% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/maskImage.R
\name{maskImage}
\alias{maskImage}
\title{Mask input image by mask image.}
\usage{
maskImage(img.in, img.mask, level = 1, binarize = FALSE)
}
\arguments{
\item{img.in}{Input image.}

\item{img.mask}{Mask or label image.}

\item{level}{Level(s) at which to mask image.  If vector or list
of values, output image is non-zero at all locations where label image
matches any of the levels specified.}

\item{binarize}{binarize the output image?}
}
\value{
An object of type antsImage.
}
\description{
Mask an input image by a mask image.  If the mask image has multiple labels,
it is possible to specify which label(s) to mask at.
}
\examples{
myimg <- antsImageRead(getANTsRData("r16"))
mask <- getMask(myimg)
myimg.mask <- maskImage(myimg, mask, 3)
seg <- kmeansSegmentation(myimg, 3)
myimg.mask <- maskImage(myimg, seg$segmentation, c(1,3))
}
\author{
Kandel BM and Avants B.
}

