% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_Operators.R
\name{mean.antsImage}
\alias{mean.antsImage}
\alias{sd.antsImage}
\title{Mean for antsImage Objects}
\usage{
\method{mean}{antsImage}(x, ..., mask)

sd.antsImage(x, ...)
}
\arguments{
\item{x}{is an object of class \code{antsImage}.}

\item{...}{additional arguments to pass to \code{\link{mean}}}

\item{mask}{binary mask of values to subset}
}
\description{
Overloaded Mean for antsImage objects
}
\details{
#' @rdname antsImagemath
#' @aliases xor,antsImage-method
setMethod(f = "xor", signature(x = "antsImage",
                             y = "antsImage"), definition = function(x) {
  xx = as.array(x)
  yy = as.array(y)
  xor(xx, yy)
  !a2
})
}
