% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscaleSVDxpts.R
\name{measure_orthogonality}
\alias{measure_orthogonality}
\title{Measure the Orthogonality of a Non-Square Matrix (scale invariant)}
\usage{
measure_orthogonality(mat)
}
\arguments{
\item{mat}{A numeric matrix.}
}
\value{
A numeric value indicating the deviation from orthogonality.
}
\description{
Computes a measure of orthogonality for a non-square matrix by comparing
the product of the transpose and the matrix to the identity matrix.
}
\examples{
A <- matrix(c(1, 0, 0, 1, 0, 1), nrow = 3, ncol = 2)
measure_orthogonality(A) # Should be close to 0 for orthonormal columns
B <- matrix(c(1, 1, 0, 1, 0, 1), nrow = 3, ncol = 2)
measure_orthogonality(B)
}
