% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscaleSVDxpts.R
\name{measure_orthogonality_gradient}
\alias{measure_orthogonality_gradient}
\title{Measure the Orthogonality Gradient of a Matrix}
\usage{
measure_orthogonality_gradient(mat)
}
\arguments{
\item{mat}{A square numeric matrix.}
}
\value{
A matrix representing the gradient of the orthogonality measure.
}
\description{
Computes the gradient of the orthogonality measure for a square matrix.
}
\examples{
A <- matrix(c(1, 0, 0, 1), nrow = 2)
measure_orthogonality_gradient(A) # Gradient should be zero for an orthogonal matrix
B <- matrix(c(1, 1, 0, 1), nrow = 2)
measure_orthogonality_gradient(B)
}
