% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MeasureMinMaxMean.R
\name{MeasureMinMaxMean}
\alias{MeasureMinMaxMean}
\title{Measure Min Max Mean}
\usage{
MeasureMinMaxMean(image, mask = NA)
}
\arguments{
\item{image}{Input image of type \code{antsImage}.}
}
\value{
List with entries \code{img.min}, \code{img.max}, \code{img.mean}.
}
\description{
Measure min, max, and mean of image.
}
\examples{
\dontrun{
  myimg <- as.antsImage(array(rnorm(1000), dim=c(10, 10, 10)))
  MeasureMinMaxMean(myimg)
}
}
\author{
Kandel BM.
}

