% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeChannels.R
\name{mergeChannels}
\alias{mergeChannels}
\title{merge images into a multiChannel antsImage}
\usage{
mergeChannels(imageList)
}
\arguments{
\item{imageList}{a list of antsImage objects to merge}
}
\value{
A multiChannel antsImage object
}
\description{
merge images into a multiChannel antsImage
}
\examples{
dims = c(30, 30)
n = prod(dims)
 r <- floor( seq(n) / (n) * 255 )
 dim(r) <- dims
 arr = r
 r <- as.antsImage(r)
 g <- r*0
 b <- r*0
 rgbImage = mergeChannels( list(r,g,b) )
 testthat::expect_error(mergeChannels(list(arr, arr)))

}
\author{
Duda, JT
}
