% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mni2tal.R
\name{mni2tal}
\alias{mni2tal}
\title{Brett's mni2tal}
\usage{
mni2tal(xin = 0)
}
\arguments{
\item{xin}{point in mni152 space.}
}
\value{
output point in approximate Talairach space.
}
\description{
mni2tal for converting from ch2/mni space to tal - very approximate.
}
\details{
This is a standard approach but it's not very accurate.
}
\examples{
mni2tal( c(10,12,14) )
}
\author{
Matthew Brett, adapted for ANTsR by B Avants
}
\references{
\url{http://bioimagesuite.yale.edu/mni2tal/501_95733_More\%20Accurate\%20Talairach\%20Coordinates\%20SLIDES.pdf},
\url{http://imaging.mrc-cbu.cam.ac.uk/imaging/MniTalairach}
}
\keyword{Talairach}

