% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ants_motion_estimation.R
\name{.motion_correction}
\alias{.motion_correction}
\title{Simple motion_correction function.}
\usage{
.motion_correction(img, fixed = NA, moreaccurate = 1, txtype = "Affine",
  verbose = FALSE, num_threads = 1, seed = NULL)
}
\arguments{
\item{img}{4D antsImage}

\item{fixed}{target fixed image}

\item{moreaccurate}{0, 1 or 2 with higher values being more accurte
(use 2 for real applications, 0 for testing)}

\item{txtype}{Transformation type}

\item{verbose}{verbosity boolean}

\item{num_threads}{will execute 
\code{Sys.setenv(ITK_GLOBAL_DEFAULT_NUMBER_OF_THREADS = num_threads)} before
running to attempt a more reproducible result.  See
\url{https://github.com/ANTsX/ANTs/wiki/antsRegistration-reproducibility-issues}
for discussion.  If \code{NULL}, will not set anything.}

\item{seed}{will execute 
\code{Sys.setenv(ANTS_RANDOM_SEED = seed)} before
running to attempt a more reproducible result.  See
\url{https://github.com/ANTsX/ANTs/wiki/antsRegistration-reproducibility-issues}
for discussion.  If \code{NULL}, will not set anything.}
}
\value{
list of outputs
}
\description{
motion corrects 4D time series imaging data
}
\note{
This function may give different results on multiple runs.
}
\examples{
set.seed(1000)
testimg<-makeImage( c(10,10,10,5),  rnorm(  5000  ) )
testimg<-iMath(testimg,"PadImage",5)
mocorr <-.motion_correction( testimg, num_threads = 1, seed = 10)
aimg_to_array = function(x) {
if (is.antsImage(x)) {
   x = as.array(x)
}
x
}
amocorr = lapply(mocorr, aimg_to_array)
mocorr2 <-.motion_correction( testimg, num_threads = 1, seed = 10)
amocorr2 = lapply(mocorr2, aimg_to_array)
testthat::expect_equal(mocorr, mocorr2)
testthat::expect_equal(amocorr, amocorr2)

# This function may give different results on multiple runs
# without setting the seed
}
\author{
Avants BB
}
