% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrvnrfs.R
\name{mrvnrfs.predict}
\alias{mrvnrfs.predict}
\title{multi-res voxelwise neighborhood random forest segmentation}
\usage{
mrvnrfs.predict(
  rflist,
  x,
  labelmasks,
  rad = NA,
  multiResSchedule = c(4, 2, 1),
  asFactors = TRUE,
  voxchunk = 60000
)
}
\arguments{
\item{rflist}{a list of random forest models from mrvnrfs}

\item{x}{a list of lists where each list contains feature images}

\item{labelmasks}{a mask (or list of masks) used to define the area to predict.
This is used to save time by contstrain the prediction in within the brain.}

\item{rad}{vector of dimensionality d define nhood radius}

\item{multiResSchedule}{an integer vector defining multi-res levels}

\item{asFactors}{boolean - treat the y entries as factors}

\item{voxchunk}{value of maximal voxels to predict at once. This value
is used to split the prediction into smaller chunks such that memory
requirements do not become too big}
}
\value{
list a 4-list with the rf model, training vector, feature matrix
and the random mask
}
\description{
Represents multiscale feature images as a neighborhood and uses the features
to apply a random forest segmentation model to a new image
}
\author{
Avants BB, Tustison NJ, Pustina D
}
