% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mrvnrfs.R
\name{mrvnrfs}
\alias{mrvnrfs}
\title{multi-res voxelwise neighborhood random forest segmentation learning}
\usage{
mrvnrfs(y, x, labelmask, rad = NA, nsamples = 1, ntrees = 500,
  multiResSchedule = c(4, 2, 1), asFactors = TRUE)
}
\arguments{
\item{y}{list of training label images, can be a factor or numeric vector
this can also be a regular old vector}

\item{x}{a list of lists where each list contains feature images}

\item{labelmask}{a mask for the features (all in the same image space)
the labelmask defines the number of parallel samples that will be used
per subject sample. two labels will double the number of predictors
contributed from each feature image.}

\item{rad}{vector of dimensionality d define nhood radius}

\item{nsamples}{(per subject to enter training)}

\item{ntrees}{(for the random forest model)}

\item{multiResSchedule}{an integer vector defining multi-res levels}

\item{asFactors}{boolean - treat the y entries as factors}
}
\value{
list a 4-list with the rf model, training vector, feature matrix
and the random mask
}
\description{
Represents multiscale feature images as a neighborhood and uses the features
to build a random forest segmentation model from an image population
}
\examples{
mask<-makeImage( c(10,10), 0 )
mask[ 3:6, 3:6 ]<-1
mask[ 5, 5 ]<-2
ilist<-list()
lablist<-list()
inds<-1:50
scl<-0.33 # a noise parameter
for ( i in inds ) {
  img<-antsImageClone(mask)
  imgb<-antsImageClone(mask)
  img[ 3:6, 3:6 ]<-rnorm(16)*scl+(i \%\% 2)+scl*mean(rnorm(1))
  imgb[ 3:6, 3:6 ]<-rnorm(16)*scl+(i \%\% 2)+scl*mean(rnorm(1))
  limg<-antsImageClone(mask)
  limg[ 3:6, 3:6 ]<-(i \%\% 2)  # the label image is constant
  ilist[[i]]<-list(img,imgb)  # two features
  lablist[[i]]<-limg
}
rfm<-mrvnrfs( lablist , ilist, mask, rad=c(0,0), multiResSchedule=1 )
rfmresult<-mrvnrfs.predict( rfm$rflist ,
   list(ilist[[2]]), mask, rad=c(0,0),
   multiResSchedule=1 )
}
\author{
Avants BB, Tustison NJ
}

