% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/N3BiasFieldCorrection.R
\name{N3BiasFieldCorrection}
\alias{N3BiasFieldCorrection}
\title{Bias Field Correction}
\usage{
N3BiasFieldCorrection(...)
}
\arguments{
\item{...}{See N3BiasFieldCorrection in ANTs}
}
\description{
Perform Bias Field Correction on the given image
}
\examples{
img<-makeImage(c(5,5),rnorm(25))
N3BiasFieldCorrection( 2 , img, img, 2 )
}
\author{
Shrinidhi KL
}

