% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/n4BiasFieldCorrection.R
\name{n4BiasFieldCorrection}
\alias{n4BiasFieldCorrection}
\title{Bias Field Correction}
\usage{
n4BiasFieldCorrection(img, mask = NA, shrinkFactor = 4,
  convergence = list(iters = c(50, 50, 50, 50), tol = 1e-07),
  splineParam = 200)
}
\arguments{
\item{img}{input antsImage}

\item{mask}{input mask, if one is not passed one will be made}

\item{shrinkFactor}{Shrink factor for multi-resolution correction,
typically integer less than 4}

\item{convergence}{List of:  \code{iters}, vector of maximum number of
iterations for each shrinkage factor, and \code{tol}, the convergence tolerance.}

\item{splineParam}{Parameter controlling number of control points in spline.
Either single value, indicating how many control points, or vector
with one entry per dimension of image, indicating the spacing in each direction.}
}
\value{
outimg Bias-corrected image
}
\description{
Perform Bias Field Correction on the given image
}
\examples{
img<-makeImage( c(50,50), rnorm(2500) )
 n4img<-n4BiasFieldCorrection(img)
}
\author{
BB Avants
}

