% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/n4BiasFieldCorrection.R
\name{n4BiasFieldCorrection}
\alias{n4BiasFieldCorrection}
\title{Bias Field Correction}
\usage{
n4BiasFieldCorrection(img, mask = NA)
}
\arguments{
\item{img}{input antsImage}

\item{mask}{input mask, if one is not passed one will be made}
}
\value{
outimg
}
\description{
Perform Bias Field Correction on the given image
}
\examples{
img<-makeImage( c(50,50), rnorm(2500) )
 n4img<-n4BiasFieldCorrection(img)
}
\author{
BB Avants
}

