% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscaleSVDxpts.R
\name{nsa_flow_retract}
\alias{nsa_flow_retract}
\title{Retraction onto the Stiefel Manifold}
\usage{
nsa_flow_retract(Y_cand, w_retract, retraction_type)
}
\description{
Retraction onto the Stiefel Manifold
}
