% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscaleSVDxpts.R
\name{nsa_flow_retract_auto}
\alias{nsa_flow_retract_auto}
\title{Retraction onto the Stiefel Manifold (soft-polar multiplicative, with adaptive inv-sqrt)}
\usage{
nsa_flow_retract_auto(
  Y_cand,
  w_retract = 1,
  retraction_type = c("polar", "soft_polar", "none"),
  eps_rf = 1e-08,
  inv_method = "diag",
  ns_iter = 1L,
  eig_thresh = 128L,
  diag_thresh = 8192L,
  verbose = FALSE
)
}
\description{
Retraction onto the Stiefel Manifold (soft-polar multiplicative, with adaptive inv-sqrt)
}
