% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscaleSVDxpts.R
\name{optimize_indicator_matrix}
\alias{optimize_indicator_matrix}
\title{Optimize Binary Indicator Matrix for Positive Entries}
\usage{
optimize_indicator_matrix(m, preprocess = FALSE)
}
\arguments{
\item{m}{A numeric matrix for which the optimization will be performed.}

\item{preprocess}{Logical. If TRUE, the function flips the sign of each row where the entries 
are predominantly negative before optimization.}
}
\value{
A binary indicator matrix `I` where each column has exactly one non-zero entry, 
        the 1's are uniformly distributed across rows, and they correspond only to positive entries of `m`.
}
\description{
This function optimizes the sum of a matrix `m * I`, where `I` is a binary indicator matrix. 
The constraint is that each column in `I` has exactly one non-zero entry and the distribution 
of 1's is uniform across rows, with 1's only set for positive entries in `m`.
}
\examples{
set.seed(123)
m <- matrix(rnorm(500), nrow = 5)
optimize_indicator_matrix(m, preprocess = TRUE)
}
