% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscaleSVDxpts.R
\name{orthogonalizeAndQSparsify}
\alias{orthogonalizeAndQSparsify}
\title{sparsify a matrix}
\usage{
orthogonalizeAndQSparsify(
  v,
  sparsenessQuantile = 0.5,
  positivity = "either",
  orthogonalize = TRUE,
  softThresholding = FALSE,
  unitNorm = FALSE,
  decomAlg = NA
)
}
\arguments{
\item{v}{input matrix}

\item{sparsenessQuantile}{quantile to control sparseness - higher is sparser}

\item{positivity}{restrict to positive or negative solution (beta) weights.
choices are positive, negative or either as expressed as a string.}

\item{orthogonalize}{run gram-schmidt if TRUE.}

\item{softThresholding}{use soft thresholding}

\item{unitNorm}{set each vector to unit norm}

\item{decomAlg}{string sets the NMF or algorithm to estimate V}
}
\value{
matrix
}
\description{
This implements a quantile based sparsification operation
}
\examples{

mat<-replicate(100, rnorm(20))
mat = orthogonalizeAndQSparsify( mat )

}
\author{
Avants BB
}
