% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perfusionregression.R
\name{perfusionregression}
\alias{perfusionregression}
\title{Perfusion Regression}
\usage{
perfusionregression(
  mask_img,
  mat,
  xideal,
  nuis = NA,
  dorobust = 0,
  skip = 20,
  selectionValsForRegweights = NULL,
  useBayesian = 0
)
}
\arguments{
\item{mask_img}{Mask image selects the voxels where CBF will be estimated.
Voxels corresponding to logical FALSE are not computed.}

\item{mat}{Matrix with a column for every time-series voxel. Number of rows
equals the number of time units in the series.}

\item{xideal}{1D time-series signal to be used a ideal or model for
regression.}

\item{nuis}{Nuisance parameters obtained from '.get_perfusion_predictors'.}

\item{dorobust}{Real value in interval from 0 to 1.  If greater than 0, then
robust regression will be performed.  A typical value would be 0.95 i.e. use
voxels with 95 percent confidence.}

\item{skip}{skip / stride over this number of voxels to increase speed}

\item{selectionValsForRegweights}{scalar function to guide parameter est.}

\item{useBayesian}{if greater than zero, use a bayesian prior w/this weight}
}
\value{
Success -- An object of type 'antsImage' containing the CBF estimate
for voxels corresponding to the mask input\cr
}
\description{
Estimate CBF using standard regression and optionally robust regression.
}
\examples{

\dontrun{
#
# cbf <- perfusionregression( mask_img, mat, xideal , nuis )
}

}
\author{
Shrinidhi KL Avants BB
}
