% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot.antsImage.R
\name{plot.antsImage}
\alias{plot.antsImage}
\title{plot.antsImage}
\usage{
\method{plot}{antsImage}(x, y, color = c("jet", "red", "blue", "green",
  "yellow"), axis = 2, slices, window.img = quantile(x[x != 0], 0.05, 0.95),
  threshold = "0.5xInf", quality = 4, outname = NA, alpha = 0.5,
  newwindow = FALSE, ...)
}
\arguments{
\item{x}{the reference image on which to overlay}

\item{y}{list of the images to use as overlays}

\item{color}{the color for the overlay , e.g c('blue','red') length
of this list should match the image list}

\item{axis}{... the axis to slice (1 , 2 or 3)}

\item{slices}{character, the slices to overlay written as 10x20x3 ( or TODO
c(5,10,25) ) where 10x20 is the range and 3 is the increment,
for multislice display}

\item{window.img}{intensity windowing}

\item{threshold}{character, we overlay values above/below this
threshold : of form LOxHI}

\item{quality}{integer quality magnification factor 1 => large (e.g.
10)}

\item{outname}{eg'figx.jpg' output name if you want to write the result to a
file}

\item{alpha}{opacity}

\item{newwindow}{boolean controlling if we open a new device for this plot}

\item{...}{other parameters}
}
\value{
output is plot to standard R window
}
\description{
Plotting an image slice or multi-slice with optional color overlay.
}
\details{
This is a plotting utility for antsImage types with a background and color
overlay option.  Useful for displaying statistical results overlaid on a
background image.
}
\examples{
\dontrun{
  mnit<-getANTsRData('mni')
  mnit<-antsImageRead(mnit,3)
  mniafn<-getANTsRData('mnia')
  mnia<-antsImageRead(mniafn,3)
  mnia<-thresholdImage(mnia,22,25)
  mnia<-smoothImage(mnia,1.5)
  mnia2<-antsImageRead(mniafn,3)
  mnia2<-thresholdImage(mnia2,1,4)
  mnia2<-smoothImage(mnia2,1.5)
  ofn<-paste(tempfile(),'.png',sep='')
  # write directly to a file
  plot( mnit, list(mnia,mnia2), slices='50x140x5',
   threshold = '0.25x1', axis=0,color=c('red','blue'), outname = ofn )
}
}
\author{
Avants BB
}

