% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.antsImage.R
\name{plot.antsImage}
\alias{plot.antsImage}
\title{Plotting an image slice or multi-slice with optional color overlay.}
\usage{
\method{plot}{antsImage}(x, y, color.img = "white", color.overlay = c("jet",
  "red", "blue", "green", "yellow", "viridis", "magma", "plasma", "inferno"),
  axis = 2, slices, colorbar = missing(y), title.colorbar, title.img,
  title.line = NA, color.colorbar, window.img, window.overlay, quality = 2,
  outname = NA, alpha = 1, newwindow = FALSE, nslices = 10,
  domainImageMap = NA, ncolumns = 4, useAbsoluteScale = FALSE,
  doCropping = TRUE, ...)
}
\arguments{
\item{x}{the reference image on which to overlay.}

\item{y}{image or list of images to use as overlays.}

\item{color.img}{color for main image.}

\item{color.overlay}{the color for the overlay , e.g c('blue','red') length
of this list should match the image list.}

\item{axis}{the axis to slice (1 , 2 or 3)}

\item{slices}{vector of slices to plot (e.g., c(10, 15, 20))}

\item{colorbar}{make colorbar?}

\item{title.colorbar}{title for colorbar}

\item{title.img}{title for main image (not displayed when file is saved with outname)}

\item{title.line}{title vertical displacement (irrelevant when a file is saved with outname)}

\item{color.colorbar}{color scale to use for colorbar}

\item{window.img}{lower and upper thresholds for display of main image}

\item{window.overlay}{lower and upper thresholds for display of overlay.
if you set \code{window.overlay = c( 0, 100 )} and \code{useAbsoluteScale=T}
then the image overlay will map into that range.  otherswise, the min and
max of the image will be used.}

\item{quality}{integer quality magnification factor 1 => large (e.g.
10)}

\item{outname}{name of output file if you want to write result to file, e.g.
\code{plot.jpg}.}

\item{alpha}{opacity}

\item{newwindow}{boolean controlling if we open a new device for this plot}

\item{nslices}{number of slices to view}

\item{domainImageMap}{this input antsImage or list contains a reference image
(\code{domainImage}) and optional reference mapping named \code{domainMap}.
these will be used to map the input image(s) to this antsImage space before
plotting. this is useful for non-standard image orientations.}

\item{ncolumns}{number of columns in plot}

\item{useAbsoluteScale}{boolean determines whether dynamic range is maximized
when visualizing overlays}

\item{doCropping}{apply cropping, defaults to \code{TRUE}}

\item{...}{other parameters}
}
\value{
output is plot to standard R window
}
\description{
This is a plotting utility for antsImage types with a background and color
overlay option.  Useful for displaying statistical results overlaid on a
background image.
}
\examples{

img <- makeImage(c(4,4), rnorm(4*4))
mask <- makeImage(c(4,4),
   as.matrix(c(0,0,0,0,
            0,1,1,0,
            0,1,1,0,
            0,0,0,0), nrow=4))
plot(img, list(mask))
\dontrun{
  mnit<-getANTsRData('mni')
  mnit<-antsImageRead(mnit)
  mniafn<-getANTsRData('mnia')
  mnia<-antsImageRead(mniafn)
  mnia<-thresholdImage(mnia,22,25)
  mnia<-smoothImage(mnia,1.5)
  mnia2<-antsImageRead(mniafn)
  mnia2<-thresholdImage(mnia2,1,4)
  mnia2<-smoothImage(mnia2,1.5)
  ofn<-paste(tempfile(),'.png',sep='')
  # write directly to a file
  plot( mnit, list(mnia,mnia2), slices=seq(50, 140, by=5),
    window.overlay = c(0.25,1), axis=2,
    color.overlay=c('red','blue'), outname = ofn )
}

}
\author{
Avants BB
}
