% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot.antsImage.R
\name{plot.antsImage}
\alias{plot.antsImage}
\title{Plotting an image slice or multi-slice with optional color overlay.}
\usage{
\method{plot}{antsImage}(x, y, color.img = "white", color.overlay = c("jet",
  "red", "blue", "green", "yellow"), axis = 2, slices,
  colorbar = missing(y), title.colorbar, title.img, color.colorbar,
  window.img = quantile(x[x != 0], c(0.05, 0.95)), window.overlay,
  quality = 4, outname = NA, alpha = 0.5, newwindow = FALSE, ...)
}
\arguments{
\item{x}{the reference image on which to overlay.}

\item{y}{image or list of images to use as overlays.}

\item{color.img}{color for main image.}

\item{color.overlay}{the color for the overlay , e.g c('blue','red') length
of this list should match the image list.}

\item{axis}{... the axis to slice (1 , 2 or 3)}

\item{slices}{vector of slices to plot (e.g., c(10, 15, 20))}

\item{colorbar}{make colorbar?}

\item{title.colorbar}{title for colorbar}

\item{title.img}{title for main image}

\item{color.colorbar}{color scale to use for colorbar}

\item{window.img}{lower and upper thresholds for display of main image}

\item{window.overlay}{lower and upper thresholds for display of overlay}

\item{quality}{integer quality magnification factor 1 => large (e.g.
10)}

\item{outname}{eg'figx.jpg' output name if you want to write the result to a
file}

\item{alpha}{opacity}

\item{newwindow}{boolean controlling if we open a new device for this plot}

\item{...}{other parameters}
}
\value{
output is plot to standard R window
}
\description{
This is a plotting utility for antsImage types with a background and color
overlay option.  Useful for displaying statistical results overlaid on a
background image.
}
\examples{
img <- makeImage(c(4,4), rnorm(4*4))
mask <- makeImage(c(4,4),
   as.matrix(c(0,0,0,0,
            0,1,1,0,
            0,1,1,0,
            0,0,0,0), nrow=4))
plot(img, list(mask))
\dontrun{
  mnit<-getANTsRData('mni')
  mnit<-antsImageRead(mnit,3)
  mniafn<-getANTsRData('mnia')
  mnia<-antsImageRead(mniafn,3)
  mnia<-thresholdImage(mnia,22,25)
  mnia<-smoothImage(mnia,1.5)
  mnia2<-antsImageRead(mniafn,3)
  mnia2<-thresholdImage(mnia2,1,4)
  mnia2<-smoothImage(mnia2,1.5)
  ofn<-paste(tempfile(),'.png',sep='')
  # write directly to a file
  plot( mnit, list(mnia,mnia2), slices=seq(50, 140, by=5),
   window.overlay = c(0.25,1), axis=2,color=c('red','blue'), outname = ofn )
}
}
\author{
Avants BB
}

