% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotPrettyGraph.R
\name{plotPrettyGraph}
\alias{plotPrettyGraph}
\title{Simple plotPrettyGraph function saves to png.}
\usage{
plotPrettyGraph(adjacencyMatrix, functionToPlot, pngfn = "graph.png",
  scaleText = 0.5, vertexSize = NA, figScale = 11, layoutmode = "eigen",
  hueval = 0)
}
\arguments{
\item{adjacencyMatrix}{igraph adjacencyMatrix}

\item{functionToPlot}{igraph node-level graph value e.g. degree,
page.rank, etc}

\item{pngfn}{filename for output png or to screen if NA}

\item{scaleText}{relative size of text to vertices}

\item{vertexSize}{cex size of vertices}

\item{figScale}{the figure will be of square size 2^figScale in pixels}

\item{layoutmode}{see gplot.layout in sna package}

\item{hueval}{controls the hue in hsv}
}
\value{
no output
}
\description{
PlotPrettyGraph given inputs from the makeGraph function.  adapted from
http://is-r.tumblr.com/.
}
\examples{
data("bold_correlation_matrix",package="ANTsR")
dmat<-data.matrix(bold_correlation_matrix)
if ( usePkg("igraph") ) {
 gg<-makeGraph( dmat, 0.1 )
 rownames(gg$adjacencyMatrix)<-colnames(bold_correlation_matrix)
 plotPrettyGraph( gg$adjacencyMatrix, gg$degree , figScale=12 , scaleText=5 )
 }
}
\author{
Avants BB, Christopher DeSante and David Sparks
}

