% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powersACTrsfMRIprocessing.R
\name{powersACTrsfMRIprocessing}
\alias{powersACTrsfMRIprocessing}
\title{Powers and ants cortical thickness processing for fMRI.}
\usage{
powersACTrsfMRIprocessing(img, fdthresh = 0.2, repeatMotionEst = 1,
  freqLimits = c(0.008, 0.09), nCompCor = 4, structuralImage = NA,
  structuralSeg = NA, structuralNodes = NA, templateMap = NA,
  smoothingSigmas = NA, verbose = FALSE)
}
\arguments{
\item{img}{input time series antsImage}

\item{fdthresh}{threshold for framewise displacement.  determines what time
frames should be interpolated. Set typically between 0.1 and 0.5.}

\item{repeatMotionEst}{number of times to repeat motion estimation.}

\item{freqLimits}{pair defining bandwidth of interest from low to high.}

\item{nCompCor}{number of compcor components to use.}

\item{structuralImage}{the structural antsImage of the brain.}

\item{structuralSeg}{a 3 or greater class tissue segmentation of the structural image.}

\item{structuralNodes}{regions of interest for network analysis, in the structural image space.}

\item{templateMap}{antsRegistration output mapping struturalImage to template space.}

\item{smoothingSigmas}{4-vector defining amount of smoothing in FWHM units}

\item{verbose}{enables visualization as well as commentary.}
}
\value{
outputs a list containing:
\itemize{
  \item{boldMat: }{Matrix of filtered BOLD data.}
  \item{boldMask: }{BOLD mask.}
  \item{nuisance: }{Nuisance variables.}
  \item{dmnBetas: }{Default mode network beta map.}
  \item{connMatPowers: }{Powers nodes connectivity matrix. Assumes template maps are to MNI space.}
  \item{connMatNodes: }{User provided nodal system connectivity matrix.}
  \item{connMatNodesPartialCorr: }{TUser provided nodal system partial correlation matrix.}
  \item{FD: }{mean framewise displacement.}
  \item{DVARS: }{signal variability.}
  \item{goodtimes: }{good time points.}
  \item{seg2bold: }{strutural segmentation in bold space.}
  \item{nodes2bold: }{strutural nodes in bold space.}
  \item{mapsToTemplate: }{invertible maps from bold to template space.}
}
}
\description{
This function leverages structural image processing based on ANTs
cortical thickness to implement Powers' functional image processing
recommendations.  The function will perform motion correction and
produce a variety of nuisance regressors.  It will also perform
spatial and temporal filtering.  In sum, these regressors mitigate
the wide variety of confounds that impact functional MRI.
}
\examples{
# this example is long-running ( perhaps 10 minutes on an OSX laptop 2016 )
\dontrun{
exrun = powersACTrsfMRIprocessing( verbose = TRUE ) # will download ex data
myid = "BBAvants" # some MRI data
pre = paste("~/rsfTest/",sep='')
fn = list.files( pre, full.names = TRUE, recursive = TRUE,
  pattern = glob2rx( paste( myid, "*rsfMRI0.nii.gz", sep='')  ) )
img  = antsImageRead( fn  )
sfn = list.files( pre, full.names = TRUE, recursive = TRUE,
  pattern = glob2rx( paste( myid, "*BrainSegmentation.nii.gz", sep='')  ) )
seg = antsImageRead( sfn )
t1fn = list.files( pre, full.names = TRUE, recursive = TRUE,
  pattern = glob2rx( paste( myid, "*BrainSegmentation0N4.nii.gz", sep='')  ) )
t1 = antsImageRead( t1fn   )
tt = powersACTrsfMRIprocessing( img, fdthresh=0.2, repeatMotionEst=1,
  structuralImage=t1, structuralSeg=seg, verbose= TRUE )
}

}
\author{
Avants BB, Duda JT
}

