% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/projectImageAlongAxis.R
\name{projectImageAlongAxis}
\alias{projectImageAlongAxis}
\title{Simple projectImageAlongAxis function.}
\usage{
projectImageAlongAxis(imageND, referenceImageNDminus1, projtype = 0,
  axis = NA)
}
\arguments{
\item{projtype}{projection type 0, 1 or 2}

\item{axis}{should be less than image dimension and greater than or equal to
0}

\item{img4d}{input image}

\item{refimg3d}{down-dimensional image to help define physical space}
}
\value{
image of n-dimensions-1 is output
}
\description{
Will turn an N-D image into an N-1-D image by summation (0), max-intensity
(1) or min-intensity (2) projection along an orthogonal axis.
}
\examples{
\dontrun{
mask4dproj<-projectImageAlongAxis( mask4d, mask3d, axis=3, projtype=0 )
}
}
\author{
Avants BB
}

