% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/quantifyCBF.R
\name{quantifyCBF}
\alias{quantifyCBF}
\title{quantifyCBF}
\usage{
quantifyCBF(perfusion, mask, parameters, M0val = NA, outlierValue = 0.02)
}
\arguments{
\item{perfusion}{input asl matrix}

\item{mask}{3D image mask (antsImage)}

\item{parameters}{list with entries for sequence and m0 (at minimimum)}

\item{M0val}{baseline M0 value (optional)}

\item{outlierValue}{trim outliers by this fractional value (optional)}
}
\value{
a list is output with 3 types of cbf images
}
\description{
Computes CBF from ASL - pasl or pcasl
}
\examples{
\dontrun{
  if (!exists("fn") ) fn<-getANTsRData("pcasl")
  # PEDS029_20101110_pcasl_1.nii.gz # high motion subject
  asl<-antsImageRead(fn)
# image available at http://files.figshare.com/1701182/PEDS012_20131101.zip
  pcasl.bayesian <- aslPerfusion( asl,
        dorobust=0., useDenoiser=4, skip=11, useBayesian=1000,
        moreaccurate=0, verbose=T, maskThresh=0.5 ) # throw away lots of data
# user might compare to useDenoiser=FALSE
  pcasl.parameters <- list( sequence="pcasl", m0=pcasl.bayesian$m0 )
  cbf <- quantifyCBF( pcasl.bayesian$perfusion, pcasl.bayesian$mask,
     pcasl.parameters )
  meancbf <- cbf$kmeancbf
  print(mean(meancbf[ pcasl.bayesian$mask==1 ]))
  antsImageWrite( meancbf , "temp.nii.gz")
  pcasl.processing <- aslPerfusion( asl, moreaccurate=0,
    dorobust=0.95, useDenoiser=NA, skip=5,  useBayesian=0 )
  # user might compare to useDenoiser=FALSE
  pcasl.parameters <- list( sequence="pcasl", m0=pcasl.processing$m0 )
  cbf <- quantifyCBF( pcasl.processing$perfusion, pcasl.processing$mask, pcasl.parameters )
  meancbf <- cbf$kmeancbf
  print(mean(meancbf[ pcasl.processing$mask==1 ]))
  antsImageWrite( meancbf , "temp2.nii.gz" )
  plot(  meancbf, slices="1x50x1" )
  }
}
\author{
Avants BB, Kandel B, Duda JT
}

