% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/randsccan.R
\name{randcca}
\alias{randcca}
\title{randomized cca function.}
\usage{
randcca(x, y, k, seed)
}
\arguments{
\item{x}{input matrix}

\item{y}{input matrix}

\item{k}{rank to use}

\item{seed}{for testing}
}
\value{
randomized svd of t(x) y is output
}
\description{
uses random matrix to estimate cca results
}
\examples{
set.seed(13)
x <- matrix(rnorm(3000), nrow=50 )
y <- x \%*\% matrix( rnorm( 60*100 ), nrow=60)
k=10
dr=randcca(x,y,k,1)
dt=svd( t(x)\%*\%y )
cor(dr$d[1:k],dt$d[1:k])
tt=2
ct=0
for ( i in 1:20 )
  ct=ct+as.numeric( dr$d[tt] < randcca(x,y[sample(1:50),],k,1)$d[tt] )
if ( ct > 0 ) stop("should not happen")
\dontrun{
# produce low rank cca matrix
K = t(dr$u)
lowAr = ( K \%*\% t(x) ) \%*\% y
K = t(dt$u)[1:k,]
lowAt = ( K \%*\% t(x) ) \%*\% y
mydecom<-sparseDecom( lowAr )
}
}
\author{
Avants BB
}

