% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsrTransform_class.R
\name{readAntsrTransform}
\alias{readAntsrTransform}
\title{readAntsrTransform}
\usage{
readAntsrTransform(filename, dimension = 3, precision = "float")
}
\arguments{
\item{filename}{filename of transform}

\item{dimension}{spatial dimension of transform}

\item{precision}{numerical precision of transform}
}
\value{
antsrTransform
}
\description{
read a transform from file
}
\examples{
\dontrun{
tx = readAntsrTransform( "yourtx.mat")
}
}
