% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeNormalizedPopulationData.R
\name{readNormalizedPopulationData}
\alias{readNormalizedPopulationData}
\title{Read normalized population from h5 file or directory}
\usage{
readNormalizedPopulationData(filename)
}
\arguments{
\item{filename}{input filename for hdf5 file.}
}
\value{
list containing demographics, imageMat, imageMask and imageBoolean
}
\description{
This function reads a file created by \code{writeNormalizedPopulationData}.
}
\author{
Avants BB
}
