% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reflectImage.R
\name{reflectImage}
\alias{reflectImage}
\title{reflectImage}
\usage{
reflectImage(img1, axis = NA, tx = NA, metric = "mattes")
}
\arguments{
\item{img1}{input object, an antsImage}

\item{axis}{which dimension to reflect across, numbered from 0 to imageDimension-1}

\item{tx}{transformation type to estimate after reflection}

\item{metric}{similarity metric for image registration.  see \code{antsRegistration}.}
}
\description{
reflects an image along its axis
}
\examples{

fi<-antsImageRead( getANTsRData("r16") , 2 )
axis = 2
asym<-reflectImage( fi, axis, "Affine" )$warpedmovout
asym<-asym-fi

}
\seealso{
\code{\link{antsRegistration}}
}
\author{
BB Avants
}
