% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/regressionNetworkViz.R
\name{regressionNetworkViz}
\alias{regressionNetworkViz}
\title{Visualize a regression result by a d3 network visualization.}
\usage{
regressionNetworkViz(mylm, sigthresh = 0.05, whichviz = "Sankey",
  outfile = "temp.html", mygroup = 0, logvals = TRUE, verbose = FALSE,
  correlateMyOutcomes = NA, corthresh = 0.9, zoom = F, doFDR = TRUE)
}
\arguments{
\item{mylm}{lm model output from bigLMStats}

\item{sigthresh}{significance threshold}

\item{whichviz}{which visualization method}

\item{outfile}{significance threshold}

\item{mygroup}{color each entry by group membership}

\item{logvals}{bool}

\item{verbose}{bool}

\item{correlateMyOutcomes}{not sure, see code}

\item{corthresh}{correlation threshold}

\item{zoom}{zooming factor}

\item{doFDR}{bool}
}
\value{
html file is output
}
\description{
Use either a force directed graph or a Sankey graph to show relationships
between predictors and outcome variables.  correlateMyOutcomes should
correspond to the outcome variables ...
}
\examples{
\dontrun{
# colnames(brainpreds)<-paste('Vox',c(1:ncol(brainpreds)),sep='')
# colnames( mylm$beta.pval )<-colnames(brainpreds)
# demognames<-rownames(mylm$beta.pval)
# regressionNetworkViz( mylm , sigthresh=0.05, outfile='temp2.html')
}
}
\author{
Avants BB
}

