\name{regressProjections}
\alias{regressProjections}
\title{Regression on image set projection}
\description{
  Perform regression on a training set of images, projected onto (provided) eigenvectors, and test on testing images.
}

\usage{
  regressProjections(input.train, input.test, demog.train, demog.test, 
           eigenvectors, mask, outcome, covariates="1", method="optimal", ...)
}

\arguments{
  \item{input.train}{Masked imaging data from training set of type \code{antsImage} or \code{matrix}.  This will typically be read from a \code{train.mha} file generated by, e.g., \code{sccan --imageset-to-matrix}.}
  \item{input.test}{Masked imaging data from testing set of type \code{antsImage} or \code{matrix}.  This will typically be read from a \code{test.mha} file generated by, e.g., \code{sccan --imageset-to-matrix}.}
  \item{demog.train}{Data frame of demographics information for training images.}
  \item{demog.test}{Data frame of demographics information for testing images.}
  \item{eigenvectors}{List of eigenvector images for dimensionality reduction.}
  \item{mask}{Mask image of type \code{antsImage}.}
  \item{outcome}{Name of outcome variable to be predicted.  Must be present in \code{demog.train} and \code{demog.test}.}
  \item{covariates}{List of names of covariates to be used for the prediction.  All names must be present in \code{demog.train} and \code{demog.test}.}
  \item{method}{Method for selecting eigenvectors.  Can be either \code{"all"} (uses all eigenvectors in \code{vector.names}) or \code{"optimal"} (uses
  BIC stepwise model selection to select eigenvectors.}
  \item{...}{Additional arguments for input to \code{glm}.  Example input would be \code{family=binomial} for classification instead of regression.}
}

\details{
\code{regressProjections} is a convenient way to perform training and testing of predictions of demographic information from imaging data.  It takes as
input demographics information, imaging data, and eigenvectors, and performs prediction of the outcome variable from the projection of the imaging data on
the eigenvectors. 
}

\value{
A list of diagnostic and statistical information generated from the prediction, including: 
  \code{stats}: Mean squared error, correlation coefficients, and p-value of prediction for training and testing data.
  \code{outcome.comparison}: Data frame comparing real vs. predicted values for testing data. 
  \code{eigenvectors}: List of eigenvectors retained in model building. 
}

\seealso{
  \code{\link{sparseDecom2}}
}

\author{
Kandel BM
}

\examples{
# generate simulated outcome
nsubjects <- 100
x1 <- seq(1, 10, length.out=nsubjects) + rnorm(nsubjects, sd=2)
x2 <- seq(25, 15, length.out=nsubjects) + rnorm(nsubjects, sd=2)
outcome <- 3 * x1 + 4 * x2 + rnorm(nsubjects, sd=1)

# generate simulated images with outcome predicted 
# by sparse subset of voxels
voxel.1 <- 3 * x1 + rnorm(nsubjects, sd=2)
voxel.2 <- rnorm(nsubjects, sd=2)
voxel.3 <- 2 * x2 + rnorm(nsubjects, sd=2)
voxel.4 <- rnorm(nsubjects, sd=3)
input   <- cbind(voxel.1, voxel.2, voxel.3, voxel.4)

# simulate eigenvectors and mask
eigvec.1     <- as.antsImage(matrix(c(1,0,0,0), nrow=2))
eigvec.2     <- as.antsImage(matrix(c(0,1,0,0), nrow=2))
eigvec.3     <- as.antsImage(matrix(c(0,0,1,0), nrow=2))
eigvec.4     <- as.antsImage(matrix(c(0,0,0,1), nrow=2))
eigenvectors <- list(eigvec.1, eigvec.2, eigvec.3, eigvec.4)
mask         <- as.antsImage(matrix(c(1,1,1,1), nrow=2))

# generate sample demographics that do not explain outcome
age <- runif(nsubjects, 50, 75)
demog <- data.frame(outcome=outcome, age=age)
# randomly divide data into training and testing 
rows.train  <- sample(1:nrow(demog), floor(nrow(demog)*2/3), replace=FALSE)
input.train <- input[rows.train, ]
input.test  <- input[-rows.train, ]
demog.train <- demog[rows.train, ]
demog.test  <- demog[-rows.train, ]

result <- regressProjections(input.train, input.test, demog.train, demog.test, 
                             eigenvectors, mask, "outcome")
}
