% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderSurfaceFunction.R
\name{renderSurfaceFunction}
\alias{renderSurfaceFunction}
\title{3D surface-based rendering of volume images.}
\usage{
renderSurfaceFunction(
  surfimg,
  funcimg,
  surfval = 0.5,
  basefval,
  offsetfval,
  smoothsval = 0,
  smoothfval = 0,
  blobrender = TRUE,
  alphasurf = 1,
  alphafunc = 1,
  outdir = "./",
  outfn = NA,
  mycol,
  physical = TRUE,
  movieDuration = 6,
  zoom = 1.1
)
}
\arguments{
\item{surfimg}{Input image to use as rendering substrate.}

\item{funcimg}{Input list of images to use as functional overlays.}

\item{surfval}{intensity level that defines isosurface}

\item{basefval}{intensity level that defines lower threshold for functional
image}

\item{offsetfval}{intensity level that defines upper threshold for
functional image}

\item{smoothsval}{smoothing for the surface image}

\item{smoothfval}{smoothing for the functional image}

\item{blobrender}{render a blob as opposed to a surface patch}

\item{alphasurf}{alpha for the surface contour}

\item{alphafunc}{alpha value for functional blobs}

\item{outdir}{output directory}

\item{outfn}{output file name}

\item{mycol}{name of color or colors}

\item{physical}{boolean}

\item{movieDuration}{in seconds}

\item{zoom}{magnification factor}
}
\value{
0 -- Success\cr 1 -- Failure
}
\description{
Will use rgl to render a substrate (e.g. anatomical) and overlay image (e.g.
functional).
}
\examples{
\dontrun{
       mnit<-getANTsRData("mni")
       mnit<-antsImageRead(mnit)
       mnia<-getANTsRData("mnia")
       mnia<-antsImageRead(mnia)
       mnit<-thresholdImage( mnit, 1, max(mnit) )
       mnia<-thresholdImage( mnia, 1, 2 )
       brain<-renderSurfaceFunction( surfimg =list( mnit ) ,
          list(mnia), alphasurf=0.1 ,smoothsval = 1.5 )
       }
}
\seealso{
\code{\link{plotBasicNetwork}}
}
\author{
Avants B, Kandel B
}
