% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorientImage.R
\name{reorientImage}
\alias{reorientImage}
\title{reorient image by its principal axis}
\usage{
reorientImage(img, axis1, axis2 = NA, doreflection = 0, doscale = 0,
  txfn = NA)
}
\arguments{
\item{img}{antsImage}

\item{axis1}{vector of size dim, might need to play w/axis sign}

\item{axis2}{vector of size dim for 3D}

\item{doreflection}{boolean}

\item{doscale}{scalar value, 1 allows automated estimate of scaling}

\item{txfn}{file name for transformation}
}
\value{
reoriented image
}
\description{
align along a specific axis
}
\examples{

fi<-antsImageRead( getANTsRData("r16"))
reofi<-reorientImage(fi,c(1,0))

}
\author{
Brian B. Avants
}
\keyword{geometry}
\keyword{image}
