% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampleImage.R
\name{resampleImage}
\alias{resampleImage}
\title{resampleImage}
\usage{
resampleImage(image, resampleParams, useVoxels = FALSE, interpType = 1)
}
\arguments{
\item{image}{input antsImage matrix}

\item{resampleParams}{vector of size dimension with numeric values}

\item{useVoxels}{true means interpret resample params as voxel counts}

\item{interpType}{one of 0 (linear), 1 (nearest neighbor),
2 (gaussian), 3 (windowed sinc), 4 (bspline)}
}
\value{
output antsImage
}
\description{
Resample image by spacing or number of voxels with various interpolators.
Works with multi-channel images.
}
\examples{

fi<-antsImageRead( getANTsRData("r16"))
finn<-resampleImage(fi,c(50,60),TRUE,0)
filin<-resampleImage(fi,c(1.5,1.5),FALSE,1)

}
\author{
Avants BB
}

