% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampleImageToTarget.R
\name{resampleImageToTarget}
\alias{resampleImageToTarget}
\title{resampleImageToTarget}
\usage{
resampleImageToTarget(
  image,
  target,
  interpType = "linear",
  imagetype = 0,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{image}{image to resample}

\item{target}{image of reference, the output will be in this space}

\item{interpType}{Choice of interpolator. Supports partial matching.
\itemize{
\item{linear}{}
\item{nearestNeighbor}{}
\item{multiLabel}{ for label images but genericlabel is preferred}
\item{gaussian}{}
\item{bSpline}{}
\item{cosineWindowedSinc}{}
\item{welchWindowedSinc}{}
\item{hammingWindowedSinc}{}
\item{lanczosWindowedSinc}{}
\item{genericLabel}{ use this for label images}
}}

\item{imagetype}{choose 0/1/2/3 mapping to scalar/vector/tensor/time-series}

\item{verbose}{print command and run verbose application of transform.}

\item{...}{additional arugment passed to \code{antsApplyTransforms} C code}
}
\value{
output antsImage resampled with target's resolution/origin/orientation/direction
}
\description{
Resample image by using another image as target reference.
This function uses antsApplyTransform with an identity
matrix to achieve proper resampling.
}
\examples{

fname <- getANTsRData("r16")
fi <- antsImageRead(fname)
fi2mm <- resampleImage(fi, c(2, 2), useVoxels = 0, interpType = "linear")
resampled <- resampleImageToTarget(fi2mm, fi)
testthat::expect_error(resampleImageToTarget(fi2mm))

}
\author{
Pustina D
}
