% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resels.R
\name{resels}
\alias{resels}
\title{Estimates image resels}
\usage{
resels(mask, fwhm)
}
\arguments{
\item{mask}{statistical value (typically the maxima of a cluster or statistical field)}

\item{fwhm}{the full width at half maxima measurement}
}
\value{
A vector of resels for dimensions 0:D
}
\description{
Utilize the estimated FWHM to find the resels per voxel
}
\details{
Interprets a given antsImage mask (binarized so as to only contain 0s and 1s) into
resolutions elements (as opposed to voxels). Doing so emphasizes the interdependent
nature of voxels when undergoing RFT based statistical analyses. Optimized for three
dimensions (2D has not yet been tested yet).
}
\examples{
mask <- getMask(antsImageRead(getANTsRData('mni')))
myresels <- resels(mask, c(1, 1, 1))

}
\author{
Zachary P. Christensen
}
\references{
Worlsey K.J., (1996) A Unified Statistical Approach for Determining Significant Signals in Images of Cerebral Activation.
}
\seealso{
rftPval, euler, rftResults
}

