% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rfSegmentation.R
\name{rfSegmentation}
\alias{rfSegmentation}
\title{A rfSegmentation function.}
\usage{
rfSegmentation(labelimg, featureimages, ntrees = 100, verbose = FALSE)
}
\arguments{
\item{featureimages}{input list of antsImage feature images - length n means
n predictors in rf}

\item{labelimage}{input antsImage labelimage --- assume non-zero entries
create a mask}
}
\value{
list of n-probability images is output where n is number of classes
}
\description{
Image segmentation via random forests.
}
\examples{
usePkg('randomForest')
img<-antsImageRead( getANTsRData('r16') ,2)
mask<-getMask( img )
segs<-Atropos( d = 2, a = img, m = '[0.2,1x1]',c = '[5,0]',  i = 'kmeans[3]', x = mask)
fimgs<-lappend( img, segs$probabilityimages )
rfsegs<-rfSegmentation( segs$segmentation, fimgs , verbose=TRUE )
plotANTsImage( rfsegs$segmentation )
# now use in atropos w/priors
# segs2<-Atropos( d = 2, a = fimgs, m = '[0.2,1x1]',c = '[5,0]',  i = rfsegs$probabilityimages, x = mask)
}
\author{
Tustison NJ, Avants BB
}

