% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfSegmentationPredict.R
\name{rfSegmentationPredict}
\alias{rfSegmentationPredict}
\title{A rfSegmentationPredict function.}
\usage{
rfSegmentationPredict(rfSegmentationModel, featureimages, mask,
  verbose = FALSE)
}
\arguments{
\item{rfSegmentationModel}{input rf model}

\item{featureimages}{input list of antsImage feature images}

\item{mask}{antsImage mask}

\item{verbose}{bool}
}
\value{
segmentation is output
}
\description{
Predict image segmentation via random forests.
}
\examples{

if ( usePkg('randomForest') ) {
img<-antsImageRead( getANTsRData("r16"))
mask<-getMask( img )
mask2<-getMask( img )
mask [ 129:255, 1:255 ]<-0
mask2[ 2:128, 1:255 ]<-0
segs<-kmeansSegmentation( img, k=3, kmask = mask)
fimgs<-list( img )
rfsegs<-rfSegmentation( segs$segmentation, fimgs , ntrees=100 )
rfseg2<-rfSegmentationPredict(  rfsegs$rfModel , fimgs , mask2 )
plot( rfseg2 )
\dontrun{
   img<-antsImageRead( getANTsRData("r16"))
   img2<-antsImageRead( getANTsRData("r64"))
   mask<-getMask( img )
   mask2<-getMask( img2 )
   segs<-kmeansSegmentation( img, k=3, kmask = mask)
   nimg<-iMath(img,"Normalize")
   fimgs<-list( nimg )
   rfsegs<-rfSegmentation( segs$segmentation, fimgs , ntrees=100 )
   mytx<-antsRegistration(fixed=img , moving=img2 ,
            typeofTransform = c('SyN') )
   fimgs2<-list( iMath(mytx$warpedmovout,"Normalize") )
   rfseg2<-rfSegmentationPredict(  rfsegs$rfModel , fimgs2 , mask )
   plot( rfseg2 )
   }
}

}
\author{
Tustison NJ, Avants BB
}

