% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ripmmarc.R
\name{ripmmarc}
\alias{ripmmarc}
\title{Rotation Invariant Patch-based Multi-Modality Analysis aRChitecture}
\usage{
ripmmarc(img, mask, patchRadius = 3, patchSamples = 1000)
}
\arguments{
\item{img}{Image to decompose}

\item{mask}{Binary mask defining regions in which to decompose.}

\item{patchRadius}{Scalar radius defining the patch size.}

\item{patchSamples}{Scalar defining the number of random patches to sample.}
}
\value{
antsImage decomposed bases
}
\description{
Patch-based and rotation invariant image decomposition.  This
is similar patch-based dictionary learning in N-dimensions.
}
\examples{
img <- antsImageRead( getANTsRData( "r16" ) )
msk <- thresholdImage( img, quantile( img[ img > 0 ] )[1], max( img ) )
ripped <- ripmmarc( img, msk, 3 )
}
\author{
Kandel BM, Avants BB
}

