% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/save.ANTsR.R
\name{save.ANTsR}
\alias{load.ANTsR}
\alias{save.ANTsR}
\title{save.ANTsR}
\usage{
save.ANTsR(filename="./ANTsRsession", objects=NA, ...)

load.ANTsR(filename)
}
\arguments{
\item{filename}{Prefix for folder to store data.}

\item{objects}{Vector of character names of objects to store.  Can be antsImages.}

\item{...}{Additional arguments to pass to \code{save}.}
}
\description{
Save and load ANTsR sessions.
}
\section{Functions}{
\itemize{
\item \code{load.ANTsR}: Load saved ANTsR session.
}}
\examples{
a <- 1
b <- c(2,3,4)
img <- antsImageRead(getANTsRData('r16'), 2)
save.ANTsR(objects=c('b', 'img'))
load.ANTsR("./ANTsRsession")
}

