% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.ANTsR.R
\name{save.ANTsR}
\alias{save.ANTsR}
\alias{load.ANTsR}
\title{save.ANTsR}
\usage{
save.ANTsR(filename = file.path(".", ".ANTsRsession"), objects = NA,
  env = as.environment(1), overwrite = FALSE, clonediskfiles = TRUE, ...)

load.ANTsR(filename=file.path('.','.ANTsRsession'), env=as.environment(1))
}
\arguments{
\item{filename}{Prefix for folder to store data.}

\item{objects}{Vector of character names of objects to store.  Can be antsImages.}

\item{env}{Environment to save from or load to.}

\item{overwrite}{logical to select whether overwriting existing data is allowed.}

\item{clonediskfiles}{logical enables the copying of disk files that are not loaded
in workspace but where character variables point. Set to TRUE if moving sessions 
from one computer to another or if you save antsRegistration() outputs. Set to 
FALSE if you save temporary sessions to continue later on the same computer.}

\item{...}{Additional arguments to pass to \code{save}.}
}
\description{
Save and load ANTsR sessions.
}
\examples{
\dontrun{ # causes problems with devtools::run_examples()
# a <- 1
# b <- c( 2, 3, 4)
# save.ANTsR(objects=c('b', 'a'))
# load.ANTsR("./.ANTsRsession")
}
}
\author{
Pustina D
}
