% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.ANTsR.R
\name{save.ANTsR}
\alias{load.ANTsR}
\alias{save.ANTsR}
\title{save.ANTsR}
\usage{
save.ANTsR(filename="./.ANTsRsession", objects=NA,
  env=as.environment(1), ...)

load.ANTsR(filename="./.ANTsRsession", env=as.environment(1))
}
\arguments{
\item{filename}{Prefix for folder to store data.}

\item{objects}{Vector of character names of objects to store.  Can be antsImages.}

\item{env}{Environment to save from or load to.}

\item{overwrite}{logical to select whether overwriting existing data is allowed.}

\item{clonediskfiles}{logical enables the copying of disk files that are not loaded
in workspace but where character variables point. This enables a later 
re-instantiation of the full workspace including antsRegistration transformation
matrices. May need to be careful when saving the entire workspace with this option
set to true.}

\item{...}{Additional arguments to pass to \code{save}.}
}
\description{
Save and load ANTsR sessions.
}
\examples{
\dontrun{ # causes problems with devtools::run_examples()
# a <- 1
# b <- c( 2, 3, 4)
# save.ANTsR(objects=c('b', 'a'))
# load.ANTsR("./.ANTsRsession")
}
}
\author{
Pustina D
}

