% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/segmentShapeFromImage.R
\name{segmentShapeFromImage}
\alias{segmentShapeFromImage}
\title{convolution-based shape identification}
\usage{
segmentShapeFromImage(img, shape, mask = NA, rad = NA, scfun)
}
\arguments{
\item{img}{antsImage}

\item{shape}{to define features}

\item{mask}{with values 1 or 0}

\item{rad}{max radius (optional)}

\item{scfun}{function to apply to feature image (optional eg abs)}
}
\value{
feature image
}
\description{
takes an image and a shape and concvoles the 2nd with the first to yield a
feature image
}
\examples{
set.seed(123)
fi<-makeImage(c(20,20),rnorm(400,mean=1,sd=0.1))
# fi<-antsImageRead( getANTsRData("r16") ,2)
mask<-getMask(fi,0.8,Inf,0)
segs<-kmeansSegmentation( fi ,3 , mask)$segmentation
segs[ segs != 1 ]<-0
shp<-labelClusters( segs,1)
shp[ shp != 1 ]<-0
fimg<-segmentShapeFromImage(fi,shp,mask)
\dontrun{
plot(fimg)
}
}
\author{
Brian B. Avants
}
\keyword{shape}
\keyword{template}

