% Generated by roxygen2 (4.0.2): do not edit by hand
\name{segmentShapeFromImage}
\alias{segmentShapeFromImage}
\title{convolution-based shape identification}
\usage{
featimg<-segmentShapeFromImage( img, shape, mask )
}
\arguments{
\item{img}{antsImage}

\item{shape}{to define features}

\item{mask}{with values 1 or 0}

\item{rad}{max radius (optional)}

\item{scfun}{function to apply to feature image (optional eg abs)}
}
\value{
feature image
}
\description{
takes an image and a shape and concvoles the 2nd with the first to yield a feature image
}
\examples{
set.seed(123)
fi<-antsImageRead( getANTsRData('r16') ,2)
mask<-getMask(fi)
segs<-Atropos( d = 2, a = fi, m = "[0.2,1x1]",c = "[5,0]",  i = "kmeans[3]", x = mask)$segmentation
segs[ segs != 1 ]<-0
shp<-labelClusters( antsImageClone(segs,'float'),10)
shp[ shp != 7 ]<-0
fimg<-segmentShapeFromImage(fi,shp,mask)
plotANTsImage(fi,func=list(fimg),
  thresh=paste(mean(fimg)+sd(fimg)*2,
  max(fimg),sep='x'))
}
\author{
Brian B. Avants
}
\keyword{shape}
\keyword{template}

