% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscaleSVDxpts.R
\name{simlr.search}
\alias{simlr.search}
\title{Perform SIMLR Search with Random Hyperparameter Sampling}
\usage{
simlr.search(
  mats,
  regs,
  nsimlr_options,
  prescaling_options,
  objectiver_options,
  mixer_options,
  sparval_options,
  ebber_options,
  pizzer_options,
  optimus_options,
  num_samples = 10,
  maxits = 100,
  nperms = 1,
  verbose = 0
)
}
\arguments{
\item{mats}{List of matrices}

\item{regs}{List of regularization matrices}

\item{nsimlr_options}{List of nsimlr options}

\item{prescaling_options}{List of prescaling options}

\item{objectiver_options}{List of objective function options}

\item{mixer_options}{List of mixer options}

\item{sparval_options}{List of sparseness quantile options 
may also pass in rand3x0.5x0.9 which will produce random uniform 
samples in that range of length 3 ( or a similar string )}

\item{ebber_options}{List of expBeta options}

\item{pizzer_options}{List of positivity options}

\item{optimus_options}{List of optimization style options}

\item{num_samples}{Number of random samples}

\item{maxits}{Maximum number of iterations in simlr.perm subroutine}

\item{nperms}{Number of permutations in simlr.perm subroutine}

\item{verbose}{boolean}
}
\value{
List containing top 3 results and a data frame with selected options
}
\description{
Perform SIMLR Search with Random Hyperparameter Sampling
}
