% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscaleSVDxpts.R
\name{simlr.search}
\alias{simlr.search}
\title{Perform SIMLR Grid Search}
\usage{
simlr.search(
  mats,
  regs,
  options_df,
  maxits = 100,
  nperms = 1,
  verbose = 0,
  FUN = rvcoef
)
}
\arguments{
\item{mats}{The input matrices for SIMLR.}

\item{regs}{The regularization options for SIMLR.}

\item{options_df}{A data frame of parameter combinations generated by `simlr.parameters`.}

\item{maxits}{The maximum number of iterations for SIMLR. Default is 100.}

\item{nperms}{The number of permutations for the significance test. Default is 1.}

\item{verbose}{The verbosity level. Default is 0.}

\item{FUN}{The function to use for the SIMLR evaluation. Default is `rvcoef`.}
}
\value{
A list containing the best SIMLR result, its significance, and the parameters.
}
\description{
This function performs a grid search over the parameter combinations provided by `options_df`.
}
