% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SmoothImage.R
\name{SmoothImage}
\alias{SmoothImage}
\title{Smooth an image}
\usage{
SmoothImage(dim, inimg, sigma, outimg)
}
\arguments{
\item{dim}{Number of dimensions of the input image}

\item{inimg}{Input image to operate on}

\item{sigma}{Smoothing factor in format 1.1x1.5x0.5 for a 3D image}

\item{outimg}{Result image, placeholder}
}
\value{
0 -- Success\cr 1 -- Failure
}
\description{
Perform smoothing on the given image with a given sigma, defined in physical
space units
}
\examples{
# different ways you might smooth images of different dimensionality
img<-makeImage(c(5,5),rnorm(25))
SmoothImage( 2 , img ,'1.2x1.5', img)
SmoothImage( 2 , img ,c(1.2,1.5), img)
}
\author{
Shrinidhi KL, Avants BB
}

