% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothImage.R
\name{smoothImage}
\alias{smoothImage}
\title{Smooth image}
\usage{
smoothImage(inimg, sigma, sigmaInPhysicalCoordinates = TRUE, FWHM = FALSE,
  max_kernel_width = 70)
}
\arguments{
\item{inimg}{Image to smooth}

\item{sigma}{Smoothing factor.  Can be scalar, in which case the same sigma
is applied to each dimension, or a vector of length \code{dim(inimg)} to
specify a unique smoothness for each dimension.}

\item{sigmaInPhysicalCoordinates}{If true, the smoothing factor is in
millimeters; if false, it is in pixels.}

\item{FWHM}{If true, sigma is interpreted as the full-width-half-max (FWHM)
of the filter, not the sigma of a Gaussian kernel.}

\item{max_kernel_width}{Maximum kernel width}
}
\value{
antsImage smoothed image
}
\description{
Smooth image or multi-channel image
}
\examples{
img <- makeImage(c(5,5), rnorm(25))
simg <- smoothImage(img, c(1.2,1.5))
}
\author{
Kandel BM, Avants BB
}

