% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscaleSVDxpts.R
\name{sparseDistanceMatrix}
\alias{sparseDistanceMatrix}
\title{Create sparse distance, covariance or correlation matrix}
\usage{
sparseDistanceMatrix(x, k = 3, r = Inf, sigma = NA,
  kmetric = c("euclidean", "correlation", "covariance", "gaussian"),
  eps = 0.000001, ncores = NA, sinkhorn = TRUE)
}
\arguments{
\item{x}{input matrix, should be n (samples) by p (measurements)}

\item{k}{number of neighbors}

\item{r}{radius of epsilon-ball}

\item{sigma}{parameter for kernel PCA.}

\item{kmetric}{similarity or distance metric determining k nearest neighbors}

\item{eps}{epsilon error for rapid knn}

\item{ncores}{number of cores to use}

\item{sinkhorn}{boolean}
}
\value{
matrix sparse p by p matrix is output with p by k nonzero entries
}
\description{
Exploit k-nearest neighbor algorithms to estimate a sparse similarity matrix.
Critical to the validity of this function is the basic mathematical
relationships between euclidean distance and correlation and between
correlation and covariance.  For applications of such matrices, one may see
relevant publications by Mauro Maggioni and other authors.
}
\examples{
\dontrun{
mat = matrix( rnorm(60), ncol=10 )
smat = sparseDistanceMatrix( mat, 2 )
r16 = antsImageRead( getANTsRData( 'r16' ) )
mask = getMask( r16 )
mat <- getNeighborhoodInMask(image = r16, mask = mask, radius = c(0,0),
  physical.coordinates=TRUE, spatial.info=TRUE )
smat = sparseDistanceMatrix( t(mat$indices), 10 ) # close points
}
}
\references{
\url{http://www.math.jhu.edu/~mauro/multiscaledatageometry.html}
}
\author{
Avants BB
}
