% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscaleSVDxpts.R, R/multiscaleSVDxpts_temp.R
\name{sparseDistanceMatrix}
\alias{sparseDistanceMatrix}
\title{Create sparse distance, covariance or correlation matrix}
\usage{
sparseDistanceMatrix(x, k = 3, r = Inf, sigma = NA,
  kmetric = c("euclidean", "correlation", "covariance", "gaussian"),
  eps = 1e-06, ncores = NA, sinkhorn = FALSE, kPackage = "FNN")

sparseDistanceMatrix(x, k = 3, r = Inf, sigma = NA,
  kmetric = c("euclidean", "correlation", "covariance", "gaussian"),
  eps = 1e-06, ncores = NA, sinkhorn = FALSE, kPackage = "FNN")
}
\arguments{
\item{x}{input matrix, should be n (samples) by p (measurements)}

\item{k}{number of neighbors}

\item{r}{radius of epsilon-ball}

\item{sigma}{parameter for kernel PCA.}

\item{kmetric}{similarity or distance metric determining k nearest neighbors}

\item{eps}{epsilon error for rapid knn}

\item{ncores}{number of cores to use}

\item{sinkhorn}{boolean}

\item{kPackage}{name of package to use for knn}

\item{x}{input matrix, should be n (samples) by p (measurements)}

\item{k}{number of neighbors}

\item{r}{radius of epsilon-ball}

\item{sigma}{parameter for kernel PCA.}

\item{kmetric}{similarity or distance metric determining k nearest neighbors}

\item{eps}{epsilon error for rapid knn}

\item{ncores}{number of cores to use}

\item{sinkhorn}{boolean}

\item{kPackage}{name of package to use for knn}
}
\value{
matrix sparse p by p matrix is output with p by k nonzero entries

matrix sparse p by p matrix is output with p by k nonzero entries
}
\description{
Exploit k-nearest neighbor algorithms to estimate a sparse similarity matrix.
Critical to the validity of this function is the basic mathematical
relationships between euclidean distance and correlation and between
correlation and covariance.  For applications of such matrices, one may see
relevant publications by Mauro Maggioni and other authors.

Exploit k-nearest neighbor algorithms to estimate a sparse similarity matrix.
Critical to the validity of this function is the basic mathematical
relationships between euclidean distance and correlation and between
correlation and covariance.  For applications of such matrices, one may see
relevant publications by Mauro Maggioni and other authors.
}
\examples{
\dontrun{
set.seed(120)
mat = matrix( rnorm(60), ncol=10 )
smat = sparseDistanceMatrix( mat, 2 )
r16 = antsImageRead( getANTsRData( 'r16' ) )
mask = getMask( r16 )
mat <- getNeighborhoodInMask(image = r16, mask = mask, radius = c(0,0),
  physical.coordinates=TRUE, spatial.info=TRUE )
smat = sparseDistanceMatrix( t(mat$indices), 10 ) # close points
testthat::expect_is(smat, "Matrix")
testthat::expect_is(smat, "dgCMatrix")
testthat::expect_equal(sum(smat), 18017)
}
\dontrun{
set.seed(120)
mat = matrix( rnorm(60), ncol=10 )
smat = sparseDistanceMatrix( mat, 2 )
r16 = antsImageRead( getANTsRData( 'r16' ) )
mask = getMask( r16 )
mat <- getNeighborhoodInMask(image = r16, mask = mask, radius = c(0,0),
  physical.coordinates=TRUE, spatial.info=TRUE )
smat = sparseDistanceMatrix( t(mat$indices), 10 ) # close points
testthat::expect_is(smat, "Matrix")
testthat::expect_is(smat, "dgCMatrix")
testthat::expect_equal(sum(smat), 18017)
}
}
\references{
\url{http://www.math.jhu.edu/~mauro/multiscaledatageometry.html}

\url{http://www.math.jhu.edu/~mauro/multiscaledatageometry.html}
}
\author{
Avants BB

Avants BB
}
