% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitChannels.R
\name{splitChannels}
\alias{splitChannels}
\title{split a multichannel antsImage}
\usage{
splitChannels(image)
}
\arguments{
\item{image}{a multichannel antsImage to split}
}
\value{
list of scalar antsImages
}
\description{
split a multichannel antsImage into a list of scalar antsImages
}
\examples{
r <- floor(seq(1:(64 * 64)) / (64 * 64) * 255)
dim(r) <- c(64, 64)
r <- as.antsImage(r)
g <- r * 0
b <- r * 0
testthat::expect_error(splitChannels(r))
rgbImage <- mergeChannels(list(r, g, b))
imgList <- splitChannels(rgbImage)
testthat::expect_length(imgList, 3)
sapply(imgList, function(x) {
  testthat::expect_s4_class(x, class = "antsImage")
})
testthat::expect_error(splitChannels("hey"))

}
\author{
Duda, JT
}
