% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsImage_class.R
\docType{methods}
\name{[,antsImage,NULL,ANY-method}
\alias{[,antsImage,NULL,ANY-method}
\title{ants image substitutions}
\usage{
\S4method{[}{antsImage,`NULL`,ANY}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{antsImage}

\item{i}{logical or i-th dimension}

\item{j}{not used or j-th dimension}

\item{drop}{method for missing data}

\item{k}{not used or k-th dimension}

\item{l}{not used or l-th dimension}

\item{value}{ok}
}
\description{
ants image substitutions
}
