% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subjectDataToGroupDataFrame.R
\name{subjectDataToGroupDataFrame}
\alias{subjectDataToGroupDataFrame}
\title{convert subject matrix data to a row in a dataframe.}
\usage{
subjectDataToGroupDataFrame(csvlist, usecol, mycolname = NA,
  datarownames = NA)
}
\arguments{
\item{csvlist}{input list of csv files eg by Sys.glob ...}

\item{usecol}{a column name or number e.g. 1 or 'Volume'}

\item{mycolname}{rename the column by this string (optional)}

\item{datarownames}{the desired now names (optional)}
}
\value{
data frame is output
}
\description{
SubjectDataToGroupDataFrames take a list of subject-level csv files of the
same type and converts them to a data frame with consistent column naming
for the values and where rows correspond to subjects.
}
\examples{

\dontrun{
# data(aal,package="ANTsR")
# csvlist<-Sys.glob("*md*csv")
# mypopulationdataframe<-subjectDataToGroupDataFrame( csvlist , "Mean" ,
#  datarownames=aal$label_name )
# should have each ROI value for each subject listed in a large data frame
}

}
\author{
Avants BB
}

