% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscaleSVDxpts.R
\name{symilr}
\alias{symilr}
\title{Symmetric multivariate, penalized image-based linear regression model (symilr)}
\usage{
symilr(dataFrame, voxmats, basisK, myFormulaK, smoothingMatrixX,
  smoothingMatrixY, iterations = 10, gamma = 1e-06, sparsenessQuantile,
  positivity = c("positive", "negative", "either"),
  initializationStrategy = list(seed = 0, matrix = NA, voxels = NA),
  repeatedMeasures = NA, verbose = FALSE)
}
\arguments{
\item{dataFrame}{This data frame contains all relevant predictors except for
the matrices associated with the image variables.}

\item{voxmats}{The named list of matrices that contains the changing predictors.}

\item{basisK}{an integer determining the size of the basis.}

\item{myFormulaK}{This is a character string that defines a valid regression
which in this case should include predictors named as \code{paste0("mildBasis",1:basisK)}}

\item{smoothingMatrixX}{allows parameter smoothing, should be square and same
size as input matrix on left side of equation}

\item{smoothingMatrixY}{allows parameter smoothing, should be square and same
size as input matrix on right side of equation}

\item{iterations}{number of gradient descent iterations}

\item{gamma}{step size for gradient descent}

\item{sparsenessQuantile}{quantile to control sparseness - higher is sparser}

\item{positivity}{restrict to positive or negative solution (beta) weights.
choices are positive, negative or either as expressed as a string.}

\item{initializationStrategy}{initialization can be seed, matrix or voxels.
seed should be a single number, voxels should be a length basisK list of
integers with value less than the number of columns in the matrix, matrix
should be a n by k matrix.  The first non-NA of these options will be used.}

\item{repeatedMeasures}{list of repeated measurement identifiers. this will
allow estimates of per identifier intercept.}

\item{verbose}{boolean to control verbosity of output}
}
\value{
A list of different matrices that contain names derived from the
formula and the coefficients of the regression model.
}
\description{
This function simplifies calculating image-wide multivariate beta maps from
that is similar to CCA.
}
\examples{

set.seed(1500)
nsub = 12
npix = 100
outcome = rnorm( nsub )
covar = rnorm( nsub )
mat = replicate( npix, rnorm( nsub ) )
mat2 = replicate( npix + 10, rnorm( nsub ) )
nk = 3
myform = paste(" ~ covar + ",
  paste0( "mildBasis", 1:nk, collapse="+" ) )  # optional covariates
df = data.frame( outcome = outcome, covar = covar )
result = symilr( df, list( vox = mat, vox2 = mat2 ), basisK = 3, myform,
  initializationStrategy = list( seed = 10, matrix = NA, voxels = NA ) )

}
\seealso{
\code{\link{milr}}
}
\author{
BB Avants.
}
