% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscaleSVDxpts.R
\name{take_abs_unsigned}
\alias{take_abs_unsigned}
\title{Take Absolute Value of Unsigned Columns}
\usage{
take_abs_unsigned(m)
}
\arguments{
\item{m}{A numeric matrix}
}
\value{
A matrix with absolute values taken for unsigned columns
}
\description{
Take Absolute Value of Unsigned Columns
}
