% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/temporalwhiten.R
\name{temporalwhiten}
\alias{temporalwhiten}
\title{Simple autocorrelation-based temporal whitening function.}
\usage{
temporalwhiten(mat, myord = 2)
}
\arguments{
\item{mat}{input matrix}

\item{myord}{integer order value}
}
\value{
matrix is output
}
\description{
Temporally whitens the input matrix using autoregression and returns the
result.
}
\examples{
mat <- matrix(c(rep(1,100),rep(0,200)),ncol=50)
wmat<-temporalwhiten( mat )
}
\author{
Avants BB
}

