% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/thresholdImage.R
\name{thresholdImage}
\alias{thresholdImage}
\title{Threshold Image}
\usage{
thresholdImage(inimg, lothresh, hithresh, inval=1, outval=0)
}
\arguments{
\item{inimg}{Input image to operate on}

\item{lothresh}{Lower edge of threshold window}

\item{hithresh}{Higher edge of threshold window}

\item{inval}{Output value for image voxels in between \code{lothresh} and \code{hithresh}}

\item{outval}{Output value for image voxels lower than \code{lothresh} or higher than \code{hithresh}}
}
\value{
antsImage
}
\description{
Threshold Image
}
\examples{
img <- makeImage(c(5,5), rnorm(25)+0.5)
imgt<-thresholdImage( img, 0.5, Inf )
}
\author{
Shrinidhi KL
}

