% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ThresholdImage.R
\name{ThresholdImage}
\alias{ThresholdImage}
\title{Threshold Image}
\usage{
ThresholdImage(dim, inimg, outimg, lothresh, hithresh, inval=0, outval=1)
}
\arguments{
\item{dim}{Dimension of the input image}

\item{inimg}{Input image to operate on}

\item{outimg}{Result image}

\item{lothresh}{Lower edge of threshold window}

\item{hithresh}{Higher edge of threshold window}

\item{inval}{Output value for image voxels in between \code{lothresh} and \code{hithresh}}

\item{outval}{Output value for image voxels lower than \code{lothresh} or higher than \code{hithresh}}
}
\value{
0 -- Success\cr 1 -- Failure
}
\description{
Threshold Image
}
\examples{
img <- makeImage(c(5,5), rnorm(25))
ThresholdImage(2, img, img, 0.5, 1e3)
}
\author{
Shrinidhi KL
}

