% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ThresholdImage.R
\name{ThresholdImage}
\alias{ThresholdImage}
\title{Theshold Image}
\usage{
ThresholdImage(...)
}
\arguments{
\item{imageDimension2|3}{Number of dimensions of the input image}

\item{inputImage}{Input image to operate on}

\item{outputImage}{Result image}

\item{thresh-low}{Lower edge of threshold window}

\item{thresh-high}{Higher edge of threshold window}

\item{inside-value}{Inside value}

\item{outside-value}{Outside value}

\item{number-of-thresholds}{Number of thresholds}
}
\value{
0 -- Success\cr 1 -- Failure
}
\description{
Perform thresholding on the given image
}
\examples{
\dontrun{
ThresholdImage( 3 , 'input_img.nii' , 'output_img.nii' , 2 , 3 )
ThresholdImage( 3 , 'input_img.nii' , 'output_img.nii' , 'Otsu' , 3 )
}
}
\author{
Shrinidhi KL
}

