% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/timeseriesN3.R
\name{timeseriesN3}
\alias{timeseriesN3}
\title{Run N3 on slices of timeseries.}
\usage{
timeseriesN3(boldimg, mask, ncorrections = c(4, 2, 2))
}
\arguments{
\item{boldimg}{input matrix}

\item{mask}{mask to use}

\item{ncorrections}{levels at which to apply n3}
}
\value{
antsImage is output
}
\description{
Repeatedly run N3 on slices of time series and return image
}
\examples{
bold<-makeImage( c(10,10,10,4), rnorm(4000) )
  mask<-getMask( getAverageOfTimeSeries( bold ) )
  boldn3<-timeseriesN3( bold, mask, c(4,2) )
}
\author{
Avants BB
}

