% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_data.R
\docType{data}
\name{tracts}
\alias{tracts}
\title{tracts}
\format{
A data frame listing the following variables.
\describe{
\item{\code{label_num}}{ Numerical label value. } 
\item{\code{label_name}}{ Shorthand anatomical value. }
}
}
\usage{
tracts
}
\description{
A data frame label numbers and names for 
the white matter tracts labels.
}
\examples{
data(tracts)
}
\references{
\url{http://www.ncbi.nlm.nih.gov/pubmed/18407524}
}
\keyword{datasets}
