% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/usePkg.R
\name{usePkg}
\alias{usePkg}
\title{Use any package.  If package is not installed, this will install from CRAN.}
\usage{
usePkg(packageName, allowInstall = FALSE)
}
\arguments{
\item{packageName}{Name of package as *string*.}

\item{allowInstall}{let the package be installed from CRAN}
}
\value{
T if package successfully loaded, F otherwise.
}
\description{
Use any package.  If package is not installed, this will install from CRAN.
}
\examples{
usePkg("randomForest")
}
\author{
Benjamin M. Kandel, BB Avants
}

