% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/usePkg.R
\name{usePkg}
\alias{usePkg}
\title{Use any package.  If package is not installed, this will install from CRAN.}
\usage{
usePkg(pkg)
}
\arguments{
\item{packageName}{Name of package as *string*.}
}
\value{
T if package successfully loaded, F otherwise. cbf image
}
\description{
Use any package.  If package is not installed, this will install from CRAN.
}
\examples{
usePkg('extremevalues')
}
\author{
Benjamin M. Kandel
}

