% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_Summary.R
\name{var}
\alias{var}
\alias{var.default}
\alias{var.antsImage}
\title{Variance generic}
\usage{
var(x, ...)

\method{var}{default}(x, ...)

\method{var}{antsImage}(x, ..., na.rm = FALSE, mask = NULL)
}
\arguments{
\item{x}{an object for which we want to compute the variance}

\item{\dots}{Any additional arguments to be passed to
\code{\link[stats]{var}}}

\item{na.rm}{a logical value indicating whether NA should be removed}

\item{mask}{is an object of class \code{antsImage}}
}
\description{
Calculates the variance of an image
}
\examples{
img <- antsImageRead(getANTsRData("r16"))
var(img)
}
