% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weingartenImageCurvature.R
\name{weingartenImageCurvature}
\alias{weingartenImageCurvature}
\title{image curvature for 2D or 3D}
\usage{
weingartenImageCurvature(image, sigma = 1, opt = "mean", labeled = FALSE)
}
\arguments{
\item{image}{antsImage}

\item{sigma}{smoothing parameter}

\item{opt}{mean by default, otherwise use string gaussian or characterize}

\item{labeled}{boolean if TRUE then assume image contains labels which denote 
the surface on which curvature is to be calculated.  in this case, the 
surface should hold integer value 1; the background should be zero;  
the interior of the surface should be greater than 1 (e.g. 2).}
}
\value{
image
}
\description{
uses the weingarten map to estimate image mean or gaussian curvature
}
\examples{
img = makeImage( c(10,10,10) , rnorm( 1000 ) )
fik <- weingartenImageCurvature( img )
\dontrun{
fi <- antsImageRead( getANTsRData( "mni" ) )
fik <- weingartenImageCurvature( fi )
}
}
\references{
Avants, B, J. Gee, and B. Avants.
Shape operator for differential image analysis
Information Processing in Medical Imaging, 2003.
}
\author{
Brian B. Avants
}
\keyword{curvature}
