% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which.R
\docType{methods}
\name{which}
\alias{which}
\alias{which,antsImage-method}
\title{Which values are TRUE}
\usage{
which(x, arr.ind = FALSE, useNames = TRUE)

\S4method{which}{antsImage}(x, arr.ind = FALSE, useNames = TRUE)
}
\arguments{
\item{x}{antsImage object}

\item{arr.ind}{logical; should array indices be returned when x is an array?}

\item{useNames}{logical indicating if the value of arrayInd() 
should have (non-null) dimnames at all.}
}
\description{
Give the TRUE indices of a logical object, allowing for array indices.
#'
}
