% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whiten.R
\name{whiten}
\alias{whiten}
\title{Simple whitening function.}
\usage{
whiten(x, k = NULL, reducex = FALSE)
}
\arguments{
\item{x}{input matrix}

\item{k}{rank to use}

\item{reducex}{reduce the input matrix to k-size subspace}
}
\value{
matrix is output
}
\description{
Whitens the input matrix using SVD and returns the result.
}
\examples{

mat <- matrix(rnorm(300), ncol = 50)
wmat <- whiten(mat)
wmat2 <- whiten(mat, 2, TRUE)

}
\author{
Avants BB
}
