% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsrTransform_class.R
\name{writeAntsrTransform}
\alias{writeAntsrTransform}
\title{writeAntsrTransform}
\usage{
writeAntsrTransform(transform, filename)
}
\arguments{
\item{transform}{antsrTransform}

\item{filename}{filename of transform (file extension is ".mat" for affine transforms)}
}
\value{
TRUE
}
\description{
write antsrTransform to disk
}
\examples{
trans= c(3,4,5)
tx = createAntsrTransform( type="Euler3DTransform", translation=trans )
writeAntsrTransform(tx,"trans.mat")
}
